return DEGs UP and DOWN list based on intersection or union of comparisons

return DEGs UP and DOWN list based on intersection or union of comparisons

DEGs_Group(
  tfit,
  lfc = NULL,
  p = 0.05,
  assemble = "intersect",
  Rank = "adj.P.Val",
  keep.top = NULL,
  keep.group = NULL,
  ...
)

Arguments

tfit

MArrayLM object generated by limma::treat()

lfc

num, cutoff of logFC for DE analysis

p

num, cutoff of p value for DE analysis

assemble

'intersect' or 'union', whether to select intersected or union genes of different comparisons, default 'intersect'

Rank

character, the variable for ranking DEGs, can be 'logFC', 'adj.P.Val'..., default 'adj.P.Val'

keep.top

NULL or num, whether to keep top n DEGs of specific comparison

keep.group

NULL or pattern, specify the top DEGs of which comparison or group to be kept

...

omitted

Value

A list of "UP" and "DOWN" genes