Specify cell target_group signature against adherent cancer cell lines.

Specify signatures against specific tissues or cell lines, and generate CCLE_tpm_new after retrieving specific tissues.

filter_non_tissue(
  data = "CCLE",
  group_col,
  target_group,
  ...,
  log = FALSE,
  q = 0.25,
  markers = NULL,
  ignore.case = TRUE,
  slot = "counts",
  ccle_tpm = NULL,
  ccle_meta = NULL
)

# S4 method for matrix,vector,character
filter_non_tissue(
  data = "CCLE",
  group_col,
  target_group,
  ...,
  log = FALSE,
  q = 0.25,
  markers = NULL,
  ignore.case = TRUE,
  slot = "counts",
  ccle_tpm = NULL,
  ccle_meta = NULL
)

# S4 method for Matrix,vector,character
filter_non_tissue(
  data = "CCLE",
  group_col,
  target_group,
  ...,
  log = FALSE,
  q = 0.25,
  markers = NULL,
  ignore.case = TRUE,
  slot = "counts",
  ccle_tpm = NULL,
  ccle_meta = NULL
)

# S4 method for data.frame,vector,character
filter_non_tissue(
  data = "CCLE",
  group_col,
  target_group,
  ...,
  log = FALSE,
  q = 0.25,
  markers = NULL,
  ignore.case = TRUE,
  slot = "counts",
  ccle_tpm = NULL,
  ccle_meta = NULL
)

# S4 method for character,character,character
filter_non_tissue(
  data = "CCLE",
  group_col,
  target_group,
  ...,
  log = FALSE,
  q = 0.25,
  markers = NULL,
  ignore.case = TRUE,
  slot = "counts",
  ccle_tpm = NULL,
  ccle_meta = NULL
)

# S4 method for missing,character,character
filter_non_tissue(
  data = "CCLE",
  group_col,
  target_group,
  ...,
  log = FALSE,
  q = 0.25,
  markers = NULL,
  ignore.case = TRUE,
  slot = "counts",
  ccle_tpm = NULL,
  ccle_meta = NULL
)

# S4 method for DGEList,character,character
filter_non_tissue(
  data = "CCLE",
  group_col,
  target_group,
  ...,
  log = FALSE,
  q = 0.25,
  markers = NULL,
  ignore.case = TRUE,
  slot = "counts",
  ccle_tpm = NULL,
  ccle_meta = NULL
)

# S4 method for ExpressionSet,character,character
filter_non_tissue(
  data = "CCLE",
  group_col,
  target_group,
  ...,
  log = FALSE,
  q = 0.25,
  markers = NULL,
  ignore.case = TRUE,
  slot = "counts",
  ccle_tpm = NULL,
  ccle_meta = NULL
)

# S4 method for SummarizedExperiment,character,character
filter_non_tissue(
  data = "CCLE",
  group_col,
  target_group,
  ...,
  log = FALSE,
  q = 0.25,
  markers = NULL,
  ignore.case = TRUE,
  slot = "counts",
  ccle_tpm = NULL,
  ccle_meta = NULL
)

# S4 method for Seurat,character,character
filter_non_tissue(
  data = "CCLE",
  group_col,
  target_group,
  ...,
  log = FALSE,
  q = 0.25,
  markers = NULL,
  ignore.case = TRUE,
  slot = "counts",
  ccle_tpm = NULL,
  ccle_meta = NULL
)

Arguments

data

'CCLE' or expression object

group_col

vector or character, to specify the group of tumor/tissue target_groups, or column name of depmap::depmap_metadata(), e.g. 'primary_disease'

target_group

pattern, specify the target_group of interest, e.g. 'colorectal'

...

params for grep() to find matched cell lines in CCLE

log

logical, if to do log transformation on expression

q

num, quantile cutoff to screen markers which are not or lowly expressed in relevant samples

markers

vector, a vector of gene names, listed the gene symbols to be filtered. Default 'NULL' means all genes

ignore.case

logical, if to ignore the case of tissue pattern

slot

character, specify which slot to use only for sce or seurat object, optional, default 'counts'

ccle_tpm

ccle_tpm data from depmap::depmap_TPM(), only used when data = 'CCLE', default NULL

ccle_meta

ccle_meta data from depmap::depmap_metadata(), only used when data = 'CCLE', default NULL

Value

a vector of gene symbols

Examples

data("NK_markers", "ccle_crc_5")
filter_non_tissue(ccle_crc_5, group_col = "cancer", target_group = "CRC",
                  markers = NK_markers$HGNC_Symbol)
#>  [1] "KIR2DL2"   "KIR2DL5A"  "KIR2DS2"   "KIR2DS4"   "KIR2DS5"   "KLRA1P"   
#>  [7] "LOC653757" "MGC24103"  "TRBC1"     "TRDC"      "TBX21"     "CST7"     
#> [13] "IL12RB2"   "NME8"      "CMA1"      "OSM"       "GRAP2"     "IL2RB"    
#> [19] "GZMH"      "BPI"       "NKG7"      "CLNK"      "CRTAM"     "CD69"     
#> [25] "IFNG"      "KLRB1"     "GZMK"      "ZAP70"     "GNLY"      "IL18R1"   
#> [31] "IL18RAP"   "CD2"       "CD160"     "FASLG"     "CCND2"     "CD244"    
#> [37] "CEACAM8"   "GPR18"     "PTGER2"    "KIR2DL1"   "TNFSF14"   "KLRD1"    
#> [43] "KLRG1"     "XCL1"      "XCL2"      "GZMA"      "CDHR1"     "NCAM1"    
#> [49] "MS4A3"     "KLRF1"     "CD96"      "GFI1"      "FCGR3B"    "EOMES"    
#> [55] "PYHIN1"    "CAMP"      "CSF2"      "SAMD3"     "DEFA4"     "KIR3DL1"  
#> [61] "NAALADL1"  "PTGDR"     "KLHL30"    "AZU1"      "CTSW"      "CD7"      
#> [67] "GPR171"    "ZNF135"    "PRF1"      "S1PR5"     "LCK"       "KLRC4"    
#> [73] "SH2D1A"    "SOCS1"     "KIR2DL4"   "NCR1"      "GZMM"      "ELANE"    
#> [79] "SH2D1B"    "CD247"     "FCGR3A"    "NCR3"      "KLRC3"     "PTPRCAP"  
#> [85] "PVRIG"     "KLRK1"     "LTA"       "LTB"       "TNF"       "APOBEC3G" 
#> [91] "KIR3DL2"   "KIR3DL3"   "KLRF2"     "MGAM"      "CCL5"      "CCL4"