Get differentially expressed genes by comparing specified groups

This function uses edgeR and limma to get 'UP' and 'DOWN' DEG lists, for multiple comparisons, DEGs can be obtained from intersection of all comparsion DEGs or by using product of p value ranks for multiple comparisons. Filter out low expressed genes and extract DE genes by using limma::voom and limma::treat, and also create an object proc_data to store processed data.

get_degs(
  data,
  group_col,
  target_group,
  normalize = TRUE,
  feature_selection = c("auto", "rankproduct", "none"),
  slot = "counts",
  batch = NULL,
  ...
)

# S4 method for DGEList,character,character
get_degs(
  data,
  group_col,
  target_group,
  normalize = TRUE,
  feature_selection = c("auto", "rankproduct", "none"),
  slot = "counts",
  batch = NULL,
  ...
)

# S4 method for matrix,vector,character
get_degs(
  data,
  group_col,
  target_group,
  normalize = TRUE,
  feature_selection = c("auto", "rankproduct", "none"),
  slot = "counts",
  batch = NULL,
  ...
)

# S4 method for Matrix,vector,character
get_degs(
  data,
  group_col,
  target_group,
  normalize = TRUE,
  feature_selection = c("auto", "rankproduct", "none"),
  slot = "counts",
  batch = NULL,
  ...
)

# S4 method for ExpressionSet,character,character
get_degs(
  data,
  group_col,
  target_group,
  normalize = TRUE,
  feature_selection = c("auto", "rankproduct", "none"),
  slot = "counts",
  batch = NULL,
  ...
)

# S4 method for SummarizedExperiment,character,character
get_degs(
  data,
  group_col,
  target_group,
  normalize = TRUE,
  feature_selection = c("auto", "rankproduct", "none"),
  slot = "counts",
  batch = NULL,
  ...
)

# S4 method for Seurat,character,character
get_degs(
  data,
  group_col,
  target_group,
  normalize = TRUE,
  feature_selection = c("auto", "rankproduct", "none"),
  slot = "counts",
  batch = NULL,
  ...
)

Arguments

data

expression object

group_col

vector or character, specify the group factor or column name of coldata for DE comparisons

target_group

pattern, specify the group of interest, e.g. NK

normalize

logical, if the expr in data is raw counts needs to be normalized

feature_selection

one of "auto" (default), "rankproduct" or "none", choose if to use rank product or not to select DEGs from multiple comparisons of DE analysis, 'auto' uses 'rankproduct' but change to 'none' if final genes < 5 for both UP and DOWN

slot

character, specify which slot to use only for DGEList, sce or seurat object, optional, default 'counts'

batch

vector of column name(s) or dataframe, specify the batch effect factor(s), default NULL

...

params for process_data() and select_sig()

Value

A list of 'UP', 'DOWN' geneset of all differentially expressed genes, and a DGEList 'proc_data' containing data after process (filtration, normalization and voom fit). Both 'UP' and 'DOWN' are ordered by rank product or 'Rank' variable if keep.top is NULL

Examples

data("im_data_6")
DEGs <- get_degs(im_data_6, group_col = "celltype:ch1",
                 target_group = "NK", gene_id = "ENSEMBL")
#>        NK-Neutrophils NK-Monocytes NK-B.cells NK-CD4 NK-CD8
#> Down             4009         3944       3146   2694   2153
#> NotSig           1476         2678       4405   4985   6183
#> Up               4926         3789       2860   2732   2075