Collect genes from MSigDB or provided GeneSetCollection.

Collect gene sets from MSigDB or given GeneSetCollection, of which the gene-set names are matched to the given regex pattern by using grep() function. By setting cat and subcat, matching can be constrained in the union of given categories and subcategories if gsc = 'msigdb'.

get_gsc_sig(
  gsc = "msigdb",
  pattern,
  cat = NULL,
  subcat = NULL,
  species = c("hs", "mm"),
  id = c("SYM", "EZID"),
  version = msigdb::getMsigdbVersions(),
  ...
)

# S4 method for GeneSetCollection,character
get_gsc_sig(
  gsc = "msigdb",
  pattern,
  cat = NULL,
  subcat = NULL,
  species = c("hs", "mm"),
  id = c("SYM", "EZID"),
  version = msigdb::getMsigdbVersions(),
  ...
)

# S4 method for character,character
get_gsc_sig(
  gsc = "msigdb",
  pattern,
  cat = NULL,
  subcat = NULL,
  species = c("hs", "mm"),
  id = c("SYM", "EZID"),
  version = msigdb::getMsigdbVersions(),
  ...
)

Arguments

gsc

'msigdb' or GeneSetCollection to be searched

pattern

pattern pass to grep(), to match the MsigDB gene-set name of interest, e.g. 'NATURAL_KILLER_CELL_MEDIATED'

cat

character, stating the category(s) to be retrieved. The category(s) must be one from msigdb::listCollections(), see details in msigdb::subsetCollection()

subcat

character, stating the sub-category(s) to be retrieved. The sub-category(s) must be one from msigdb::listSubCollections(), see details in msigdb::subsetCollection()

species

character, species of interest, can be 'hs' or 'mm'

id

a character, representing the ID type to use ("SYM" for gene SYMBOLs and "EZID" for ENTREZ IDs)

version

a character, stating the version of MSigDB to be retrieved (should be >= 7.2). See msigdb::getMsigdbVersions().

...

params for grep(), used to match pattern to gene-set names

Value

A GeneSet object containing all matched gene-sets in MSigDB

Examples

data("msigdb_gobp_nk")
get_gsc_sig(
  gsc = msigdb_gobp_nk,
  pattern = "natural_killer_cell_mediated",
  subcat = "GO:BP",
  ignore.case = TRUE
)
#> GeneSetCollection
#>   names: GOBP_NATURAL_KILLER_CELL_MEDIATED_IMMUNITY, GOBP_NATURAL_KILLER_CELL_MEDIATED_CYTOTOXICITY_DIRECTED_AGAINST_TUMOR_CELL_TARGET, ..., GOBP_POSITIVE_REGULATION_OF_NATURAL_KILLER_CELL_MEDIATED_CYTOTOXICITY (18 total)
#>   unique identifiers: AP1G1, ARRB2, ..., KLRC4-KLRK1 (67 total)
#>   types in collection:
#>     geneIdType: SymbolIdentifier (1 total)
#>     collectionType: BroadCollection (1 total)