Show the summary info of available cell types in PanglaoDB.

Show the name and number of each cell type in PanglaoDB. Help users know which subset(s) marker list(s) could be retrieved by PanglaoDB.

list_panglao_types(organ)

Arguments

organ

character, specify the tissue or organ label to list cell types

Value

a vector of available cell types of the organ in PanglaoDB

Examples

list_panglao_types(organ = "Immune system")
#>  [1] "B cells (110)"                        
#>  [2] "B cells memory (66)"                  
#>  [3] "B cells naive (69)"                   
#>  [4] "Basophils (82)"                       
#>  [5] "Dendritic cells (133)"                
#>  [6] "Eosinophils (30)"                     
#>  [7] "Gamma delta T cells (66)"             
#>  [8] "Macrophages (153)"                    
#>  [9] "Mast cells (162)"                     
#> [10] "Megakaryocytes (48)"                  
#> [11] "Monocytes (102)"                      
#> [12] "Myeloid-derived suppressor cells (17)"
#> [13] "Natural killer T cells (24)"          
#> [14] "Neutrophils (80)"                     
#> [15] "NK cells (98)"                        
#> [16] "Nuocytes (12)"                        
#> [17] "Plasma cells (86)"                    
#> [18] "Plasmacytoid dendritic cells (58)"    
#> [19] "Red pulp macrophages (12)"            
#> [20] "T cells (107)"                        
#> [21] "T cytotoxic cells (8)"                
#> [22] "T follicular helper cells (13)"       
#> [23] "T helper cells (62)"                  
#> [24] "T memory cells (63)"                  
#> [25] "T regulatory cells (21)"