Merge markers list into one.

Merge markers collected from different DB into one 'GeneSet' object, saved a data.frame in json format under longDescription with 'TRUE' and '-' to indicate which DB each gene is from, this can be shown via jsonlite::fromJSON().

merge_markers(...)

Arguments

...

GeneSet or GeneSetCollection object to be merged

Value

A GeneSet class of union genes in the given list

Examples

data("msigdb_gobp_nk")
Markers <- merge_markers(msigdb_gobp_nk[1:3])
jsonlite::fromJSON(GSEABase::longDescription(Markers))
#>      Gene HP_ABNORMAL_NATURAL_KILLER_CELL_MORPHOLOGY
#> 1   AP3B1                                       TRUE
#> 2    CD70                                       TRUE
#> 3  DNMT3B                                       TRUE
#> 4   GATA2                                       TRUE
#> 5   IKBKG                                          -
#> 6   IL2RG                                       TRUE
#> 7    JAK3                                       TRUE
#> 8  KNSTRN                                       TRUE
#> 9    LYST                                       TRUE
#> 10  NLRC4                                       TRUE
#> 11   PGM3                                       TRUE
#> 12 PIK3CD                                       TRUE
#> 13   PRF1                                          -
#> 14  PRKCD                                          -
#> 15   RAG1                                       TRUE
#> 16   RAG2                                       TRUE
#> 17 SH2D1A                                          -
#> 18  SP110                                       TRUE
#> 19 STXBP2                                          -
#> 20 UNC13D                                          -
#> 21  WIPF1                                          -
#> 22   XIAP                                          -
#>    HP_ABNORMAL_NATURAL_KILLER_CELL_PHYSIOLOGY
#> 1                                        TRUE
#> 2                                           -
#> 3                                           -
#> 4                                           -
#> 5                                        TRUE
#> 6                                        TRUE
#> 7                                           -
#> 8                                           -
#> 9                                           -
#> 10                                       TRUE
#> 11                                          -
#> 12                                          -
#> 13                                       TRUE
#> 14                                       TRUE
#> 15                                          -
#> 16                                          -
#> 17                                       TRUE
#> 18                                          -
#> 19                                       TRUE
#> 20                                       TRUE
#> 21                                       TRUE
#> 22                                       TRUE
#>    HP_REDUCED_NATURAL_KILLER_CELL_COUNT
#> 1                                  TRUE
#> 2                                  TRUE
#> 3                                  TRUE
#> 4                                     -
#> 5                                     -
#> 6                                  TRUE
#> 7                                     -
#> 8                                  TRUE
#> 9                                     -
#> 10                                 TRUE
#> 11                                 TRUE
#> 12                                 TRUE
#> 13                                    -
#> 14                                    -
#> 15                                    -
#> 16                                    -
#> 17                                    -
#> 18                                    -
#> 19                                    -
#> 20                                    -
#> 21                                    -
#> 22                                    -