Split cells according to specific factors

Gathering cells to make the pool according to specific factors, and randomly assign the cells from the pool to pseudo-sample with the randomized cell size. (min.cells <= size <= max.cells)

pseudo_sample_list(data, by, min.cells = 0, max.cells = Inf)

Arguments

data

matrix or data.frame or other single cell expression object

by

a vector or data.frame contains factor(s) for aggregation

min.cells

num, default 0, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default Inf, the maximum size of cells aggregating to each pseudo-sample

Value

A list of cell names for each pseudo-sample

Examples

counts <- matrix(abs(rnorm(10000, 10,10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(subset = rep(c("A","B"),50),
                   level = rep(1:4, each = 25))
rownames(meta) <- 1:100
scRNA <- Seurat::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_sample_list(scRNA, by = c("subset","level"),
                   min.cells = 10, max.cells = 20)
#> $level
#> $level$`1`
#>  [1] "48" "20" "76" "30" "18" "70" "22" "94" "42" "60" "82" "52"
#> 
#> $level$`2`
#>  [1] "84"  "26"  "100" "36"  "56"  "4"   "96"  "62"  "90"  "44" 
#> 
#> $level$`3`
#>  [1] "54" "68" "50" "24" "72" "80" "8"  "86" "6"  "40" "14" "66"
#> 
#> $level$`4`
#>  [1] "2"  "10" "12" "16" "28" "32" "34" "38" "46" "58" "64" "74" "78" "88" "92"
#> [16] "98"
#> 
#> 
#> $subset
#> $subset$`1`
#>  [1] "99" "85" "11" "69" "29" "79" "17" "73" "91" "59" "31" "95"
#> 
#> $subset$`2`
#>  [1] "55" "1"  "25" "71" "67" "53" "45" "63" "43" "57" "33"
#> 
#> $subset$`3`
#>  [1] "9"  "97" "49" "87" "61" "15" "81" "27" "65" "77"
#> 
#> $subset$`4`
#>  [1] "3"  "5"  "7"  "13" "19" "21" "23" "35" "37" "39" "41" "47" "51" "75" "83"
#> [16] "89" "93"
#> 
#>