Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10, 10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(
  subset = rep(c("A", "B"), 50),
  level = rep(1:4, each = 25)
)
rownames(meta) <- 1:100
scRNA <- Seurat::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA,
  by = c("subset", "level"),
  min.cells = 10, max.cells = 20
)
#>         A.1_1     B.1_1     A.2_1     B.2_1     A.3_1     B.3_1     A.4_1
#> 1   105.91176 110.42702 155.72733  99.90910 210.77277 161.14869 142.24226
#> 2   106.00023 147.41837 101.20695 121.97934 100.07721 104.22823 145.27347
#> 3   208.20891 130.70789 127.85601 183.40237 163.83196 119.90250 171.79139
#> 4   131.23962 136.53124 118.76028 150.15218 134.27037 107.23935 181.24316
#> 5   126.61486 172.13618 199.39078 172.04638 122.88489 185.38641 160.40936
#> 6   134.33846 128.42261  81.94381 164.06651 133.30050 120.12776 148.74271
#> 7   148.30546 125.86232  85.59916 189.29447 221.81124 119.08957 128.01334
#> 8   146.87152 138.11738 145.59271 139.81761 192.27869  97.82896 157.37949
#> 9   158.00773 155.51552 155.80225 137.81084 159.29092 136.82731 203.87340
#> 10  137.72646 100.75796 115.99173 162.42077 129.81715 165.46959 163.82209
#> 11  112.90143 174.58821 111.95486 152.76161 168.22174 144.68457 132.01344
#> 12  160.82751 128.92682 150.81430 170.25132 195.85063 144.86176 119.25919
#> 13  148.61913 128.63753 158.36599 165.74383 162.89555 164.37542 106.84201
#> 14  137.03840 147.99668 145.26191 126.92195 131.71397 135.20403 144.01278
#> 15  187.15623 131.58797  72.63771 153.66145 149.45266 144.22387 110.78561
#> 16  120.71013 151.81705 159.48723 172.07079 163.74061 120.25867 174.23975
#> 17  153.03068 140.26743 141.11670 200.87978 147.83118 127.52483 113.39986
#> 18  126.07531 124.91760 164.88965 189.81047 160.37015 142.65819 192.21743
#> 19  156.74928 178.77210 135.14335 167.59813 184.92739 112.74272 148.34471
#> 20  188.03870  96.81727 110.81639 147.97431 149.02454 105.90943 125.96023
#> 21  179.43313 159.87605 134.92588 170.05228 140.30340 145.32332 129.79092
#> 22  149.21920 120.79703 180.86749 180.83962 140.48870 175.90758 138.67522
#> 23  145.37386 121.58304 135.44017 126.39009 121.93586 111.74362 129.32558
#> 24  164.41698 153.05245 105.58512 188.18143 123.78753 127.33173 189.23267
#> 25  180.10365 158.76302 133.16359 179.18791 147.94068 162.10854 126.08967
#> 26  128.55720 121.31901 150.49842 152.68118 159.15857  96.85964 108.91424
#> 27  138.34327 145.32930 159.54385 108.75014 128.16561 111.28785 145.40552
#> 28  134.93959 146.85046 107.98389 136.17190 138.15638 131.28817 156.19435
#> 29  156.01075 156.36843 150.75093 143.24332 167.80359 150.66412 134.22274
#> 30  124.80094 123.78939 140.23992 137.81060 159.25063 118.89526 124.03832
#> 31  190.37866 171.03239 138.13907 131.87937 152.58801 140.57055 122.92634
#> 32  182.10148 128.58939 167.10662 172.97964 111.78932 163.17960 172.54343
#> 33  169.78082 151.44339 127.91172 102.85404 121.42860  95.30608 115.85687
#> 34  191.34577 199.20735 134.96091 133.16652 175.76301 130.18426 150.02424
#> 35  129.73342 165.95347 114.36104 148.09839 152.88106 122.90250 103.68992
#> 36  178.78169 100.73153 154.97782 140.36042 209.69195 153.95435 139.41973
#> 37  115.56067 138.05353 125.52904 151.22730 148.84068  85.91801 122.69166
#> 38  135.88681 112.24229 116.15911 152.96993 124.25536 147.72467 134.41624
#> 39  188.77069 151.92047 123.69099  88.79346  94.31779 113.99300 167.29779
#> 40   94.14044 121.06053 116.67648 175.29122 162.64616 144.00890 159.55962
#> 41  156.89299  93.31268 162.79964 170.71756 181.01362 115.55790 123.71551
#> 42  152.37948 150.93221 173.21055 169.68559 145.71363 137.03163 119.35803
#> 43  165.10556 105.22412 186.92333 134.56674 185.73051 179.69479 181.37354
#> 44  140.67688  99.44932 121.88032 155.29901 138.63042 133.75127 123.96100
#> 45  135.42602 141.11117 183.59256 176.63372 138.40331 126.57600 112.62061
#> 46  137.08155 138.36076 150.38482 128.68998 148.76050 110.85141  85.35942
#> 47  136.05306 167.81784 166.18843 142.57926 140.95294 100.87458 106.41150
#> 48  196.77008 128.14698 111.40870 145.00777 175.93092 216.06494 186.38593
#> 49  169.26335  93.32145 159.84171 157.62881 159.87159 111.95983 115.65153
#> 50  123.16823 164.13977 122.58283 157.95917 178.97696 182.87857 134.71275
#> 51  115.87324 146.37802 145.76365 126.34509 174.89261 166.98746 112.03460
#> 52  115.31447 114.68550 130.97585 118.53422 173.81467 120.78035 139.44888
#> 53  134.61946 112.71875 153.84558 180.67600 116.56604 222.24101 168.96568
#> 54  125.46157 146.38683 132.73498 164.47653 149.66041 107.16486  85.78384
#> 55  190.98562 148.38304 122.85945 131.32794 180.98884 162.40508 111.39557
#> 56  190.49598 135.85220 134.44715 155.36249 152.13559 156.53479 111.94918
#> 57  159.14242 151.77313 190.03612 184.04244 147.27109 166.56812 150.12040
#> 58  122.81487 141.47936 142.53718 108.68591 166.08696 198.88605 142.97106
#> 59  178.14959 173.96234 192.12717 167.32634 186.95416 172.72522 124.61639
#> 60  111.57262 142.97234 151.80735 125.69987 140.69785 106.38119 129.03764
#> 61  179.74690  83.35381  98.18648 206.40746 172.43914 112.77920 142.59234
#> 62  155.15904  86.52679 136.68214 129.52290 139.10412 135.40661 189.87821
#> 63  140.80345 166.97024 101.89124  79.96980 164.52413 148.46827 174.96357
#> 64  141.52425 203.09023 129.70545 166.70018 171.77700 134.33004  89.70566
#> 65  120.87924 174.51196 158.67953 144.63972 138.75410 138.17448 126.07355
#> 66   91.52151 148.05415 156.64397 125.49262 137.24197 147.43694 107.93966
#> 67  188.79890 104.85025 161.00582 138.05899 135.55546 158.14230 137.25484
#> 68  132.58372 116.77321 172.71157 169.35136 192.93094 102.43099 158.53804
#> 69  168.53940 122.11936 162.16585 153.63971 143.82239  96.85905 153.21139
#> 70  151.19526 199.43215 114.83124 145.00839 127.16915  87.64389 133.31117
#> 71  149.08833 148.68745 175.77670 133.39222 177.31061  96.61203 155.59550
#> 72  136.27335 151.62919 133.34158 148.56650 152.08391 102.98783 175.78517
#> 73  167.60765 158.69247 133.68261 148.35684 168.62654 110.26962 174.32784
#> 74  139.74617 118.58081 178.58074 159.71300 170.01393 134.65444 139.20531
#> 75  164.52050 106.86867 120.24235 133.56451 151.16300 130.30863 139.13922
#> 76  144.57515 120.76356 180.23050 176.33918 174.51550 127.07078 161.71993
#> 77  165.14964 162.69874 151.25528 176.74318 170.00252 131.07203 144.49581
#> 78  135.80540 106.70303 142.92947 206.70068 178.62024 141.57285 155.96611
#> 79  143.30121 144.14284 135.72516 178.44511 143.49356 124.19537 114.36051
#> 80  133.68474 145.10129 194.53654 103.40083 143.50972 126.82633 139.41081
#> 81  137.00520 149.12954 123.20336 154.43080 134.51911 141.21258  78.38973
#> 82  134.57129 113.91629 191.79568 117.92653 114.66522 138.30311 143.55949
#> 83  190.61963 101.29319 110.67995 116.34742 132.82159 124.00169 114.98809
#> 84  158.80430 161.54654 134.53722 203.58682 102.61839 141.45948 129.94709
#> 85  164.94246 185.63819 222.87825 170.07353 148.28082 175.71812 150.81833
#> 86  258.88380 195.52649  95.99967 161.55509  96.66332 126.03359 123.98822
#> 87  136.03487 121.37706 168.55767 128.27978 146.04605 154.04452 100.97997
#> 88  126.09347 182.86631 158.52636 152.20678 142.49093 134.52390 104.97569
#> 89  208.39445 128.32704 107.21256 125.30336 155.05634 174.08407 106.16508
#> 90  120.68076 187.30059 128.06646 109.85259 169.05571 168.85879 143.19400
#> 91  126.65627 162.59056 122.75685 189.12158 110.12944 152.00831 197.32462
#> 92  159.58502 107.09251 119.94502 109.52449 127.99615 156.41751 155.42231
#> 93  246.29957 146.63745 167.25674 140.97440 155.54959 120.69606 102.84222
#> 94  183.01157 109.38049 154.81736 163.33793 158.81560 145.46819 157.88180
#> 95  157.77835 118.45149  75.53369 164.41971 176.79544 108.99972 167.85574
#> 96  123.88769  96.59859 135.08836 114.79898 152.17591 115.04677 125.45383
#> 97  131.74028 135.77598 105.92576 146.58026 118.78940 154.29525 185.65868
#> 98  125.20520 140.23872 177.24670 119.76509  88.43644  94.20591 174.26785
#> 99  153.75691 173.49211 133.98172 160.96738 132.29124 174.49535 112.09459
#> 100 103.09679 171.88165 146.87039 174.05368 193.08475 158.23316 129.62633
#>         B.4_1
#> 1   209.77036
#> 2   227.70320
#> 3   171.27631
#> 4   112.57398
#> 5   201.51292
#> 6   135.18954
#> 7   160.49214
#> 8    85.66387
#> 9   123.67483
#> 10  123.04768
#> 11  159.37932
#> 12  156.40308
#> 13  132.06926
#> 14  197.47444
#> 15  143.02287
#> 16  186.81124
#> 17  148.08722
#> 18  137.24049
#> 19  133.86999
#> 20  142.87543
#> 21  199.26742
#> 22  151.83791
#> 23  107.13738
#> 24  124.34732
#> 25   82.66175
#> 26  123.68091
#> 27  167.48466
#> 28  162.56377
#> 29  140.13246
#> 30  139.41128
#> 31  153.41071
#> 32  120.61083
#> 33  169.74816
#> 34  107.32848
#> 35  146.89121
#> 36  184.71261
#> 37  213.37964
#> 38  165.71661
#> 39  129.36849
#> 40  167.60885
#> 41  180.03616
#> 42  114.66379
#> 43  122.13964
#> 44  145.87340
#> 45  210.58680
#> 46  152.73808
#> 47  126.50228
#> 48  202.25646
#> 49  126.94451
#> 50  150.91379
#> 51  114.54251
#> 52  157.48404
#> 53  164.39835
#> 54  200.44426
#> 55  178.68343
#> 56  155.31976
#> 57  163.57047
#> 58  152.70957
#> 59  107.03276
#> 60  128.91776
#> 61  174.77588
#> 62  130.65940
#> 63  140.10556
#> 64  172.60744
#> 65  193.29116
#> 66  155.07999
#> 67  144.61375
#> 68  112.88425
#> 69  169.33751
#> 70  164.38746
#> 71  163.18167
#> 72  237.03887
#> 73  133.53539
#> 74  136.86243
#> 75  118.97868
#> 76  204.36700
#> 77  183.21646
#> 78  144.96527
#> 79  125.42674
#> 80  147.27168
#> 81  143.00222
#> 82  166.42174
#> 83  202.49828
#> 84  175.89736
#> 85  152.72314
#> 86  128.18264
#> 87  145.08966
#> 88  173.00538
#> 89  135.18462
#> 90  208.93791
#> 91  151.11363
#> 92  150.40042
#> 93  161.42606
#> 94  195.42906
#> 95  171.85000
#> 96  188.26891
#> 97  144.63852
#> 98  152.30631
#> 99  144.34139
#> 100 167.83235