DE analysis pipeline

Standard DE analysis by using edgeR and limma::voom pipeline

de_analysis(
  dge,
  group_col,
  target_group,
  normalize = TRUE,
  group = FALSE,
  filter = c(10, 10),
  plot = FALSE,
  lfc = 0,
  p = 0.05,
  markers = NULL,
  gene_id = "SYMBOL",
  slot = "counts",
  batch = NULL,
  summary = TRUE,
  ...
)

Arguments

dge

DGEList object for DE analysis, including expr and samples info

group_col

character, column name of coldata to specify the DE comparisons

target_group

pattern, specify the group of interest, e.g. NK

normalize

logical, if the expr in data is raw counts needs to be normalized

group

logical, TRUE to separate samples into only 2 groups: `target_group`` and 'Others'; FALSE to set each level as a group

filter

a vector of 2 numbers, filter condition to remove low expression genes, the 1st for min.counts (if normalize = TRUE) or CPM/TPM (if normalize = FALSE), the 2nd for samples size 'large.n'

plot

logical, if to make plots to show QC before and after filtration

lfc

num, cutoff of logFC for DE analysis

p

num, cutoff of p value for DE analysis and permutation test if feature_selection = "rankproduct"

markers

vector, a vector of gene names, listed the gene symbols to be kept anyway after filtration. Default 'NULL' means no special genes need to be kept.

gene_id

character, specify the gene ID target_group of rownames of expression data when markers is not NULL, could be one of 'ENSEMBL', 'SYMBOL', 'ENTREZ'..., default 'SYMBOL'

slot

character, specify which slot to use for DGEList, default 'counts'

batch

vector of character, column name(s) of coldata to be treated as batch effect factor, default NULL

summary

logical, if to show the summary of DE analysis

...

omitted

Value

MArrayLM object generated by limma::treat()

Examples

dge <- edgeR::DGEList(counts = Biobase::exprs(mastR::im_data_6),
                      samples = Biobase::pData(mastR::im_data_6))
de_analysis(dge, group_col = "celltype.ch1", target_group = "NK")
#> No group or design set. Assuming all samples belong to one group.
#>        NK-Neutrophils NK-Monocytes NK-B.cells NK-CD4 NK-CD8
#> Down             2530         2201       1673   1494   1099
#> NotSig            737         1711       2738   2978   3727
#> Up               2805         2160       1661   1600   1246
#> An object of class "DGEList"
#> $counts
#>                 GSM1479438 GSM1479439 GSM1479440 GSM1479441 GSM1479442
#> ENSG00000000419         41         18         12         14         20
#> ENSG00000000457         32          9         24         19         16
#> ENSG00000000938       2804       1711        105          1         56
#> ENSG00000001084          2         34         14         20         24
#> ENSG00000001167        160         74         75         57         67
#>                 GSM1479443 GSM1479499 GSM1479500 GSM1479501 GSM1479502
#> ENSG00000000419         15         48         23         16         24
#> ENSG00000000457         14         34          8         20         24
#> ENSG00000000938        560       2085       1722        156          1
#> ENSG00000001084         30          2         35         17         21
#> ENSG00000001167         48        200         80         54         51
#>                 GSM1479503 GSM1479504 GSM1479506 GSM1479507 GSM1479508
#> ENSG00000000419         20         18         39         24         18
#> ENSG00000000457         17         16         40          9         18
#> ENSG00000000938         43        657       2406       1672        150
#> ENSG00000001084         20         38          2         35         19
#> ENSG00000001167         56         52        194         72         65
#>                 GSM1479509 GSM1479510 GSM1479511 GSM1479520 GSM1479521
#> ENSG00000000419         19         25         21         45         19
#> ENSG00000000457         18         12         15         33          7
#> ENSG00000000938         22        176        573       2963       1529
#> ENSG00000001084         20         24         36          4         28
#> ENSG00000001167         51         54         59        197         65
#>                 GSM1479522 GSM1479523 GSM1479524 GSM1479525
#> ENSG00000000419         20         20         30         21
#> ENSG00000000457         17         18         14         15
#> ENSG00000000938        217          2         63        494
#> ENSG00000001084         17         15         19         30
#> ENSG00000001167         62         58         63         53
#> 6067 more rows ...
#> 
#> $samples
#>                  group lib.size norm.factors  title geo_accession
#> GSM1479438 Neutrophils  1421890    0.5840488 lib226    GSM1479438
#> GSM1479439   Monocytes   982216    0.9428708 lib227    GSM1479439
#> GSM1479440     B.cells   778374    1.0757535 lib228    GSM1479440
#> GSM1479441         CD4   715624    1.1080012 lib229    GSM1479441
#> GSM1479442         CD8   802285    1.0720010 lib230    GSM1479442
#>                           status submission_date last_update_date type
#> GSM1479438 Public on Jan 06 2015     Aug 14 2014      May 15 2019  SRA
#> GSM1479439 Public on Jan 06 2015     Aug 14 2014      May 15 2019  SRA
#> GSM1479440 Public on Jan 06 2015     Aug 14 2014      May 15 2019  SRA
#> GSM1479441 Public on Jan 06 2015     Aug 14 2014      May 15 2019  SRA
#> GSM1479442 Public on Jan 06 2015     Aug 14 2014      May 15 2019  SRA
#>            channel_count source_name_ch1 organism_ch1 characteristics_ch1
#> GSM1479438             1     Whole Blood Homo sapiens             age: 32
#> GSM1479439             1     Whole Blood Homo sapiens             age: 32
#> GSM1479440             1     Whole Blood Homo sapiens             age: 32
#> GSM1479441             1     Whole Blood Homo sapiens             age: 32
#> GSM1479442             1     Whole Blood Homo sapiens             age: 32
#>            characteristics_ch1.1 characteristics_ch1.2
#> GSM1479438   cellcount: 14737500 celltype: Neutrophils
#> GSM1479439    cellcount: 2000000   celltype: Monocytes
#> GSM1479440    cellcount: 1012333     celltype: B-cells
#> GSM1479441    cellcount: 1071990         celltype: CD4
#> GSM1479442    cellcount: 1011154         celltype: CD8
#>                   characteristics_ch1.3          characteristics_ch1.4
#> GSM1479438 collectiondate: June 26 2012 diseasestatus: Healthy Control
#> GSM1479439 collectiondate: June 26 2012 diseasestatus: Healthy Control
#> GSM1479440 collectiondate: June 26 2012 diseasestatus: Healthy Control
#> GSM1479441 collectiondate: June 26 2012 diseasestatus: Healthy Control
#> GSM1479442 collectiondate: June 26 2012 diseasestatus: Healthy Control
#>            characteristics_ch1.5 characteristics_ch1.6 characteristics_ch1.7
#> GSM1479438           donorid: 44             gender: F              index: 5
#> GSM1479439           donorid: 44             gender: F             index: 23
#> GSM1479440           donorid: 44             gender: F              index: 6
#> GSM1479441           donorid: 44             gender: F             index: 25
#> GSM1479442           donorid: 44             gender: F             index: 12
#>            characteristics_ch1.8      characteristics_ch1.9
#> GSM1479438        race: Hispanic samplename: 44_Neutrophils
#> GSM1479439        race: Hispanic   samplename: 44_Monocytes
#> GSM1479440        race: Hispanic      samplename: 44_Bcells
#> GSM1479441        race: Hispanic        samplename: 44_CD4T
#> GSM1479442        race: Hispanic        samplename: 44_CD8T
#>            characteristics_ch1.10             characteristics_ch1.11
#> GSM1479438             smoker: -- time since last flare (months): --
#> GSM1479439             smoker: -- time since last flare (months): --
#> GSM1479440             smoker: -- time since last flare (months): --
#> GSM1479441             smoker: -- time since last flare (months): --
#> GSM1479442             smoker: -- time since last flare (months): --
#>                 characteristics_ch1.12             characteristics_ch1.13
#> GSM1479438 time since steroid dose: -- time since symptom onset (yrs): --
#> GSM1479439 time since steroid dose: -- time since symptom onset (yrs): --
#> GSM1479440 time since steroid dose: -- time since symptom onset (yrs): --
#> GSM1479441 time since steroid dose: -- time since symptom onset (yrs): --
#> GSM1479442 time since steroid dose: -- time since symptom onset (yrs): --
#>               characteristics_ch1.14 molecule_ch1
#> GSM1479438 years since diagnosis: --    total RNA
#> GSM1479439 years since diagnosis: --    total RNA
#> GSM1479440 years since diagnosis: --    total RNA
#> GSM1479441 years since diagnosis: --    total RNA
#> GSM1479442 years since diagnosis: --    total RNA
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     extract_protocol_ch1
#> GSM1479438 Fresh blood samples were collected from healthy subjects and subjects diagnosed type 1 diabetes, amyotrophic lateral sclerosis, and sepsis, as well as multiple sclerosis patients before and 24 hours after the first treatment with IFN-beta. At the time of blood draw, an aliquot of whole blood was collected into a Tempus tube (Invitrogen), while the remainder of the primary fresh blood sample was processed to highly pure populations of neutrophils, monocytes, B cells, CD4 T cells, CD8 T cells, and natural killer cells. RNA was extracted from each of these cell subsets, as well as the whole blood samples, and processed into RNA sequencing (RNAseq) libraries (Illumina TruSeq). Sequencing libraries were analyzed on an Illumina HiScan, with a target read depth of ~20M reads. Reads were demultiplexed, mapped to human gene models (ENSEMBL), and tabulated using HTSeq. Read count data were normalized by the TMM procedure (edgeR package).
#> GSM1479439 Fresh blood samples were collected from healthy subjects and subjects diagnosed type 1 diabetes, amyotrophic lateral sclerosis, and sepsis, as well as multiple sclerosis patients before and 24 hours after the first treatment with IFN-beta. At the time of blood draw, an aliquot of whole blood was collected into a Tempus tube (Invitrogen), while the remainder of the primary fresh blood sample was processed to highly pure populations of neutrophils, monocytes, B cells, CD4 T cells, CD8 T cells, and natural killer cells. RNA was extracted from each of these cell subsets, as well as the whole blood samples, and processed into RNA sequencing (RNAseq) libraries (Illumina TruSeq). Sequencing libraries were analyzed on an Illumina HiScan, with a target read depth of ~20M reads. Reads were demultiplexed, mapped to human gene models (ENSEMBL), and tabulated using HTSeq. Read count data were normalized by the TMM procedure (edgeR package).
#> GSM1479440 Fresh blood samples were collected from healthy subjects and subjects diagnosed type 1 diabetes, amyotrophic lateral sclerosis, and sepsis, as well as multiple sclerosis patients before and 24 hours after the first treatment with IFN-beta. At the time of blood draw, an aliquot of whole blood was collected into a Tempus tube (Invitrogen), while the remainder of the primary fresh blood sample was processed to highly pure populations of neutrophils, monocytes, B cells, CD4 T cells, CD8 T cells, and natural killer cells. RNA was extracted from each of these cell subsets, as well as the whole blood samples, and processed into RNA sequencing (RNAseq) libraries (Illumina TruSeq). Sequencing libraries were analyzed on an Illumina HiScan, with a target read depth of ~20M reads. Reads were demultiplexed, mapped to human gene models (ENSEMBL), and tabulated using HTSeq. Read count data were normalized by the TMM procedure (edgeR package).
#> GSM1479441 Fresh blood samples were collected from healthy subjects and subjects diagnosed type 1 diabetes, amyotrophic lateral sclerosis, and sepsis, as well as multiple sclerosis patients before and 24 hours after the first treatment with IFN-beta. At the time of blood draw, an aliquot of whole blood was collected into a Tempus tube (Invitrogen), while the remainder of the primary fresh blood sample was processed to highly pure populations of neutrophils, monocytes, B cells, CD4 T cells, CD8 T cells, and natural killer cells. RNA was extracted from each of these cell subsets, as well as the whole blood samples, and processed into RNA sequencing (RNAseq) libraries (Illumina TruSeq). Sequencing libraries were analyzed on an Illumina HiScan, with a target read depth of ~20M reads. Reads were demultiplexed, mapped to human gene models (ENSEMBL), and tabulated using HTSeq. Read count data were normalized by the TMM procedure (edgeR package).
#> GSM1479442 Fresh blood samples were collected from healthy subjects and subjects diagnosed type 1 diabetes, amyotrophic lateral sclerosis, and sepsis, as well as multiple sclerosis patients before and 24 hours after the first treatment with IFN-beta. At the time of blood draw, an aliquot of whole blood was collected into a Tempus tube (Invitrogen), while the remainder of the primary fresh blood sample was processed to highly pure populations of neutrophils, monocytes, B cells, CD4 T cells, CD8 T cells, and natural killer cells. RNA was extracted from each of these cell subsets, as well as the whole blood samples, and processed into RNA sequencing (RNAseq) libraries (Illumina TruSeq). Sequencing libraries were analyzed on an Illumina HiScan, with a target read depth of ~20M reads. Reads were demultiplexed, mapped to human gene models (ENSEMBL), and tabulated using HTSeq. Read count data were normalized by the TMM procedure (edgeR package).
#>            taxid_ch1
#> GSM1479438      9606
#> GSM1479439      9606
#> GSM1479440      9606
#> GSM1479441      9606
#> GSM1479442      9606
#>                                                                 data_processing
#> GSM1479438 fastq files were created from .bcl files via Illumina's CASAVA 1.8.2
#> GSM1479439 fastq files were created from .bcl files via Illumina's CASAVA 1.8.2
#> GSM1479440 fastq files were created from .bcl files via Illumina's CASAVA 1.8.2
#> GSM1479441 fastq files were created from .bcl files via Illumina's CASAVA 1.8.2
#> GSM1479442 fastq files were created from .bcl files via Illumina's CASAVA 1.8.2
#>                                                                       data_processing.1
#> GSM1479438 reads were aligned to Hg19 via Omicsoft sequence aligner (OSA) version 2.0.1
#> GSM1479439 reads were aligned to Hg19 via Omicsoft sequence aligner (OSA) version 2.0.1
#> GSM1479440 reads were aligned to Hg19 via Omicsoft sequence aligner (OSA) version 2.0.1
#> GSM1479441 reads were aligned to Hg19 via Omicsoft sequence aligner (OSA) version 2.0.1
#> GSM1479442 reads were aligned to Hg19 via Omicsoft sequence aligner (OSA) version 2.0.1
#>                                              data_processing.2
#> GSM1479438 gene counts were generated by HTSeq version 0.5.4p3
#> GSM1479439 gene counts were generated by HTSeq version 0.5.4p3
#> GSM1479440 gene counts were generated by HTSeq version 0.5.4p3
#> GSM1479441 gene counts were generated by HTSeq version 0.5.4p3
#> GSM1479442 gene counts were generated by HTSeq version 0.5.4p3
#>                                              data_processing.3
#> GSM1479438 TMMs were generated with Bioconductor package EdgeR
#> GSM1479439 TMMs were generated with Bioconductor package EdgeR
#> GSM1479440 TMMs were generated with Bioconductor package EdgeR
#> GSM1479441 TMMs were generated with Bioconductor package EdgeR
#> GSM1479442 TMMs were generated with Bioconductor package EdgeR
#>             data_processing.4
#> GSM1479438 Genome_build: Hg19
#> GSM1479439 Genome_build: Hg19
#> GSM1479440 Genome_build: Hg19
#> GSM1479441 Genome_build: Hg19
#> GSM1479442 Genome_build: Hg19
#>                                                                                    data_processing.5
#> GSM1479438 Supplementary_files_format_and_content: tab delimited file contains TMM normalized counts
#> GSM1479439 Supplementary_files_format_and_content: tab delimited file contains TMM normalized counts
#> GSM1479440 Supplementary_files_format_and_content: tab delimited file contains TMM normalized counts
#> GSM1479441 Supplementary_files_format_and_content: tab delimited file contains TMM normalized counts
#> GSM1479442 Supplementary_files_format_and_content: tab delimited file contains TMM normalized counts
#>            platform_id    contact_name                  contact_email
#> GSM1479438    GPL15456 Scott,,Presnell SPresnell@benaroyaresearch.org
#> GSM1479439    GPL15456 Scott,,Presnell SPresnell@benaroyaresearch.org
#> GSM1479440    GPL15456 Scott,,Presnell SPresnell@benaroyaresearch.org
#> GSM1479441    GPL15456 Scott,,Presnell SPresnell@benaroyaresearch.org
#> GSM1479442    GPL15456 Scott,,Presnell SPresnell@benaroyaresearch.org
#>            contact_department           contact_institute contact_address
#> GSM1479438 Systems Immunology Benaroya Research Institute 1201 Ninth Ave.
#> GSM1479439 Systems Immunology Benaroya Research Institute 1201 Ninth Ave.
#> GSM1479440 Systems Immunology Benaroya Research Institute 1201 Ninth Ave.
#> GSM1479441 Systems Immunology Benaroya Research Institute 1201 Ninth Ave.
#> GSM1479442 Systems Immunology Benaroya Research Institute 1201 Ninth Ave.
#>            contact_city contact_state contact_zip.postal_code contact_country
#> GSM1479438      Seattle            WA                   98101             USA
#> GSM1479439      Seattle            WA                   98101             USA
#> GSM1479440      Seattle            WA                   98101             USA
#> GSM1479441      Seattle            WA                   98101             USA
#> GSM1479442      Seattle            WA                   98101             USA
#>            data_row_count  instrument_model library_selection library_source
#> GSM1479438              0 Illumina HiScanSQ              cDNA transcriptomic
#> GSM1479439              0 Illumina HiScanSQ              cDNA transcriptomic
#> GSM1479440              0 Illumina HiScanSQ              cDNA transcriptomic
#> GSM1479441              0 Illumina HiScanSQ              cDNA transcriptomic
#> GSM1479442              0 Illumina HiScanSQ              cDNA transcriptomic
#>            library_strategy
#> GSM1479438          RNA-Seq
#> GSM1479439          RNA-Seq
#> GSM1479440          RNA-Seq
#> GSM1479441          RNA-Seq
#> GSM1479442          RNA-Seq
#>                                                                  relation
#> GSM1479438 BioSample: https://www.ncbi.nlm.nih.gov/biosample/SAMN02990322
#> GSM1479439 BioSample: https://www.ncbi.nlm.nih.gov/biosample/SAMN02990323
#> GSM1479440 BioSample: https://www.ncbi.nlm.nih.gov/biosample/SAMN02990324
#> GSM1479441 BioSample: https://www.ncbi.nlm.nih.gov/biosample/SAMN02990325
#> GSM1479442 BioSample: https://www.ncbi.nlm.nih.gov/biosample/SAMN02990331
#>                                                      relation.1
#> GSM1479438 SRA: https://www.ncbi.nlm.nih.gov/sra?term=SRX680546
#> GSM1479439 SRA: https://www.ncbi.nlm.nih.gov/sra?term=SRX680547
#> GSM1479440 SRA: https://www.ncbi.nlm.nih.gov/sra?term=SRX680548
#> GSM1479441 SRA: https://www.ncbi.nlm.nih.gov/sra?term=SRX680549
#> GSM1479442 SRA: https://www.ncbi.nlm.nih.gov/sra?term=SRX680550
#>            supplementary_file_1 age.ch1 cellcount.ch1 celltype.ch1
#> GSM1479438                 NONE      32      14737500  Neutrophils
#> GSM1479439                 NONE      32       2000000    Monocytes
#> GSM1479440                 NONE      32       1012333      B-cells
#> GSM1479441                 NONE      32       1071990          CD4
#> GSM1479442                 NONE      32       1011154          CD8
#>            collectiondate.ch1 diseasestatus.ch1 donorid.ch1 gender.ch1
#> GSM1479438       June 26 2012   Healthy Control          44          F
#> GSM1479439       June 26 2012   Healthy Control          44          F
#> GSM1479440       June 26 2012   Healthy Control          44          F
#> GSM1479441       June 26 2012   Healthy Control          44          F
#> GSM1479442       June 26 2012   Healthy Control          44          F
#>            index.ch1 race.ch1 samplename.ch1 smoker.ch1
#> GSM1479438         5 Hispanic 44_Neutrophils         --
#> GSM1479439        23 Hispanic   44_Monocytes         --
#> GSM1479440         6 Hispanic      44_Bcells         --
#> GSM1479441        25 Hispanic        44_CD4T         --
#> GSM1479442        12 Hispanic        44_CD8T         --
#>            time.since.last.flare..months..ch1 time.since.steroid.dose.ch1
#> GSM1479438                                 --                          --
#> GSM1479439                                 --                          --
#> GSM1479440                                 --                          --
#> GSM1479441                                 --                          --
#> GSM1479442                                 --                          --
#>            time.since.symptom.onset..yrs..ch1 years.since.diagnosis.ch1
#> GSM1479438                                 --                        --
#> GSM1479439                                 --                        --
#> GSM1479440                                 --                        --
#> GSM1479441                                 --                        --
#> GSM1479442                                 --                        --
#> 19 more rows ...
#> 
#> $original_counts
#>                 GSM1479438 GSM1479439 GSM1479440 GSM1479441 GSM1479442
#> ENSG00000000003          0          0          0          2          1
#> ENSG00000000005          0          0          0          0          0
#> ENSG00000000419         41         18         12         14         20
#> ENSG00000000457         32          9         24         19         16
#> ENSG00000000460          4          5          9          4          6
#>                 GSM1479443 GSM1479499 GSM1479500 GSM1479501 GSM1479502
#> ENSG00000000003          0          0          0          0          3
#> ENSG00000000005          0          0          0          0          0
#> ENSG00000000419         15         48         23         16         24
#> ENSG00000000457         14         34          8         20         24
#> ENSG00000000460          6          4          6         12         10
#>                 GSM1479503 GSM1479504 GSM1479506 GSM1479507 GSM1479508
#> ENSG00000000003          2          0          0          0          0
#> ENSG00000000005          0          0          0          0          0
#> ENSG00000000419         20         18         39         24         18
#> ENSG00000000457         17         16         40          9         18
#> ENSG00000000460         12         10          7          4         12
#>                 GSM1479509 GSM1479510 GSM1479511 GSM1479520 GSM1479521
#> ENSG00000000003          2          1          0          0          0
#> ENSG00000000005          0          0          0          0          0
#> ENSG00000000419         19         25         21         45         19
#> ENSG00000000457         18         12         15         33          7
#> ENSG00000000460          7          7          7          3          5
#>                 GSM1479522 GSM1479523 GSM1479524 GSM1479525
#> ENSG00000000003          0          2          1          0
#> ENSG00000000005          0          0          0          0
#> ENSG00000000419         20         20         30         21
#> ENSG00000000457         17         18         14         15
#> ENSG00000000460          9          7          7          6
#> 50040 more rows ...
#> 
#> $vfit
#> $targets
#>                  group lib.size norm.factors  title geo_accession
#> GSM1479438 Neutrophils 830453.1    0.5840488 lib226    GSM1479438
#> GSM1479439   Monocytes 926102.8    0.9428708 lib227    GSM1479439
#> GSM1479440     B.cells 837338.6    1.0757535 lib228    GSM1479440
#> GSM1479441         CD4 792912.3    1.1080012 lib229    GSM1479441
#> GSM1479442         CD8 860050.3    1.0720010 lib230    GSM1479442
#>                           status submission_date last_update_date type
#> GSM1479438 Public on Jan 06 2015     Aug 14 2014      May 15 2019  SRA
#> GSM1479439 Public on Jan 06 2015     Aug 14 2014      May 15 2019  SRA
#> GSM1479440 Public on Jan 06 2015     Aug 14 2014      May 15 2019  SRA
#> GSM1479441 Public on Jan 06 2015     Aug 14 2014      May 15 2019  SRA
#> GSM1479442 Public on Jan 06 2015     Aug 14 2014      May 15 2019  SRA
#>            channel_count source_name_ch1 organism_ch1 characteristics_ch1
#> GSM1479438             1     Whole Blood Homo sapiens             age: 32
#> GSM1479439             1     Whole Blood Homo sapiens             age: 32
#> GSM1479440             1     Whole Blood Homo sapiens             age: 32
#> GSM1479441             1     Whole Blood Homo sapiens             age: 32
#> GSM1479442             1     Whole Blood Homo sapiens             age: 32
#>            characteristics_ch1.1 characteristics_ch1.2
#> GSM1479438   cellcount: 14737500 celltype: Neutrophils
#> GSM1479439    cellcount: 2000000   celltype: Monocytes
#> GSM1479440    cellcount: 1012333     celltype: B-cells
#> GSM1479441    cellcount: 1071990         celltype: CD4
#> GSM1479442    cellcount: 1011154         celltype: CD8
#>                   characteristics_ch1.3          characteristics_ch1.4
#> GSM1479438 collectiondate: June 26 2012 diseasestatus: Healthy Control
#> GSM1479439 collectiondate: June 26 2012 diseasestatus: Healthy Control
#> GSM1479440 collectiondate: June 26 2012 diseasestatus: Healthy Control
#> GSM1479441 collectiondate: June 26 2012 diseasestatus: Healthy Control
#> GSM1479442 collectiondate: June 26 2012 diseasestatus: Healthy Control
#>            characteristics_ch1.5 characteristics_ch1.6 characteristics_ch1.7
#> GSM1479438           donorid: 44             gender: F              index: 5
#> GSM1479439           donorid: 44             gender: F             index: 23
#> GSM1479440           donorid: 44             gender: F              index: 6
#> GSM1479441           donorid: 44             gender: F             index: 25
#> GSM1479442           donorid: 44             gender: F             index: 12
#>            characteristics_ch1.8      characteristics_ch1.9
#> GSM1479438        race: Hispanic samplename: 44_Neutrophils
#> GSM1479439        race: Hispanic   samplename: 44_Monocytes
#> GSM1479440        race: Hispanic      samplename: 44_Bcells
#> GSM1479441        race: Hispanic        samplename: 44_CD4T
#> GSM1479442        race: Hispanic        samplename: 44_CD8T
#>            characteristics_ch1.10             characteristics_ch1.11
#> GSM1479438             smoker: -- time since last flare (months): --
#> GSM1479439             smoker: -- time since last flare (months): --
#> GSM1479440             smoker: -- time since last flare (months): --
#> GSM1479441             smoker: -- time since last flare (months): --
#> GSM1479442             smoker: -- time since last flare (months): --
#>                 characteristics_ch1.12             characteristics_ch1.13
#> GSM1479438 time since steroid dose: -- time since symptom onset (yrs): --
#> GSM1479439 time since steroid dose: -- time since symptom onset (yrs): --
#> GSM1479440 time since steroid dose: -- time since symptom onset (yrs): --
#> GSM1479441 time since steroid dose: -- time since symptom onset (yrs): --
#> GSM1479442 time since steroid dose: -- time since symptom onset (yrs): --
#>               characteristics_ch1.14 molecule_ch1
#> GSM1479438 years since diagnosis: --    total RNA
#> GSM1479439 years since diagnosis: --    total RNA
#> GSM1479440 years since diagnosis: --    total RNA
#> GSM1479441 years since diagnosis: --    total RNA
#> GSM1479442 years since diagnosis: --    total RNA
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     extract_protocol_ch1
#> GSM1479438 Fresh blood samples were collected from healthy subjects and subjects diagnosed type 1 diabetes, amyotrophic lateral sclerosis, and sepsis, as well as multiple sclerosis patients before and 24 hours after the first treatment with IFN-beta. At the time of blood draw, an aliquot of whole blood was collected into a Tempus tube (Invitrogen), while the remainder of the primary fresh blood sample was processed to highly pure populations of neutrophils, monocytes, B cells, CD4 T cells, CD8 T cells, and natural killer cells. RNA was extracted from each of these cell subsets, as well as the whole blood samples, and processed into RNA sequencing (RNAseq) libraries (Illumina TruSeq). Sequencing libraries were analyzed on an Illumina HiScan, with a target read depth of ~20M reads. Reads were demultiplexed, mapped to human gene models (ENSEMBL), and tabulated using HTSeq. Read count data were normalized by the TMM procedure (edgeR package).
#> GSM1479439 Fresh blood samples were collected from healthy subjects and subjects diagnosed type 1 diabetes, amyotrophic lateral sclerosis, and sepsis, as well as multiple sclerosis patients before and 24 hours after the first treatment with IFN-beta. At the time of blood draw, an aliquot of whole blood was collected into a Tempus tube (Invitrogen), while the remainder of the primary fresh blood sample was processed to highly pure populations of neutrophils, monocytes, B cells, CD4 T cells, CD8 T cells, and natural killer cells. RNA was extracted from each of these cell subsets, as well as the whole blood samples, and processed into RNA sequencing (RNAseq) libraries (Illumina TruSeq). Sequencing libraries were analyzed on an Illumina HiScan, with a target read depth of ~20M reads. Reads were demultiplexed, mapped to human gene models (ENSEMBL), and tabulated using HTSeq. Read count data were normalized by the TMM procedure (edgeR package).
#> GSM1479440 Fresh blood samples were collected from healthy subjects and subjects diagnosed type 1 diabetes, amyotrophic lateral sclerosis, and sepsis, as well as multiple sclerosis patients before and 24 hours after the first treatment with IFN-beta. At the time of blood draw, an aliquot of whole blood was collected into a Tempus tube (Invitrogen), while the remainder of the primary fresh blood sample was processed to highly pure populations of neutrophils, monocytes, B cells, CD4 T cells, CD8 T cells, and natural killer cells. RNA was extracted from each of these cell subsets, as well as the whole blood samples, and processed into RNA sequencing (RNAseq) libraries (Illumina TruSeq). Sequencing libraries were analyzed on an Illumina HiScan, with a target read depth of ~20M reads. Reads were demultiplexed, mapped to human gene models (ENSEMBL), and tabulated using HTSeq. Read count data were normalized by the TMM procedure (edgeR package).
#> GSM1479441 Fresh blood samples were collected from healthy subjects and subjects diagnosed type 1 diabetes, amyotrophic lateral sclerosis, and sepsis, as well as multiple sclerosis patients before and 24 hours after the first treatment with IFN-beta. At the time of blood draw, an aliquot of whole blood was collected into a Tempus tube (Invitrogen), while the remainder of the primary fresh blood sample was processed to highly pure populations of neutrophils, monocytes, B cells, CD4 T cells, CD8 T cells, and natural killer cells. RNA was extracted from each of these cell subsets, as well as the whole blood samples, and processed into RNA sequencing (RNAseq) libraries (Illumina TruSeq). Sequencing libraries were analyzed on an Illumina HiScan, with a target read depth of ~20M reads. Reads were demultiplexed, mapped to human gene models (ENSEMBL), and tabulated using HTSeq. Read count data were normalized by the TMM procedure (edgeR package).
#> GSM1479442 Fresh blood samples were collected from healthy subjects and subjects diagnosed type 1 diabetes, amyotrophic lateral sclerosis, and sepsis, as well as multiple sclerosis patients before and 24 hours after the first treatment with IFN-beta. At the time of blood draw, an aliquot of whole blood was collected into a Tempus tube (Invitrogen), while the remainder of the primary fresh blood sample was processed to highly pure populations of neutrophils, monocytes, B cells, CD4 T cells, CD8 T cells, and natural killer cells. RNA was extracted from each of these cell subsets, as well as the whole blood samples, and processed into RNA sequencing (RNAseq) libraries (Illumina TruSeq). Sequencing libraries were analyzed on an Illumina HiScan, with a target read depth of ~20M reads. Reads were demultiplexed, mapped to human gene models (ENSEMBL), and tabulated using HTSeq. Read count data were normalized by the TMM procedure (edgeR package).
#>            taxid_ch1
#> GSM1479438      9606
#> GSM1479439      9606
#> GSM1479440      9606
#> GSM1479441      9606
#> GSM1479442      9606
#>                                                                 data_processing
#> GSM1479438 fastq files were created from .bcl files via Illumina's CASAVA 1.8.2
#> GSM1479439 fastq files were created from .bcl files via Illumina's CASAVA 1.8.2
#> GSM1479440 fastq files were created from .bcl files via Illumina's CASAVA 1.8.2
#> GSM1479441 fastq files were created from .bcl files via Illumina's CASAVA 1.8.2
#> GSM1479442 fastq files were created from .bcl files via Illumina's CASAVA 1.8.2
#>                                                                       data_processing.1
#> GSM1479438 reads were aligned to Hg19 via Omicsoft sequence aligner (OSA) version 2.0.1
#> GSM1479439 reads were aligned to Hg19 via Omicsoft sequence aligner (OSA) version 2.0.1
#> GSM1479440 reads were aligned to Hg19 via Omicsoft sequence aligner (OSA) version 2.0.1
#> GSM1479441 reads were aligned to Hg19 via Omicsoft sequence aligner (OSA) version 2.0.1
#> GSM1479442 reads were aligned to Hg19 via Omicsoft sequence aligner (OSA) version 2.0.1
#>                                              data_processing.2
#> GSM1479438 gene counts were generated by HTSeq version 0.5.4p3
#> GSM1479439 gene counts were generated by HTSeq version 0.5.4p3
#> GSM1479440 gene counts were generated by HTSeq version 0.5.4p3
#> GSM1479441 gene counts were generated by HTSeq version 0.5.4p3
#> GSM1479442 gene counts were generated by HTSeq version 0.5.4p3
#>                                              data_processing.3
#> GSM1479438 TMMs were generated with Bioconductor package EdgeR
#> GSM1479439 TMMs were generated with Bioconductor package EdgeR
#> GSM1479440 TMMs were generated with Bioconductor package EdgeR
#> GSM1479441 TMMs were generated with Bioconductor package EdgeR
#> GSM1479442 TMMs were generated with Bioconductor package EdgeR
#>             data_processing.4
#> GSM1479438 Genome_build: Hg19
#> GSM1479439 Genome_build: Hg19
#> GSM1479440 Genome_build: Hg19
#> GSM1479441 Genome_build: Hg19
#> GSM1479442 Genome_build: Hg19
#>                                                                                    data_processing.5
#> GSM1479438 Supplementary_files_format_and_content: tab delimited file contains TMM normalized counts
#> GSM1479439 Supplementary_files_format_and_content: tab delimited file contains TMM normalized counts
#> GSM1479440 Supplementary_files_format_and_content: tab delimited file contains TMM normalized counts
#> GSM1479441 Supplementary_files_format_and_content: tab delimited file contains TMM normalized counts
#> GSM1479442 Supplementary_files_format_and_content: tab delimited file contains TMM normalized counts
#>            platform_id    contact_name                  contact_email
#> GSM1479438    GPL15456 Scott,,Presnell SPresnell@benaroyaresearch.org
#> GSM1479439    GPL15456 Scott,,Presnell SPresnell@benaroyaresearch.org
#> GSM1479440    GPL15456 Scott,,Presnell SPresnell@benaroyaresearch.org
#> GSM1479441    GPL15456 Scott,,Presnell SPresnell@benaroyaresearch.org
#> GSM1479442    GPL15456 Scott,,Presnell SPresnell@benaroyaresearch.org
#>            contact_department           contact_institute contact_address
#> GSM1479438 Systems Immunology Benaroya Research Institute 1201 Ninth Ave.
#> GSM1479439 Systems Immunology Benaroya Research Institute 1201 Ninth Ave.
#> GSM1479440 Systems Immunology Benaroya Research Institute 1201 Ninth Ave.
#> GSM1479441 Systems Immunology Benaroya Research Institute 1201 Ninth Ave.
#> GSM1479442 Systems Immunology Benaroya Research Institute 1201 Ninth Ave.
#>            contact_city contact_state contact_zip.postal_code contact_country
#> GSM1479438      Seattle            WA                   98101             USA
#> GSM1479439      Seattle            WA                   98101             USA
#> GSM1479440      Seattle            WA                   98101             USA
#> GSM1479441      Seattle            WA                   98101             USA
#> GSM1479442      Seattle            WA                   98101             USA
#>            data_row_count  instrument_model library_selection library_source
#> GSM1479438              0 Illumina HiScanSQ              cDNA transcriptomic
#> GSM1479439              0 Illumina HiScanSQ              cDNA transcriptomic
#> GSM1479440              0 Illumina HiScanSQ              cDNA transcriptomic
#> GSM1479441              0 Illumina HiScanSQ              cDNA transcriptomic
#> GSM1479442              0 Illumina HiScanSQ              cDNA transcriptomic
#>            library_strategy
#> GSM1479438          RNA-Seq
#> GSM1479439          RNA-Seq
#> GSM1479440          RNA-Seq
#> GSM1479441          RNA-Seq
#> GSM1479442          RNA-Seq
#>                                                                  relation
#> GSM1479438 BioSample: https://www.ncbi.nlm.nih.gov/biosample/SAMN02990322
#> GSM1479439 BioSample: https://www.ncbi.nlm.nih.gov/biosample/SAMN02990323
#> GSM1479440 BioSample: https://www.ncbi.nlm.nih.gov/biosample/SAMN02990324
#> GSM1479441 BioSample: https://www.ncbi.nlm.nih.gov/biosample/SAMN02990325
#> GSM1479442 BioSample: https://www.ncbi.nlm.nih.gov/biosample/SAMN02990331
#>                                                      relation.1
#> GSM1479438 SRA: https://www.ncbi.nlm.nih.gov/sra?term=SRX680546
#> GSM1479439 SRA: https://www.ncbi.nlm.nih.gov/sra?term=SRX680547
#> GSM1479440 SRA: https://www.ncbi.nlm.nih.gov/sra?term=SRX680548
#> GSM1479441 SRA: https://www.ncbi.nlm.nih.gov/sra?term=SRX680549
#> GSM1479442 SRA: https://www.ncbi.nlm.nih.gov/sra?term=SRX680550
#>            supplementary_file_1 age.ch1 cellcount.ch1 celltype.ch1
#> GSM1479438                 NONE      32      14737500  Neutrophils
#> GSM1479439                 NONE      32       2000000    Monocytes
#> GSM1479440                 NONE      32       1012333      B-cells
#> GSM1479441                 NONE      32       1071990          CD4
#> GSM1479442                 NONE      32       1011154          CD8
#>            collectiondate.ch1 diseasestatus.ch1 donorid.ch1 gender.ch1
#> GSM1479438       June 26 2012   Healthy Control          44          F
#> GSM1479439       June 26 2012   Healthy Control          44          F
#> GSM1479440       June 26 2012   Healthy Control          44          F
#> GSM1479441       June 26 2012   Healthy Control          44          F
#> GSM1479442       June 26 2012   Healthy Control          44          F
#>            index.ch1 race.ch1 samplename.ch1 smoker.ch1
#> GSM1479438         5 Hispanic 44_Neutrophils         --
#> GSM1479439        23 Hispanic   44_Monocytes         --
#> GSM1479440         6 Hispanic      44_Bcells         --
#> GSM1479441        25 Hispanic        44_CD4T         --
#> GSM1479442        12 Hispanic        44_CD8T         --
#>            time.since.last.flare..months..ch1 time.since.steroid.dose.ch1
#> GSM1479438                                 --                          --
#> GSM1479439                                 --                          --
#> GSM1479440                                 --                          --
#> GSM1479441                                 --                          --
#> GSM1479442                                 --                          --
#>            time.since.symptom.onset..yrs..ch1 years.since.diagnosis.ch1
#> GSM1479438                                 --                        --
#> GSM1479439                                 --                        --
#> GSM1479440                                 --                        --
#> GSM1479441                                 --                        --
#> GSM1479442                                 --                        --
#> 19 more rows ...
#> 
#> $E
#>                 GSM1479438 GSM1479439 GSM1479440 GSM1479441 GSM1479442
#> ENSG00000000419   5.643067   4.320208   3.899971  4.1927460   4.575057
#> ENSG00000000457   5.290395   3.358682   4.870825  4.6201673   4.261899
#> ENSG00000000938  11.721555  10.851800   6.977214  0.9197275   6.037684
#> ENSG00000001084   1.589956   5.219279   4.114096  4.6923171   4.832215
#> ENSG00000001167   7.594457   6.329923   6.494520  6.1802551   6.294321
#>                 GSM1479443 GSM1479499 GSM1479500 GSM1479501 GSM1479502
#> ENSG00000000419   4.344363   5.806401   4.720269   4.348242  4.8657561
#> ENSG00000000457   4.248148   5.315013   3.253143   4.661400  4.8657561
#> ENSG00000000938   9.520737  11.232666  10.915969   7.593866  0.8360088
#> ENSG00000001084   5.320904   1.528416   5.315428   4.433131  4.6773110
#> ENSG00000001167   5.990080   7.853947   6.496597   6.072032  5.9375468
#>                 GSM1479503 GSM1479504 GSM1479506 GSM1479507 GSM1479508
#> ENSG00000000419   4.712787   4.477929   5.432015   4.741275   4.414937
#> ENSG00000000457   4.484518   4.312870   5.468084   3.374493   4.414937
#> ENSG00000000938   5.798178   9.629323  11.360954  10.834356   7.439103
#> ENSG00000001084   4.712787   5.535263   1.450162   5.276312   4.490886
#> ENSG00000001167   6.175414   5.982721   7.731860   6.306474   6.238907
#>                 GSM1479509 GSM1479510 GSM1479511 GSM1479520 GSM1479521
#> ENSG00000000419   4.611609   4.965249   4.620666   5.641953   4.427209
#> ENSG00000000457   4.535660   3.936680   4.148597   5.200247   3.048697
#> ENSG00000000938   4.818060   7.756348   9.358051  11.667245  10.720651
#> ENSG00000001084   4.683758   4.907534   5.384226   2.304083   4.974696
#> ENSG00000001167   6.012707   6.061008   6.089219   7.759867   6.175229
#>                 GSM1479522 GSM1479523 GSM1479524 GSM1479525
#> ENSG00000000419   4.632546   4.692640   5.208811   4.684634
#> ENSG00000000457   4.404277   4.544541   4.136054   4.212566
#> ENSG00000000938   8.039865   1.657016   6.266758   9.208196
#> ENSG00000001084   4.404277   4.289284   4.563476   5.189107
#> ENSG00000001167   6.240778   6.205453   6.266758   5.999836
#> 6067 more rows ...
#> 
#> $weights
#>          [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 23.52843 18.48717 16.13976 17.30505 19.65489 16.58688 23.87302 18.14619
#> [2,] 21.81077 11.50456 17.79285 17.63470 15.67051 14.85962 22.16243 11.33712
#> [3,] 15.99580 17.33898 28.92827 11.33712 28.15291 22.83036 15.84055 17.49110
#> [4,] 11.33712 22.01323 16.35964 17.27292 19.05162 21.33154 11.33712 21.69631
#> [5,] 28.45129 28.03019 27.06651 25.71295 26.94764 25.02412 28.27501 27.79304
#>          [,9]    [,10]    [,11]    [,12]    [,13]    [,14]    [,15]    [,16]
#> [1,] 15.85911 17.84322 18.90439 17.35320 24.32060 18.39060 16.44432 17.36002
#> [2,] 17.48626 18.17182 14.90869 15.52740 22.60091 11.43514 18.11891 17.68869
#> [3,] 29.00207 11.33712 27.56602 22.36070 15.64512 17.38263 28.83397 11.33712
#> [4,] 16.07418 17.81118 18.22607 22.03148 11.33712 21.92405 16.66949 17.32775
#> [5,] 26.82649 26.17149 26.23093 25.70079 28.02760 27.96524 27.31005 25.75954
#>         [,17]    [,18]    [,19]    [,20]    [,21]    [,22]    [,23]    [,24]
#> [1,] 19.25672 17.82953 24.28634 18.29810 16.02801 17.30298 19.34119 17.41836
#> [2,] 15.24790 15.95287 22.56914 11.36871 17.67316 17.63267 15.32948 15.58457
#> [3,] 27.83933 22.07950 15.65981 17.42484 28.96007 11.33712 27.90424 22.32211
#> [4,] 18.60834 22.45537 11.33712 21.83542 16.24599 17.27085 18.69976 22.08878
#> [5,] 26.56539 26.10726 28.04630 27.89798 26.97126 25.71120 26.64520 25.75578
#> 6067 more rows ...
#> 
#> $design
#>            NK Neutrophils Monocytes B.cells CD4 CD8
#> GSM1479438  0           1         0       0   0   0
#> GSM1479439  0           0         1       0   0   0
#> GSM1479440  0           0         0       1   0   0
#> GSM1479441  0           0         0       0   1   0
#> GSM1479442  0           0         0       0   0   1
#> 19 more rows ...
#> 
#> 
#> $tfit
#> $coefficients
#>                  Contrasts
#>                   NK-Neutrophils NK-Monocytes NK-B.cells      NK-CD4     NK-CD8
#>   ENSG00000000419      -1.095206  -0.01680197  0.2110171 -0.05809146 -0.3303203
#>   ENSG00000000457      -1.088967   0.97055237 -0.3572824 -0.41337101  0.0262364
#>   ENSG00000000938      -2.066211  -1.40087125  1.9171064  7.37217240  2.9648155
#>   ENSG00000001084       3.639531   0.16131804  0.9968626  0.77122334  0.6032395
#>   ENSG00000001167      -1.718854  -0.31097227 -0.2460316 -0.06739113 -0.1839269
#> 6067 more rows ...
#> 
#> $stdev.unscaled
#>                  Contrasts
#>                   NK-Neutrophils NK-Monocytes NK-B.cells    NK-CD4    NK-CD8
#>   ENSG00000000419      0.1576995    0.1676062  0.1731017 0.1696399 0.1655715
#>   ENSG00000000457      0.1654286    0.1950812  0.1738364 0.1738042 0.1802781
#>   ENSG00000000938      0.1643141    0.1597551  0.1407226 0.1822446 0.1418917
#>   ENSG00000001084      0.1828308    0.1510240  0.1633341 0.1603945 0.1574262
#>   ENSG00000001167      0.1364293    0.1367530  0.1378117 0.1393668 0.1383734
#> 6067 more rows ...
#> 
#> $sigma
#> [1] 1.0207698 0.6732638 3.3887698 0.8461538 0.6360073
#> 6067 more elements ...
#> 
#> $df.residual
#> [1] 18 18 18 18 18
#> 6067 more elements ...
#> 
#> $cov.coefficients
#>                 Contrasts
#> Contrasts        NK-Neutrophils NK-Monocytes NK-B.cells NK-CD4 NK-CD8
#>   NK-Neutrophils           0.50         0.25       0.25   0.25   0.25
#>   NK-Monocytes             0.25         0.50       0.25   0.25   0.25
#>   NK-B.cells               0.25         0.25       0.50   0.25   0.25
#>   NK-CD4                   0.25         0.25       0.25   0.50   0.25
#>   NK-CD8                   0.25         0.25       0.25   0.25   0.50
#> 
#> $pivot
#> [1] 1 2 3 4 5 6
#> 
#> $rank
#> [1] 6
#> 
#> $Amean
#> ENSG00000000419 ENSG00000000457 ENSG00000000938 ENSG00000001084 ENSG00000001167 
#>        4.749181        4.373652        7.965056        4.328704        6.437080 
#> 6067 more elements ...
#> 
#> $method
#> [1] "ls"
#> 
#> $design
#>            NK Neutrophils Monocytes B.cells CD4 CD8
#> GSM1479438  0           1         0       0   0   0
#> GSM1479439  0           0         1       0   0   0
#> GSM1479440  0           0         0       1   0   0
#> GSM1479441  0           0         0       0   1   0
#> GSM1479442  0           0         0       0   0   1
#> 19 more rows ...
#> 
#> $contrasts
#>              Contrasts
#> Levels        NK-Neutrophils NK-Monocytes NK-B.cells NK-CD4 NK-CD8
#>   NK                       1            1          1      1      1
#>   Neutrophils             -1            0          0      0      0
#>   Monocytes                0           -1          0      0      0
#>   B.cells                  0            0         -1      0      0
#>   CD4                      0            0          0     -1      0
#>   CD8                      0            0          0      0     -1
#> 
#> $df.prior
#> [1] 4.585584
#> 
#> $s2.prior
#> [1] 0.8251309
#> 
#> $s2.post
#> [1] 0.9979457 0.5287807 9.3197235 0.7381381 0.4899055
#> 6067 more elements ...
#> 
#> $df.total
#> [1] 22.58558 22.58558 22.58558 22.58558 22.58558
#> 6067 more elements ...
#> 
#> $t
#>                  Contrasts
#>                   NK-Neutrophils NK-Monocytes NK-B.cells     NK-CD4     NK-CD8
#>   ENSG00000000419      -6.952034   -0.1003498   1.220290 -0.3427920 -1.9970831
#>   ENSG00000000457      -9.052453    6.8417284  -2.826396 -3.2707098  0.2001352
#>   ENSG00000000938      -4.119061   -2.8723805   4.462527 13.2507168  6.8444597
#>   ENSG00000001084      23.170066    1.2432774   7.103783  5.5965687  4.4600928
#>   ENSG00000001167     -18.000107   -3.2488434  -2.550636 -0.6908555 -1.8990510
#> 6067 more rows ...
#> 
#> $p.value
#>                  Contrasts
#>                   NK-Neutrophils NK-Monocytes   NK-B.cells       NK-CD4
#>   ENSG00000000419   4.830559e-07 9.209521e-01 2.349437e-01 7.349249e-01
#>   ENSG00000000457   5.673866e-09 6.199504e-07 9.674042e-03 3.413958e-03
#>   ENSG00000000938   4.314141e-04 8.703347e-03 1.845372e-04 3.916390e-12
#>   ENSG00000001084   3.070262e-17 2.265173e-01 3.435909e-07 1.146976e-05
#>   ENSG00000001167   6.927190e-15 3.596825e-03 1.801611e-02 4.966960e-01
#>                  Contrasts
#>                         NK-CD8
#>   ENSG00000000419 5.800431e-02
#>   ENSG00000000457 8.431659e-01
#>   ENSG00000000938 6.161205e-07
#>   ENSG00000001084 1.856517e-04
#>   ENSG00000001167 7.040557e-02
#> 6067 more rows ...
#> 
#> $treat.lfc
#> [1] 0
#> 
#>