Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10, 10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(
  subset = rep(c("A", "B"), 50),
  level = rep(1:4, each = 25)
)
rownames(meta) <- 1:100
scRNA <- Seurat::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA,
  by = c("subset", "level"),
  min.cells = 10, max.cells = 20
)
#>         A.1_1     B.1_1     A.2_1     B.2_1     A.3_1     B.3_1     A.4_1
#> 1   162.68358 136.12225 134.52429 173.68242 151.51099 169.17612 145.43654
#> 2   151.27620 165.06225 131.68096 142.00570 137.91246 146.75946 115.95824
#> 3   131.18324 148.65388 156.80502 142.15385 152.70510 133.08713 157.30062
#> 4   154.55403 128.01320 131.76456 135.21552 159.43147 144.31896 130.94038
#> 5   184.72860 168.79638 114.24179 159.65146 150.49497 165.30848 155.84682
#> 6   152.67941 128.39910 167.27178 173.83655 166.74254 135.60706 133.78724
#> 7   169.50575 171.33773 147.24066 156.90302 219.66044 155.32981 155.67493
#> 8   136.29427 118.18376 196.41790 167.48971 162.12999 136.35596 121.08635
#> 9   149.67811 211.27416 141.02062 142.32783 152.22916 144.51714 175.11409
#> 10  170.05617 117.60180 110.89038 186.25239 181.46988 154.70926  99.90459
#> 11  168.58617 129.84050 154.24700 175.46248 151.51992 132.63148  94.14665
#> 12  189.90078 150.67149  88.75369 185.16111 114.39378 153.01290 164.80199
#> 13  155.42993 159.93274 127.84133 143.19787 174.64971  98.87976 141.39506
#> 14  152.57293 154.18846 125.12722 185.65761 189.73801 122.06085 105.82803
#> 15  154.78671 148.60767 193.75446  99.03954 137.42471  69.92650 162.24908
#> 16   80.18271 113.73111 188.65246 146.55139 126.43037 147.46774 122.11737
#> 17  150.73544 110.33691 116.47516 190.51173 214.25623 173.59567 105.94574
#> 18  119.87342 130.71180 143.05280 155.78547 143.13676  89.54248 122.55498
#> 19  124.33469 173.29762 124.75167 170.72634 175.02018 122.38818 173.52536
#> 20  201.33215 121.88494 173.89773 103.55361 122.81688 185.77560 171.25984
#> 21  142.85222 116.86383 149.68688 129.89918 181.33142 212.42698 128.64230
#> 22  116.17903 136.14568 177.41376 142.41221 150.54039 132.56154 167.40109
#> 23  125.75037 119.28305 182.77045 149.48632 136.48720  84.33838 127.38594
#> 24  139.41809 143.59679 185.72934 123.84700 196.83082 166.64849 154.24230
#> 25  200.05783 123.42282 154.33486 113.46590  97.68065 180.79158 146.16700
#> 26  155.24968 137.76379 141.11625 168.85977 163.37546 108.52148 136.08755
#> 27  169.43835 190.87124 134.90744 136.67638 144.56203 130.37815 191.74718
#> 28  166.81697 160.66284 149.67759 135.92054 165.06697 177.90088 138.40333
#> 29  130.10074 148.34966  68.89793 128.28857  84.19148 124.92686 139.38568
#> 30  192.58804 149.90752  91.78567 163.82507 192.64296 143.13197 113.68345
#> 31  171.90204 136.26616 140.10206 106.31941 138.43633 144.18148 176.91732
#> 32  150.93944 142.99695  90.48080 141.63275  99.24911 140.91034  89.01790
#> 33  122.93294 148.93024 121.04854 163.48104 111.37548 108.58043 177.38242
#> 34  219.61458 151.05559 164.28887 178.77546 113.47902 145.48917  81.06789
#> 35  104.97552 138.03448  91.56173 169.39365 142.87030 130.01257 101.21000
#> 36  123.30956 126.60158 139.28228 109.99627 141.09206 164.80872 186.05438
#> 37  193.13349 145.76585 194.99522 133.87610 120.86115 119.42864 136.91869
#> 38  141.88742 176.90007 113.76279 171.31643 130.32303 106.46540 175.26218
#> 39  167.15804 130.06069 140.61831 131.93276 168.04880 130.79667 198.31537
#> 40  194.92958 158.54434  78.99859 154.50763 171.08198 145.48865 176.01158
#> 41  121.86236 151.42306 143.45705 154.96723 117.98800 130.73992 126.49036
#> 42  168.21427 128.21552  77.81086 123.64682 180.06818 189.80156 154.73921
#> 43  128.06533  98.82991 144.53465 121.43603 133.12923 142.96861 143.86932
#> 44  135.80722 153.28210 137.79954  98.02724 146.65567 125.71577 121.50642
#> 45  141.66276 173.23611 134.25337 191.63703 132.46370 167.22479 129.34513
#> 46  204.85966 147.79408 125.68780 230.94767 152.72819 122.54355 147.16371
#> 47  147.52993 145.38598 149.80772 175.14099 163.20238 168.68897 177.54648
#> 48  184.61295 128.02223 177.92069 197.65344 134.93907 133.69987 167.55531
#> 49  138.47845 175.47605 124.15704  97.98906 108.92389 185.45354 121.87240
#> 50  117.18000 150.44193 151.44891 163.92585 199.30765 122.91095 152.85904
#> 51  154.93663 172.52410 147.10250 114.01544 143.94686 197.73096 149.92690
#> 52  109.74190 139.35904 184.56021 172.88482 171.23511 139.00832 147.05661
#> 53  158.81682 118.26057 121.76345 169.05720 151.82510 156.07622 160.51659
#> 54  174.82309 183.08128 136.50378 131.96274 161.26129 119.73084 186.71421
#> 55  103.51921  93.44037 164.83284 113.02855 137.82732 179.56688 158.02442
#> 56  140.89713 147.32007 145.39210 156.87644 133.78679 120.25537 202.40384
#> 57  134.02594 158.04816 124.66270 167.81493 165.82737 141.35054  65.47046
#> 58  176.95216 175.12089 203.59929 192.85033 167.28064 112.89900 130.60004
#> 59  210.97416 158.95213 148.91241 116.37873 135.87351 162.09732 153.16568
#> 60  211.61333 106.38122 104.14002 140.61584 112.41679 208.34239 155.82029
#> 61  166.77393 134.68593 124.31276 157.00950 130.26400 130.17055 132.34167
#> 62  159.61621 174.01855 108.09814 137.93622 184.45088 120.77301 174.78743
#> 63  145.21941 161.05014 133.36780 175.39010 111.77781 148.84862 155.96120
#> 64  112.84048 142.84025 124.76440 142.46509 114.39845 160.15844  94.89968
#> 65  131.11481 146.69649 180.48275 138.16237 140.58685 212.84301 131.78299
#> 66  151.79709  75.11254 158.56421 136.06473 126.92494 142.21807 128.10515
#> 67  167.93933 150.26303 152.04338 141.08242 165.23958 147.76556 132.44259
#> 68  172.36044 134.82455 127.39067 117.45647 155.24803 169.32331 148.99547
#> 69  132.84020 132.10055 154.59065 151.21481 122.72427 167.38376 134.58308
#> 70  145.42442 145.82371 113.60939 118.77164 149.31600 128.74860 127.46369
#> 71  172.18792 142.85589 108.32598 126.65237 131.00409 126.37627 175.56696
#> 72  157.21962 126.88021 151.52404 186.89896 136.40229 120.23469 120.34956
#> 73  160.84055 211.68568 164.81865 137.27057 138.72290 155.06082 144.60735
#> 74  178.84792 168.12984 165.56803 102.52441 191.43563 158.73490 172.92168
#> 75  166.61727 167.63298 149.82606 157.28168 125.24666 163.24449 165.10408
#> 76  146.11293 178.00404 198.37759 186.93136 159.87212 173.17158 186.54133
#> 77  169.34473 111.83889 132.92470 159.68586 183.40924 162.95682 101.68573
#> 78  126.33176 155.33331 112.22892 150.53554 187.75773 104.17598 153.67130
#> 79  136.52392 127.70218 138.70292 153.67720 133.69276 150.03042 119.54256
#> 80   93.62544 152.96670 132.40091 180.87891 149.55116 147.03622 133.58833
#> 81  135.10740 161.71733 140.52609  75.62111 115.73402 170.08905 152.06076
#> 82   92.73828  99.67662 162.86527 160.01463 159.31513 104.75111 131.52481
#> 83  123.15426 176.70423  98.44435 153.38220 154.69197 144.23604 154.67473
#> 84  143.03856 128.90775 147.78615 133.86750 194.44294  96.34680 174.61552
#> 85  161.66677 130.91906 133.55258 150.69231 123.44441 132.73937 206.64772
#> 86  120.62217 124.40175 152.71242 114.96418 147.20777 142.17811 143.66298
#> 87  194.32096 156.70382 171.74332 106.77446 157.72664 168.18249 103.05683
#> 88  193.39926 152.41555 126.92569 155.79006 151.71556 170.79745  99.15764
#> 89  113.35088 153.56562 150.78656 119.76701 158.86347 104.67469 141.42259
#> 90  124.93932 112.48300 154.19116 119.72665 147.14001 135.04696 153.05274
#> 91  149.33282 142.93816  84.80178 166.03918 148.55972 148.94718 103.46597
#> 92  198.21239 163.06710 165.05169 140.73729 112.06006 132.91558 177.30476
#> 93  144.57772 103.90892 111.92247 112.10933 185.06717 182.90534 129.81936
#> 94  139.34276 169.72747 176.40969 124.74648 159.15314 132.19757 119.94353
#> 95  192.89025 144.92163 143.16440 162.42079 136.10495 108.20969 105.66424
#> 96  141.29194 170.31977 108.13543 185.05814 116.56382 157.49461 115.79937
#> 97  141.66124 145.58515 110.64144 166.12594 111.53336  95.81524 154.42895
#> 98  138.38599 118.43199 203.07604 159.63122 151.17486 156.05031 134.21657
#> 99  131.90337 155.25413 119.96301 165.71264 170.27998  99.22573 134.64676
#> 100 159.00360  72.65571 120.27558 147.33633 123.45019 144.74358 177.81780
#>         B.4_1
#> 1   151.68823
#> 2   198.21668
#> 3   168.29050
#> 4   125.35533
#> 5   181.06161
#> 6   133.22835
#> 7    97.52377
#> 8   183.36215
#> 9   153.34464
#> 10  194.57745
#> 11   94.29790
#> 12  131.10082
#> 13  121.39100
#> 14  180.13584
#> 15  130.56292
#> 16  159.93497
#> 17  128.60167
#> 18  144.80524
#> 19  128.96281
#> 20  136.40870
#> 21  121.47676
#> 22  144.36501
#> 23  188.61639
#> 24  141.16664
#> 25  175.95725
#> 26  144.06812
#> 27  126.05842
#> 28  227.58384
#> 29  155.62779
#> 30  165.18352
#> 31  170.65598
#> 32  109.45274
#> 33  141.86041
#> 34  186.84955
#> 35  145.49649
#> 36  112.75703
#> 37  180.95336
#> 38  141.04361
#> 39  134.19500
#> 40  147.37453
#> 41  157.18488
#> 42  151.61152
#> 43  104.30919
#> 44  104.29198
#> 45  166.13624
#> 46  192.17606
#> 47  156.55067
#> 48  165.03721
#> 49  165.43506
#> 50  110.40747
#> 51  107.26560
#> 52  204.59034
#> 53  162.68803
#> 54  140.70488
#> 55  180.84911
#> 56  176.88421
#> 57  123.93774
#> 58  158.28513
#> 59  163.05043
#> 60  121.16142
#> 61  202.20962
#> 62  131.03268
#> 63  168.12675
#> 64  140.99860
#> 65  169.53868
#> 66  165.11296
#> 67  183.68987
#> 68  118.55882
#> 69  119.47209
#> 70   98.02286
#> 71  113.14204
#> 72  162.80948
#> 73  169.98052
#> 74  149.28784
#> 75  202.79074
#> 76  151.11534
#> 77  135.79119
#> 78  145.25710
#> 79  145.78128
#> 80  127.57248
#> 81  157.78972
#> 82  145.42684
#> 83  143.19328
#> 84  123.28770
#> 85  176.11581
#> 86  116.34449
#> 87  128.27560
#> 88  207.88805
#> 89  120.22796
#> 90  162.44071
#> 91  158.73169
#> 92  120.11548
#> 93  168.47158
#> 94  190.85095
#> 95  142.69555
#> 96  138.99451
#> 97  242.27678
#> 98  131.86063
#> 99  132.22312
#> 100 160.75733