Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10, 10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(
  subset = rep(c("A", "B"), 50),
  level = rep(1:4, each = 25)
)
rownames(meta) <- 1:100
scRNA <- SeuratObject::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA,
  by = c("subset", "level"),
  min.cells = 10, max.cells = 20
)
#>        A.1_1     B.1_1     A.2_1     B.2_1     A.3_1     B.3_1     A.4_1
#> 1   135.0482 130.35931 104.20373 175.29378 208.73813 122.80602 122.94309
#> 2   145.0101 155.19135 147.26388 139.99157 121.57895 102.62741 146.11542
#> 3   152.0521 177.76620 122.87936 142.91563 185.30757 139.50626 138.83827
#> 4   142.6634 140.14486  96.59135 167.55987 151.38459 144.02991 142.93123
#> 5   157.0629 114.76685 133.88375 152.71888 157.79216 144.90456 156.46448
#> 6   185.4597 126.45673 135.10897 161.34751 107.69317 109.10560 109.25001
#> 7   145.1351 122.70631 139.65706 170.06866 115.51948 100.99433 131.71334
#> 8   161.7921 191.13172 137.40893 138.26771 116.65561 111.72766 135.55774
#> 9   150.9870 161.13497 131.82622 130.92899 148.72359 196.33733 114.67232
#> 10  162.3730 179.70487 134.63521 115.45021 160.36683 121.26682  94.31434
#> 11  167.0505  89.34870  95.15999 177.53700  94.40111 172.34672 194.46043
#> 12  150.2398 158.67600 164.41020 109.94129 129.99821 159.83864 153.49909
#> 13  182.4147 117.07708 166.92289 155.45462 148.45956 109.17392 154.64398
#> 14  200.9290  93.11517 116.55672 139.21904 140.53945 145.24917 146.98777
#> 15  137.0111 118.31611 124.62653 108.59913 169.60599 161.97946  79.08457
#> 16  164.2733 142.19493 153.13669 147.33151 117.18080 142.77278 154.11747
#> 17  151.2793 123.65855 169.48788 162.66615 176.99494 125.48092 119.75620
#> 18  158.8249 107.98052  99.93794 194.27746 123.83424 156.69925 143.34302
#> 19  163.6427 108.80553 118.06094 147.46634 185.62540 110.69866 150.96485
#> 20  167.1243 103.54846 124.38164 143.30207 124.78175 176.03781 148.77685
#> 21  128.4156 184.31322 104.55473 174.33752 172.69803 141.19271  88.47739
#> 22  170.0473 173.73358 165.55320 164.61975 167.13803 135.94745 121.00733
#> 23  168.5482 124.53071 167.12876 180.25051 176.87647 145.11952 109.63661
#> 24  163.3860 155.14819  83.03183 148.79133 144.74591 121.92493 128.12249
#> 25  178.3064 147.13086 144.84653 204.89117 158.79289 147.61640 188.16686
#> 26  198.2172  97.18453 151.13142 166.32510 117.98047 125.98989 124.33086
#> 27  175.5575 116.31433  93.17267 148.84758  85.64464 137.62751 132.42341
#> 28  155.2296 172.51937 107.06400 165.33761 163.79094 114.15673 174.78181
#> 29  129.8864 143.38165 142.57933 116.59204  98.72576 125.54992 170.27686
#> 30  185.8181 161.28422 187.23609 151.58151 130.91723 145.96466 177.87081
#> 31  169.0251 174.45875 149.13208 150.07164 122.16289 148.16762 137.16442
#> 32  167.1712 170.74255 133.59975 146.92343 172.61066 128.69749 102.10744
#> 33  208.9566 134.52117 160.48870 121.24605 136.12404 137.85605 159.65184
#> 34  153.2789 149.30146 130.49479 180.95858 164.27096 177.20770 104.33730
#> 35  145.4958 143.18803 143.39723 206.34083 129.43133 149.06984 105.41151
#> 36  141.7267 120.69774 125.62377 109.72732 144.82170 129.44134 145.57652
#> 37  144.7128  70.03742 120.47587 180.93637 150.45666 124.43756 179.71045
#> 38  171.6328 160.48598 141.09244 150.88102 130.84865  99.83242 125.37217
#> 39  106.1625 144.93248 165.12388  96.21906 174.90369 182.07478 112.63471
#> 40  183.3874  56.91040 113.53565 169.61940 147.11203 102.23515 141.03062
#> 41  141.6983 117.01775 196.58812 150.01680 165.78413 181.82893 116.88896
#> 42  199.1750 114.87600 149.38856 194.29093 167.81701 160.31202 124.15580
#> 43  159.9496 178.22714 189.29087 172.76945 162.74378 120.17614 140.19338
#> 44  201.0663 105.49849 198.31188  78.45490 166.34701 136.74650 181.38470
#> 45  170.8917 125.99085 124.20969 157.96691 177.26322 167.04549 146.45038
#> 46  103.6032 162.33523 139.82038 184.71651  96.46023 113.41934 110.01165
#> 47  159.0001 171.33512 124.14692 145.73533 196.15397 102.37885 144.76847
#> 48  185.6436 130.78833 168.86986 153.43394 165.99420  87.77639 153.22520
#> 49  153.4603 113.22425 120.54894 108.08586 140.77293 173.89096 118.10780
#> 50  173.9738 125.66146 127.25999 147.50522 135.20473 151.78256 122.19338
#> 51  126.3803 151.71834 125.78911 174.77440 149.43418 165.57846 147.92962
#> 52  157.5550 161.86278 129.07497 126.69653 142.24435 128.86606 121.80240
#> 53  128.8939 106.62659 132.52059 159.47442 183.59941 125.32933 139.04211
#> 54  142.8149 186.24409 169.00213 154.08617 130.53582 145.90037 104.38177
#> 55  165.5208 190.38274 133.56865 171.67104 180.52666 135.92596 179.56422
#> 56  151.4000 162.56933  91.67519 139.63309 234.14995 149.18777 163.42606
#> 57  178.6848 149.67472 154.17379 178.04970 124.32750 111.68375 154.11564
#> 58  155.1955 142.16730 120.23211 126.34012 185.16537 165.17787 164.58406
#> 59  147.3723 141.75571 140.91461 156.79676 122.39708 185.48104 122.13547
#> 60  139.8277 132.02439 194.50797 142.34435 144.95671 138.13132 145.40197
#> 61  153.5108 151.74507 193.64793 125.72421 121.12513 140.76705 112.94754
#> 62  128.0191 127.14713 171.44820 117.82552 145.64008 171.32843  88.05406
#> 63  188.3953 131.41014 165.01116 150.65351 158.64237 125.79347 165.95006
#> 64  193.8205 171.42201 101.50105 136.81442 107.20292 171.74142 155.28011
#> 65  134.6171 151.43175 109.80695 167.91266 122.58909 186.46781 123.24062
#> 66  125.6109 122.24558 165.76146 184.14516 151.47632 117.90196  91.98289
#> 67  219.7669 142.10270 136.21054 167.62963 137.79133 121.67173 126.84073
#> 68  173.6675 122.13465 156.45856 154.59394 140.98345 117.79761 117.45566
#> 69  111.9982 119.30807 167.44744 165.57545 130.42769 168.41554 123.12218
#> 70  156.7083 147.62144 139.81810 129.55727 144.86566  89.59351  57.29407
#> 71  165.8961 139.50039 143.52752 180.93541 125.28853 151.19408 121.76300
#> 72  193.9750 162.54059 143.37627 168.77996 165.99356 128.11944 184.97432
#> 73  170.8084 129.22802 113.92726 149.71042 179.20101  98.51960 106.25803
#> 74  162.4068 160.19024 130.01295 151.26848 150.69955 124.32388 175.41642
#> 75  167.6358 179.14858 123.87153 192.41952 204.40515 124.89970  86.47614
#> 76  166.0006 122.81755 157.99404  83.38530 135.50612 149.13526 179.60516
#> 77  165.7981 118.91010 183.31846 132.16334 166.51580 154.66234 137.69070
#> 78  114.6144 122.15571 172.39184 168.51160 205.38242 158.59017 100.24210
#> 79  127.0972 112.06556  93.69500 141.60914 182.31137 156.23512 161.83677
#> 80  107.2772 123.50774 113.52572 152.73411 214.89690  98.83708 140.00380
#> 81  170.5633 134.01760 187.10227 147.94419 159.31236 109.91689 147.68905
#> 82  163.9096 110.98767 137.61527 166.71521 195.40389 146.29923 134.81261
#> 83  170.5900 129.63648 179.83803 141.07391 139.98171 148.28692 142.56690
#> 84  129.8355 195.64888 108.19603 127.34752 156.70930 182.37663 193.62635
#> 85  189.1220 141.84673 114.51156 142.10548 109.37090 153.54886 120.53536
#> 86  149.7252 118.14469 124.03701 144.61138 131.51946 157.38001 124.97077
#> 87  138.5495 138.83416 141.16464 155.58844 165.27269 134.86678 158.28111
#> 88  137.7846  82.61091 131.69436 137.82673 124.01388 102.18314 153.75047
#> 89  143.1370 163.56677 119.67983 158.81935 132.78259 159.23589 151.22414
#> 90  205.3637 161.34690 116.02766 143.63057 128.51016 135.57957  88.91586
#> 91  173.3124 118.20977 180.06576 173.84089 168.64165 167.11301 157.80283
#> 92  178.5371  88.29190 140.77457 177.43233 155.62970 104.28806 137.83847
#> 93  138.3526 102.97588 121.26129  90.99351 155.26398 152.42525 157.30167
#> 94  177.2855 141.62605 143.57666 144.56760 151.81862 178.56421 173.27543
#> 95  152.0126 169.75292 136.26436 179.82587 116.63960 103.50518 125.52268
#> 96  157.6602 132.17186 161.02046 193.37617 106.27218  97.75241 178.57864
#> 97  166.6258 109.05343 121.93655 198.28901 134.14702 185.71371 176.33047
#> 98  100.2521 113.78969 156.36075 145.21099 164.79923 128.32934 130.19999
#> 99  152.2526 120.82280 144.30096 151.01130 131.73153 148.08733 195.72492
#> 100 114.7334 152.64959  86.67653 150.04693 175.92130 134.86353 194.53257
#>         B.4_1
#> 1   184.32928
#> 2   113.54756
#> 3   156.92504
#> 4   140.21024
#> 5   151.35443
#> 6   174.02369
#> 7   190.52648
#> 8   138.91046
#> 9   131.04888
#> 10  141.74985
#> 11  143.83358
#> 12  145.54876
#> 13  210.12746
#> 14  130.69688
#> 15  153.81314
#> 16  188.71569
#> 17  143.81393
#> 18  207.92018
#> 19  152.17493
#> 20  130.02607
#> 21  199.00737
#> 22  163.81920
#> 23  167.77794
#> 24  162.88349
#> 25  111.51875
#> 26  133.41464
#> 27  199.97554
#> 28  142.84580
#> 29  143.81190
#> 30  140.26792
#> 31  132.24232
#> 32  149.71575
#> 33  189.28331
#> 34  209.61009
#> 35  139.34442
#> 36  155.57936
#> 37  165.84561
#> 38  141.76266
#> 39  132.84317
#> 40  147.41196
#> 41  127.34958
#> 42  123.86174
#> 43  160.52094
#> 44  138.26170
#> 45  139.69464
#> 46  174.97221
#> 47  152.05329
#> 48  191.07361
#> 49  165.53367
#> 50  128.30342
#> 51  123.44549
#> 52  120.44471
#> 53  182.13170
#> 54  235.45467
#> 55  138.21745
#> 56   90.19197
#> 57  200.81685
#> 58  137.88415
#> 59  211.11717
#> 60  142.76489
#> 61  157.40250
#> 62  142.74756
#> 63  119.69992
#> 64  166.08864
#> 65  106.45523
#> 66  128.37936
#> 67  143.22843
#> 68  113.07520
#> 69  114.12482
#> 70  144.86399
#> 71  186.05150
#> 72  129.90225
#> 73  156.83717
#> 74  191.94725
#> 75  118.28437
#> 76  168.95486
#> 77  130.27920
#> 78  132.85548
#> 79  136.80856
#> 80  117.87955
#> 81  121.58162
#> 82  131.75001
#> 83  165.45431
#> 84  131.60218
#> 85  156.85883
#> 86  160.81114
#> 87  107.73516
#> 88  139.29721
#> 89  199.09750
#> 90  160.74609
#> 91  125.93784
#> 92  155.00900
#> 93  119.84436
#> 94  231.65629
#> 95  191.53450
#> 96  130.45728
#> 97  126.18182
#> 98  139.77796
#> 99  133.48988
#> 100 188.97875