Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10, 10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(
  subset = rep(c("A", "B"), 50),
  level = rep(1:4, each = 25)
)
rownames(meta) <- 1:100
scRNA <- SeuratObject::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA,
  by = c("subset", "level"),
  min.cells = 10, max.cells = 20
)
#>         A.1_1     B.1_1     A.2_1     B.2_1     A.3_1     B.3_1     A.4_1
#> 1   183.79007 139.87620 137.90239 241.17208 146.25007  93.81776 166.71688
#> 2   189.66046 146.96515 158.99114 154.51683 135.46361 122.87095 122.96125
#> 3   166.56170 124.44236 180.48513 204.42111 150.59686 187.99431 141.43896
#> 4    72.23017 188.36113 112.42997 196.28971 152.95862 128.42459 121.84145
#> 5   117.07991 117.87185  94.73239 172.52731 116.50811 138.62464 166.43287
#> 6   153.48619 127.41547 160.91970 162.98735 162.13818 114.31287 111.56226
#> 7   175.45635 129.80870 191.74446 176.28932 143.24939 150.07965 186.28121
#> 8   138.47124 129.79502 127.20676 162.08956 131.88479 136.38873 168.75980
#> 9   123.33478 184.87857 128.60397 178.35994 130.23435 114.21721 119.01295
#> 10  132.39969 146.91017 155.18628 171.19126 164.55613 115.45617 144.11715
#> 11  116.40207 106.30800 169.86986 152.52448 190.06721  99.44899 148.95524
#> 12  148.22224 157.70392 121.62769 162.70755 139.12957 150.55835 140.38058
#> 13  170.59985 120.26616 133.36683 143.61555 154.84850 121.03905 152.86988
#> 14  132.80933 127.77739 153.85576 129.46251 138.67097 112.41789 167.35364
#> 15  143.72796 123.35722 158.79769 179.91304 193.33685 114.80155 204.86604
#> 16  196.26129 184.29168 148.71549 131.55356 196.39297 125.98282 122.15544
#> 17  136.78843 130.67935 159.80998 120.07736 172.55266 167.03564 115.86831
#> 18  136.40450 139.52304 107.38106 140.64180 104.07882 151.78165 149.46659
#> 19  147.08426 124.94801 125.74622  98.84090 198.03782 105.97510 143.26453
#> 20  132.19592 126.64152 133.44174 158.19442 129.37019 153.01717 123.05112
#> 21  150.36237 170.26803  97.78723 176.56680 122.33319 134.59913 130.19447
#> 22  150.72895 199.27803 103.03077 151.12703 193.51354 118.79150 149.97620
#> 23  145.91020 109.51535 122.81177 127.70246 209.67684 119.51873 106.72862
#> 24  155.96723 112.12585 139.20377 111.83155 131.18655 109.62616 126.90149
#> 25   79.81401 160.03288 126.32221 175.37215 158.64483 140.16179  96.67015
#> 26  176.18750 116.47333 131.53127 135.87663 153.06482 159.14625 121.46250
#> 27  143.99360 173.94257 187.10672 116.09111 202.29107 156.87831 108.35370
#> 28  186.05550  81.62937 110.86520 188.82724 166.96726  76.48206 181.35504
#> 29  101.22709 108.30562 123.00355 182.60994 130.62735 152.95470 175.39576
#> 30  126.30411 126.47594 148.58065 198.61919 180.78609 124.52846 144.92498
#> 31  145.10710 122.15660 171.93469 157.21437 171.28677 135.93147 124.59628
#> 32   94.67944 110.60162 179.68661  93.68696 134.98088 134.75257  72.65418
#> 33  202.04444  64.76613 145.20426 208.32655 206.97327 158.62057 153.52643
#> 34  127.10785 103.51083 121.51644 190.63524 166.64193 108.68162 115.59605
#> 35  148.18780  97.68812 142.71464 151.26051 146.78097 137.15149 147.02588
#> 36  188.47074 128.88527 151.37963 175.85698 203.76391 171.16727 126.75757
#> 37  138.55160 117.97830 117.50245 113.11280 139.78855 144.24600 195.91061
#> 38  114.18737 139.61644  99.46692 191.83313 130.14949 174.76297 103.27663
#> 39  153.35930 125.28453 148.50561 140.16875  95.95088 166.06833 168.23912
#> 40  142.51538 120.88845 116.60048 124.43127 134.28930 113.45765 182.35999
#> 41  156.75603 115.79994 175.65548 133.09313 144.37190 159.93512  97.10712
#> 42  154.30416 207.86652 115.51824 197.34190 150.65787 111.09932 139.88185
#> 43  137.03289 214.68814 123.75833  96.23100 146.35574 139.12964 146.58011
#> 44  130.67801 168.05181 146.41048 138.06768 142.93483 175.01188 167.46576
#> 45  117.33551 132.42474 145.47803 170.42759 135.29098 149.14154 124.91579
#> 46  148.97806 140.03254 136.11742 179.72998 147.18275 134.40141 151.62969
#> 47  148.78823 135.14748  87.85595 169.71344 130.02012 131.72857 134.12622
#> 48  116.80434  74.07358 114.57244 103.65561 154.00498 142.87669 147.50045
#> 49  179.60382 128.41915 130.65253 167.92699 124.59733 115.10989 128.82673
#> 50  102.57610  98.02300 150.76475 155.81913 113.20761 140.84952 161.45296
#> 51  119.49546 139.84968 134.45197 181.54682 177.40606 133.70164 172.90970
#> 52  143.26057 199.53619 158.47027 196.14061 147.47814 140.52600 151.69008
#> 53  115.63873 171.30326 142.91872 167.59290 182.30698 134.42542 118.25905
#> 54  197.96737 187.04895 111.38316 166.12079 163.69394 137.65199 177.98617
#> 55  190.43217 149.10434 138.31215 165.09286 136.98338 107.23972 183.97212
#> 56  154.34743 155.29700 189.67423  86.54259 102.39412 134.46356 194.18351
#> 57  135.59444 124.27580 126.17889 133.31618 133.62096 162.89203 182.84244
#> 58  150.72332 159.19419 124.05723 119.88759 154.69244  79.27657 145.19822
#> 59  129.15645 110.32245 114.55354 100.71709 158.31408 118.47202 158.30475
#> 60  171.43405 138.07929 115.52084 210.77729 161.76421 146.90684 145.31646
#> 61  148.68720 157.14826  96.97669 197.41205 129.44870 151.51967 128.20227
#> 62  160.35004 149.84286  97.14879 107.13024 162.63561 103.27719 102.47938
#> 63  179.45830  93.87317 141.90914 163.84741 106.39487 142.76713 178.00261
#> 64  145.43738 117.46202 122.96674 136.73259 119.17827 104.88727 128.43856
#> 65  180.74113  84.95350 114.47821 135.57882 166.26662 156.17172 133.63743
#> 66  120.48810 160.48911 199.04443 179.11446 167.17946 105.85130 123.16561
#> 67  138.20944 145.34900 126.64869 162.44537 136.83822 133.66752 195.09119
#> 68  155.99050 122.20858 103.82398 150.22550 192.57061 116.32368 123.88745
#> 69  124.45082 119.83080 169.91277 168.62544 141.82455 212.57919 125.66493
#> 70  114.84712  99.09789 158.17926 187.81926 172.04691 129.14794  96.40810
#> 71   84.68267 129.02633 149.62102 144.34243 130.45363 132.78966 108.84377
#> 72  112.46601 109.40762  92.87685 136.85571 139.14022 141.03900 145.05276
#> 73  169.32720 135.18384 152.68332 148.85024 131.57369 152.42844 133.18096
#> 74   96.19066 144.82707 129.13947 147.32860 166.15830 172.01249 162.06337
#> 75  184.11002 198.46829 156.35578 123.21286 105.06082 107.15775 146.90302
#> 76  151.24673 185.61098 133.20526 141.09615 147.98507 133.89629 136.63669
#> 77  156.01015 134.56820 139.63952 186.12690 153.78124 169.76792 135.52004
#> 78  154.87999  66.58040 200.16181 195.24947 153.63861 142.00002 165.48478
#> 79  155.51192 133.90906 147.76832 154.73400 131.77251 158.33838 122.36306
#> 80  177.17523 159.71107 120.66128 180.93909 148.81170 115.79775 132.06234
#> 81  177.82241 155.88409 215.82732 131.16774 103.59614 175.11993 150.30072
#> 82   83.20809 108.82105 149.46916 120.93605 153.35715 127.25148 119.64715
#> 83  186.76907 112.38947 193.49902 123.15905 162.67286 167.67543 151.30399
#> 84  173.37540 114.09578  98.96005 145.47182 119.75799 148.87151 138.57614
#> 85  181.43366 165.82632 135.43412 146.21908 179.58963 120.63917 120.52662
#> 86  181.74651 145.75397 118.44095  93.47015 152.29199 177.62983 122.56547
#> 87  165.40933 184.14502 178.47901 158.57459 178.47477 159.12212 148.33646
#> 88  144.63540  88.59621 169.04636 162.27721 148.47655 121.53008 120.88847
#> 89  148.43011 135.04010 145.85368 158.68044 157.74681 176.41675 116.66889
#> 90  182.20782 151.37900  99.29330 127.10863 140.08060 112.97787 150.52270
#> 91  169.80199 142.32678 134.68899 181.89448  88.74948 184.97918 125.33184
#> 92  159.90897 153.50238 115.47109 205.40319 146.76623 135.66158 120.29182
#> 93  192.73219 108.17627 141.43051 151.52085 140.39469 139.72087 147.26731
#> 94  171.62145 151.96545 192.83970 160.94221 134.86046  98.15048 107.34909
#> 95  212.44305 148.51892 150.51626 207.30370 107.93381 139.18568 121.02498
#> 96  136.41775 117.33274 204.72005  94.12204 146.28311 177.24479  96.66243
#> 97   99.85513 158.19567 104.17420 171.41975 165.76907 149.24230 155.33782
#> 98  176.81151 136.72836 157.14244 123.33818 141.90079 133.61424 129.51569
#> 99  136.10050 124.78728 143.38228 103.37273 137.86487 189.59193 147.56385
#> 100 128.87716 109.45026 108.41150 161.55556 166.40721 127.95972 105.16565
#>         B.4_1
#> 1   158.83692
#> 2   234.07383
#> 3   160.06078
#> 4   111.65471
#> 5   128.47124
#> 6   184.42081
#> 7   182.92510
#> 8   174.19770
#> 9   164.44972
#> 10  181.16083
#> 11  143.08104
#> 12  111.80259
#> 13  167.27078
#> 14  160.05508
#> 15  143.17420
#> 16  157.14225
#> 17  190.51807
#> 18  171.64086
#> 19  204.75186
#> 20  186.83970
#> 21  198.69126
#> 22  116.79888
#> 23   94.04751
#> 24  127.13957
#> 25  153.52961
#> 26  125.02695
#> 27  135.36391
#> 28  180.27305
#> 29  181.17469
#> 30  134.88560
#> 31  140.52331
#> 32  161.68541
#> 33  172.31663
#> 34  145.36460
#> 35  204.37101
#> 36  160.70135
#> 37  136.28837
#> 38  127.66588
#> 39  199.45777
#> 40  207.84263
#> 41  170.06505
#> 42  184.71222
#> 43  154.62174
#> 44  161.40353
#> 45  162.15599
#> 46  187.10821
#> 47  146.22721
#> 48   95.94990
#> 49  101.77511
#> 50  164.97008
#> 51  133.11817
#> 52   90.29693
#> 53  149.08732
#> 54  159.45631
#> 55  132.52114
#> 56  148.97666
#> 57  152.47234
#> 58  198.39973
#> 59  158.88836
#> 60  143.07524
#> 61  151.10782
#> 62  110.43357
#> 63  189.05710
#> 64   95.78748
#> 65  168.18288
#> 66  118.17740
#> 67  206.94681
#> 68  135.56877
#> 69  150.99233
#> 70  164.08435
#> 71  102.25649
#> 72  123.22506
#> 73  194.42796
#> 74   83.26980
#> 75  148.36022
#> 76  165.17523
#> 77  160.42792
#> 78  152.36305
#> 79  136.49150
#> 80  159.19230
#> 81  128.87072
#> 82  152.36686
#> 83  168.02719
#> 84  126.28323
#> 85  145.33253
#> 86  141.50716
#> 87  122.64563
#> 88  165.51662
#> 89  143.88513
#> 90  151.39547
#> 91  127.07774
#> 92  170.27331
#> 93  157.81370
#> 94  136.62102
#> 95  130.39273
#> 96  139.50241
#> 97  148.59667
#> 98  150.57686
#> 99  136.04289
#> 100 109.85400