Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10, 10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(
  subset = rep(c("A", "B"), 50),
  level = rep(1:4, each = 25)
)
rownames(meta) <- 1:100
scRNA <- SeuratObject::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA,
  by = c("subset", "level"),
  min.cells = 10, max.cells = 20
)
#>         A.1_1     B.1_1     A.2_1     B.2_1     A.3_1     B.3_1     A.4_1
#> 1   174.68603 127.33075 113.68616 127.54049 133.50264 149.49056 101.36087
#> 2    86.62215 150.27200 133.19768 102.89766 152.86165 177.90272 150.81274
#> 3   118.55710 176.21915 115.83498 141.45992 153.60275 137.07946 139.23730
#> 4   149.98353 147.99919  84.30184 153.93266 124.83840 110.97086 167.18753
#> 5   133.66636 104.48738 158.11738 165.70987  92.97646 134.10991 136.96015
#> 6   178.82714 163.35371  96.85254 167.25774 177.93151 172.37388 150.67907
#> 7   162.09758 111.41874 110.52065 102.28941 212.13564 155.00507 139.40977
#> 8   144.70828 112.48269 185.56124 152.92671 147.85429 107.79755 126.46732
#> 9   109.78674 139.33193 132.38684 135.67844 188.67483 108.15565 165.77617
#> 10  167.20877 166.42406 104.08548 125.83319  99.83862 132.00050 139.45172
#> 11  133.73218 135.24538  99.70768 158.47142 170.67470 139.92846 149.66115
#> 12  174.89849  95.77171 182.64353 152.52028 144.71131 103.91986 124.78726
#> 13  131.89685 146.24328 186.11033 145.59521 154.36473 103.77764 135.04765
#> 14  161.94622  96.80347 119.36322 157.02092 176.42619 121.21456 145.74769
#> 15  176.88270 137.79812 121.49882 147.79347 176.80268 142.73215 142.02423
#> 16  128.45757 149.81430 116.79806 119.82720 148.73490  96.14145 106.89868
#> 17  169.09702  91.01783 118.55626 165.13854 136.48717 105.19635 160.26854
#> 18   97.92230 118.74011 176.49970 192.05071 212.34152 147.01056 173.25976
#> 19  158.55101 109.40297 140.14494 134.26961 157.69460 135.01505 107.65157
#> 20  157.17573 142.53366 133.80834 129.59872 151.37677 116.39222 145.38793
#> 21  138.34805 123.41980 117.20207 172.22622 120.72596 123.20563 130.34184
#> 22  237.49718 179.82043  94.75165 184.16323 160.08669 105.17799 158.23223
#> 23  193.17881 134.69044 147.36983 142.78999 101.60456 113.68241 212.45246
#> 24  108.81937 181.41933  91.09720 133.30583 140.89366 122.95047 135.50989
#> 25  194.03821 158.89350 113.70965 139.33607 100.01720 127.35333 140.33964
#> 26  168.14246  74.74144 146.27480 150.86072 183.86016 126.99632 138.18832
#> 27  150.24433 130.79123 158.28417  91.88316 121.00631 191.03995 154.58416
#> 28  192.48021 145.40537 155.00230 142.04680 101.36811 144.95084 138.07735
#> 29  113.45988  97.24391 121.40480 153.03442 168.30140 133.72767  98.12116
#> 30  175.40666 105.67449 119.00369 185.05483 140.09600 128.75217 158.77597
#> 31   84.33614 153.32234 137.79636 163.11221 154.10880 139.15282 171.55130
#> 32  181.31690 124.61797 167.52365 141.72124 129.31371 156.79237 182.87123
#> 33   98.17337 110.26595 116.84114 165.05671 190.58470 121.46552 191.09043
#> 34  173.60747 199.80584  94.63868 100.56796 116.39750 123.89035 180.25090
#> 35  112.79187 118.03371 128.53337 124.53186 117.17642 160.30602 165.59400
#> 36  141.67647 147.85190 153.64347 146.48311 111.57729 191.31566 175.51726
#> 37  139.82336 111.53482 160.77024 186.64706 117.62812 119.54697 162.12368
#> 38  168.08407 113.28720 146.99397 188.38553 171.23540 119.96317 126.16571
#> 39  149.98580 138.78138 146.85847 169.42414 170.13694 108.56441 105.18106
#> 40  157.26890 158.36401 134.13431 155.21678 140.32103 153.79557 200.73951
#> 41  147.15806 170.11608 177.74587 136.03764 101.62914 128.37742 138.83916
#> 42  165.29081 145.62572 165.72866 167.90501 153.36773 136.92388 109.51556
#> 43  186.13815 137.10913 180.84270 126.15465  66.47790 116.40646 116.34832
#> 44  218.28819 119.77605 153.83642 187.72925 122.16258 102.80363 170.06034
#> 45  190.60381 155.83732 134.62675 148.00128 136.85420 143.26847 194.12863
#> 46  141.98692 153.74042 136.00915 131.80693 151.81738 134.06912 116.83427
#> 47  178.58597 103.91806 191.56072 111.31425 166.44718  99.37897 160.49833
#> 48  175.72236 170.21355 128.39969 147.20380 128.55443 130.51132 102.98726
#> 49  141.68000 198.19081 103.43683 168.20496 144.97823 134.26182 121.03077
#> 50  116.97308 117.97666 135.97111 167.19846 124.62836 127.49973 162.06591
#> 51  127.37385 188.49188 141.38336 186.32610 122.61315 159.52191 155.28917
#> 52  193.45733 129.99437 102.90907 192.34383 143.33430 137.23861 169.46393
#> 53  175.11969 133.34146 146.88874 136.86633 159.10573 152.23222 131.68873
#> 54  122.11527 141.84157  97.08676 220.00812 188.15636 156.43344 131.29246
#> 55  147.30855 135.98909 116.48912 155.00549 140.14200 135.45628 140.80028
#> 56  116.61367 164.48451 174.70557 130.48865 109.41289 156.30180 135.98666
#> 57  185.69772 178.52844 134.25356 183.00777 219.30885 123.61293 115.52363
#> 58  165.36605  88.23602 139.06016 121.06244 189.14383 158.26071  94.67912
#> 59  154.32577 110.17882 145.19189 162.34842 135.50190 189.00450 151.52121
#> 60  170.89412 121.11735 105.70986 180.35782 155.40763 114.55623 127.97828
#> 61  175.58719 184.60808 161.21279  88.18836 156.35595  87.02370 123.58821
#> 62  179.25409 128.71500 128.03053 121.23574 188.44087 155.41891 141.27144
#> 63  195.21396 144.85745 134.42268 151.85214 167.49631 167.26451 153.99372
#> 64  161.63944 159.35537 153.55467 184.50154  94.55769 146.72037 166.81612
#> 65  139.74547 174.07803 149.32291 123.15272 140.50036 143.23394 144.49096
#> 66   94.68625 163.29513 144.09599  96.59733 182.66713 130.65566 175.38887
#> 67  140.72377 179.55446 117.15057 149.40374 127.20692 171.77514 145.67671
#> 68  160.29989 159.18322 170.19582 190.55232 153.39437 108.07640 159.10022
#> 69  127.94603 126.07122 111.61067 152.57819 207.66797 194.13709 139.15771
#> 70  143.09085 138.29185 140.35910 193.74759 182.21040 120.48223  85.12456
#> 71  145.38776  99.56714 137.60586 164.60337 156.45806  74.90092 134.24521
#> 72  149.31969 110.42963 173.62650 122.88160 121.20093  90.44692 149.56092
#> 73   92.83734 132.77797 110.68850 131.45755 189.66827 137.89819  49.03054
#> 74  152.88425 153.79901 156.54500 208.97500 130.76896 152.71814 150.41601
#> 75  135.98590 161.16395 141.28470 132.22985 159.95709 144.48795 146.36334
#> 76  165.12046 155.30572 123.44481 142.47932 145.45067 106.04753 142.90398
#> 77  161.21661 106.88478 135.08798 177.57854 129.46783 128.98776 143.77639
#> 78  124.76961 131.59796 141.58911 133.38356 160.74207 155.94691 144.96975
#> 79  198.99765 173.20335 121.47839 167.04842 113.28937 160.22205  97.69227
#> 80  203.36620 161.31302 117.56905 125.98926 132.46883 112.65266 160.30742
#> 81  144.91109 105.93049 129.84898 156.71715 140.63140 169.53165 132.08041
#> 82  156.95734 136.57909 107.03419 136.93232 130.68695 195.57459 109.02909
#> 83  120.48552 130.92064 169.61919 146.51028 161.32950 184.71270  81.62995
#> 84  171.64906 145.10802 146.59045 122.52780 150.81071 152.39995 145.09560
#> 85  133.05185 102.25463  97.04801 140.92651 194.06482  96.78357 156.52122
#> 86  150.15278 142.63476 142.21017 150.87098 189.48571 117.06748 200.55558
#> 87  124.79319 164.16372 182.77937 125.26971 119.20338 191.44263 119.08693
#> 88  141.15525 117.49913 121.33013 159.08592 159.67943 102.15350 115.41009
#> 89  147.18791 139.31040 157.45579 161.76644 133.25039 159.60271 132.04099
#> 90  108.19192 137.65801 143.96316 159.30671 185.47032 178.05917 193.34443
#> 91  210.76829 155.27737 204.36356 155.05446 170.90047 177.76111 200.08978
#> 92  153.55664 127.03861 113.25240 113.88407 173.07129 120.35794 157.05753
#> 93  164.06837 131.48857 169.05265 162.18829 167.22826 142.70813 114.45246
#> 94  189.01607 158.00869 103.41333 179.92350 118.31593 102.17061 178.78475
#> 95  164.65321 132.75301 189.96538 118.91702 175.38628 115.94578 154.06343
#> 96  153.70260  99.77318 138.09612 125.79003  86.69976 143.02976 130.53423
#> 97  176.16038 120.82208 146.73426 212.56698 140.15894 129.06413 142.96255
#> 98  141.14264 189.06208 156.43417 144.78825 132.52135 113.86810 171.50832
#> 99  144.94401 142.86954 169.32161 181.27977 133.31207 175.28606 134.66197
#> 100 151.38887 143.56410  93.08200 124.30684 182.33428 135.03512 122.50815
#>         B.4_1
#> 1   151.89303
#> 2    96.12289
#> 3   195.79990
#> 4   149.32521
#> 5   109.16684
#> 6   171.12917
#> 7   143.15727
#> 8   114.84731
#> 9   137.96867
#> 10  102.83864
#> 11  183.21279
#> 12  143.89764
#> 13  154.87116
#> 14  181.98910
#> 15  166.80482
#> 16  198.25225
#> 17  138.99199
#> 18  144.29044
#> 19   99.08740
#> 20  169.48272
#> 21  115.52903
#> 22  148.72223
#> 23  170.18656
#> 24  149.35763
#> 25  176.40611
#> 26  156.36915
#> 27  163.57180
#> 28  126.12137
#> 29  147.05445
#> 30  143.01257
#> 31  142.72576
#> 32  148.19894
#> 33  254.91335
#> 34  119.53316
#> 35  170.50509
#> 36  214.08981
#> 37  144.95842
#> 38  157.24098
#> 39  153.55086
#> 40  131.01834
#> 41  128.50087
#> 42  163.27402
#> 43  231.88977
#> 44  176.65490
#> 45   99.79022
#> 46  155.20137
#> 47  138.12330
#> 48  165.02584
#> 49  146.61445
#> 50  137.92375
#> 51  127.77538
#> 52  187.07938
#> 53  127.92309
#> 54  125.56296
#> 55  167.07083
#> 56  200.16981
#> 57  109.37322
#> 58  159.94818
#> 59  131.55626
#> 60  175.60601
#> 61  151.50097
#> 62  162.16552
#> 63  154.38470
#> 64  244.92484
#> 65  170.03838
#> 66  167.91692
#> 67  121.02232
#> 68  147.22587
#> 69  131.99207
#> 70  151.34855
#> 71  188.59741
#> 72  157.67515
#> 73  159.77045
#> 74  157.33351
#> 75  135.64072
#> 76  139.49023
#> 77  163.07957
#> 78  190.70899
#> 79  152.20542
#> 80  155.33766
#> 81  150.84674
#> 82  142.87104
#> 83  138.02527
#> 84  184.29050
#> 85  100.48958
#> 86  168.36898
#> 87  134.30712
#> 88  134.95280
#> 89  211.78465
#> 90  149.21713
#> 91  168.36120
#> 92  138.53442
#> 93  140.39325
#> 94  150.45220
#> 95  201.01415
#> 96  193.72938
#> 97  165.69660
#> 98  155.33615
#> 99  152.22462
#> 100 138.27268