Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10,10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(subset = rep(c("A","B"),50),
                   level = rep(1:4, each = 25))
rownames(meta) <- 1:100
scRNA <- Seurat::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA, by = c("subset","level"),
               min.cells = 10, max.cells = 20)
#>         A.1_1     B.1_1     A.2_1     B.2_1     A.3_1     B.3_1     A.4_1
#> 1   116.31400 193.80553 104.26374 135.56577 143.80040 125.34069 157.11561
#> 2   150.11830 139.35480 168.42508 158.25905 183.77839 108.24653 136.92760
#> 3   163.36370 144.98404 123.86190 150.16533  98.16942 116.26572 137.72924
#> 4    92.57391 145.07585 149.01986 118.81534 140.77502 191.66359 130.20289
#> 5   118.19221 196.42855 174.45114 128.51322 182.83667 150.21396 145.49176
#> 6   164.64440 138.57203 132.76945 183.70306 145.90778 111.56494 160.30843
#> 7   170.13272 111.17658 135.24506 152.04796 124.37125 128.14396 189.15458
#> 8   135.81538 187.04623 143.87384 168.61528 140.89602 159.74360 147.82195
#> 9   156.02759 112.41761 101.54895 112.71178 163.52870 124.02447 137.05311
#> 10  148.55663 136.67781 117.70422 132.04340 128.36303 155.93249 120.77676
#> 11  147.37855 146.57582 134.20042 154.09449 224.57601 158.11281  93.98070
#> 12  125.36781 123.02736 166.67401 151.33587 168.33483 184.13701 136.13753
#> 13  124.54646 107.23752 127.13551 153.09532 115.23965 115.30652 134.85306
#> 14  163.72859 158.28162 163.28338 209.90804 136.94713 176.37718 165.93936
#> 15  120.22701 113.28105 163.74970 107.70892 155.67277  87.88474 149.50367
#> 16  137.81574 112.90415 189.05538 126.64069 217.27219 167.90982 211.79353
#> 17  131.91614 136.02783 141.38329 135.16321 153.17983 110.93675 126.61138
#> 18  129.60567 128.96617 149.68031 173.83417 173.22632 130.17926 134.80611
#> 19  171.55261 149.93523 186.09427 140.88551 137.64758  83.45722 118.64676
#> 20  131.45777 149.18577 112.76186 141.44448 186.16685 133.14406 161.57713
#> 21  149.80831 138.86409 108.90924 152.30904 213.06155 136.51916 139.43122
#> 22  116.97683  95.85985 158.52102 154.66667 113.87159 156.62482 150.02739
#> 23  182.77501 135.59552 124.87269 169.50364 125.59100 108.00339 174.54884
#> 24  165.74105 155.30377 190.06052 128.23564 165.26022 145.99158 145.01780
#> 25  140.93647 114.19192 155.23020 150.60721 180.76681 117.69801 106.27374
#> 26  194.89418 133.69019 136.22319 135.97372 152.40949 139.14366 142.19628
#> 27  164.36620 123.78715 142.62623 187.68918 143.13754 130.34671 169.02629
#> 28  191.04578 119.88196 126.08444 163.59606 147.40180 121.69841  98.73990
#> 29  196.87715 183.96780 146.86255 162.35796 183.87237 140.97121 138.48660
#> 30  115.23412 160.46819 100.98239 218.83123 161.23953 104.02007 136.81322
#> 31  149.83104 148.05965 177.18590 149.42525 158.76740 179.41185 190.00245
#> 32  151.89761 129.14660 127.95105 224.07558 164.28129 161.26224 125.39593
#> 33  189.69525 124.14401 152.56203 169.28018 131.01331 157.02687 132.12058
#> 34  154.42103 167.69089 131.41644 152.44009 174.60006 153.37258 187.14670
#> 35  123.51667 103.58288 153.61589 138.50957 118.28317 171.41331 147.91120
#> 36  180.75392 115.30614 170.59926 191.91472 117.24297 125.95018  94.75832
#> 37  187.65581  98.27983 118.95905 160.97343 174.13812 110.23462 143.86041
#> 38  147.50841 109.28589 118.42764 171.57077 154.20524 137.62105 105.27876
#> 39  113.71991 154.73517  85.43455 152.29732 158.73077 140.12878 135.51947
#> 40  148.50417 132.93561 141.83522 165.79727 159.64473 135.33079 125.10984
#> 41  112.31216 100.10813 142.96706 179.52423 125.76600 145.49725 135.09231
#> 42  186.12863 164.09251 147.87522 166.09015 128.16367 161.29217  91.31038
#> 43  160.85870 165.06318 161.85898 149.62374 154.33399 141.34024 114.17112
#> 44  171.51375 112.75881 123.84414 133.91396 157.11139  93.00611 131.19634
#> 45  120.71657 135.42958 158.43165 150.49374 134.66021 126.79685 128.05851
#> 46  147.11009 147.97758 137.55537 174.33958 178.39411 146.29877 137.13185
#> 47  150.96065 107.75802 106.86939 174.16729 198.98607 159.86319 129.16088
#> 48  122.40200 173.85546  93.84393 149.21268 108.43827 198.41858 102.56123
#> 49  145.71060 184.43604 161.82637 160.59175 126.13743 120.63289 139.78262
#> 50  138.71062 166.20584 166.64288 135.13081 113.20288 132.65929 139.35055
#> 51  156.96957 156.52124 159.82640 107.45981 139.76443 109.04789  84.97956
#> 52  126.64147 137.13108 154.23209 185.63091 162.25532 124.44340 122.91995
#> 53  151.13144 196.04993 163.80275 129.26295 101.00265 137.79969 119.57466
#> 54  148.99327 141.13613 136.34441 155.17666 189.10630 135.99441 101.44308
#> 55  182.08683 152.60110 131.94845 159.50094 172.37356 134.54900 147.06575
#> 56  183.86982 161.21500 140.58877 185.67734 128.65141 161.89569 150.06740
#> 57  135.72308 177.18155 118.54441 197.44264 143.27535 105.41047 140.02272
#> 58  118.55039 144.11957  71.08985 164.59954 113.00701  71.27692 153.44003
#> 59  189.75409 143.26175 129.31321 111.08754 161.18426 181.49891 173.33022
#> 60  173.91002 172.19206 177.12880 171.01066 162.52999 129.18761 126.40850
#> 61  179.29746 158.73724 134.05999 151.80031 126.89028 161.64477 152.04033
#> 62  144.59465 107.66553 127.69458 141.18148 152.34058 148.30579 147.59300
#> 63  170.49519 125.34278 123.24675 143.68052 118.09088 184.19185 121.28881
#> 64  183.75903 123.93828 143.40189 154.80437 120.48499 146.80417 142.33913
#> 65  192.80492 154.37753 123.10157 182.42221 153.25881 168.64960 120.94441
#> 66  164.15247 187.02961 125.93707 146.20615 205.41728 128.47550  98.83194
#> 67  150.13340 105.51825 171.60253 150.06498 188.76311 138.83832 126.25774
#> 68  167.28617 102.55451 135.20952 135.74095 151.51748 136.55450 135.11492
#> 69  117.13060 140.78150 167.59540 150.89234 155.30199 131.44335 174.75584
#> 70  163.45967 115.80666 128.94005 120.81982 162.94300 173.28061 144.69767
#> 71  129.73990 104.07129 153.05995 126.25019 123.56288 119.81716 134.48185
#> 72  134.25119 146.67943 149.92144 175.47843 218.21671 129.66561 139.84293
#> 73  149.66940 138.17193 142.68524 169.74829 175.28508 147.32769 155.63462
#> 74  122.94739 146.40416 122.16019 138.60088 175.28493 128.60782 177.20132
#> 75  147.56703 155.03214 137.42701 105.23349 132.94125 167.54636 125.66715
#> 76  150.67715 139.79988 209.09299 101.01064 190.41498 178.87115 144.12422
#> 77  158.59715 110.84802 130.54550 166.40325 163.83651  92.06279 111.13580
#> 78  150.02698 127.99221 124.62616 127.88662 100.17900 157.95184 148.22805
#> 79  116.41852 149.98859 113.92559 150.75295 130.35023 140.69964 154.10293
#> 80  143.56056 121.13050 106.97054 164.25079 150.97523 109.00866 114.28974
#> 81  141.24784 179.74988 113.93700 169.01194 147.13138 132.80939 175.48514
#> 82  185.58853 126.16705 104.78670 203.96150 187.67926 100.18006 124.76816
#> 83  167.69264 118.60048 161.96728 139.34381 142.35628 141.06775 128.84129
#> 84  191.21949 118.80412 134.08690 183.55366 159.52757 137.37953 140.27540
#> 85  115.13351 173.24466  92.42364 107.20504 163.31815 140.37023 133.81435
#> 86  117.75885 159.81794 142.25457 200.88043 157.35768  99.71251 173.61293
#> 87  197.32159 165.92298 147.75262 121.01416 186.76731 101.70300 174.44373
#> 88  113.59457 144.28876 108.77469 155.91460 225.46709 133.63882 188.36792
#> 89  138.87860 149.02102 134.47538  85.13207 155.23267 159.54096 160.58591
#> 90  143.11800 165.42636 140.55102 150.36017 129.34828 150.88735  84.07515
#> 91  205.02670 149.36689 108.85698 152.95311 150.46445 113.52899 166.95133
#> 92  185.63797 113.85574 142.68019 113.93734 176.20092 154.00161 134.17364
#> 93  125.20703 136.28206 143.14836 163.77143 138.49396 210.28791 139.49616
#> 94  178.03948 119.08548  95.33924 168.72930 186.09913 142.70247 157.33956
#> 95  168.06517 159.89085 121.47983 135.88168 146.49502  95.94127  94.54484
#> 96   96.48222 129.63035 129.57461 136.18176 193.09185 144.91405 140.64394
#> 97  141.86112 122.63591 137.64916 159.36211  98.82294 137.84152 119.75005
#> 98  161.27096 147.48869 148.75635 194.16094 161.49557 117.12054 176.83770
#> 99  103.11399 109.80443 112.08152 112.64427 112.65079 142.76755 188.22233
#> 100 150.33893 121.77646 119.29472 157.85405 185.64132 138.80662 146.44265
#>         B.4_1
#> 1   108.65081
#> 2   229.86555
#> 3    99.59084
#> 4   172.44769
#> 5   188.19155
#> 6   144.31406
#> 7   164.38633
#> 8   113.46375
#> 9   174.49486
#> 10  145.78893
#> 11  133.29569
#> 12  130.81694
#> 13  129.75609
#> 14  171.90392
#> 15  164.21558
#> 16  135.35670
#> 17  129.68735
#> 18  115.63725
#> 19  152.37431
#> 20  134.44596
#> 21  198.89771
#> 22  167.71871
#> 23  139.79542
#> 24  139.67091
#> 25  120.03217
#> 26  188.37805
#> 27  133.78741
#> 28  164.24534
#> 29  127.60278
#> 30  132.37737
#> 31  147.93191
#> 32  205.15999
#> 33  154.51663
#> 34  146.99129
#> 35  149.26009
#> 36  129.74153
#> 37  139.68914
#> 38  169.82673
#> 39  112.52449
#> 40  152.74021
#> 41  151.92690
#> 42  117.33356
#> 43  157.98251
#> 44  139.04054
#> 45  124.67655
#> 46  150.98623
#> 47  115.58441
#> 48  112.50027
#> 49  115.99470
#> 50  179.73655
#> 51  136.00767
#> 52  127.93666
#> 53  180.51723
#> 54  138.64941
#> 55  110.32430
#> 56  175.15502
#> 57  124.43902
#> 58  176.69121
#> 59  126.24710
#> 60  110.18147
#> 61  109.10352
#> 62  148.71493
#> 63  147.12953
#> 64  134.99982
#> 65  138.71736
#> 66  140.95030
#> 67  206.58200
#> 68  211.40721
#> 69  183.96853
#> 70  146.31164
#> 71  161.50029
#> 72  162.94732
#> 73  168.15421
#> 74  188.56134
#> 75  166.27911
#> 76  222.35013
#> 77  192.72388
#> 78   99.96953
#> 79  155.61801
#> 80  176.40641
#> 81  120.29007
#> 82  210.18107
#> 83  173.26663
#> 84  191.80764
#> 85  171.85377
#> 86  135.95604
#> 87  238.11804
#> 88  185.18357
#> 89  174.32534
#> 90  173.41278
#> 91  146.71994
#> 92  140.83349
#> 93  187.71127
#> 94  156.14824
#> 95  108.62901
#> 96  201.26668
#> 97  156.77116
#> 98  172.78892
#> 99  168.09869
#> 100 111.64985