Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10, 10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(
  subset = rep(c("A", "B"), 50),
  level = rep(1:4, each = 25)
)
rownames(meta) <- 1:100
scRNA <- Seurat::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA,
  by = c("subset", "level"),
  min.cells = 10, max.cells = 20
)
#>         A.1_1     B.1_1     A.2_1     B.2_1     A.3_1     B.3_1     A.4_1
#> 1   123.59201  98.04128 181.51775 115.63350 127.84046 141.60634  95.36373
#> 2   167.87599 111.62761 130.21276 153.98970 146.09368 111.83248 127.92581
#> 3   130.66947 151.09197 162.81838 148.58181 144.16595 122.43653 129.64181
#> 4   131.68781 129.88866 141.54695 163.26356 151.50079 170.36965 171.78285
#> 5   168.33784 172.93749 167.01800 184.69430 173.39262 185.52447 136.19919
#> 6   128.72518 157.56073 129.87571 214.35863 168.79683 171.19924 170.57216
#> 7   139.72189 132.65263 126.92034 112.09509 124.76952 125.04482 149.30838
#> 8   138.26966 116.67406 133.59154 133.39550 122.14016 173.28743 224.92777
#> 9    93.66905 124.20674 129.57388 170.59134 174.94685 132.19209 172.67766
#> 10  138.56245 143.45058 119.86673 170.24989 147.90062 217.71900  92.28940
#> 11  167.02969 130.22280  96.23659 146.19316 178.27875 199.92033 150.55274
#> 12  140.10434 120.64342 162.91404 152.70516 137.00342 136.62424 167.53419
#> 13  191.45524 138.14222 151.74575 150.90769 158.00160 104.94233 110.88455
#> 14  158.06115 158.86946 164.76756 147.54409 115.77470 176.77544 160.60817
#> 15  172.33961 141.44101 122.37072 133.94742 190.45170 135.18646 122.36170
#> 16  126.46427 160.96149 118.37630 197.52786 153.30774 159.52114 120.40665
#> 17  189.73963 147.70243 152.49815 152.82659 185.21284 154.42211 163.78148
#> 18  149.15954 149.65420 130.46970 110.57992 153.28238 170.16136 105.64967
#> 19  165.38348 140.48443 103.98483 189.31340 181.15665 190.84210 183.05523
#> 20  150.56139  83.81064 134.21676 149.91892 208.85604 149.73872 156.85718
#> 21  136.45176 153.59739 135.21070 204.49410 167.07272 116.46285 165.69594
#> 22  142.94997 159.70740 149.47689  90.33052 137.32471  81.61476 133.31427
#> 23  125.29541 129.59859 139.24744 170.92667 158.17078 172.67355 113.38487
#> 24  104.42020 174.83989 162.19368 209.28418 173.04317 130.96860 110.38110
#> 25  113.06051 145.24064 139.61922 144.41085 121.72725  87.62359 180.55210
#> 26   92.68345 131.79646 123.79763 151.04162 122.07022 143.05666 160.91899
#> 27  166.99791 141.30437 154.43365 163.05705 125.00386 168.01323 132.85280
#> 28  206.03718 117.90480 181.19516 117.63025 112.73361 180.60742 131.26387
#> 29  173.53644 154.84956 106.05627 157.70828 149.19693 124.16261  84.74573
#> 30  153.70371 159.37049 136.03098 140.24925 169.25523 137.39448 148.27493
#> 31  165.08640 133.81597 170.01443 169.30584 146.77639 166.33672 152.69772
#> 32  118.95934 170.37623 110.01885 126.48879 157.15207 124.20138 194.06752
#> 33  209.61223 120.54464 113.88979 119.87135 157.31435 173.28568 114.06590
#> 34  182.30698 123.54197 161.64046 141.61230 147.04789  85.21920 112.38889
#> 35  179.25285 159.68659 122.36030 198.28692 177.17097 134.42582 131.90018
#> 36  150.15548  91.95671 111.31081 126.46143 166.29095 129.61796 107.89515
#> 37  138.23409 118.80087 116.47965 153.16238 110.43880 104.61221 150.42053
#> 38  211.11940 158.89254 114.20708 131.72791 169.75151 121.13212 113.13181
#> 39  197.75736 181.22435 126.42546 132.34957 125.00863 143.11906 131.59461
#> 40  170.53252 116.43966 162.65639 242.29976 114.94587 204.33389 135.09176
#> 41  188.82632  98.14322 113.35964 161.32874 122.45653 113.55621 125.14178
#> 42  123.73835 115.87242 140.15317 132.59278 132.85408 162.77807 123.53555
#> 43  187.43585 139.97640 142.67547 160.72268 157.10611 102.20805 130.40114
#> 44  155.47120 147.48001 137.03708 126.81806 169.28862 197.24083 135.38174
#> 45  131.70490 119.07722 146.38009 172.23881 147.64552  86.18046 141.71721
#> 46  131.39422 215.97196 146.63381 165.59996 173.93201 104.17029 161.08134
#> 47  144.71646 155.07184 180.92337 134.40121 178.40779 108.47662 100.09488
#> 48  162.85760  91.99051 121.15406 154.26648 147.89699 137.07971 150.69514
#> 49  169.53509 146.55367 108.27417 176.22146 111.98499 137.38286 135.14320
#> 50  163.38831 115.62404 117.91720 149.42380 134.91613 106.66556 170.02843
#> 51  102.77397 167.48911 161.37746 152.57411 131.55730 169.59959 193.02981
#> 52  147.91462 182.94936 139.66415 154.96861 184.42414 113.05181 106.08263
#> 53  128.94271 142.40825 105.15222 166.06250 230.74765 144.60432 175.97774
#> 54  179.41579 115.82148 158.64612 110.26889 105.29743 137.16876 131.48382
#> 55  135.03135 109.54083 114.47131 136.79091 163.40116 203.10716 123.06205
#> 56  192.20848 133.90603 168.61711 129.48946 144.07780 115.82958 145.85244
#> 57  154.95374 109.06361 178.05163 135.88529 177.21935 120.56999 132.66426
#> 58  185.72637 134.06200 111.22382 127.36313 190.36577 158.94061 117.26821
#> 59  162.80599  88.83178 121.56762 211.96142 141.54714 159.03032 106.63516
#> 60  150.55698 200.10727 108.61795 141.94771 190.33670 158.80028 130.31622
#> 61  158.54485 146.02534  99.80171 124.14440 150.82554 136.24751 175.25466
#> 62  167.77554 148.84355 137.22519 182.30841 144.91653 180.41513 110.20192
#> 63  186.41283 135.09226 134.62032 178.40976 180.52789 132.67424 166.75892
#> 64  157.06412 159.61614 140.78403 158.99383 133.50826 139.81859 204.09587
#> 65  101.43674 170.30497 104.89169 206.03987 137.23514 124.35233 178.71803
#> 66  111.30563 155.06595 127.55501 119.40309 139.01472 119.79711 139.14758
#> 67  124.50722 148.69634 134.55034 124.35149 141.75343 142.12180 136.43876
#> 68  109.06674 115.80121 156.70061  91.96683 104.53159 132.56095 171.27058
#> 69  137.89893 117.10419 136.94757 184.04025 190.67770  97.12747 120.48870
#> 70  186.99676 159.24103 102.74478 150.57800 123.08218 181.03547 132.03290
#> 71  141.92754 112.70756 108.98379 100.71391 103.22543 161.12655 133.29861
#> 72  130.29253 109.54478 134.91912 153.59207 155.57763 120.85345 114.07955
#> 73  123.17685 134.07471 114.21718 103.88185 145.52045 152.11295 143.04522
#> 74  141.66417 173.04186  95.87543 148.03782 120.00158 150.51459 194.59561
#> 75  136.07167 145.85089 116.46614 158.29996 130.75390 104.14031 108.79393
#> 76  131.80105  98.76022 160.48419 139.03297 135.81756 162.97778 132.04517
#> 77  173.67139 117.42854 118.80958 184.70036 191.26998 141.96417  93.49326
#> 78  183.37259 131.92214 152.59363 122.70838 158.36673 168.51980 100.97780
#> 79  168.74458 180.57748 147.16166 117.43905 181.79198 134.03177 112.66167
#> 80  160.05001 141.93404 145.64026 160.90029 160.16618 177.33446 156.63305
#> 81  129.57955 154.28862 142.91304 171.77199 159.05282 143.05894 170.34601
#> 82  155.56171 146.96331 111.19667 157.13012 112.63645 164.94943 164.68915
#> 83  203.53478 134.79227 148.92676 158.62285 144.65668 128.72723  72.27174
#> 84  156.36544 141.32613 156.63033 139.20474 238.90160 118.53272 131.26026
#> 85  143.69647 180.06142 112.54155 168.66183 143.37859 125.93526 153.83204
#> 86  136.91460 146.91946 114.17331 144.39012 167.18928 169.36569 141.61460
#> 87  150.90477 180.63611  97.50502 150.29373 196.09570 134.41871 140.43690
#> 88  126.72261 166.34201 124.61072 166.05760 131.67558 128.60551 108.15059
#> 89  141.50144 181.23253 126.72938 148.93408 124.92108 132.36491 154.42741
#> 90  160.82118 105.47323 119.99211 168.53568 177.59429 147.01113 137.14817
#> 91  109.59151 109.12598 156.04030 129.60388 163.70986  85.96923 178.56212
#> 92  163.55213 145.47254 141.68361 141.98408 114.67568 121.53240 164.41325
#> 93  129.98082 118.75802 118.07980 120.86143 159.83350 120.14872 115.64783
#> 94  171.67762 143.88259 183.96694 146.52270 157.10992 125.17243 128.99555
#> 95  140.57795 132.72514 170.81644 125.15138  89.33805 152.32510 177.75208
#> 96  123.49407 153.27844 186.36569 108.37673 103.86024 148.02384  84.95215
#> 97  122.67349 128.94230 139.69251 133.80958 132.43902 127.25848 185.53207
#> 98  175.32090 141.81217 167.36676 122.41930 185.65192 140.66527 140.02921
#> 99  132.96006 134.85989 132.38319 116.64387 163.20656 129.00697  82.23534
#> 100 119.91921 167.31854 134.42337 133.29633 132.03816 157.48498 128.49139
#>         B.4_1
#> 1   114.71024
#> 2   178.69895
#> 3   209.49355
#> 4   161.71837
#> 5   178.13219
#> 6   108.64173
#> 7   141.33843
#> 8   114.30516
#> 9   135.21033
#> 10  152.76713
#> 11  161.89333
#> 12  159.69931
#> 13  133.53619
#> 14  193.10213
#> 15  175.77470
#> 16  152.01249
#> 17  166.19496
#> 18  119.12501
#> 19  194.78020
#> 20  112.87069
#> 21  143.48416
#> 22  104.69817
#> 23  128.56918
#> 24  166.62259
#> 25  161.92924
#> 26  146.20886
#> 27  138.30623
#> 28  157.55120
#> 29  139.72767
#> 30  177.80611
#> 31  160.45128
#> 32  172.90846
#> 33  161.16933
#> 34  141.64510
#> 35  149.41839
#> 36  158.08425
#> 37  157.11993
#> 38  141.89189
#> 39  118.99392
#> 40  158.12953
#> 41  143.32826
#> 42  125.32485
#> 43  147.07274
#> 44  152.76659
#> 45  117.05277
#> 46  142.31266
#> 47  110.99583
#> 48   88.19826
#> 49  126.00033
#> 50  152.65249
#> 51  226.01546
#> 52  156.51386
#> 53  157.21430
#> 54  133.33613
#> 55  129.55731
#> 56  145.71572
#> 57  128.95756
#> 58  144.18060
#> 59  147.48140
#> 60  253.54050
#> 61  165.29252
#> 62  166.47345
#> 63  136.82173
#> 64  157.78111
#> 65  136.10515
#> 66  156.22974
#> 67  200.19104
#> 68  156.74730
#> 69  147.27645
#> 70  127.77367
#> 71  192.37636
#> 72  142.08500
#> 73  139.18215
#> 74  108.64378
#> 75  165.25146
#> 76  185.06101
#> 77  151.52694
#> 78  130.71060
#> 79  148.13101
#> 80  117.41397
#> 81  153.55849
#> 82  144.06124
#> 83  205.98070
#> 84  135.78192
#> 85  204.07744
#> 86  145.33974
#> 87  164.55703
#> 88  158.59554
#> 89  174.10379
#> 90  136.05719
#> 91  133.60218
#> 92  106.29213
#> 93  197.25808
#> 94  191.08835
#> 95  155.97487
#> 96  175.22595
#> 97  142.83013
#> 98  130.19848
#> 99  118.22585
#> 100 137.87824