Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10, 10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(
  subset = rep(c("A", "B"), 50),
  level = rep(1:4, each = 25)
)
rownames(meta) <- 1:100
scRNA <- SeuratObject::CreateSeuratObject(counts = counts, meta.data = meta)
#> Warning: Data is of class matrix. Coercing to dgCMatrix.
pseudo_samples(scRNA,
  by = c("subset", "level"),
  min.cells = 10, max.cells = 20
)
#>        A.1_1     B.1_1     A.2_1     B.2_1     A.3_1     B.3_1     A.4_1
#> 1   139.8384 111.55458 107.35186 143.68834 124.42652 147.72141 170.44506
#> 2   131.4885 134.87377 166.67795 165.77052 143.68210 170.85570 150.70509
#> 3   126.5073 141.32515 144.38825 150.71325 179.63404 165.64364 177.45607
#> 4   153.3528  91.02960 108.61328 174.04426 153.90823 192.39153  80.61644
#> 5   182.6625 133.13289 194.89728 119.02481 187.70613  91.82668 186.64988
#> 6   147.3761 115.11032 134.89602 147.45441 101.60993 129.23670 156.65983
#> 7   128.4434 213.51685 121.26351 120.15210 170.72898 129.27134  93.44136
#> 8   140.6072 112.54275 186.98780 209.02781 146.27250 146.11814 147.15583
#> 9   217.2099 117.01875 202.21725 189.35180 164.33986 139.20729 171.45748
#> 10  189.9579  99.45938 138.27893 104.74523 173.23637 116.09263 137.98918
#> 11  153.9598 163.42902 123.49714 118.61299 128.95438 126.74826 146.51908
#> 12  155.4010  63.98523 136.92159 233.48888 149.40716 176.19462 130.34258
#> 13  140.6873 103.80325 187.79440 144.81737 125.99921 192.58906 149.25573
#> 14  130.0712 205.80349 138.18661 151.26526 173.12511 146.51057 169.43272
#> 15  125.7020 125.81416 129.54773 112.36941 131.00741 179.96809 113.66395
#> 16  144.5074 187.94467 140.33158 122.58036 179.33307 103.33738 149.89358
#> 17  173.4256 141.07598  95.41463 165.31697 180.65981 164.57982 165.96054
#> 18  130.3655 116.20579 157.66343 196.76735 134.25917 140.20406 115.02765
#> 19  119.5506 180.70887 129.46808 157.68715 126.47183 126.20650 121.47245
#> 20  188.6185 109.55208 160.41669 160.00333 123.85548 110.17028 138.73517
#> 21  155.2111 134.89032 101.41938 160.24904 152.27508 133.26305 113.55648
#> 22  139.1129 129.75239 191.19791 138.39735 188.00512 105.48175  91.68409
#> 23  102.4263  99.70393 124.09768 137.69232 162.43889  96.66954 147.50602
#> 24  129.7944 165.15392 167.36011 143.24731 139.72187 135.23012 107.66612
#> 25  256.1190 104.74902 148.50585 105.96566 170.03041 167.88806 205.41953
#> 26  143.5029 136.59973 171.74506 159.75371 144.09458 125.22359 147.05961
#> 27  119.1153 109.77083 162.53606 149.51261 184.41049 157.39903 120.10717
#> 28  163.4483 171.07243 147.41867 210.88412 160.78351 174.13003 133.57001
#> 29  172.8605 119.31483 163.95773 179.31483 107.89739 135.28992 114.44614
#> 30  113.5093 125.76339 159.60997 129.15467 191.95331 159.91144 122.48829
#> 31  133.2000 100.92780 170.44210 159.66328 144.44032 143.98844 123.76475
#> 32  131.9174  84.62902 126.72930 182.48225 142.36811 158.65640 114.82587
#> 33  114.5268 134.68175 197.04318 160.88404 180.12506 207.01105 139.83466
#> 34  137.8706 169.63580 156.98804 150.52291 151.87344 104.65680 168.64789
#> 35  170.0503 160.23572 156.10468 171.13142 166.06353  81.60186 121.66658
#> 36  140.6279 117.85817 153.32955 181.91571  95.69643 125.56895 156.36814
#> 37  160.4048 115.52583 130.79595 173.97672 144.99437 131.41140 177.08438
#> 38  149.5111 101.10053  91.33136 128.45346 171.92035 133.86392 174.04967
#> 39  114.4019 128.02835  99.71471 173.50913 136.38982  99.27717 162.92624
#> 40  175.1700 164.97842 174.83490 111.10618 184.15774 180.56952 182.63416
#> 41  100.1892 116.58451 159.09408 137.75667 130.00169 124.03567 126.92960
#> 42  155.7043 140.14541 147.06177 173.13578 150.56954 147.09817 107.55443
#> 43  198.6974 176.37720 117.71389 161.97550 133.67893 135.91243 155.12147
#> 44  192.8717 123.07055 138.19937 215.49787 118.29841 160.95782 120.36154
#> 45  191.3662 106.32544 142.89612 153.13031 127.28017 178.21922 135.95753
#> 46  139.3717 122.30525 187.67372 150.70129 139.19651 130.39796 128.33429
#> 47  131.2980 127.12835 128.86145 129.38547 154.44938 159.09476 119.34225
#> 48  107.5694 211.01261  77.50564 160.25541 133.08935 100.87421 138.23525
#> 49  125.7597 133.80620  83.31201 170.27252 129.76601 162.76823 131.90537
#> 50  102.3514 108.42855 207.61097 161.68062 193.16228 167.26259 137.29790
#> 51  130.8600 157.06389  90.02597 157.29193  86.49744 115.73863 155.08884
#> 52  157.9186 130.24549 172.83662 131.81624 191.99884 158.75131 133.84037
#> 53  164.6312 175.79915 159.94467 121.17608 143.67359 158.02148 132.64086
#> 54  117.7692 137.77267 133.42297 141.30890 135.38697 179.27828 124.88769
#> 55  106.5864 174.70580 132.21946 143.99324 166.75373 141.59124  95.26182
#> 56  168.0458 191.94465 104.40085 158.91679 176.47924 158.75382 155.59008
#> 57  198.6347 136.90699 103.52567 129.57842 102.77497 116.37414 100.72493
#> 58  152.8312 165.42220 159.64823 175.34190 148.03260  99.25402 159.21807
#> 59  148.2442 145.89430 184.27661 170.13189 163.23681 125.95744 174.16025
#> 60  195.6842 169.03246 159.93288 148.24149 182.54702 161.02893 137.46596
#> 61  171.4905 153.74534  90.81970 109.38815 126.56497 131.95869 128.59891
#> 62  128.0343 149.71837 145.48924 138.80212 192.72399 102.46950 171.80061
#> 63  147.7028  92.63553  88.32647 130.96960 146.63786 146.53622 122.00247
#> 64  184.8743 186.77901 101.50762 161.80034 134.70343 165.97828  82.98951
#> 65  170.5660 146.12659 128.50558 188.66498 152.85635 169.86357 189.86742
#> 66  192.6795 177.55343 155.10043 123.00092 176.96463 114.74878 189.95503
#> 67  160.3313 113.95384 191.63719 171.29674 120.46423 170.03670 107.09406
#> 68  163.9410  94.31192 143.68181 155.21739 184.08555 138.76138  97.66230
#> 69  135.9708 117.07436 172.86552 107.10523 108.84516  90.39318 111.94586
#> 70  168.9193 105.66552 113.49336 202.95920 132.52186 139.99816  97.47600
#> 71  151.6902 151.19084 147.24903 141.15777 167.30802 146.00357 160.87801
#> 72  141.8681 207.98248 159.93025 121.40675  96.78638 127.99886 132.69393
#> 73  147.0336 146.70768 189.94374 166.83911 119.92552 102.95866 129.28144
#> 74  146.1407 129.25769 136.90265 138.10783 149.89737 149.16760 172.79314
#> 75  148.5161 206.78369 192.87242 144.74245 150.87365 109.93430  98.21656
#> 76  117.6831 112.67107 126.84876  79.31688 121.86346 138.74519 120.27616
#> 77  147.1288 141.98128 107.84413 148.27420 130.05743 208.52746 103.90289
#> 78  133.1494 116.84530 169.73095  99.15492 170.14295 129.15530 123.20145
#> 79  209.9020 114.57765 165.00623 181.70151 103.15598 194.33224 132.61367
#> 80  107.6037 134.37084 207.35527 117.24333 117.77712 126.59215 166.99270
#> 81  179.3972 163.97474 171.71563 139.06789 198.79433 113.34909 169.45172
#> 82  118.5715  67.78247 138.27016 156.79548 143.26023 120.79135 165.27228
#> 83  118.5681 100.45385 105.83303 135.31004 167.17309 161.99539  99.60512
#> 84  145.2694 116.75806 141.62567 121.91330 177.74621 135.39543 112.94654
#> 85  139.3558 112.36682 179.48145 112.66058 169.00129 167.83895 194.39127
#> 86  160.0052 102.18638 104.00457 128.67458  83.91995 145.44545 128.04988
#> 87  134.8358 132.60670 147.51137 145.62623 100.31255 151.28042 148.28074
#> 88  107.1164 137.52397  63.11366 139.69120 198.32632 149.48081 108.81440
#> 89  140.8406 115.43666  92.45571 163.39245 171.85119 135.01998 149.42337
#> 90  203.8098 125.31933 106.98639 158.36917 217.10806 142.41991 153.73851
#> 91  160.5022 131.15772 117.15730 137.07469 160.44083 103.05726 203.96756
#> 92  149.0310  93.42562 129.79130 161.31474 187.03406 151.38523 141.28463
#> 93  126.7677 210.49749 206.51259 128.93049 174.18238 144.76375 128.75013
#> 94  153.8496 129.18853 111.33756 212.84922 173.19474 155.63360 201.64657
#> 95  134.7531 187.65088 121.52642 149.02518 158.45809 188.37843 145.23853
#> 96  164.8808 181.16367 161.94207 190.08317  92.93216 143.75490 155.96344
#> 97  128.5695 149.24105 168.24867 112.18264 151.46872 115.17001 110.22292
#> 98  183.3367 121.72759 128.22289 110.66976 162.83146 154.58781 146.67753
#> 99  186.2797 149.87510  99.51693 135.32988 139.34927 195.92076 130.16032
#> 100 135.3186 181.85991 114.06121 157.83416 135.96957 126.83673 101.25802
#>         B.4_1
#> 1   158.49892
#> 2   134.63334
#> 3   155.55719
#> 4   153.28957
#> 5   148.17942
#> 6   166.18151
#> 7   167.25153
#> 8   206.22147
#> 9   163.65606
#> 10  125.99658
#> 11  182.34513
#> 12  142.09590
#> 13  189.14550
#> 14  127.37219
#> 15  108.98059
#> 16  155.02482
#> 17   99.94314
#> 18  193.35853
#> 19  151.92245
#> 20  164.18991
#> 21  153.69891
#> 22  138.70912
#> 23  161.74752
#> 24  174.96031
#> 25  194.51848
#> 26  141.85251
#> 27  182.82584
#> 28  119.15022
#> 29  220.88779
#> 30  166.68868
#> 31  136.23759
#> 32  126.79635
#> 33  106.14310
#> 34  149.46515
#> 35  123.88609
#> 36  136.59908
#> 37  149.66203
#> 38  243.41959
#> 39  157.27787
#> 40   83.49168
#> 41  140.88756
#> 42  131.64812
#> 43  186.39991
#> 44  117.30780
#> 45  146.78543
#> 46  175.38413
#> 47  181.79696
#> 48  130.14582
#> 49  153.47887
#> 50  103.78551
#> 51  175.18139
#> 52  138.14131
#> 53  146.60358
#> 54  130.79649
#> 55  142.55796
#> 56  179.03840
#> 57  148.44245
#> 58  133.04075
#> 59  111.17146
#> 60  143.18292
#> 61  238.90670
#> 62  122.94031
#> 63  118.91186
#> 64  169.17452
#> 65  142.64796
#> 66  145.66296
#> 67  137.20188
#> 68  138.58083
#> 69  127.46610
#> 70  122.35749
#> 71  111.58981
#> 72  213.06504
#> 73  161.31976
#> 74  114.11675
#> 75  132.91803
#> 76  201.38780
#> 77  118.82929
#> 78  163.03268
#> 79  182.46332
#> 80  161.93347
#> 81  135.48553
#> 82  159.17417
#> 83  175.26589
#> 84  138.34334
#> 85  138.04075
#> 86  161.29310
#> 87  133.49780
#> 88  136.97576
#> 89  126.29045
#> 90  235.72253
#> 91  229.93564
#> 92  191.64828
#> 93  113.82935
#> 94  137.34025
#> 95  171.59219
#> 96  169.44703
#> 97  177.86841
#> 98  156.40595
#> 99  159.17250
#> 100 186.99325