Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10,10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(subset = rep(c("A","B"),50),
                   level = rep(1:4, each = 25))
rownames(meta) <- 1:100
scRNA <- Seurat::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA, by = c("subset","level"),
               min.cells = 10, max.cells = 20)
#>        A.1_1     B.1_1     A.2_1     B.2_1     A.3_1     B.3_1     A.4_1
#> 1   168.6726 167.15361 140.50193 121.04303 159.38640 146.25589 165.30896
#> 2   160.0913 109.83119 183.96403 162.92653 184.45399 137.21175 143.25293
#> 3   159.1602 122.80269  97.30097 200.80127 139.95844 129.82415  93.88533
#> 4   110.8695 130.78303 168.11215 156.07098 131.93888 140.49188 113.77651
#> 5   149.2944 123.74158 118.80285 166.70700 121.76201 144.58473 138.90699
#> 6   155.6124 103.77998 125.30816 134.26210 201.28508 128.14867 102.10934
#> 7   198.8720 133.94146 166.41827 122.06912 115.36985 164.88914  90.15386
#> 8   188.6998 153.01499 102.49346 134.91277 107.27438 160.37674 117.94136
#> 9   153.0277 155.81186 170.55736 144.33511 146.77052 145.34889 179.24179
#> 10  138.3008 104.49165 162.16759 122.91151 113.63970 138.56639 116.11592
#> 11  173.4665 155.12702 139.80582 187.57894 110.61765 127.54091 114.72044
#> 12  185.2531 106.23531 177.33293 140.16484 133.54550 137.04997 138.73110
#> 13  150.7156 113.37124 161.41112 138.69436 160.20417 105.96739 150.01438
#> 14  166.0056 137.18558 164.44328 155.60495 148.45391 136.20637 130.64382
#> 15  181.9814 119.41457 127.11388 159.45768 176.26258 120.95514 207.03013
#> 16  128.8443 113.35777 182.21109 136.89726 113.88297 111.75851 113.31817
#> 17  165.8445 121.25255  85.45127 172.09541 119.11039 197.61705 149.89988
#> 18  202.9630  64.29046 132.65861 135.99325 142.73027 182.09546 169.39410
#> 19  201.5485 125.74987 111.31950 138.95322 134.35614 155.63174 136.97672
#> 20  172.4605 131.24112 132.17440 136.37903 107.64468 172.67580 147.72111
#> 21  136.6158 148.62836 121.81876 205.13986 141.24415 211.92370 183.08659
#> 22  165.0009 162.04022 134.79870 146.20343 140.92246  99.40128 155.45134
#> 23  126.9521 165.76969 122.24189 149.46169 180.65445 139.79523 131.76829
#> 24  172.5630 157.91276 171.22305 206.12986 156.14605  94.62134 170.27983
#> 25  158.0163  92.93096 119.12824 221.51317 151.43802 126.42460 105.59310
#> 26  176.4969 103.39293 147.57917 160.32495 188.71089 120.33989 147.66491
#> 27  152.4804 136.21090 130.84596 114.21911 127.04094 148.66110 116.11385
#> 28  194.8143 156.06356 132.80764 143.13515 174.32457 130.74487 107.69534
#> 29  161.8225 160.63079  88.39733 117.64810 167.73505 150.69268 100.44739
#> 30  137.9202 115.59239 199.63424 159.86964 137.19302 145.32417 177.98344
#> 31  165.8228 187.00098 106.55773 124.43441 178.73415 171.19391 147.97799
#> 32  168.6522 124.54963 124.63158 164.39816 115.98253 143.74882 112.36411
#> 33  129.2390 172.65890 133.66714 168.26742  99.83031 134.49519 133.01184
#> 34  178.5257 101.69445 177.69403 216.95960 155.20341 107.71485 120.07386
#> 35  146.0059 153.71330 137.51004 115.92327 121.87615 155.33653 144.29605
#> 36  126.3273 158.48015 107.90104 138.89017 165.82576 148.40104 132.38675
#> 37  153.3465 205.61291 144.93996 142.89128 186.30041 162.97287 124.57924
#> 38  137.5364 228.31688 106.18571 205.85369 155.61983 161.67547 152.57961
#> 39  137.7651 136.84379 139.88089 191.61832 128.48591 147.60646 142.78341
#> 40  154.5691  80.58365 133.90440 176.51861 131.45131 117.93416  78.11380
#> 41  149.6105 117.79929 150.39201 164.74268 186.67547 152.53095 144.98848
#> 42  172.6235 196.18870 151.81145 118.75683 168.60796  93.11945 182.93918
#> 43  127.9170 196.41894 131.77827 141.11680 105.74039 167.21700 133.45795
#> 44  164.5612 136.12474 127.93186 152.58539 132.57685 126.70012 130.97746
#> 45  126.3026 184.80894 111.80249 166.90478 113.52145 119.12117 129.64140
#> 46  190.0670 128.00600  95.16143 172.42337 105.10218 130.50838 153.63066
#> 47  126.2561 132.71364 144.64173 172.94304 171.37640 177.31645 136.00913
#> 48  140.9496 121.72004 150.78421 193.15796 219.30127 141.96482 142.51154
#> 49  185.6602 146.24766 118.29474 126.65722 176.63437 142.18255 112.43884
#> 50  143.0218 143.43078 120.45715 144.18504 140.10490 121.24162 184.20299
#> 51  180.4504 149.46237 145.93314 164.02201 179.55998 151.79415 142.05426
#> 52  122.1280 144.40778 109.17951 199.24922  98.63161 106.41825 128.53263
#> 53  158.4300 196.47848 183.69239 138.05882 109.76710 123.48355 121.16551
#> 54  127.9935 169.83314 149.83892 130.83704 161.04587 126.00571  93.29950
#> 55  131.3908  84.07451 132.42834 194.72518 160.02191 167.07383 153.50817
#> 56  168.3145 114.72973 140.09032  96.08341 189.05503 161.10831 118.35665
#> 57  168.9842 106.42300 121.96413 187.22835 155.97364 153.23935 124.36176
#> 58  163.7749 146.47504 109.39103 116.15516 194.13490 100.87630 129.50596
#> 59  126.5712 137.26027 141.76526 108.64430 141.79569 147.01979 113.93973
#> 60  215.8579 166.28056 130.75153 233.67212 143.65372 131.60806 161.54526
#> 61  139.8095 111.49893 183.32910 137.35425 172.27189 115.70936 142.64799
#> 62  175.9576 203.76214 150.61499 183.21687 166.30541 141.78405  90.17495
#> 63  157.9771 124.59968 202.73646 134.36427 191.37261 136.08425 137.16730
#> 64  112.0444 153.09837 160.20832 127.91247 180.02209 118.21415 154.29029
#> 65  181.3040 100.37686 149.45657 128.39948 112.24034 173.02246 176.47213
#> 66  177.2784 169.57406 139.96477 120.21138 202.27167 159.99550 108.90612
#> 67  113.5900 173.96714 138.79088 170.16052 186.58657 112.17871 118.15845
#> 68  132.5046 131.35463 151.83324 154.85522 215.24101 167.20173 165.32560
#> 69  166.9559 178.66881 142.15642 172.25750 144.06822 116.05626 139.25616
#> 70  203.9614 111.52843 148.20811 161.03678 158.30822 167.87485 140.80241
#> 71  151.5505 141.96562 120.86322 116.34879 114.71805 111.28291 134.61085
#> 72  147.6092 130.53891 111.77790 175.53155 111.33879 153.58383 172.69369
#> 73  121.8882 156.52126 142.38203 145.64699 128.14107 124.55456 157.47157
#> 74  178.0516 119.93887 150.88611 153.05476 160.75836 140.89779 118.70545
#> 75  146.9490 124.16078 151.38594 153.03571 105.22980 153.45943 152.98285
#> 76  163.5984 110.71020 129.04067 158.52782 165.89956 116.68333 129.29328
#> 77  131.4612 158.27481 164.19672 180.67782 122.56319 142.64362 162.18519
#> 78  165.3873 172.49262 123.59188 172.62222 155.94160 107.29185 109.73061
#> 79  124.8403 127.04861 134.91639 114.78869 165.08622 185.60473 114.03716
#> 80  124.5950 155.68334 182.58391 151.45647 159.80933 181.70224 200.05191
#> 81  213.5225 140.31719 166.89162 138.50960 144.22913 111.44584 117.29249
#> 82  154.0033 147.06779 120.20159  94.29336 136.80346 137.78094 176.08730
#> 83  135.4103 127.58317 117.91933 217.76462 138.40989 152.40752 147.10221
#> 84  136.9012 108.84588 140.27902 144.15817 180.76334 136.96278 143.66259
#> 85  178.7541 141.75025 148.87327 164.66414 151.72091 167.00998 170.65846
#> 86  117.9313 217.88325 144.39670 143.65944 214.11253 112.91161 128.83401
#> 87  148.8563 123.72280 148.79339 145.14157 162.66185 115.17294 104.04986
#> 88  175.9961 135.08707 199.22338 190.47458 133.97028  67.01034 124.98181
#> 89  196.5383 162.25839 142.46443 150.27706 146.24688 183.61390 133.31654
#> 90  188.7932 108.81617  76.79189 152.21044 173.62618 160.41893 147.28633
#> 91  193.1964  95.20526 111.38236 135.17150 138.01566 159.11938 158.82877
#> 92  143.3456 110.56416 147.04463 130.84586 122.28640 157.05477 150.62802
#> 93  174.0327 179.22070 153.01649 156.15334 148.47944 123.62562 112.72128
#> 94  104.4114 121.28676 118.30360 170.06474 138.38638 136.41756 142.87318
#> 95  136.9150 145.42294 166.17279 188.60875 118.93904 148.27659  97.85425
#> 96  234.3396 170.97968 126.37679 140.50629  96.75725 132.17815 124.07195
#> 97  197.0098 173.02962 142.86351 187.26721 152.61609 154.69034 122.01276
#> 98  140.9355 117.54534 170.08386 109.71864 133.14562 147.94633 107.07071
#> 99  125.6660 153.82975 162.17162 127.66196 185.76931 138.41205 169.48143
#> 100 175.5506 155.91319 173.30332 176.86611 129.68969 166.14228 190.71303
#>         B.4_1
#> 1   175.02183
#> 2   169.29259
#> 3   136.88014
#> 4   142.34799
#> 5   145.73188
#> 6   171.02014
#> 7   151.84392
#> 8   134.79169
#> 9   102.60861
#> 10  124.15581
#> 11  184.28741
#> 12  171.67074
#> 13  104.78543
#> 14  127.46809
#> 15  176.90111
#> 16  124.39964
#> 17  157.97531
#> 18  159.78719
#> 19   95.51520
#> 20  140.12890
#> 21  178.67388
#> 22  186.43564
#> 23  205.55897
#> 24  105.17786
#> 25  141.63427
#> 26  107.10161
#> 27  144.08464
#> 28  226.90095
#> 29  152.51151
#> 30  168.53530
#> 31  177.81279
#> 32  140.89346
#> 33  202.85781
#> 34  231.56244
#> 35  141.66144
#> 36  102.26197
#> 37  116.07522
#> 38  159.23932
#> 39  181.49057
#> 40  198.26725
#> 41  130.53332
#> 42  115.42627
#> 43  157.74765
#> 44  151.09964
#> 45  121.73710
#> 46  134.21628
#> 47  116.61875
#> 48  154.67467
#> 49  133.99867
#> 50  175.70955
#> 51  147.62777
#> 52  134.32421
#> 53  149.51392
#> 54  159.15181
#> 55  165.67017
#> 56  168.47652
#> 57  170.90675
#> 58  145.86334
#> 59  170.75055
#> 60  175.52378
#> 61   99.98230
#> 62   91.24043
#> 63  145.38971
#> 64  114.21575
#> 65  118.41553
#> 66  140.18280
#> 67  179.89235
#> 68  105.32237
#> 69  160.92870
#> 70  122.43939
#> 71  166.86735
#> 72  177.61917
#> 73  140.98244
#> 74  155.85839
#> 75  135.28478
#> 76  112.59105
#> 77  113.68614
#> 78  119.49219
#> 79  142.25719
#> 80  164.15007
#> 81  151.22257
#> 82  182.83573
#> 83  166.79672
#> 84  152.28604
#> 85  170.99984
#> 86  210.20263
#> 87  123.04244
#> 88  142.71233
#> 89  122.78642
#> 90  153.05216
#> 91  142.34271
#> 92  123.21177
#> 93  164.94705
#> 94  162.54251
#> 95  149.84606
#> 96  145.85124
#> 97  218.89181
#> 98  120.64778
#> 99  133.08537
#> 100 131.22198