Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10, 10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(
  subset = rep(c("A", "B"), 50),
  level = rep(1:4, each = 25)
)
rownames(meta) <- 1:100
scRNA <- SeuratObject::CreateSeuratObject(counts = counts, meta.data = meta)
#> Warning: Data is of class matrix. Coercing to dgCMatrix.
pseudo_samples(scRNA,
  by = c("subset", "level"),
  min.cells = 10, max.cells = 20
)
#>         A.1_1     B.1_1     A.2_1     B.2_1    A.3_1     B.3_1     A.4_1
#> 1   223.70533 159.37037 132.35178 165.54931 237.8143  94.30530 112.81703
#> 2   142.09014 136.68601 158.67542 169.43934 172.3260 149.87242 157.92574
#> 3   116.10675 178.88338 150.04720 136.94680 138.1821 172.94121 114.24117
#> 4   194.80086 139.47874  91.68807 165.02032 158.0605 143.91964 130.61910
#> 5   125.86370 139.38585 212.05769 102.15165 164.6988 106.13603 149.49380
#> 6   121.71042  95.47504  97.56403 105.76511 156.9708 196.25764 133.64582
#> 7   160.66232 104.18761 137.75727 150.51667 197.8011 145.64203 180.37138
#> 8   192.30825 124.73265 128.77350 122.19962 172.6685  95.29508 151.74297
#> 9   170.86261 123.41791 198.14599 165.06314 175.0165 170.12690 185.83441
#> 10  125.07749 172.29327 129.52770 165.51727 120.1720 111.05458 114.25820
#> 11  163.15356 136.07083 145.81698 113.78929 206.2401 143.42667 142.12612
#> 12  118.19569 139.64598 136.84480 183.01740 151.3674 185.59193 114.16975
#> 13  161.46651 132.17734 146.17588 167.42102 162.7697 141.22145 133.08709
#> 14  167.39938 110.65322 169.18768 154.08207 168.0561 116.06191  92.90594
#> 15  190.77687 122.17480 115.39117 184.04192 164.9808 123.62490 101.18855
#> 16  171.77922 121.70631 127.92507 170.58512 146.9586 118.65808 183.39426
#> 17  167.23353 149.22596 142.95500 117.81782 160.1103 129.62827 134.91882
#> 18  126.88696 176.91780 172.53613 155.54057 152.5744 142.87871 155.48188
#> 19  164.06340 125.40732 143.34818 183.18997 183.9136 134.71135 151.51818
#> 20  179.96325 154.34492 115.71874 199.03418 160.4703 157.29231 109.44270
#> 21  200.16405 116.23708 107.27109 160.79410 125.0549 127.81534 182.97365
#> 22  117.81898 122.62415 136.11009 121.35762 146.1488 122.59328 124.70569
#> 23   69.57662 110.57038  99.54289 142.79402 108.3960 155.33520 109.02556
#> 24  169.50516 206.44924 111.42884 171.19464 156.6559 154.07733 134.72364
#> 25  121.88413 139.83723 116.15246 132.29167 166.6306 143.67261 117.23656
#> 26  145.11652 124.22076 166.04194 197.20626 118.2041  82.49064  94.84878
#> 27  135.09526 159.59369 144.68159 179.89918 117.7551 143.68881 124.98303
#> 28  177.61788  95.46001 155.87014 142.08539 157.9195 141.38008 171.49419
#> 29  150.74088 141.90585 111.94395 146.92643 178.8512 135.89430 133.86455
#> 30  158.44479 130.47791 172.73844 195.74838 157.3746 172.00879 130.56340
#> 31  151.30096  68.46236 161.48609 144.94457 121.5595 140.01773 108.01800
#> 32  147.81491 135.12146 137.33438 134.90468 137.8247 110.34359  84.83025
#> 33  146.79721 170.30747 100.67400 161.06326 159.4530 135.81522 129.21948
#> 34  189.29042 119.86524 137.65105 171.91903 123.8290 136.20029 148.59766
#> 35  112.02136 131.33001 168.88572 129.11732 135.4630 142.49550  92.02071
#> 36  155.76075 109.95435 173.65762 182.86162 123.5004 132.02475 172.07030
#> 37  193.19020 128.87677 154.02069 190.15411 164.4965 146.23773  86.36857
#> 38  129.71914 170.69849 120.61624 142.15315 186.9138 209.65185 126.05150
#> 39  119.15050 176.48562 144.04911 171.54947 189.2626 109.91064 132.82814
#> 40  128.81372 123.94566 107.09679 171.81159 143.9806 107.39507 169.12158
#> 41  139.97698 158.96206 126.40014 147.57356 162.5288 152.82788 158.36642
#> 42  135.74326  91.33245 154.69373 157.05090 177.1895 146.82904 154.86840
#> 43  115.90209 172.91812 181.25556  97.75905 167.6910 183.59245 132.35852
#> 44  117.37187 113.91596 112.06288 112.37367 216.3532 172.76802  95.55427
#> 45  114.15480 169.91465 217.67308 148.13376 206.6574 136.34676 158.38127
#> 46  173.05118 149.31280 162.91653 125.25079 137.3823 114.91072 134.72699
#> 47  167.32634 102.85175 147.68156 171.77579 199.2674 130.11747 133.88742
#> 48  189.18511 178.30624 132.27161 114.95862 147.2197 125.91289  96.58058
#> 49  109.88459 139.53899 126.60443 146.94785 178.0408 127.12209 142.96079
#> 50  102.04682 122.16320 142.71231 160.32366 126.1764 153.83924 200.04090
#> 51  161.64736 197.01119 151.96896 128.47387 143.8004 131.26146 134.30664
#> 52  149.84606 165.21940 175.08170 200.78878 109.3777 154.42368 133.46499
#> 53  123.04335 111.88641 153.93945 174.64102 165.1166  99.06724 126.79038
#> 54   99.51780 137.83086 103.80043 155.68445 157.0213 165.82272 126.73754
#> 55  142.16532 123.53415 112.43463 124.41473 134.3439 108.50817 101.74987
#> 56  123.88599 130.50071 118.61762 150.00400 101.7055 180.92794  80.87294
#> 57  110.27950 123.76329 104.93945 146.77487 103.5377 175.47678 119.39461
#> 58  149.11305 137.59184 191.06860 167.72770 153.1372 161.08060 177.80099
#> 59  143.47903 165.60744 113.25171 150.93224 171.3059 188.65791 121.27839
#> 60  120.77357 149.15851 163.08082 144.90673 151.4952 152.02918 123.09484
#> 61  147.45839 137.67727 123.54212 109.91518 147.5642 104.13552 108.26433
#> 62  140.50255 149.38572 141.71784 188.69169 134.6169  93.27244 149.61347
#> 63  162.48349 153.98517 128.38471 161.21088 177.3547 171.60473 185.69595
#> 64  167.19940 188.72179 138.30686 111.47795 130.8704 105.49707 120.21312
#> 65  156.99944 117.73578 142.75516 111.53741 128.5902  94.76461 231.54184
#> 66  114.85552 142.22060 197.46648 162.57749 159.4488 147.32503 165.71031
#> 67  146.82720 129.57881 121.26218 194.37119 119.3993 118.25902 119.42182
#> 68  188.68405 102.78177 145.65146 172.66486 144.7578 145.24189 107.42464
#> 69  100.21499 130.39290 190.78833 129.31171 148.5384 154.33327 141.35263
#> 70  131.79466 123.31301 166.10675 173.51915 125.5193 149.75756 148.24045
#> 71  110.95071 126.21411 135.92916 188.19762 180.3791 158.59320  94.11211
#> 72   97.73214 115.24840 125.39262 183.80099 235.5659 123.72537 130.89392
#> 73  175.36210 135.87773 182.98038 175.14013 127.8262 173.06960 107.18742
#> 74  125.12313 149.48144 134.42802 167.55895 150.9381 105.96924 141.23994
#> 75  165.95599 178.58471 117.33843 142.96710 176.8071 187.05018 112.63546
#> 76  183.10474 111.99439 174.21038 175.81600 193.7043 145.23052 151.97205
#> 77  134.95122 157.78781 157.44218 117.26441 182.6065 102.54470 144.84961
#> 78  134.84667 161.88893 146.96784 144.75970 155.1668 105.83165 154.01555
#> 79  163.02407 172.76576 139.63659 150.95320 176.1340 131.69189 121.37867
#> 80  137.90289 194.60764 202.02372 141.56057 177.9063 212.77388 123.22959
#> 81  156.99447 124.54281 124.00662 187.37073 173.6241 163.21611 142.63468
#> 82  149.18235  94.71276 149.98994 134.43637 166.2337 154.23050 139.56496
#> 83  202.10951 146.94386 155.96088 148.87445 181.5071 187.75333 138.51989
#> 84  133.70073 134.09184  83.20992 161.76122 104.1366 141.75164  94.23013
#> 85  162.51634 178.16876 161.30856 172.08638 106.5034 226.78972 153.15694
#> 86  181.74598 174.58159 170.81099 174.09392 105.8402 135.52464 147.38886
#> 87  112.47535 170.97718 197.96384 144.34909 174.0286 143.74511 160.22452
#> 88  174.62511 185.91495 128.32089 111.11485 108.5851 128.85577 124.02023
#> 89  153.37345 152.16987  60.98881 142.77955 126.0465 163.69371  99.85685
#> 90  176.86550 124.96973 143.94763 176.62285 138.7651 117.68612 100.77327
#> 91  181.65745 157.98513 153.80317 142.67602 194.3300 126.63215 142.43322
#> 92  147.91799 159.30615 122.52182 178.23905 157.7291 118.57872 127.86833
#> 93  148.13881 100.13607 164.56599  99.45615 163.6483 153.33957 186.24240
#> 94  133.79389 132.36650 155.86223 123.13626 180.9238  90.23512 140.13400
#> 95  145.09825 140.73365 148.27321 138.28787 129.5878 141.25066 144.52695
#> 96  183.39506 142.67267 182.40669 122.59148 128.0082 137.91797 175.37519
#> 97  165.55367 205.11864 136.06286 160.32769 138.1094 181.68545 147.61928
#> 98  164.39289 130.69904 107.76214 139.58184 148.1837 150.45901 152.39470
#> 99   83.93642 122.14576 148.06022 163.03799 166.5829 127.96142 130.82377
#> 100 147.01250 126.45051 118.18439 117.79432 138.4811 149.95137 138.86036
#>         B.4_1
#> 1   173.06278
#> 2   143.46510
#> 3   166.25450
#> 4   149.36091
#> 5   188.38129
#> 6   121.29152
#> 7   176.31319
#> 8   196.76226
#> 9   156.25663
#> 10  155.07638
#> 11  129.21850
#> 12  167.68293
#> 13  156.53900
#> 14  163.84647
#> 15  113.47761
#> 16  178.30743
#> 17  177.83611
#> 18  140.25616
#> 19  135.94117
#> 20  185.22871
#> 21  159.80559
#> 22   86.41023
#> 23  168.94523
#> 24  118.94518
#> 25  199.13456
#> 26  166.44383
#> 27  153.37149
#> 28  169.50940
#> 29  114.40327
#> 30  147.90866
#> 31  125.24778
#> 32  165.29151
#> 33   87.67412
#> 34  128.04806
#> 35  156.47831
#> 36   97.09518
#> 37  165.87034
#> 38  120.74353
#> 39  122.63706
#> 40  154.19982
#> 41  146.95866
#> 42  119.29450
#> 43  131.86825
#> 44  156.37219
#> 45   73.42707
#> 46  215.56829
#> 47  125.25682
#> 48  159.77641
#> 49  153.21330
#> 50  145.37009
#> 51  199.92777
#> 52  181.41614
#> 53  157.71728
#> 54  113.26942
#> 55  124.11208
#> 56  137.66082
#> 57  173.44059
#> 58  171.69728
#> 59  145.36929
#> 60  151.12073
#> 61  138.75339
#> 62  199.01715
#> 63  128.02965
#> 64  162.05200
#> 65  146.87294
#> 66  139.27641
#> 67  176.28986
#> 68  169.64541
#> 69  110.27633
#> 70  158.99647
#> 71  150.34824
#> 72  202.26434
#> 73  167.58758
#> 74  183.97365
#> 75  160.39011
#> 76  122.85919
#> 77  120.43152
#> 78  151.49237
#> 79  151.03626
#> 80  162.85458
#> 81  102.88088
#> 82  191.96227
#> 83   98.86832
#> 84  126.64361
#> 85  150.40208
#> 86  107.41059
#> 87  141.30393
#> 88  127.49545
#> 89   89.53715
#> 90  139.31321
#> 91  130.50813
#> 92  122.30942
#> 93  169.44876
#> 94  182.92801
#> 95  133.78734
#> 96  142.64289
#> 97  134.03510
#> 98  168.76050
#> 99  148.65586
#> 100 177.58862