Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10,10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(subset = rep(c("A","B"),50),
                   level = rep(1:4, each = 25))
rownames(meta) <- 1:100
scRNA <- Seurat::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA, by = c("subset","level"),
               min.cells = 10, max.cells = 20)
#>         A.1_1     B.1_1     A.2_1    B.2_1     A.3_1     B.3_1     A.4_1
#> 1   114.75905 120.39973  89.06582 166.7077 156.69414 104.87367 179.79009
#> 2   205.88578  87.31616 154.18142 178.8229 154.96896 129.78499 188.77282
#> 3   217.62870 153.16733 144.69254 157.3426 197.36947 175.31754 167.95490
#> 4   207.13378 146.05388 170.89768 160.7545 128.99835 146.43960 141.30968
#> 5   128.38853 158.51266 165.51404 108.1722 182.02624 130.88499 179.78190
#> 6   150.00062 157.85671 107.32420 124.9730 144.94172 155.68813  83.63427
#> 7   178.92676 157.08829 180.38958 179.5418 116.90395 152.28349 117.90507
#> 8   189.16628  98.48586 129.01729 102.9626 141.51005 133.42325 176.11462
#> 9   129.66293 167.78963 187.86645 131.8696 171.45796 171.08627 155.08158
#> 10   83.19759 112.17834 100.16715 183.7284 182.09379 118.27631 136.24689
#> 11  131.49884 118.88867 178.74139 155.0327 146.45854 130.78994  83.58002
#> 12  108.39687 108.20839 113.31687 179.2614 140.64420 133.62572 126.45235
#> 13  123.81047 138.77239 132.31570 210.1162 187.49798 137.95923 151.97490
#> 14  186.79697 106.65391 140.27146 108.6964 183.72606 166.43225 187.15729
#> 15  158.61041 146.87737 158.59954 174.9272 177.86083 182.95758  92.45733
#> 16  214.02014 122.07736 174.34645 128.9623 154.01156 140.76328  73.99506
#> 17  176.43363 120.63840 181.34919 190.0700 154.34205 181.34037 129.20477
#> 18  119.73830 126.11457 187.95972 146.9796 208.22348 184.89900 137.21763
#> 19  175.75877 131.41608 111.00465 165.6122 111.65066 116.37022 156.28087
#> 20  148.07259 139.20873  88.27210 196.8261 147.83995 193.16997 122.01480
#> 21  185.33772 113.82264 167.13322 179.1781 111.61921 186.13341 160.30098
#> 22  126.88511 115.98089 209.25198 150.7929 152.23831 141.41657  71.46323
#> 23  145.22524 155.70752 129.62338 131.3374 165.30236 136.28083 185.66584
#> 24  126.58065 119.61875 152.31785 162.2325 157.85643 160.58580  88.73436
#> 25  149.89139 157.45778 142.17308 181.6448 114.26941 134.52716 144.36438
#> 26  134.66877 123.25902 152.01271 148.0370 125.66498 183.78229 131.31036
#> 27  131.01799 129.09125 135.02242 140.9236 184.60179 165.88967 150.14469
#> 28  152.00046 171.71474 130.06431 180.5712 183.25859 181.81960 132.46514
#> 29  173.80930 106.39895 190.19224 113.3858 140.49610 148.10735  97.49861
#> 30  154.64247 160.91057 124.92583 154.7740 139.37273 141.98991 148.50635
#> 31  138.84419 144.29063 161.96225 174.7581 146.99244 168.21430 122.23170
#> 32  173.78342 134.54133 144.55841 156.6226 133.21360 120.68905 162.79606
#> 33  141.71415 145.28334 209.15707 168.5349 149.67544 140.00845 129.76406
#> 34  188.27389 134.52534 130.52720 168.6584 169.23812 139.51541 151.68694
#> 35  208.32070 126.21428 152.89361 196.0257 168.41844 118.25293 123.26466
#> 36  137.32197 121.33453 171.65722 183.5089 143.35517 157.17804 132.01495
#> 37  151.69226 168.39151 122.16247 187.4930 133.25408 134.22798 120.20581
#> 38  210.86683 157.50676 151.65303 153.1612 155.12318 147.71090 113.39426
#> 39  129.07814 144.98465 107.53147 143.3905 206.09704 148.60667 226.64293
#> 40  140.07106 193.62041 133.56194 139.2272 117.94790 133.95034 171.85689
#> 41  161.23411 152.41309 150.82440 191.4454 196.71097 144.02246 175.47485
#> 42   86.82360 115.93556 145.55718 163.7002 132.73004 142.66111 188.76160
#> 43  146.42523 131.49535 132.37946 178.7878 128.40832 127.27050 147.02151
#> 44  136.37981 154.32110 154.28769 101.8788 167.25311 143.61363 140.90006
#> 45  133.07348 101.42604 103.82643  99.6597 221.61380 122.79565 137.37093
#> 46  135.37102 131.44853 143.70837 120.5887 105.50154 163.84402 118.27477
#> 47  153.32309 174.84297 153.90880 146.5482 118.10098 168.21240 132.25225
#> 48   97.59005 117.40214 144.52949 163.7819 125.13122 135.65155 188.68077
#> 49  162.74001 139.57047 108.59590 137.0038 119.44040  82.88384 196.04968
#> 50  140.27566 138.29176 203.77106 133.7036 132.69842 145.43424 104.99588
#> 51  150.80007 117.30341 176.75946 245.1814 197.03405 145.31247 121.51714
#> 52  151.81452 110.21463 120.13460 133.4281 129.45282 118.53583 137.59041
#> 53  136.09396 175.78322 194.75838 166.8011 170.53140 132.14422 146.04412
#> 54  164.09016 137.52747 100.59945 124.7653 140.11015 175.03224 152.61719
#> 55  164.98637 133.67834 184.07260 235.9722 102.38339  95.30155 134.04451
#> 56  198.71540 152.17149 187.00437 135.9192 139.45673 134.08977 164.20760
#> 57  127.52586 117.77382 164.30009 199.4042 150.04686 104.86545 118.15003
#> 58  132.41299 115.67813 137.64475 152.0847 162.96010 121.53987 160.73343
#> 59  202.35333 131.37892 159.96505 137.5848  89.77503 186.49622 182.18508
#> 60  132.67597 102.40492 166.30421 128.8398 196.11266 161.61493  82.98828
#> 61  123.59651 190.72232 185.08652 148.3453 129.98270 219.36786 124.22959
#> 62  168.93830 111.33758 178.65334 167.8635 160.06310 157.27149 151.58117
#> 63  148.80525 177.28321  79.56150 159.2789 177.18453 101.84409 121.48983
#> 64  149.51825 140.13577 175.10617 167.2383 120.23542 164.85915 141.24322
#> 65  132.01463 114.09000 172.97598 142.0114 138.97813  96.02673 129.80107
#> 66  144.42938 159.23543 109.06883 165.4284 105.39274 134.99663  98.30062
#> 67  115.04482 127.53461 145.06168 174.6235 163.17161 149.62737 102.87350
#> 68  140.86269 106.21179 173.12178 147.7036 105.51124 120.99558 114.70237
#> 69  167.32710 119.84015 141.29764 203.5629 197.73008 140.42847 154.72180
#> 70  193.30517 110.68344 133.17852 132.2168 138.48513 146.72285  70.64461
#> 71  107.83999 160.31852 161.38654 116.7287 110.43766 181.63499 185.66976
#> 72  182.27421  84.88359 195.88446 148.8094 138.95279 162.82131 144.85245
#> 73  163.93816 106.43321 121.04475 147.3578 149.86591 211.07705 171.86100
#> 74  100.83410 140.38168  74.10347 153.3071 184.47268 122.30029  91.97941
#> 75   79.21062 122.20220 124.42888 173.8805 109.17028 163.30830 166.25943
#> 76  200.01601 127.04220  89.17015 111.6096 145.29267 138.65597 109.40761
#> 77  151.75117 133.42181 136.39215 103.7019 100.92382 134.98753 165.92588
#> 78  171.01288 144.15835 176.32668 133.0110 130.77087 139.15196 138.54277
#> 79  178.62610 182.28563 134.65628 124.6726 139.34902 154.71815 154.71226
#> 80   96.61319 164.08637 158.12603 184.8925 194.73217 136.57109 130.39523
#> 81  194.05012 121.14393 151.02262 138.8407 175.90357 176.08356 130.21404
#> 82  171.06675 126.88409 110.33444 125.3881 211.29486 174.29832 127.51573
#> 83  164.02580 160.98500 132.38669 181.4691 173.53106  98.32093 119.62269
#> 84  136.89804 171.57677 133.51913 139.9355 210.93654 100.59529 184.10892
#> 85  141.87431 148.75864 196.04966 119.1211 140.03175  83.18074 116.85097
#> 86  138.35617 110.12337 136.79740 132.2749 174.79692 108.83813 147.87281
#> 87  165.18613 126.31325 123.08925 154.0043 163.63924 133.45752 141.10849
#> 88  182.76397 147.49862 100.05765 143.7651 145.02082 122.20434 131.77889
#> 89  177.24924 171.29544 141.59831 176.2853 139.90391 148.47659 118.35576
#> 90  123.05567 103.39267 102.24276 153.9175 155.74649  99.01084 160.61151
#> 91  142.45702 157.10431 168.17465 169.1146 166.65346 120.21136 119.59436
#> 92  143.14913 182.02348 131.46680 135.0326 120.55320 159.42275 103.71743
#> 93  174.68485 119.98795 116.89403 109.8070 102.58882 169.86309 112.83062
#> 94  229.94227 145.14404 125.55655 118.7245 150.84669 114.10171 205.71480
#> 95  157.44968  94.01808 151.77374 119.2183 144.17179 107.45744  85.50910
#> 96  116.43101 145.26069 146.18740 148.5693 166.09888 113.37795 145.34365
#> 97  124.59629 104.41076 114.15821 171.3952 167.46137 105.62348 122.67751
#> 98  106.00680 134.39985 178.00069 149.1332 126.80444 135.55859 161.41851
#> 99  151.48341 169.22029 167.84619 148.3804 178.85893 117.57654 123.56172
#> 100 153.76338 105.87793 120.87272 149.7171 147.52352 151.10736 125.50830
#>         B.4_1
#> 1   145.73331
#> 2   124.88531
#> 3   117.56235
#> 4    89.48424
#> 5   143.85270
#> 6   166.11511
#> 7   129.91249
#> 8   111.81083
#> 9   141.96601
#> 10  162.33838
#> 11  129.21344
#> 12  136.02957
#> 13  141.65924
#> 14  136.32657
#> 15  110.34416
#> 16  155.52705
#> 17  168.94981
#> 18  169.35610
#> 19  163.28503
#> 20  123.60936
#> 21  101.56289
#> 22  152.51404
#> 23  155.43565
#> 24  138.75186
#> 25   84.87082
#> 26  128.98292
#> 27  181.01845
#> 28  156.99585
#> 29  191.75228
#> 30  201.21756
#> 31  154.10545
#> 32  137.54804
#> 33  146.86966
#> 34  210.46298
#> 35  195.63143
#> 36  163.24511
#> 37  140.35862
#> 38   91.37368
#> 39  175.43219
#> 40  174.25948
#> 41  138.19986
#> 42  176.88852
#> 43  149.21832
#> 44  121.80872
#> 45  151.58671
#> 46  155.57373
#> 47  117.44503
#> 48  140.20990
#> 49  192.88320
#> 50  143.99717
#> 51  180.08574
#> 52  161.33726
#> 53  153.08871
#> 54  174.51637
#> 55  168.81660
#> 56  160.77327
#> 57  146.63424
#> 58  180.58593
#> 59  129.91498
#> 60  150.75532
#> 61  161.47765
#> 62  137.62318
#> 63  190.50007
#> 64  147.39391
#> 65  154.64335
#> 66  135.41220
#> 67  164.64964
#> 68  149.87644
#> 69  117.96180
#> 70  161.23078
#> 71  166.76790
#> 72  139.67007
#> 73  131.85574
#> 74  163.57569
#> 75  204.35353
#> 76  114.65797
#> 77  187.72272
#> 78  157.24973
#> 79  170.33900
#> 80  162.35452
#> 81  119.60897
#> 82  132.76881
#> 83  155.45407
#> 84  104.59427
#> 85  178.85747
#> 86  141.11648
#> 87  154.80664
#> 88  150.62069
#> 89  150.70857
#> 90  184.81231
#> 91  105.48014
#> 92  162.21611
#> 93  134.77369
#> 94   95.21347
#> 95  158.56070
#> 96  189.98737
#> 97  131.91144
#> 98  159.48811
#> 99  162.11563
#> 100 143.87949