Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10, 10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(
  subset = rep(c("A", "B"), 50),
  level = rep(1:4, each = 25)
)
rownames(meta) <- 1:100
scRNA <- Seurat::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA,
  by = c("subset", "level"),
  min.cells = 10, max.cells = 20
)
#>        A.1_1     B.1_1     A.2_1     B.2_1     A.3_1     B.3_1     A.4_1
#> 1   147.1503 138.49750 140.08538 143.19052 155.70997 124.94657 135.79267
#> 2   176.6041 120.64991 136.91828 141.42459 183.64145 150.42821 151.52675
#> 3   161.2738 206.21607 118.88772 158.76564 113.91974 132.17271 122.22740
#> 4   182.4120 131.96648 103.05749 133.26305 131.74969 121.91963 129.94311
#> 5   122.8174 154.93575 156.12390 161.20134 139.70426 160.47233 114.27212
#> 6   145.9252 148.04154 118.13945 151.10803 152.07578 102.42176 128.61768
#> 7   163.6031  87.54240 121.23184 122.97206 120.98543 124.18301 145.20331
#> 8   209.7148 140.13556 186.92438 167.78798 128.14005 159.15237 143.92773
#> 9   214.1832 158.79986  89.68076 122.39121 162.33434 128.91377 131.78682
#> 10  171.3008 135.17067 147.22455 225.11620 206.85537 141.00166 148.49081
#> 11  129.0880 117.23453 108.18030 180.06677 138.89569 163.56644 190.48060
#> 12  123.4907  93.36534 145.48816 157.47884 147.45743 126.94192 158.56466
#> 13  139.8300 138.42649 140.31250 196.40380 169.31918 145.68873 113.54929
#> 14  169.6167 146.46768 162.67599 165.71036 108.42201 130.06183  96.88001
#> 15  132.1373 127.28618 153.96120 171.03191 147.43537 153.69577 127.52756
#> 16  187.3098 116.65513 173.36781 162.39855 142.73771 164.89859 227.84102
#> 17  209.5733 113.23216 138.21386 124.28725 143.49879 193.12564 134.07688
#> 18  160.1513 106.28255 155.77960 227.91811 156.24558 116.21238 109.88711
#> 19  153.9058 152.52786 117.94928 114.08643 152.30596 144.87138 181.43255
#> 20  181.9298  99.77095 114.48060 188.98550 209.81817 178.97980 104.99004
#> 21  191.5568 122.43490 112.87898 134.38063 157.37205 138.10828 141.33961
#> 22  180.7688 134.32696 149.49902 184.41864 131.46332  83.88335 202.61518
#> 23  173.6951 122.86641 107.10170 122.16238 109.26920 126.45609 122.63089
#> 24  206.3781  87.76855 122.63358 147.15144 142.57861 119.48409 131.71073
#> 25  131.0300 138.13216 113.82743 160.92025 112.32849 101.07328 121.91332
#> 26  203.0166 125.86431 178.84784 133.87157 133.49115 110.94529 150.93489
#> 27  119.2286 162.95058 115.26245 176.73305 136.21995 159.37682 181.45359
#> 28  178.2984 134.01567 153.09597 127.22476 133.29506 103.98948 126.11144
#> 29  157.0113 146.67378 154.56130 159.49438 159.17770 160.01783 146.06986
#> 30  205.5129 133.61460 196.55352 161.77359  95.17941 123.43512 110.77715
#> 31  195.9857 122.78736 132.73724 170.55388 181.52874 200.53092 116.48913
#> 32  142.3830 174.94216 127.85134 144.33370 194.68890 168.42754 190.86464
#> 33  132.1778 140.99684 144.80464 136.81146 157.94257 101.06220  94.82346
#> 34  150.8018 143.60267 117.89456 149.94599 104.89483 149.18720 157.66436
#> 35  147.2062 110.61155 146.61409 186.79813 182.11461 138.09460  95.20742
#> 36  148.8788 214.60415 150.96115 154.80443 112.77744 219.34987 160.77585
#> 37  166.9496 131.14155 171.58821 123.94695 201.06121 156.38121 113.99681
#> 38  113.6826 162.25060 145.67044 186.44086 118.03686 124.42620 141.64624
#> 39  182.7408 103.43799 142.44038 159.30622 166.14863 129.90443  82.76664
#> 40  165.0025 140.44966 154.88533 135.50783 145.65726 146.74301  87.70447
#> 41  148.1301 161.38856 125.37088 196.14767 187.21736 143.91221 121.09159
#> 42  175.5762 141.82052 177.63214 171.32837 197.01585 102.59039 132.51311
#> 43  201.1082 124.47456 134.57786 130.57596 159.53308 138.67975 110.96969
#> 44  143.9199 181.98100 148.89590 163.48487 169.27319 160.52798 157.11499
#> 45  178.0546 103.70880 124.87171 150.08223 120.74778 159.98804 125.65108
#> 46  143.8243 140.82876 113.64811  92.57965 167.93105  90.68376 143.80382
#> 47  165.3517 140.21667 135.29677 238.50908 133.43228 159.65769 153.87047
#> 48  182.9454 137.16892 135.72637 160.89965 177.28476 206.36351 165.69280
#> 49  103.0122 142.15540 115.35474 130.79416 150.28711 170.17152 123.48687
#> 50  199.7566 147.60294 135.99290 212.08740 165.37317 172.42118 109.37064
#> 51  175.5849 167.61395 106.83165 206.03198 141.18453 121.15112 147.42013
#> 52  156.5958 126.67985 152.62410 203.72205 144.36560 109.95716 106.25453
#> 53  150.1569  84.44430 142.96445 172.78333 101.23274 142.12889 188.12541
#> 54  152.1927 104.61959 113.10008 158.84282 143.10928  78.59430 135.71740
#> 55  131.4865 147.47872 129.35652 180.66415 139.59403 134.62837 145.73871
#> 56  115.1528 135.32250 150.39608 171.30063 145.82654 147.50822 141.35802
#> 57  175.9258 165.08309  96.75602 158.51604 179.48547 175.20792 137.39820
#> 58  203.6403 146.95536 117.78500 142.44346 156.39712  93.80030 159.92049
#> 59  155.9038 148.53520 167.65250 142.71339 161.48385  96.48807 126.38415
#> 60  188.9839 159.38697 150.99834 144.69507 138.87444 175.63558 158.07788
#> 61  110.1870 171.83687 126.71399 188.05710 150.76119 107.50529 201.18219
#> 62  108.6105 134.78593 165.23893 158.18894 155.97263 121.27536 181.00304
#> 63  120.1238 171.81985 135.46706 142.34812 145.87407 124.58284 141.00461
#> 64  162.7745 177.05147 110.86297 147.23456 123.57539 152.49693 128.30087
#> 65  137.2187 185.24251 147.46343 194.02331 154.33540 166.16195 119.76069
#> 66  158.2793 142.36406 125.16614 101.40322 117.97807 103.20838 169.76996
#> 67  180.1756 130.09511 136.77326 135.75397 180.05551  64.63350 132.95375
#> 68  165.3482 146.53349 137.43098  93.48857 143.73044 133.60523  90.64133
#> 69  152.8177 172.65776 111.17364 106.95671 135.32945 113.57690 117.92914
#> 70  178.8908 109.06893 149.90504 123.98682 191.76135 164.94695 152.86997
#> 71  161.0886 155.98200 133.69419 166.66059 139.42036 152.40445 111.80973
#> 72  118.6117 100.31926 156.19236 192.64077 136.61290 139.66745 207.37517
#> 73  203.4810 234.40183  88.70524 160.52999 150.25307 115.10692 187.31243
#> 74  127.4740 122.31653 117.58243 145.42774 133.52953 166.53163 126.07116
#> 75  175.1499 143.52482 157.17924 199.06952 143.20065 168.33287  98.74954
#> 76  139.5144 126.54013  92.60861 163.76324 193.90068 157.76837 135.96169
#> 77  135.8708 157.01712 176.03103 169.07272 192.90077 179.95421 175.03974
#> 78  139.9806 156.75102 113.46972 149.37664 100.32784 133.82612 148.80500
#> 79  161.0303 172.51608 129.05112 169.16311 136.61179 199.97129 193.37108
#> 80  189.7146 105.36196 128.60488 170.92281 147.43531 119.18521 118.11828
#> 81  146.6124 203.15131  87.22395 144.89423 153.68301 149.65341 157.27600
#> 82  123.8303 166.27290 177.08626 140.85249 132.46418 173.80482 171.02591
#> 83  152.1974 105.77755 132.96606 143.22052 133.01990 196.40843 117.63404
#> 84  110.9826 193.15539 157.30422 151.11125 145.25606 125.61955 153.62959
#> 85  170.9019 197.74768 135.28293 148.15417 134.88565 104.15615 132.42802
#> 86  149.5363 236.42707 122.91562 162.39279 161.13606 136.64326 171.94941
#> 87  102.6951 144.90956 173.71338 203.73528 171.04991 123.93941 109.67769
#> 88  168.0395  73.46041 130.42109 122.08311 182.27635 153.84067 189.77181
#> 89  141.8060 106.77642 154.02513 144.63091 134.18323 119.54131 159.79567
#> 90  164.0104 194.80749 142.73293 146.52724 129.61210 173.22048 152.98006
#> 91  144.4085 176.34288 188.34868 160.27053 184.71421 149.60300 156.67810
#> 92  215.4807 126.83865 141.42673 179.18486 158.89776  72.76622 165.28023
#> 93  130.9943 124.14789 102.26767 165.26903 170.79782 202.38922 149.76301
#> 94  164.5302 112.90650 139.77644 163.08273 132.40426 128.70751 122.31172
#> 95  184.6269 136.29841 153.16044 136.94002 199.76723 140.72453 122.17036
#> 96  176.6697 109.68755 141.63941 149.25121 104.19130 124.17574 151.26526
#> 97  161.1620 153.23769 142.53802 121.03426 168.44496 151.27662  95.72871
#> 98  140.0873 177.69128 126.33318 156.44739 205.43321 143.65687 151.88035
#> 99  216.3934 153.01640 104.62107 156.66988 129.93083 168.07548 110.55345
#> 100 114.1868 128.60025 152.95928 136.47870 185.96430 150.00132  93.29867
#>         B.4_1
#> 1   146.08027
#> 2   119.42136
#> 3   147.76115
#> 4   155.22352
#> 5   130.48365
#> 6   129.51698
#> 7   124.23451
#> 8   100.15455
#> 9   131.25179
#> 10   94.11744
#> 11  126.95376
#> 12  113.81715
#> 13  162.35462
#> 14  140.48747
#> 15  185.04269
#> 16  118.60064
#> 17  129.92253
#> 18  122.96631
#> 19  158.32725
#> 20  121.02027
#> 21  155.94624
#> 22  165.00526
#> 23  189.05450
#> 24  115.61137
#> 25  126.26174
#> 26  196.62986
#> 27  160.08531
#> 28  147.59981
#> 29  148.36430
#> 30  163.99939
#> 31  177.24860
#> 32  128.32891
#> 33  123.28535
#> 34   90.64954
#> 35  120.75460
#> 36  181.98485
#> 37  171.94511
#> 38  106.69891
#> 39  172.63188
#> 40  153.91466
#> 41  178.84910
#> 42  141.62080
#> 43  155.37745
#> 44  113.68571
#> 45  149.35303
#> 46  127.97574
#> 47  172.50275
#> 48  199.18145
#> 49  128.04415
#> 50  131.52116
#> 51  160.73407
#> 52  114.97293
#> 53  144.62528
#> 54  212.42966
#> 55  109.06574
#> 56  149.54124
#> 57  121.28798
#> 58  143.11957
#> 59  152.27455
#> 60  195.64526
#> 61  167.62986
#> 62  148.58683
#> 63  159.48561
#> 64  151.95101
#> 65  187.26549
#> 66  191.53462
#> 67  133.51348
#> 68  118.38886
#> 69  147.11203
#> 70  146.52383
#> 71  209.07561
#> 72  119.73074
#> 73  147.36496
#> 74  187.25506
#> 75  165.14363
#> 76  176.26583
#> 77  132.37474
#> 78  179.08297
#> 79   84.07785
#> 80  122.60792
#> 81  118.08357
#> 82  127.34273
#> 83   87.34759
#> 84  162.92935
#> 85  170.43108
#> 86  146.56748
#> 87  160.25063
#> 88  185.24222
#> 89  103.07833
#> 90  154.55693
#> 91  174.17091
#> 92  153.10349
#> 93  104.64392
#> 94  165.52357
#> 95  171.49160
#> 96  144.02168
#> 97  127.58525
#> 98  198.34786
#> 99  146.37335
#> 100 191.00597