Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10, 10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(
  subset = rep(c("A", "B"), 50),
  level = rep(1:4, each = 25)
)
rownames(meta) <- 1:100
scRNA <- Seurat::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA,
  by = c("subset", "level"),
  min.cells = 10, max.cells = 20
)
#>         A.1_1     B.1_1     A.2_1     B.2_1     A.3_1     B.3_1     A.4_1
#> 1   158.51251 105.00211 148.49870 160.82385 141.72361 141.44172 139.44705
#> 2   152.60985  99.49997 169.66238 113.62723 141.86424 150.17804 113.87831
#> 3   119.20780 158.72698 154.97619 133.16153 148.54658 130.11879 142.77981
#> 4   152.33282 122.92195  62.93860 202.73269 169.62708  90.89131 160.76893
#> 5   139.17059 116.51522 158.91039 175.31091 163.76329  86.00016 155.78309
#> 6   121.08847 133.78375 113.26042  99.28629 158.19386 105.65143 109.16725
#> 7   165.67583 108.52704 157.95706 184.72681 153.81045 109.44087 153.03837
#> 8   115.96058 111.88208  74.43141 164.75475 154.14128 174.64707 137.43878
#> 9   109.07510 196.14793 156.37224 132.21840 155.94713 137.54255 103.18439
#> 10  130.13146 152.69219 160.16756 151.64430 120.39076 158.01190 149.00860
#> 11  139.39357 151.75595 120.62502 144.06274 112.32125 125.63397 122.70687
#> 12  164.08583 131.38092 132.30722 176.98022 130.77408 117.05349 142.03495
#> 13  152.39333 118.97928 207.56824 131.40371 165.38746 155.34803 174.85531
#> 14  106.14244 147.88561 139.45578 141.04658 165.48982 139.51699 165.64355
#> 15  122.86354 120.56632 116.21855 169.99097 169.87240 113.59820  97.04946
#> 16  123.94859 172.27203 119.72167 113.57211 133.06398 131.93126 109.69736
#> 17   90.80574 154.53646 117.29554 130.99370 103.48217 123.11679 159.40185
#> 18  140.21442 153.89948 104.20114 126.64114 187.04983 157.55295 132.92211
#> 19  154.28947  95.51986 118.60501 115.25016 155.62017 114.58327 160.04558
#> 20  192.11902 172.80203 160.52358 188.83046 217.98343 137.21609 104.87568
#> 21  158.39901 110.14452 149.24033 129.56019 173.64059 190.19104 113.33217
#> 22  140.66473 108.88798 145.51938 123.09264 126.31374 135.53222 174.12427
#> 23  161.98406 156.10649 162.53220 159.93735 137.61642 168.29260 158.15483
#> 24  169.22654 148.22766 139.18294 134.92251 120.80599 113.19547 114.64011
#> 25  114.17972 134.09165 130.21622 156.56086 102.27079  85.70275 175.49870
#> 26  135.91501 183.44799 115.39602 171.82778 139.47179 184.44085 178.70996
#> 27  185.26572 144.28681 115.27504 100.89594 185.29060 117.58977 122.79619
#> 28  160.57662 179.55461 128.97074 163.36470 146.04960 114.16338 123.35574
#> 29  193.97927 118.06395 181.18077 149.29968 155.64276 146.08627 135.49579
#> 30  195.28462 197.62354 144.65228 144.73497 198.98538 177.29159  83.02130
#> 31  165.73513  70.37334 120.58895 189.41483 170.32899 172.04729 154.28509
#> 32  150.54810 137.72166 131.84175  99.32699 189.36880 131.05774 105.67987
#> 33  112.93678 128.83945 148.55955  94.82819 148.88044 129.10444 171.40472
#> 34   86.89105 163.23084 148.83846 106.36830 158.21628 148.67371 174.50576
#> 35  163.90796 151.06842 140.55654 181.22601 180.51615 117.32637 131.86846
#> 36  214.33547 119.47518 122.22614 155.36815  92.04362 128.99334  86.28176
#> 37  191.83523 150.63193  99.65563 147.89453 122.48301 155.40235 139.66712
#> 38  174.97189 136.94421 125.45190 187.20305 157.59591 118.27996 161.61302
#> 39  168.45689 143.07930 123.38443 156.58396 161.80445 142.79394 118.94966
#> 40   91.26626 116.71874 136.68123 113.69454 152.21024 172.45750 166.32450
#> 41  159.17771 134.60249 114.25836 171.61077 172.20840 106.67892 143.64694
#> 42   83.56903 164.93120 136.89960 136.04346 166.89407 130.69849 185.59710
#> 43  169.94498 171.13542 160.45322 157.33833 143.18256 120.72988 156.75675
#> 44  135.04985 127.35969 205.67367 109.32492 143.02899 140.34533 196.37379
#> 45  190.11759 103.33245 136.21813 133.87490 182.93329 105.75943  88.20468
#> 46  165.96795 115.34985 148.63042 125.81935 126.96299 120.75115 139.38105
#> 47  127.99695 159.84226 150.93196 185.35529 148.69195  82.11019 117.67580
#> 48  195.59277 153.66867 139.76634 135.03871 188.50213  67.29432 116.41589
#> 49  130.08302 127.12088 156.61955 132.67673 156.75527 104.28487 131.10874
#> 50  139.04014  84.02051 167.41711 176.93936 220.97473 177.70875  98.78693
#> 51  145.43888 118.84519 127.36147 152.06814 106.90196 116.10675 145.40061
#> 52  168.22707 135.21534 149.56235 159.58990 135.63749 153.18553 103.24752
#> 53  129.28442 127.39722 102.97673 140.27684 144.58842 153.60613  86.38455
#> 54  188.35266 157.19380 112.62098  84.29209 167.78548 154.08934 137.46434
#> 55  132.26678 151.63699 144.23386 159.64558 164.04188 160.14291 104.55122
#> 56  160.79862 163.34947 169.86556 174.38823 168.24048 121.81077 100.75996
#> 57  119.62465 128.96768 142.73161 142.35954 119.32707 179.82550 131.49061
#> 58   91.35840 134.96257 100.69728 154.20044 124.15272 148.71953  85.30675
#> 59  187.07053  96.46396 138.55230 144.07533 203.77251 136.85732 114.90865
#> 60  129.53126 146.86900 131.88140 129.17726 149.06280 148.48931 113.94071
#> 61  129.54005 107.60509 102.94498 209.29893 139.00123 204.27331 151.33782
#> 62  129.32723 186.09183 130.95092  92.98971 102.50282 147.41975 156.40730
#> 63  139.43036 137.64592 129.97247 136.69924 133.93348 134.52847 130.51233
#> 64  109.71205 147.94384  87.62251 165.73034 112.30469 140.69594 101.45336
#> 65  136.58065  78.41808  82.15589 119.73700 181.75192 173.16944 170.21065
#> 66  146.24040 160.25290 132.76793 134.08622 128.71395 235.20854 135.27126
#> 67  132.04355 116.50576 130.71225 192.14966 161.32681 184.88310 185.15866
#> 68  167.45937 136.19589 151.08334 151.39703 170.64085 120.45959 138.09228
#> 69  116.51663 146.06722 106.01783 145.39484 161.92189 112.86156  89.82189
#> 70  176.30930 156.44255 197.58472 142.10362 132.31246 141.39373 113.30064
#> 71  158.38171 112.85564 148.66655 153.77026 135.94444 133.60012 116.24162
#> 72  160.54276 139.58945 175.58573 117.80426 145.77976 163.74132 171.85551
#> 73  123.28179  96.07166  89.91679 145.01413 186.39224 130.16952 168.07248
#> 74  167.07373 163.18243 224.73829 147.45422 149.64547 152.95302 129.56477
#> 75  194.19648 132.04129 141.90979 161.09237 181.98817 150.08212 150.01730
#> 76  190.21177 132.42438 116.40496 124.47553 141.47566 198.46801 121.32440
#> 77  215.00799 168.78090 134.14673 165.40950 205.58551 100.50134 112.22440
#> 78  134.05062 125.03990 180.31109 175.47740 118.16976 142.34197 146.05229
#> 79  151.77632 155.02106 146.89474 132.04158 142.57613 138.29899 206.83110
#> 80  134.08932 130.62177 153.64198 140.77893  64.58297 112.18169 117.17051
#> 81  134.16285 105.29957 137.91576 150.26116 148.12091 172.99083 143.25704
#> 82  178.06977 160.58322 136.78593 172.91803 117.17156 178.85509 101.61013
#> 83  145.33480 157.46294 114.93449 164.50600 201.76520  90.46410 185.17587
#> 84  127.08474 152.12156 112.09271 136.49763 138.79554 174.62831 112.65398
#> 85  135.64017 177.21875 183.07642 149.25177 134.63970 132.59553 114.06191
#> 86  132.49825 104.02970 166.04058 190.13317 113.38300 119.47826 162.93851
#> 87  141.80711 150.41532 113.68131 130.93992 155.70838 199.12903 123.01875
#> 88  171.43045 120.55755 130.49839 152.83181 149.89222 106.82919 105.62467
#> 89  195.38762 158.01502 165.45474 174.95096 155.42767 181.36998 147.06907
#> 90  128.44995 128.29747 162.10686 220.19908 161.45045 159.31074 106.08260
#> 91  124.69830 115.06376 129.29854 133.73991 183.23252 115.84959 134.30923
#> 92  161.05923 153.46265  99.99277 120.60034 171.91543 132.50854 129.73966
#> 93  142.53245 102.92277 188.75631 183.54839 148.78609 155.12960 112.48188
#> 94  166.98683 191.46529 139.96126 143.13179 119.65181  89.10960  87.48594
#> 95  151.10124 151.65121 185.43980 103.91030 163.56147  93.50304 191.66633
#> 96   96.51665 180.09728 114.52228 154.00658 124.00880 167.72735 107.02569
#> 97  174.88615 151.72667 175.74524 103.72826 233.09037 170.69185 144.26630
#> 98  139.69038 106.51407 158.30474 118.09251 143.95946 158.71163 118.95032
#> 99  157.14428  82.95407  88.19351 148.36121 155.20564 127.80178 129.83267
#> 100 169.36427 171.41907  91.59125  94.25254 124.78246 109.81764 156.45447
#>         B.4_1
#> 1    98.98744
#> 2   101.23227
#> 3   237.54195
#> 4    75.30628
#> 5   108.88347
#> 6   162.16317
#> 7   149.05438
#> 8   134.94933
#> 9   146.21247
#> 10  134.43152
#> 11  116.53802
#> 12  129.75747
#> 13  204.19629
#> 14  230.52304
#> 15  120.74352
#> 16  108.96869
#> 17  166.24380
#> 18  142.20429
#> 19  145.58943
#> 20  182.71903
#> 21  148.78914
#> 22  185.07412
#> 23  130.76290
#> 24  114.42699
#> 25  115.16407
#> 26  151.72461
#> 27  146.12386
#> 28  135.60555
#> 29  144.67695
#> 30  165.44949
#> 31  122.29827
#> 32  159.66915
#> 33  160.14704
#> 34  134.09620
#> 35  146.86951
#> 36  125.65811
#> 37  128.01092
#> 38  149.59430
#> 39  129.61600
#> 40  145.93102
#> 41  145.68089
#> 42  107.13582
#> 43   93.85701
#> 44  178.06602
#> 45  159.45763
#> 46  142.64155
#> 47  141.97620
#> 48  182.81508
#> 49  105.29078
#> 50  129.05290
#> 51  132.16240
#> 52  162.81891
#> 53  143.83665
#> 54   90.58205
#> 55  161.28160
#> 56  176.97320
#> 57   91.62434
#> 58  154.60254
#> 59  150.49804
#> 60  179.74148
#> 61  196.69258
#> 62  178.45847
#> 63  178.77577
#> 64  168.96068
#> 65  135.78617
#> 66  142.88742
#> 67  104.53959
#> 68  145.93723
#> 69  191.50498
#> 70  124.24857
#> 71  176.24682
#> 72  177.28851
#> 73  190.71632
#> 74  127.94143
#> 75  142.22440
#> 76  150.67005
#> 77  155.46195
#> 78  155.17789
#> 79  184.15626
#> 80  169.86056
#> 81  194.44213
#> 82  161.66920
#> 83  183.10897
#> 84  144.98464
#> 85  172.70974
#> 86  220.59767
#> 87  141.68314
#> 88  139.15237
#> 89  121.70948
#> 90  141.08091
#> 91  133.31082
#> 92  131.99725
#> 93  199.18722
#> 94  129.23201
#> 95  109.42148
#> 96  168.68996
#> 97  133.26134
#> 98  194.40970
#> 99  106.74975
#> 100 142.31269