Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10, 10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(
  subset = rep(c("A", "B"), 50),
  level = rep(1:4, each = 25)
)
rownames(meta) <- 1:100
scRNA <- Seurat::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA,
  by = c("subset", "level"),
  min.cells = 10, max.cells = 20
)
#>         A.1_1     B.1_1     A.2_1     B.2_1     A.3_1     B.3_1     A.4_1
#> 1    87.18084 153.80011 168.98585 177.47163 132.54996 118.35708 175.55189
#> 2   195.98750 165.62785 175.03746 124.40591  97.01701 120.55698 148.76347
#> 3   143.92683 149.96070 142.81543 141.48331 153.12543 130.56916 171.02057
#> 4   148.47655 205.88249 176.57880 142.15957  93.91945 141.63462 179.81629
#> 5   180.65219 114.54029 102.87510 193.60673 194.81121 103.54356 116.85854
#> 6   171.41036 135.23481 123.86744 166.97874 129.43964 118.87017  84.26084
#> 7   153.02710 156.14688 123.50244 143.92408 168.04473 116.11392 144.95812
#> 8   150.46565 139.41363 147.05528 157.23993 147.46795 137.18085 156.04137
#> 9   164.72309 140.80908 129.55877 122.68786 201.03705 124.39603 172.56438
#> 10  226.45996 134.28978 157.76900  96.79798 143.62291 125.35008  81.12193
#> 11  176.06704 119.15809 160.53620 161.09574 140.42550 142.17848  96.53239
#> 12  100.95746 162.16882  97.58199 201.35637 136.07861 142.95924 134.66584
#> 13   68.29792 147.88186 126.98787 178.34505 219.49049 119.00135 194.35834
#> 14  183.74936 123.56027 156.45117 120.71074 160.65608 127.11920 149.86136
#> 15  187.12329 119.09766 126.82143 228.31310 169.38973 141.31757 139.44794
#> 16  143.01002 168.34976 172.63282 154.44406 191.07647 159.12450 168.64032
#> 17  132.88782 155.30657 156.55162 146.48955 136.62033 157.79550 175.35083
#> 18  102.05373 161.79149 202.15757 177.38956 114.13726 137.92382 104.26081
#> 19  174.83699 140.65545  88.10175 149.32101 133.25262  95.59666 169.39984
#> 20  133.18788 157.73243 112.05441 138.86953 140.78837  93.71551 118.28600
#> 21  109.97297  96.22882 127.97040 180.81460 135.22122  75.02852 135.85074
#> 22  148.36813 122.74428 116.35600 197.22059 158.65358 148.74714 147.10688
#> 23  126.54264 157.92656 216.36803 117.35155 217.36839 183.33612 109.43622
#> 24  184.69471 131.42359 102.57871 131.49254 136.67068 108.81416 137.52146
#> 25  191.20128 172.20443 147.93921 161.52984 219.18881 154.58629 148.79064
#> 26  114.52156 172.94745 147.02506 175.25181 140.90250 152.81203  74.28084
#> 27  122.62401 227.79313 154.34794 150.88957 161.25359 172.07627 120.55162
#> 28  186.73881 143.20964 102.88751  97.18572 122.56089 123.24158 137.29225
#> 29  151.13010  82.99041 158.08776 142.30808 112.72489 157.39147 109.23201
#> 30   87.94681 117.15834 121.44489 167.49853 111.07802 106.13000 128.97285
#> 31  180.72441 142.45104 145.59070 177.53797 180.79437 117.06844 100.10059
#> 32  158.93024 103.56234 137.81770 147.31225 132.95433 156.40557 133.14691
#> 33  185.70440 101.03387 166.04815 169.44740 140.97928 140.64931 129.56389
#> 34  133.71613 107.53877 141.47123 191.85764 160.60031 159.32175 180.55046
#> 35  106.53226 133.52914 163.40978 123.56275 179.07644 167.86335  88.06956
#> 36  138.95164 126.74601 148.74272 206.01908 137.68124 150.82682 180.21797
#> 37  124.14694 137.09336 183.67287  92.59590 158.50220 100.99562 159.90803
#> 38  137.27927 114.49721 190.88397 148.82267  82.92165 139.45781 127.60800
#> 39  125.17224 134.20865 151.64534 219.79505 148.40901 163.67092 111.31878
#> 40  159.35956 160.26129 144.03438 153.10794 108.37786 154.68973 152.45486
#> 41  123.73618 132.97753 139.50421 199.16217 164.15374  96.95759 136.18655
#> 42  167.44219 171.15732 174.57854 157.47634 144.90147 175.34788 172.51787
#> 43  123.85869 113.22813  84.84151 134.33673 123.38788 118.09779  92.02316
#> 44  181.11558 107.86817 178.17941 137.16939 147.94908 144.08141 154.80045
#> 45  130.65262 159.38418 165.04524 155.29658 154.59940 204.58573 125.60639
#> 46  168.58739 177.96707 130.13287 181.24023 142.13057 142.61471 157.86094
#> 47  124.62802 118.41888 103.16316 144.47999 181.58897 150.62789 134.47853
#> 48  184.00922 121.18060 147.15919 171.61270 151.55642 108.03559 157.06208
#> 49  196.74642 163.28689 115.11537 129.90476 197.09929  93.39290 135.43756
#> 50  151.42386 141.10599 134.94184 186.04989 116.22592 149.62614 130.57039
#> 51  101.47298 139.74164 140.39021 155.78045 129.93577 122.93957 132.88810
#> 52  111.04975 128.08027 158.29077 176.90383 203.18105 173.32013 102.91191
#> 53  160.71006 118.32323 175.47037 119.50904 131.41063 125.80986 118.28576
#> 54  124.52703 121.56056 108.53479 140.56056 119.75339 148.27444 132.61749
#> 55  114.10603 109.97454 154.29809 160.81747 123.58958 130.69787 143.91456
#> 56  157.81852 110.91422 130.05652 196.12154 155.29656 130.04238 117.98430
#> 57  130.70686 133.82359 157.13913 128.75563 141.32740 120.84459 163.23819
#> 58   91.90422 206.36813 123.47140 165.81141 113.72048 128.13172 131.01885
#> 59  157.38000 160.74279 132.91200 136.73387 129.66401 155.44190 168.86318
#> 60  141.73404 118.63712 130.31476 157.11589 169.49139 143.72879 170.00205
#> 61  131.31977 154.40720 170.26747 132.65100 204.31029 118.66596 107.34685
#> 62  158.85830 132.11675 112.33446 111.77150 148.97745 154.68022 110.53539
#> 63  185.54884 137.40397 165.60418 176.98476 149.22157 141.39924 178.59106
#> 64  167.45014 143.10791 118.62912 104.38570 138.37753 129.19639 121.08745
#> 65   98.90198 179.51772 140.15119 166.87887 182.72821 120.56584 113.07736
#> 66  192.77904 120.73838 132.07288  83.63395 153.40564 149.61102 153.56519
#> 67  169.61557 158.52124  81.75070 150.29654 178.53689 136.34299 129.40891
#> 68  145.29269 171.11820  90.52265 137.59555 115.80674 184.06604 150.50590
#> 69   95.36488 106.30376 148.52935 151.38458 120.29199 135.67946 150.62823
#> 70  190.24032 131.46758 169.02643 151.60201 185.19018 148.60322 111.29955
#> 71  155.41444 114.99268 138.66168 180.74032 159.88469 144.62405 146.71991
#> 72  159.29970 132.96161 159.18253 145.83784 150.90610 128.05710 164.41681
#> 73  136.97376 121.79205 140.97671 106.69202 225.38179 121.40431 145.18719
#> 74  115.69938 188.13295 128.78606 121.31307 136.12669 147.98430 106.82122
#> 75  114.16746 115.09955 110.24826 145.62497 168.92036 126.30884 176.59415
#> 76  123.82991 131.21613 125.62229 128.49144 130.97779 131.05831 156.53403
#> 77  164.65882 103.53623  98.73085 153.91542 183.42401 153.92618 124.77191
#> 78  168.10362 106.54908 131.66597 187.91301 130.41537 134.78812 184.79600
#> 79  151.83520 209.42981 157.02037 180.35214 157.88742 189.63886 134.86598
#> 80  153.10302 150.19023 112.38023 148.29711 182.46280 136.07637 113.87453
#> 81  111.92028 162.71577 154.28472 141.26001 214.63513 128.97046 118.20971
#> 82  143.63091  98.89672 109.75396 120.02117 113.87753 119.69738 145.39414
#> 83  144.12598 188.45535 126.37477 135.68582 161.60492 102.66087 120.29897
#> 84  152.71255 105.47269  65.72801 150.61662 173.55273 145.72388 145.53711
#> 85  103.24380 120.66180 174.29316 147.31275 147.14142 162.22333 133.02733
#> 86  158.51450 188.29297  93.48072 179.23827 198.81276 133.94740 138.11217
#> 87  117.29172 195.77355 167.82728 155.87921 119.98458 164.41357 189.98930
#> 88  170.30108 173.12568 128.35338 152.66484 137.25317 215.28529 160.52180
#> 89  129.42615 162.14027 163.00964 164.04012  95.20194 118.97086 178.04071
#> 90  188.53256 110.02709 123.37971 140.78320 119.16236 128.96690 186.92585
#> 91  144.33596 142.39064 173.02542 107.70272 140.12900 138.95467 138.93953
#> 92  183.00118 124.74807 129.43926 166.22708 235.72407  88.09197 141.24951
#> 93  145.28793 140.02647 119.16155 121.14524 157.45311 124.60159 128.18999
#> 94  151.63138 139.26387 146.56756 147.40570 180.78427 164.21615 173.85290
#> 95  177.58763 115.64134 132.14782 179.58716 204.88763 132.05118 126.35855
#> 96  192.73924 130.81959 118.50826 137.98646 161.75882  96.74138 156.50636
#> 97  177.38388 127.06090 104.96733 136.06053 154.02454 171.47941 141.96192
#> 98  186.93324  76.44315 174.04032 127.54379 138.79597 104.33344 134.00835
#> 99  172.85754 100.86531 135.49005 169.38569 116.09657 126.80337 110.69409
#> 100 160.70370 169.02355 115.44276 126.26384 107.81672 153.67453 174.84193
#>         B.4_1
#> 1   140.96396
#> 2   150.87923
#> 3   144.62262
#> 4   131.64819
#> 5   142.60804
#> 6   182.01909
#> 7   172.87382
#> 8   219.33419
#> 9   174.21321
#> 10  153.05127
#> 11  160.45950
#> 12  160.75418
#> 13  160.33143
#> 14  190.41370
#> 15   95.91773
#> 16  150.81542
#> 17  148.37538
#> 18  117.10667
#> 19  139.89644
#> 20  127.60291
#> 21  117.56938
#> 22  149.89944
#> 23  198.94994
#> 24  148.78321
#> 25  162.38844
#> 26  140.39735
#> 27  130.32931
#> 28  126.65851
#> 29  150.63018
#> 30  135.35997
#> 31  122.34649
#> 32  163.05680
#> 33  181.96120
#> 34  145.34669
#> 35  129.42539
#> 36  154.08469
#> 37   87.41537
#> 38  118.87811
#> 39  143.06690
#> 40  141.51673
#> 41  141.24016
#> 42  152.32253
#> 43  166.09892
#> 44  155.56694
#> 45  144.58645
#> 46  125.61966
#> 47  129.93435
#> 48  172.84937
#> 49  145.89646
#> 50  174.14008
#> 51  185.67282
#> 52  162.14719
#> 53  181.45847
#> 54  179.90340
#> 55  141.12589
#> 56  146.92022
#> 57  125.71514
#> 58  129.31880
#> 59  153.82865
#> 60  195.28877
#> 61  148.36639
#> 62  139.65274
#> 63  162.91708
#> 64  176.36341
#> 65  119.40132
#> 66  149.11243
#> 67   90.07199
#> 68  171.29596
#> 69  114.00478
#> 70  129.14963
#> 71  145.84825
#> 72  159.16052
#> 73  126.50746
#> 74  165.34471
#> 75  182.48737
#> 76  141.04992
#> 77  147.31412
#> 78  151.03782
#> 79  156.96140
#> 80  140.71233
#> 81  173.09420
#> 82  127.68619
#> 83  147.26960
#> 84  138.49711
#> 85  178.25366
#> 86  123.74905
#> 87  159.00425
#> 88  142.88547
#> 89  117.85846
#> 90  165.13516
#> 91  128.76546
#> 92  164.74817
#> 93  123.27079
#> 94  140.86997
#> 95  132.80650
#> 96  121.31377
#> 97  119.95538
#> 98  122.43885
#> 99  157.81197
#> 100 125.86126