Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10,10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(subset = rep(c("A","B"),50),
                   level = rep(1:4, each = 25))
rownames(meta) <- 1:100
scRNA <- Seurat::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA, by = c("subset","level"),
               min.cells = 10, max.cells = 20)
#>         A.1_1     B.1_1     A.2_1     B.2_1     A.3_1     B.3_1     A.4_1
#> 1   150.97921 153.32405 155.36909 183.02717 125.58866 142.27868  98.68335
#> 2   146.81476 145.23751 155.84777 210.16858 153.78143 109.39171 163.08886
#> 3   163.92445 150.24527 191.41357 171.81881 180.27761 181.24472 108.69739
#> 4   106.49132 122.04034 161.54231 197.71263 139.57639 119.13842 162.00843
#> 5   139.37241 125.63345 100.74988 198.55351 197.13176  80.02105 119.00180
#> 6   198.02878 185.92335 125.48409 124.01159 113.09827 192.80305 140.64028
#> 7   151.82188 145.05614 140.80661 138.53989 171.80465 113.18903 133.99752
#> 8   155.83786 157.61857 116.56124 131.24094  92.90015 126.06101 181.22878
#> 9   139.85678 114.82727 144.19925  92.57022 164.71344 151.55677  89.14116
#> 10  168.58847 167.25029 121.09789 114.07176 147.51893 156.34955 148.67702
#> 11  184.52063 160.34410 130.56294 135.96107 100.11634 119.97935 155.70944
#> 12  159.83014  94.67355 134.28124 115.27667 179.86324 102.03690 127.07767
#> 13  167.19738 154.76127 135.09402 132.39641 122.65407 119.30580  98.46690
#> 14  179.38478 149.10977 120.84761 141.13917 144.22122 172.01550 166.78745
#> 15  130.44453 145.87923 128.47080 176.05901 130.12864 106.27558 116.15578
#> 16  109.01622 149.87945 167.38199 130.84146 128.29900 146.80606 184.42384
#> 17  140.26924 131.55124  67.80755 145.11108 119.01032 131.71099 134.83608
#> 18  199.87188 146.85984 168.87814 168.38232 157.16914 145.75681 153.46575
#> 19  177.69250 135.21225 152.58216 135.16817 114.50454 111.89907 123.38596
#> 20  190.73820 133.41652 134.12292 184.20512 195.92261 155.38400 179.11755
#> 21  135.94918 122.46149 125.33947 129.76727 146.58440 168.83923 132.34722
#> 22  169.83527  78.21503 104.41445 196.97057 167.78149 154.14349 101.44437
#> 23  161.09878 111.87194 138.23069 130.49476 130.64188 144.20390 154.18692
#> 24  148.25652 170.77342 124.84758 166.65578 127.17208 137.39490 152.11113
#> 25  129.19965 155.34101 160.89668 123.35686 163.25901  65.61724 134.49619
#> 26  172.22929 111.47792 110.01631 185.03489 188.19426 145.16112 119.85482
#> 27  131.36923 138.08362 157.20251 163.82576 119.81147 173.83758 152.56836
#> 28  155.29290 143.76646 116.88867 134.48508 110.10213 119.48143 120.31591
#> 29  120.77567 161.04950 136.02868 129.23128 154.77148 149.81324 144.46849
#> 30  168.37781 104.31758 129.67552 186.86084 105.79273 127.98384 183.49648
#> 31  170.24107 191.51004 150.12157 139.55032 146.83385  94.59832 145.82271
#> 32  122.44675 120.57996 107.77525 103.45916 129.14917  86.49771 202.98303
#> 33  176.11463 136.14760 244.25974 146.45187 123.72144 127.57881 149.99832
#> 34  172.21629  77.39951 168.82102 147.62276 139.48080 157.05679 178.44519
#> 35  181.80373 146.08983 166.11920 141.25883 150.94722 192.70718 135.21864
#> 36  102.09721 185.08882 136.60025 147.40457 155.54595 132.67506 125.71887
#> 37  122.87181  84.28975 187.74440 143.89959 143.11084 115.65707 162.21830
#> 38  161.02950 134.63808 150.17925 148.96301 113.89048 116.46827 163.16966
#> 39  132.99733 120.63183 214.86303 184.38984  79.00980 179.58666 130.12728
#> 40  125.04725 155.39719 104.53198 151.07973 110.48383 154.50731 156.23069
#> 41  162.93855 176.64614  82.54334 160.56650 108.28033 106.66360 173.56589
#> 42  131.38866  99.07766 132.17384 215.98214 147.80603 149.17135 160.97023
#> 43  125.59602 149.61701 169.13849 155.51368 175.36276 167.38000 155.33736
#> 44  162.89753 143.96687  93.22791 162.62643 150.05987 107.85394 189.46682
#> 45  183.51016 151.10060 189.87868 116.81373 140.98850  86.16717 115.58313
#> 46  118.90098  77.82623 114.08385 114.28156 155.93645 117.18866 105.16174
#> 47  241.71335 134.96442 141.84654 193.69038 136.24651 142.67521 113.16582
#> 48  204.08283 129.21363 161.52865 153.46064 150.07469 143.17300 144.34520
#> 49  136.06353 130.17608 120.35013 133.25847  91.74950 102.87323 138.17782
#> 50  191.79538 119.91758 122.51984 131.10406 125.30488 140.74460 169.18786
#> 51  193.64921 166.11917 179.47492 177.49714 150.74207  82.81842 103.95312
#> 52  141.28617 141.25872 112.14047 112.07770 163.71426 201.75062 137.07980
#> 53  131.18107 128.92606 119.40909 136.98918 177.17760 150.14514 128.69668
#> 54  164.08324 157.18308 139.72121 170.28428 180.27695 121.42946 146.51654
#> 55   78.34457 170.75921 125.89117 121.57213 192.24890 118.89591 153.21726
#> 56  141.51150 131.70664 158.42503  72.11105 122.42803 140.17103 193.00678
#> 57  170.72030 123.59231 125.38771 103.88505 155.55313 128.73735 152.02980
#> 58  134.26802 161.76277 180.24554 174.42468  97.45215 192.48103 172.72534
#> 59  132.27456 109.84143 117.68311 135.88620 161.01334 154.14899 126.77000
#> 60  134.83144 119.20584 128.17450 145.36772 124.26360 130.06073 137.70866
#> 61  194.21692 200.37573 148.48507 145.34364 178.73471 150.04261 116.82531
#> 62  127.49468 160.20858 162.99465 138.21215 190.98250  95.48097 107.97606
#> 63  122.09538 124.36703 169.84663 133.45864 178.43675 149.34272 147.28314
#> 64  186.70312 193.88481 147.13363 156.63528 135.94930 140.47347 143.87383
#> 65  156.31520 133.39883 141.67873 132.30907 106.83553 153.00823 112.89284
#> 66  174.57189 138.52619 139.32947 183.01426 137.77004 151.89805 158.36174
#> 67  236.91255 129.87648 152.93686 153.61956 240.02732 113.59611 187.28665
#> 68  148.47225 102.16086 169.45116 161.24621 125.68562 161.12103 111.04848
#> 69  129.15689 112.11384 144.60713 155.42667 125.59325 136.11250 124.51413
#> 70  186.41193 120.16430 147.27226 151.02882 176.54050 127.05465 150.14766
#> 71  192.94992 135.62614 139.05073  92.19307 109.08826 200.43118 156.80740
#> 72  158.67605 138.90309 144.91685 153.56368 133.20813 146.82750 159.00653
#> 73  145.38413 130.98329 123.94577 102.58359 170.41765 190.17124 121.62306
#> 74  144.16144 129.25184 209.39522 105.10647 138.60107 124.31792 144.37540
#> 75  158.74252 120.87604  88.59808 175.83191 122.63801 166.45047 148.93867
#> 76  151.67385 190.75016 120.35973 127.72582 111.54151 140.11775 124.85633
#> 77  103.27931 122.55300 185.40713 185.76363 189.26763  99.86409 144.64993
#> 78  128.70044 120.58590 149.89388 135.63684 108.17823 160.18616 182.82670
#> 79  121.57493 155.61206 144.79010 182.94665 131.60924 155.47938 140.56491
#> 80  145.05937 144.28475  96.04968  89.20592 124.18509 194.25686 155.38263
#> 81  156.46521 185.10746 126.57999 156.36583 168.09059 140.41584 156.82041
#> 82  140.81357 127.88358 147.49300 274.64409 137.57321 154.26522 114.76467
#> 83  103.99492 123.22175  95.36266 157.24134 159.59523 153.01198 168.32392
#> 84  205.32603 164.68827 201.00742 150.51199 165.28544  93.68108 127.02671
#> 85  204.40571 155.79508 158.45668 171.06580 186.81364 158.89498 109.44330
#> 86  163.79242 152.57610 121.86548 185.17501 135.97268 143.90387 116.12979
#> 87  143.43187 211.99577 124.17260 130.38534 153.43165 155.04862 143.70509
#> 88  179.34895 107.32263 110.45238 226.51772 142.93133 160.82470 157.31534
#> 89  116.68979 100.06510 158.37070 175.56582  92.64996 103.66826 199.97450
#> 90  138.01762 155.50459 106.35697 148.66205  93.67452  93.88391 125.91249
#> 91  125.65081 129.87272 176.57780 177.98415 174.34883 132.94910 124.14930
#> 92  220.00461 121.66294 123.66719 185.65036 128.49749 108.40979 118.91684
#> 93  108.59782 141.72897 145.45000 179.94311 183.04598 151.24911 155.35939
#> 94  134.16265 139.17959 136.15915 204.25575 164.56123 172.08897 127.83958
#> 95  133.10282 140.22259 155.89216 129.90568 174.41149 183.49164 126.43974
#> 96  201.00260 133.83633 118.05022 140.75158 175.54614 148.54022 123.14350
#> 97  156.22058 161.86616 147.97702 169.70624 155.66478 123.25061 140.06113
#> 98  142.29444 123.51179 143.97684 178.23928 141.32057 129.37670 119.35738
#> 99  167.56144 155.02551 129.02934 166.96994 130.21239 123.22397 128.18873
#> 100 132.65945 170.60599 127.93877 158.68814 120.47905 163.30622 120.60853
#>         B.4_1
#> 1   203.79006
#> 2   179.67060
#> 3   188.22440
#> 4   153.59418
#> 5   141.73309
#> 6   147.76191
#> 7    87.05437
#> 8   120.44560
#> 9   169.12388
#> 10  172.26326
#> 11  168.88526
#> 12  218.55343
#> 13  181.44413
#> 14  199.87322
#> 15  170.45183
#> 16  188.24780
#> 17  115.34801
#> 18  175.51927
#> 19  209.30601
#> 20  113.36082
#> 21  221.61736
#> 22  133.63367
#> 23  118.51183
#> 24  129.95748
#> 25  135.89598
#> 26  166.81986
#> 27  141.89964
#> 28  174.98402
#> 29  120.74859
#> 30  134.86719
#> 31  130.52254
#> 32  149.42544
#> 33  147.98511
#> 34  196.29118
#> 35  150.09614
#> 36  157.06725
#> 37  157.70956
#> 38  126.82313
#> 39  194.62687
#> 40  112.86067
#> 41  119.31364
#> 42  139.47836
#> 43   82.09109
#> 44  114.67305
#> 45  192.95565
#> 46  132.35177
#> 47  178.73189
#> 48  144.59146
#> 49  163.95444
#> 50  180.95801
#> 51  117.70566
#> 52  124.24011
#> 53  154.59136
#> 54  171.99911
#> 55  167.18140
#> 56  173.41582
#> 57  158.60256
#> 58  101.47763
#> 59  214.71667
#> 60  121.78712
#> 61  127.28228
#> 62  198.56507
#> 63  156.21646
#> 64  114.26481
#> 65  209.37238
#> 66  130.79152
#> 67  154.62605
#> 68  168.40456
#> 69  167.59121
#> 70  177.90960
#> 71  150.68192
#> 72  226.65435
#> 73  160.64616
#> 74  128.43245
#> 75  153.42026
#> 76  151.43810
#> 77  133.47192
#> 78  169.19993
#> 79  109.42164
#> 80  150.96016
#> 81  166.29050
#> 82  153.03539
#> 83  168.71839
#> 84  170.12802
#> 85  115.70277
#> 86  152.42121
#> 87  130.78170
#> 88  113.61436
#> 89  160.32929
#> 90  187.74651
#> 91  132.01930
#> 92  172.29686
#> 93  146.46085
#> 94  124.04324
#> 95  125.61792
#> 96  146.43168
#> 97  106.64675
#> 98  164.80636
#> 99  192.75157
#> 100 187.25160