Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10, 10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(
  subset = rep(c("A", "B"), 50),
  level = rep(1:4, each = 25)
)
rownames(meta) <- 1:100
scRNA <- Seurat::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA,
  by = c("subset", "level"),
  min.cells = 10, max.cells = 20
)
#>         A.1_1     B.1_1     A.2_1     B.2_1     A.3_1     B.3_1     A.4_1
#> 1    99.09972 136.09700 173.27152 123.80209 166.50155  98.29692 116.84719
#> 2   154.76861  91.73171 127.66743 139.36974 133.28674 172.90262 107.92130
#> 3   131.86373 152.45927 163.15747 147.32141 159.46444 204.09964 151.74487
#> 4   141.72394 199.09573 158.82265 169.52683 127.28015 157.73104 119.93046
#> 5   143.63333 165.33764  93.85933 132.56985 167.56417 140.38169 121.72479
#> 6   132.09645 142.46097 147.17236 176.96559 158.35256 121.32619 144.80917
#> 7   160.65414 111.50344 137.15417 165.84561 198.92688 129.93444 130.68700
#> 8   171.91490 173.31772 146.85541 222.12209 171.17936 169.41966 103.22942
#> 9   115.15114 186.90666 173.65214 168.77415 197.69672 119.39245 146.84859
#> 10  201.30546 127.96953 130.19565 143.97943 189.14576 125.77214 145.71800
#> 11  178.26608 138.06893 134.40719 140.28534 178.92922 117.95085 131.19368
#> 12  116.02427 170.83439 128.58185 117.89942 175.33569 160.58298 100.57535
#> 13  121.94240 194.91071 158.90857 148.56808 134.04065 135.25381 163.20331
#> 14  102.55410 152.32987 213.73535 144.62592  86.30611 148.87449 130.85490
#> 15  159.86331 121.26898 120.27264 165.84775 125.55722 129.76703 140.19007
#> 16  120.51429 132.91764 165.71963 141.40187 123.47089 134.85583 180.42450
#> 17  109.77245 150.37287 143.49970 131.92755 164.49572 151.73047 157.11667
#> 18  212.04559 195.01802 149.97900 178.20036 155.97578 120.39155 155.09324
#> 19  132.49157  78.81199 106.61264 160.69591 171.64462 123.08100 140.11036
#> 20  183.00665 140.84823 180.65296 126.14739 179.68283 157.75956 148.04535
#> 21  147.27975 103.38045 108.34280 147.41438 145.61899  86.37771 153.49002
#> 22  154.84587 143.70713 113.73464 127.80458 141.20317 136.26851 117.93750
#> 23  172.13793 109.12365  93.90063 135.60163 147.12722 103.44497 151.29113
#> 24  142.05347 202.55323 129.92037 104.62036 168.80743 165.16861 165.10630
#> 25  120.06297 142.41800 156.96341 189.21675 162.33445  96.97028 142.25868
#> 26  136.55221 153.00542 154.59512 167.01108 108.26413 182.46975 125.04695
#> 27  127.48965 107.13052 137.51923 146.63839 126.33621  68.07551  79.92200
#> 28  140.54437 127.75061 248.34475 195.22997 178.22190 139.18828 147.39963
#> 29  141.05612 164.85780 162.25251 133.08018 156.93500 132.85621  89.73503
#> 30  141.94012 145.52819 138.04776 182.94500 185.35391 153.08381  86.73233
#> 31  151.33516 161.67072 213.17577 139.84028 142.49264 169.05789 127.07801
#> 32  133.54573 109.84131 133.51951 139.31486 146.30505 122.76235 151.12271
#> 33  191.28375 174.15485 128.65021 131.71576 110.02482 136.46302 127.56584
#> 34  118.15917 164.41505 119.62768 158.17996 181.15976 147.79943 180.82043
#> 35  160.82884 153.07897  83.92370 169.71792 181.16074 135.72770 136.33652
#> 36  132.28424 133.38786 167.70901 174.85172 146.31056  79.23078 132.34379
#> 37  135.46603 108.14508 123.52799 197.18804 162.31439 163.68251 194.75548
#> 38  184.76950 126.26208 134.39391 149.12126 186.53564 139.50642 169.25212
#> 39  167.55579  92.59446 146.24125 200.91519 136.58050 145.14408 137.58160
#> 40  215.01107 111.52361 131.55039 136.80793 152.87177 166.43790 108.73936
#> 41  110.59312 151.89491 174.30335 157.53694 143.80260 111.41955 125.87978
#> 42  181.60617 131.14292 123.66215 173.15481 121.52863 122.42942 150.42680
#> 43  149.93078 127.73493 115.24062 101.68568 167.77682 124.93053 194.14688
#> 44  124.70230 141.92772 152.32240 135.84228 173.89613 147.70821 100.36696
#> 45  134.03148 158.75129 127.22220 164.78142 127.32691 123.99856 122.58284
#> 46  197.60362 143.60062 174.53800 184.17561 186.34951 165.67240 140.69987
#> 47  168.25887 156.96169 101.95234 114.60297 120.87320 146.77130 118.61415
#> 48  118.16210 101.75427 131.71472 203.80909 139.07653 167.89373 128.02312
#> 49  123.76271 137.61001 126.75473 203.19742 180.79268 171.23010  95.13733
#> 50  300.72525 141.32514 139.03248  83.93765 195.51318 123.19377 110.19920
#> 51  181.32578 112.19977 132.14854 218.84919 203.78544 149.08503 133.22403
#> 52  185.75310 161.35284 170.34632 110.34764 176.15509 161.94065 190.65682
#> 53  146.93495 152.19909 154.11370 168.10362 122.31892 147.47252 187.78259
#> 54  161.89315 119.51080 147.29468 163.35041 156.82360 149.42685 147.09391
#> 55  154.32084 218.63060 117.19720 115.19550 173.35555 144.05412 132.03401
#> 56  144.17733 164.35338 139.45857 128.24582 140.09293 129.95069 125.88186
#> 57  176.95966 197.37779 139.18306 129.93796 123.81997 117.34819 151.19133
#> 58  185.15574 173.62611 129.00738 153.51821 153.67371 132.41615 171.92249
#> 59  192.81451 171.30127 129.27430 169.28422 125.13693 135.13995 135.73852
#> 60  104.90286 137.79452 142.93601 191.97689 138.97477 104.08418 116.78358
#> 61  149.99656 132.50786 120.24912 198.57148 160.32359 155.04953 167.94719
#> 62  111.17176 119.25726 128.79607 134.23216 158.25932 129.67973 121.25054
#> 63  109.98779 194.37424 121.39241 161.42869 160.49094 125.64251 171.17261
#> 64  140.52207 130.36163 136.77812 164.09027 172.32950 117.46433 154.93853
#> 65  133.29816 154.53676 134.54907 254.97648 166.04797 123.71857 171.51597
#> 66  151.14956 181.45778 208.18589 185.31242 184.63262 160.39247 203.98273
#> 67  146.06243 155.76942 148.63458 166.03838 193.80308 128.04490 109.04942
#> 68  132.06260 120.00006 132.95156 154.89472 139.33084 163.53392 117.36701
#> 69  148.53200 122.23311 172.54300 154.16677 125.69190 117.83139 158.84311
#> 70  155.82356 109.78096 117.16225 152.23250 147.48123 138.45135 134.92245
#> 71  109.01060  89.36753 102.30207 228.55982 175.82156 132.32177 118.43638
#> 72  101.82124 127.08378 193.25584 118.74316 118.41085 177.49756 139.95313
#> 73  188.91173 147.25118 190.77280 160.63294 192.23444 133.57216  76.88537
#> 74  135.70969 133.54777 159.48085 146.09691 135.57159 129.20834 150.17025
#> 75  180.89028 137.13286 123.35244 138.08187 138.20568 168.68240 159.95106
#> 76  157.81676 147.35233 122.74992 119.69017 180.22197 101.97386 130.66085
#> 77  152.17962 123.04578 131.64828 144.49323 136.02885 172.86653 162.27205
#> 78  154.74247 124.35965 133.88649 135.48165 119.53149 123.84876 161.99275
#> 79  159.96183 130.22470  75.34602 184.12984 148.69456 179.52875 127.79471
#> 80  189.65521 145.01298 133.20981 144.09881 123.83234 133.24381  77.66639
#> 81  162.89375 157.89206 184.36107 137.34034 136.80514 137.38035 124.58995
#> 82  131.67176 138.51704 111.79982 146.79856 157.33597  98.98207 164.45639
#> 83  183.05872 202.22508 181.66055 115.94183 143.26965 200.34793 147.94758
#> 84  153.06224 153.38564 124.90873 186.48754 186.09980 121.03630 132.39430
#> 85  112.85637 141.23421 127.73212 120.62155 158.30235 103.27867 112.98931
#> 86  141.37532 133.76316 136.11299 118.13146 166.45601 158.30117 127.46795
#> 87  105.08220 125.75352 152.50902 147.89335 107.57043 124.47721 135.66217
#> 88  154.63371 110.48857 130.90564 134.07227 179.15912 190.53426 149.51824
#> 89  205.06034 133.03377 189.79715 188.43669 108.30848 133.90315 153.81810
#> 90  153.74699 109.32028 144.94880 127.40049 163.86843 169.40508 113.05036
#> 91  110.48912  78.46379 116.41581 150.68370 140.95031 132.54361 111.61511
#> 92  154.53631 108.76700 131.23780 189.70443 185.08319 146.98506 115.64913
#> 93  183.76554 153.97328 152.78321 142.29932 147.31626 106.35782 168.51072
#> 94  117.06384 110.18320 123.33100 181.50334 135.15399 134.89019 120.77356
#> 95  155.56907 110.39236 182.88085  87.70223 179.17864 197.23355 202.39138
#> 96  171.60676 109.00005 151.27518 136.20106 159.60389  94.48214 220.96488
#> 97  136.86574 105.78188 115.16786 161.67330 101.28860 136.93075 140.52169
#> 98  148.55479 103.98811 126.23402 178.92593 147.42921 153.94731 103.49979
#> 99  164.37707 123.84850 159.44943 179.86527 165.74961 134.73705 156.98398
#> 100 106.35159 140.07349 175.66524 133.21535 154.73139 176.95189 132.58780
#>         B.4_1
#> 1   161.22643
#> 2   144.62055
#> 3   129.92245
#> 4   124.62406
#> 5   152.26421
#> 6   161.37333
#> 7   167.67650
#> 8   137.79604
#> 9   111.53370
#> 10   82.10046
#> 11   80.34325
#> 12  142.36135
#> 13  175.08001
#> 14  171.89300
#> 15  132.84262
#> 16  148.96802
#> 17  184.92208
#> 18  148.36653
#> 19  130.17885
#> 20   79.62582
#> 21  113.17431
#> 22  157.21634
#> 23  142.16733
#> 24  170.26715
#> 25  154.50337
#> 26  180.90398
#> 27  166.66706
#> 28  122.88208
#> 29  179.26073
#> 30  189.44145
#> 31  146.78567
#> 32  132.85636
#> 33  205.43869
#> 34  144.24191
#> 35  121.33383
#> 36  183.02786
#> 37  163.63263
#> 38  190.75695
#> 39  179.70708
#> 40  154.03210
#> 41  165.68200
#> 42  157.22975
#> 43  131.85719
#> 44  161.34286
#> 45  141.83117
#> 46  171.44335
#> 47  192.11782
#> 48  173.70518
#> 49  194.93230
#> 50  142.43958
#> 51  147.48870
#> 52  136.25680
#> 53  190.32506
#> 54  172.64696
#> 55  101.64266
#> 56  144.08307
#> 57  130.65603
#> 58  156.66352
#> 59  131.19713
#> 60  170.14184
#> 61  139.14506
#> 62  170.77629
#> 63  107.12796
#> 64  221.14186
#> 65  203.88506
#> 66  152.80100
#> 67  179.53223
#> 68  128.47414
#> 69  201.45898
#> 70  178.35743
#> 71  230.72462
#> 72  204.48059
#> 73  146.85854
#> 74  174.64972
#> 75  182.00439
#> 76  193.27235
#> 77  140.37898
#> 78  200.86662
#> 79  109.51296
#> 80  192.76581
#> 81  155.26675
#> 82  189.77367
#> 83  152.44538
#> 84  170.01184
#> 85  206.17514
#> 86  209.31187
#> 87  207.48460
#> 88  156.38959
#> 89  162.64743
#> 90  119.33840
#> 91  135.06016
#> 92  155.35567
#> 93  206.31224
#> 94  158.15286
#> 95  135.88428
#> 96  202.42269
#> 97  193.31131
#> 98  116.10168
#> 99  168.12367
#> 100 101.02181