Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10, 10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(
  subset = rep(c("A", "B"), 50),
  level = rep(1:4, each = 25)
)
rownames(meta) <- 1:100
scRNA <- SeuratObject::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA,
  by = c("subset", "level"),
  min.cells = 10, max.cells = 20
)
#>        A.1_1     B.1_1     A.2_1     B.2_1     A.3_1     B.3_1     A.4_1
#> 1   172.3441 129.20501  83.74130 137.72274 143.59945 180.49935 159.79656
#> 2   143.7950 144.70569 102.45553 148.49297 166.64601 136.52338  72.47620
#> 3   124.9206  88.92646 135.89142 156.37680 155.43416  97.03735  93.97661
#> 4   136.7287 171.75771 110.61749 146.62028 138.92260  89.35853 117.56948
#> 5   157.4649 122.51916 133.09478 149.29131 107.76046 141.96490 162.56707
#> 6   161.1422 199.74567 169.23725 122.04381 133.81399 103.86871  95.68036
#> 7   143.3122 166.60707 138.87994 202.40455 148.02904 114.86379  87.14183
#> 8   172.4072 139.20483 100.74695 168.05885 171.04084 177.31797 131.05969
#> 9   154.2089 129.78358  91.41899 200.46380 162.72164 220.73236 190.34938
#> 10  128.1335 180.27431 156.91152 151.19108 163.42150 132.24827 149.18474
#> 11  122.5365 170.21500 138.82570 142.98387 151.35800 135.83045 172.81386
#> 12  224.6409 171.90098 124.79820 122.83212 139.05725 124.08265 162.90003
#> 13  129.8690 163.35535 130.67554 156.06889 150.94659 155.09471 139.56349
#> 14  146.8067 179.76271 123.03852 143.83989 151.76155 171.04168 162.55728
#> 15  125.7444 134.57295 143.38868 100.38384 147.81038 106.98287 101.82765
#> 16  138.5906 181.82512 186.04672 173.72952 144.99305  94.21547 155.48655
#> 17  152.1237 121.09883 126.79030 193.91372 134.51365 116.62108 115.17150
#> 18  142.7016 189.79629 158.27542 159.86634 163.42250 184.27378 116.47058
#> 19  120.7000 108.29242 142.93956 133.15696 137.77611 189.92870 188.77334
#> 20  146.5586 159.47040 117.47843 150.31782 156.75733 130.45185 190.06465
#> 21  150.1517 157.27981 120.02246 167.25642  98.42553 165.12646 147.40308
#> 22  161.1689 176.26662 101.28122 149.50123 133.19137 141.82851 152.26143
#> 23  161.8722 150.31769 162.62094 134.20079 151.49595 139.53038 109.67031
#> 24  119.4453 127.79841 109.35537 145.31936 198.07323 208.59923  95.81662
#> 25  182.5947 170.25542  90.06952 173.98868 151.65977 143.86202 137.61899
#> 26  198.2563 120.39422 101.59093 164.66434 136.91390 190.35170  73.07813
#> 27  116.2255 117.61128  98.18287 196.93662 166.59614 117.46534 159.33833
#> 28  138.8963  96.43677 153.92280 118.46658 128.70364 200.97341 210.57050
#> 29  103.3408 154.01627 131.72381 127.09728 124.55495 107.64022 126.06644
#> 30  199.2351 115.21871 194.27567 157.77569 143.91609  94.74726 124.75184
#> 31  185.6980  92.05285 172.10684 144.55345 152.57041 135.25760 178.28446
#> 32  166.8441 149.41066 108.45446 135.61339 160.12255 133.94539 131.90635
#> 33  184.3814 127.42663 131.95334 162.84928 139.84415 119.57142 198.01669
#> 34  236.1236 163.13701 141.89501 147.77888 155.04598 122.73350 159.45690
#> 35  168.1990 182.98865 130.08322 160.58028 103.94725 105.85121 122.73722
#> 36  186.3880 133.57345 149.45061 137.04597 118.42761 177.57878 152.16003
#> 37  138.4554 152.01552  93.55364 154.64145 114.36902 110.94496 154.50411
#> 38  181.1296 155.97069 173.97131 136.77400 136.34011 155.02417 207.30001
#> 39  111.5157 163.70977 163.93942 206.91836 142.02738  91.77534 128.24694
#> 40  148.7660 194.81450 193.02175 157.75860 127.97137 221.98550  94.23739
#> 41  128.5895 109.39841 127.52090 134.01995 181.26865 131.47950 156.46581
#> 42  111.0487  73.37392 129.16596 184.71785 182.74270 152.43528 181.39762
#> 43  148.2061 141.54431 151.50823 159.04877 129.30489 118.71786 131.72486
#> 44  145.9750 180.12940 145.86418 158.72992 140.53309 111.17845 123.75823
#> 45  221.2435 139.22349 154.52295 155.16914 219.92480 116.14972 152.09842
#> 46  139.9289 111.10090 135.43774 167.12054 135.60434 141.55103 142.33174
#> 47  178.6296 150.58415 113.95473 138.51511 119.56372 172.10864 183.50956
#> 48  119.3542 156.12658 133.95326 135.94989 149.08668 144.38544 125.67594
#> 49  136.4043 182.68448 171.80624 201.75665 169.59677 136.35311 112.80688
#> 50  176.4292 153.98575 140.39863 176.34344 115.04658 136.19562 163.79592
#> 51  198.5250 167.17846 147.66775 212.30001 159.84382 132.10255 103.17637
#> 52  153.1247 129.58768 134.34774 170.54038 106.62488 109.02916 197.37683
#> 53  152.3330 139.14395 177.81266 156.10628 237.16355 106.26759 138.31715
#> 54  195.3498 140.75389 123.23966 137.86918 118.12041  96.05777 141.99535
#> 55  145.3155 166.25222 192.71166 135.35134 165.26217 134.59068 179.46211
#> 56  128.0379 134.62841 131.11891 127.05440 204.36216 211.54438 128.78806
#> 57  119.9272 167.47596 174.47388 156.85289 157.18458 135.84754 122.82750
#> 58  163.8983 121.43808 151.45901 135.18493  96.28866 152.84892  99.10389
#> 59  126.6301  52.86058  98.29103 171.38723 116.35431 164.34207 131.15881
#> 60  124.8793 176.61600 189.33508 140.76384 144.65096 136.31024  95.69395
#> 61  146.5709 149.09925  90.83612 179.87679 136.83715 137.31782 191.41106
#> 62  154.9886 139.24237 130.31907 170.03380 159.31357 135.47609 124.18405
#> 63  220.1999 115.03445 135.96266 134.63331 173.11829 150.97595 207.05145
#> 64  138.5693 139.19770 131.59747 115.49449 121.50466 155.59237 146.79750
#> 65  172.2420 183.73722 142.74466 161.79178 209.70471 144.46043 133.87184
#> 66  205.2008 101.79127 131.54571 193.18061 175.96936 164.80929 101.21205
#> 67  143.4084 141.41391 124.39057  91.60484 168.44825 123.50105 128.72319
#> 68  147.1513 140.34126 117.78975 143.49790 128.01589 172.49900 173.14259
#> 69  145.8607 117.22902 153.16667 189.74320 106.79356 159.66401 147.33177
#> 70  147.6952 128.03324 176.72735 123.33491 187.78443 139.22786 109.60389
#> 71  131.5479 145.90448  80.26545 202.77238 154.74502 190.60928 128.35594
#> 72  181.9091 139.13183 150.66826 117.25071 143.11707 142.03238 122.77804
#> 73  158.3556 147.92369 151.27668 159.63118 135.35075  94.82064 122.72343
#> 74  179.6408 147.63826 131.55204 121.86789 166.16839 129.43194 130.27561
#> 75  121.9665 172.32510 138.71576 106.70451 161.34221 114.47701 167.43407
#> 76  185.6623 142.44385 174.69195 177.77786 168.07047 154.14238 165.10249
#> 77  182.0700  77.74698 145.90461 135.61908 170.80186 161.68126 164.35637
#> 78  190.8491 162.29132 124.85378 181.20296 204.48037  91.33651 146.78609
#> 79  150.7428 108.69870 146.12076 138.59656 149.95043 167.93065 150.89676
#> 80  190.8248 179.10100 172.08083 116.25312 135.20471  95.66470 139.04934
#> 81  113.8919 215.77029 120.89042 132.65823 119.81014 178.03527 145.43618
#> 82  190.5970 171.76856 144.41281 176.62346 128.71425 143.82548 123.50556
#> 83  104.0275 179.26904 110.93765 127.68235 146.57626 122.80984 132.72542
#> 84  110.0134 171.59222 126.82375 177.37297 147.52567 177.69272 121.46751
#> 85  180.8844 171.07926 125.04639 155.56445  99.67253 122.82107  98.34363
#> 86  156.4218 188.75985 122.65713 179.54838 176.61029 183.29750 165.77704
#> 87  176.2821 111.11007 216.67619 185.60910 179.79609 138.75979 106.74114
#> 88  139.1259 157.71226 126.06923 201.81072 185.04136 135.16563 117.05162
#> 89  136.9147 163.79169 116.70493 161.97329 140.12075 119.17449 185.37301
#> 90  139.4015 161.80660 137.85196 122.14654 133.19992  88.04138 135.58568
#> 91  153.8494 146.46378 138.15978 173.28093 167.79443 168.42713 147.19991
#> 92  163.0227 157.42571 187.03876 109.79517 141.78918 130.85375 133.54998
#> 93  172.6625 122.96560 141.24019 196.36715 135.89419 152.98875 170.86689
#> 94  173.3864 165.04749 124.51850 133.16427 193.53469 153.11642  96.84832
#> 95  161.0546 135.87804 134.32074 149.24211 131.53598  93.22062 136.61623
#> 96  164.4917 134.92498 136.45468 177.17778 220.33537 181.37795 136.57286
#> 97  142.1177 146.44525 119.24159 145.85839 206.67338 177.16934 182.28247
#> 98  156.4927 143.95295 156.69677 110.90501 109.65501 158.43122 151.40857
#> 99  159.2662 133.28213 182.01896 153.38282 142.75464 141.87094 128.62363
#> 100 153.4575 133.04834 148.33415 166.59525 184.94408 136.12085 131.15279
#>         B.4_1
#> 1   156.32688
#> 2   162.81089
#> 3   126.86168
#> 4   148.70880
#> 5   149.01313
#> 6   191.66031
#> 7    94.06503
#> 8   121.99792
#> 9   108.99411
#> 10  167.39739
#> 11  167.39986
#> 12  179.70202
#> 13  175.92365
#> 14  152.06161
#> 15  141.37178
#> 16  158.10234
#> 17  146.72979
#> 18  177.38127
#> 19  185.48057
#> 20  194.47775
#> 21  193.88901
#> 22  135.28293
#> 23  162.50443
#> 24  138.96621
#> 25  170.03042
#> 26  210.47410
#> 27  155.78024
#> 28  136.03161
#> 29  136.20671
#> 30  122.68183
#> 31  165.83258
#> 32  193.91354
#> 33  175.63586
#> 34  111.67615
#> 35  152.07704
#> 36  185.97069
#> 37  175.69096
#> 38  194.74914
#> 39  158.33269
#> 40  129.71394
#> 41  170.90224
#> 42  140.88214
#> 43  136.06847
#> 44  187.25832
#> 45  139.28800
#> 46  109.25881
#> 47  137.19931
#> 48  137.30792
#> 49  151.19258
#> 50  105.78864
#> 51  141.89157
#> 52  172.32842
#> 53  156.28173
#> 54  213.70482
#> 55  152.03676
#> 56   93.32976
#> 57  116.37420
#> 58  132.08497
#> 59  147.07548
#> 60  151.42410
#> 61  168.34097
#> 62  164.56699
#> 63  100.98276
#> 64  199.89013
#> 65  182.27378
#> 66  179.46118
#> 67  168.30352
#> 68  182.94343
#> 69  129.98827
#> 70  162.20694
#> 71  113.12486
#> 72  163.08509
#> 73  163.69222
#> 74  182.03966
#> 75  125.20996
#> 76  160.14334
#> 77  144.55522
#> 78  159.57665
#> 79  124.87439
#> 80  120.67896
#> 81  215.36293
#> 82  106.64331
#> 83  143.14583
#> 84  135.55976
#> 85  122.83025
#> 86  188.46635
#> 87  166.47605
#> 88  170.12008
#> 89  158.57899
#> 90  151.23639
#> 91  173.04651
#> 92  177.10035
#> 93  180.80692
#> 94  143.77715
#> 95   97.04590
#> 96   93.86106
#> 97  120.94768
#> 98  252.31477
#> 99  121.22165
#> 100 125.86716