Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10, 10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(
  subset = rep(c("A", "B"), 50),
  level = rep(1:4, each = 25)
)
rownames(meta) <- 1:100
scRNA <- SeuratObject::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA,
  by = c("subset", "level"),
  min.cells = 10, max.cells = 20
)
#>         A.1_1     B.1_1     A.2_1     B.2_1     A.3_1     B.3_1     A.4_1
#> 1   169.74199 109.80274 128.57945 205.55691 164.61004 117.31029 109.25133
#> 2   159.06060  94.77390 117.96432 163.69364 157.55214 165.37706 162.92506
#> 3   157.67859 165.91262 144.29867 148.61261 163.53614 167.76513 122.55706
#> 4   100.86299 156.43795 123.76292 119.95502 186.59517 155.93184 125.61207
#> 5   169.73304 138.50345 144.30191 149.54358 144.62503 144.95923 145.53278
#> 6   170.26362 176.89430 117.77392 180.67851 167.26972 128.31886 134.57186
#> 7   132.13725 136.14668 123.88262 141.21128 184.16389 102.37315 124.88860
#> 8   151.94964  98.50835 161.92604 120.17618 195.03276  82.11226 117.86812
#> 9   152.19573 147.80078  98.46060 178.91086 150.92592 141.54848 155.03024
#> 10  118.46017 163.80053  98.52449 149.53945 174.65091 133.68914 149.95514
#> 11   96.64139 185.97536 190.68515 194.40867 149.35763 155.10697 113.28786
#> 12  143.43073 150.96578 135.55219 129.08375 172.22501 130.56906 103.81718
#> 13  157.01527 146.56780  84.21621 146.49621 134.33160 150.85186 142.52808
#> 14  198.06807 107.54409 165.08676 138.78515 167.22022 143.03146 130.65238
#> 15  195.88336 167.96354 125.23690 150.02868 184.59441 166.56843 157.04704
#> 16  159.22820 112.44613 144.02191 159.75744 146.78667 174.02883 144.57802
#> 17  157.03248 208.53662  63.77214 151.10532 129.78894 149.81314 147.45366
#> 18  162.13723 143.51142 154.80325 162.39033 158.61303 124.11703 128.12987
#> 19  163.67374 103.96132 114.09927 120.79887 180.88040 143.36605 132.32602
#> 20  173.02754 131.38417 181.05174 207.13383 114.95395 173.97140 171.34992
#> 21  157.65877 160.39934 144.57388 151.93929 185.58325 177.49504 142.07966
#> 22  165.42390 179.24214 144.56533 162.86240 121.51074 118.86364 128.92770
#> 23  177.16873 154.34815 157.95716 123.79516 198.64540 167.65763 152.73372
#> 24  146.02623 149.84840 130.28855 142.47715 132.25802 132.99738 158.73107
#> 25   95.71125 140.06009 198.84503 144.37919 183.89934 109.88580 181.39426
#> 26  171.24609 174.02889 164.13735 103.15158 170.76797 123.24233  96.45615
#> 27  206.62730 107.47005 134.04455 177.74759 198.19901 144.07305 101.55418
#> 28  163.79774 192.37507 102.59305 161.57466 161.78176 123.92583 104.28842
#> 29  153.28983 185.73580  96.55823 156.79357 147.11072 113.54548 162.75553
#> 30  119.88382 135.60905 118.87434 138.03973 167.42361 124.94878 194.06231
#> 31  184.83311 139.66930 173.61094 141.86978 115.00908 121.90866 115.11115
#> 32  157.24449 174.18569 195.36789 187.03127 127.21207 115.96328 143.09568
#> 33  138.45255 145.65591 204.02410 146.22067 133.04731 111.48031  99.11800
#> 34   85.67776 142.18706 160.63253 133.95703 148.99794 132.99373 166.73097
#> 35  123.41624 128.80058 170.80173 157.88889 168.62515 154.54455 144.92124
#> 36  119.03706 196.69727 148.45566 118.29196 146.90498 102.85833 145.78222
#> 37  175.48111  95.73105 155.09554 218.82161  81.69902 156.72443 125.37853
#> 38  159.93813 133.51941 106.67732 186.20512 168.10423  94.64404 166.39775
#> 39  105.30502 137.83983 101.42776 143.88396 137.07185 141.94888 154.61036
#> 40  136.47104 166.43101 129.26888 124.74542 159.62485 170.53132 119.30517
#> 41  128.99563 133.12382 101.42455 111.93916 110.00722 174.96377 119.37586
#> 42  120.67516 144.59857 129.95465 180.49624 152.85620  96.16441 110.97800
#> 43  133.92379 211.66138 172.10476 170.80433 146.80298 164.26863 104.30014
#> 44  171.98736 156.58707 110.92360 187.05852 143.59788 135.70306 217.69071
#> 45  177.92815 128.10874 122.33058 118.17475 190.94667  98.26869 114.00230
#> 46  170.79278 137.10955 113.73234 206.29713 137.01801 106.08030 182.49674
#> 47  150.38727 131.03358 148.52791 119.03779 170.34854 112.52699  91.33613
#> 48  162.74814 131.58808 138.59220 154.29003 163.73805 137.23755 123.69339
#> 49  134.26039 183.35008 148.76747 165.49888 127.28788 170.10400 142.63834
#> 50  142.64754 167.67602 173.73423 116.92307 112.25410 131.92707 195.26885
#> 51  200.89296 129.72550 184.59931 175.06109 130.24131 154.65695 148.47368
#> 52  140.68799 128.59961 110.72488 191.71786 185.32720 144.58185 155.34538
#> 53  148.69714 132.85310 147.93863 157.44017 151.08430 136.02282 172.14523
#> 54  138.82074 167.00598 196.32316 130.89531 131.31319 116.37197 145.21022
#> 55  197.22787 119.80369  92.40606 101.83091 126.34642 177.08969 165.43485
#> 56  117.51122 135.92891 106.10862 199.50272 150.21514 109.14597 144.10504
#> 57  164.30761 170.18750 173.99413 179.35434 170.66781 181.40993 110.21467
#> 58  135.01195 112.24845 110.86963 143.06993 181.81246 139.28160 122.21576
#> 59  161.93143 174.21095  96.17327 126.66968 167.54637 164.18817 115.56667
#> 60  153.66829 120.20631 134.24837 144.87063 183.72612 170.36123 133.65268
#> 61  156.60160 151.63926  60.10097 154.14913 110.37889  89.85451 120.88994
#> 62  184.14593  86.49798 166.48880 119.94491 165.52155 131.62081 123.68668
#> 63  132.96994 140.92546 113.19501 183.38193 160.36066 149.77475 134.72756
#> 64  163.11928 159.89596 159.97462 140.94465 118.50806 130.19700 129.48142
#> 65  155.79897 170.35290 113.14106 121.03403 158.03287 152.74382 147.70912
#> 66  134.18006 138.89556 107.01824 144.05701 118.65767 113.54632 169.34920
#> 67  150.10286 108.36374 123.22610 155.02646 117.86886 118.40960 135.07237
#> 68  173.48038  98.65994 157.31924 166.33114 142.15736 163.03006 126.15783
#> 69  150.75436 159.12143 171.56885 130.96837 138.48315 127.02009 170.72463
#> 70  192.16350 155.54035 149.14322 162.25880 149.89733 144.57822 173.37676
#> 71  159.44988  97.83706 163.25802 140.11070 165.29181 173.32353  68.62162
#> 72  127.87151 140.04936 144.16416 171.00464 166.13967 162.08495 172.68337
#> 73  122.26240 134.99364 123.81328 122.71155 129.80025 165.96762 123.71381
#> 74  215.97885  87.56026 138.47148 152.50971 182.02548 122.87226 123.13074
#> 75  125.17158 154.57642 110.29221 176.62105 162.95526 169.95543 105.74567
#> 76  147.17267 164.04379 177.07678 129.04043 197.13923 137.79326 133.68985
#> 77  172.47772 164.56810 130.70369 188.38967 181.58392 206.99924 150.10644
#> 78  109.46147 137.50301 188.75428 134.11647 166.73741 125.37202 124.33527
#> 79  190.14623 164.18913 142.93777 151.60063 128.32588 150.04680 142.17250
#> 80  108.87817 158.76806 136.52830 137.60525 152.16148 178.24586 120.58715
#> 81  108.16411 134.95927  86.10835 112.48876 113.48074 124.21249 155.87954
#> 82  139.43049 129.28967 165.72161 152.66844 166.32899 142.62595 129.00509
#> 83  195.72229 112.19018 167.05195 156.88241 134.23107 131.29022 169.71723
#> 84  174.01693 154.20091 145.52092 157.83516 176.28680 107.46116  99.40647
#> 85  140.73561 101.28174 179.96963 126.68120 185.58500 140.04946 195.90194
#> 86  197.85310 152.91288 153.06139 185.76157 167.87142  77.70607 126.61616
#> 87  128.21475 164.06313 165.34070 148.87110 140.54871 175.13536 122.12590
#> 88  152.23213 144.13990 138.28974 133.91487 157.28672 159.64727 149.72485
#> 89  147.48818 176.53650 128.10284 121.18249 138.39503 152.23315 135.49410
#> 90  190.22289 159.01997 127.29698  92.63289 108.84738 186.91182 233.29406
#> 91  151.34416  97.99516 133.43219 162.47412 138.40608 161.03948 121.61513
#> 92  158.34609 141.25273 198.31330 142.98718 222.37934 151.14385 169.01611
#> 93  186.00861 164.04862 120.91901 143.81213 135.01107  92.28043  79.65493
#> 94  122.00961 136.09043 100.65658  89.65001 148.76398 133.07827 127.29247
#> 95  113.99761 103.00268 171.48323 150.67725 248.66858 111.84729 114.28395
#> 96  124.83929 149.98760 141.92842 176.32317 170.87189 123.97069 135.73312
#> 97  186.36018 128.02906 167.67656 165.32073 155.48548 173.87148 150.14363
#> 98   98.10471 117.59549 149.87266 140.32939 150.16472 122.15466 111.85472
#> 99  154.17670 128.02605 139.46021 166.98434 133.92616 170.76577 179.83886
#> 100 152.56781 158.70362 181.53322 130.92261 134.82187 144.91068 160.22684
#>         B.4_1
#> 1   137.99660
#> 2   162.28478
#> 3   121.15653
#> 4   164.34457
#> 5   152.40026
#> 6   168.52260
#> 7   143.84829
#> 8   121.02027
#> 9   169.32367
#> 10  121.60497
#> 11  134.72738
#> 12  128.64415
#> 13  115.63969
#> 14  137.80569
#> 15  159.57799
#> 16  125.09529
#> 17  118.01644
#> 18  115.88882
#> 19  128.91492
#> 20  175.01828
#> 21  153.36797
#> 22  165.36996
#> 23  160.68873
#> 24  157.35617
#> 25  141.58689
#> 26  154.44648
#> 27  144.46499
#> 28   90.55613
#> 29  179.86416
#> 30  165.98969
#> 31  164.54030
#> 32  127.84625
#> 33  120.10174
#> 34  200.49347
#> 35  113.45829
#> 36   66.18245
#> 37  167.37741
#> 38  173.34038
#> 39  149.73562
#> 40  136.01209
#> 41  152.32034
#> 42  138.84322
#> 43  178.15728
#> 44  169.55989
#> 45  159.64343
#> 46  176.69117
#> 47  126.15631
#> 48  140.04568
#> 49  189.89163
#> 50  165.83932
#> 51  112.07443
#> 52  188.04636
#> 53  115.83255
#> 54  138.00122
#> 55  114.70454
#> 56  148.56731
#> 57  156.43950
#> 58  183.94111
#> 59  184.76445
#> 60  120.63601
#> 61  108.23068
#> 62  164.79702
#> 63  127.13366
#> 64  103.35466
#> 65  183.79451
#> 66  162.41272
#> 67  152.93606
#> 68  157.98149
#> 69   64.97510
#> 70  142.56926
#> 71  158.04927
#> 72  100.70776
#> 73  137.46078
#> 74  126.34283
#> 75  150.16289
#> 76  140.83178
#> 77  143.65060
#> 78  137.44837
#> 79  119.23895
#> 80  135.33638
#> 81  134.00968
#> 82  167.97844
#> 83  144.91794
#> 84  140.32096
#> 85   97.19976
#> 86  140.35764
#> 87  144.28134
#> 88  128.96552
#> 89  133.25579
#> 90  157.88085
#> 91  149.26521
#> 92  125.01885
#> 93  190.11393
#> 94  149.58865
#> 95  148.23878
#> 96  159.37437
#> 97  128.88132
#> 98  152.20952
#> 99  133.75152
#> 100 114.91497