Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10, 10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(
  subset = rep(c("A", "B"), 50),
  level = rep(1:4, each = 25)
)
rownames(meta) <- 1:100
scRNA <- Seurat::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA,
  by = c("subset", "level"),
  min.cells = 10, max.cells = 20
)
#>         A.1_1     B.1_1     A.2_1     B.2_1     A.3_1     B.3_1     A.4_1
#> 1    86.10782 127.61148 116.47864 142.30072 178.79201 149.68022 117.87598
#> 2   128.04583 195.21470 139.77897 168.49649 113.56114 118.16806 128.87359
#> 3   156.47887 140.40988 153.54273 126.77250 199.61488 191.75455 130.80810
#> 4   115.79519 185.22592 154.09755 139.34602 201.63301 164.17379  92.10909
#> 5   154.01390 120.11694 108.05563 141.40972 173.92024 140.76248 128.52985
#> 6   153.29788 153.84118 101.38693 132.82774 167.10071 173.00563 158.68382
#> 7   103.96123 139.64548 100.09671 168.71978 135.37860 113.51812 165.98059
#> 8   193.93913 134.46828 124.30874 192.37616 128.91724 105.61585 139.53403
#> 9   227.05140 137.96390 188.89554 113.20822 169.16668 125.88435 160.05909
#> 10  173.92543 128.86448 136.08868 211.07149 195.65237 131.04560 104.04118
#> 11  135.10533 133.55103 155.24679 137.40159 129.97127  96.99311 151.47408
#> 12  140.90011 127.62530 123.85611 133.58579 113.28824 133.72592 120.56337
#> 13  132.95493 123.44621 184.88663 163.34290 138.32797  95.02126 215.94446
#> 14   77.13009 188.82805  87.12602 149.91263 154.08974 150.23526 163.48870
#> 15  153.87305 129.78437 115.46806 148.58974 136.91010 132.62526 134.18856
#> 16  123.81562 119.78236 117.08415 149.82419 143.17743 163.87208 122.90723
#> 17  141.04884 115.51814 146.43633 154.78356 138.57549 118.78973 179.50500
#> 18  164.48436 159.73325 166.06016 174.66636 121.00278  96.22592 108.71000
#> 19  151.70860 170.95899 163.03927 152.35292 112.71514 115.92169 119.27730
#> 20  160.17681 110.22552 129.56736 128.96172 200.74258 109.59663 175.98808
#> 21  120.64272 111.92123 135.97181 110.02904  99.80399  75.30130 104.53101
#> 22  150.31343 117.61442  93.34595 166.90837 158.54177 168.71591 123.63193
#> 23  139.93554  76.19670 170.99776 111.49386 212.76886 124.15567 136.52598
#> 24  132.90447 110.24895 149.97132 191.86448 113.00206 156.64311 149.47514
#> 25  184.43740 146.01290 180.07808 146.36055 154.47635 124.41924 110.46096
#> 26  153.08147 154.51538 184.63104 141.14467 159.95192 153.62228 150.67848
#> 27  175.39722 147.30603 142.44397 182.76677 120.71019 139.11193 159.92880
#> 28  152.96535 143.67666 112.12842  90.81616 184.94549 163.58672 171.71942
#> 29  136.87733 136.72443  96.29101 102.55119 149.19861 143.45862 120.31502
#> 30  156.92855 107.51639 121.23277 197.32116 124.46273 141.67565 145.31302
#> 31  196.74913 237.82762 164.35847 145.37779 122.01537 148.09594 123.59378
#> 32  162.99994 151.97156 151.54349 114.68608 157.67420 126.06043 143.69359
#> 33  144.20765 122.83863 109.72149 148.11170 172.47195 131.55776  97.29061
#> 34  168.56663 167.46019 107.82840 146.06836 133.14742 148.88753 115.69285
#> 35  144.92362 141.35452 192.04048 144.32199  89.41590 132.55643 181.22892
#> 36  179.07150 130.92421 121.88575 123.69768 120.21441 128.16540 143.01132
#> 37  214.56073 128.99914 126.08794 168.95539 142.54056 172.77838 111.28591
#> 38  165.40116 125.66867 127.05680 128.67828 155.14622 120.46215 142.32187
#> 39  136.61700 166.71254 121.13307 180.50711 219.17853 153.64244 131.38837
#> 40  132.37435 102.80998 174.26470 153.86462 139.74127 163.61033 133.43577
#> 41  157.16349 123.54230 193.39045 137.13904 126.23116 181.44436 135.69029
#> 42  108.65610 135.63268 161.80620 123.74405 172.35803 130.90765 192.93531
#> 43  138.03232 155.17216 175.95061 193.36864 183.00812 156.11353 169.75881
#> 44  157.43191 183.93201 171.50854 136.24901 159.94168 185.00344 146.92194
#> 45  140.11265  96.15024 171.06203  85.52324 294.20130 147.16225 161.11140
#> 46  177.27959 186.29428 115.19577 172.25265 161.78062 100.27936 180.72585
#> 47  188.27653 152.03455 116.91274 163.83779 151.90512 140.43662 120.60600
#> 48  149.89191 138.61215 129.92198 162.03992 188.44080 165.82488 151.84909
#> 49  185.90681 126.63514 197.12227 155.07702 205.27770 139.48674 150.17421
#> 50  200.49926 114.76556 143.49228 165.71392 210.11045 183.79119 150.89331
#> 51  176.18137  78.92406 174.41941 123.59633 140.90688  86.51919 171.51028
#> 52  105.52208  88.99994 140.28624 120.39301 140.08931 111.52610 143.85670
#> 53  118.63987 166.92884 151.42535 231.66698 186.18470  97.54261 107.98865
#> 54  176.56989 122.96148 132.31264 160.94239 177.28238 140.48004 174.39548
#> 55  139.28116 142.81295  93.07655 130.17072 130.08573 112.71007 172.36923
#> 56  128.43802 168.16559 139.26555 126.77044 201.87825 158.32543 150.61910
#> 57  174.90028 159.62607 161.69054 140.69606 127.88292 181.56839 122.31484
#> 58  134.89135 166.83069 137.83309 153.71577 157.19095 125.23833 117.42898
#> 59  113.23146 118.50950 103.77374 146.06640 132.85217 144.44672 152.79716
#> 60  171.25242 141.67665 140.53415 110.30392 111.33416 146.19681 110.98277
#> 61  166.91866  92.30003 123.75415  96.00057 158.23164  66.35770 121.33618
#> 62  164.91605 116.97951 112.57743 209.57026 209.26722 135.48589 150.68597
#> 63  148.03606 149.50126 147.56518 168.42880 167.50852 158.76832 185.39590
#> 64  182.95672 127.71543 214.62396 141.25132 191.38886 140.15119 134.62911
#> 65   92.15578 156.32266 147.82306 144.15256 134.11937 174.80635 158.97087
#> 66  115.62582 151.19090  97.75940 102.91489 177.59534 174.96470  94.23503
#> 67  159.16620 170.84832 129.87455 150.06851 170.69887 106.52570 171.48980
#> 68  140.83140 106.35014 139.87225 135.44066 213.61426 147.64687 157.38226
#> 69  185.31910  92.25709 189.85995 141.68487 135.92315  96.19506 157.55143
#> 70  172.17498 165.30492 109.23563 179.30624 127.59548 160.05865 133.14963
#> 71   91.08928 120.62716  76.94086 129.72323 100.36810  90.32450 127.43099
#> 72  191.97392 103.13151  94.01063 139.51362 156.83947 145.70871 133.71553
#> 73  133.16774 133.69785 154.72672 185.05561 167.30108 136.84307 160.08481
#> 74  118.37784  80.20024 104.94137 145.13630 121.93577 147.20279  95.46234
#> 75  171.11611 182.81516 127.19222 175.45478 211.90156 115.16494 131.60215
#> 76  122.51686 109.52375 139.59573 209.34843  96.77239 149.17542  95.98086
#> 77  166.07273 112.65393 170.04103 142.16018 188.99593  97.54427 117.92402
#> 78  151.97962 176.22210 134.13338 167.05630 177.80019 109.07082 143.84439
#> 79  175.53034 161.34761 126.83169 186.43932 180.13443 115.16860 127.80233
#> 80  170.81018 146.43243 127.86957 176.00652 191.13185 138.59414 144.98642
#> 81  120.83852 167.93090 147.24038 159.97769 164.85767  91.20629  76.08427
#> 82  153.05978 106.74290 107.92401 104.78767 157.97876 138.37676 110.20413
#> 83  133.46682  76.09584 114.61887 109.66750 119.74543 121.04856 147.98525
#> 84  108.69149 124.11869 155.72109 138.12368 142.03898 171.02186 121.03976
#> 85  137.86570  94.96583 148.75471 181.48759 175.47693 143.58621 147.74878
#> 86  134.75614 153.56233 158.29096 156.63017 159.65053 207.28897  86.47179
#> 87  170.76389 188.46023 187.17773 164.38951 144.27752 131.34227 147.47393
#> 88  199.07235 114.27545 193.77437 149.56694 179.67777 145.28575 115.43943
#> 89  112.48363  76.06246 164.63364 144.25063 174.59810 132.13823 139.60537
#> 90  148.78762 131.43681 157.47861 188.39605 147.84013  72.99921 109.61814
#> 91  178.35336 133.72124 142.82605 168.60710 142.85472 135.21257 110.72049
#> 92  162.81867 148.85574 145.97019 127.12209 134.27011 172.66327 139.78637
#> 93  145.11327 108.01148 132.64483 134.67169 143.69924 142.10988  99.20474
#> 94  139.12000 120.46163 132.65453 130.34327 162.77483 166.01762 148.32355
#> 95  174.50110 106.67489 143.23741 117.94914 164.32701 153.82839 122.96279
#> 96  140.46186 140.84782 140.14492 151.20970 172.87944 123.31262 121.09738
#> 97  135.64354 134.05990  96.69109 160.73380 162.73792 122.95084 135.97296
#> 98  148.23850 129.90004 175.90108  95.35018 157.19341 142.72539 112.74966
#> 99  111.80199 104.64814 123.75889 156.64443 146.70506 119.38022  81.46141
#> 100 138.32758 152.96215 131.05611 146.72284 210.20101 141.77347 196.31570
#>         B.4_1
#> 1   166.18948
#> 2   134.21564
#> 3   128.45049
#> 4   161.37011
#> 5   157.03223
#> 6   179.60314
#> 7   158.72946
#> 8   177.31461
#> 9   237.52302
#> 10  123.91126
#> 11  138.66548
#> 12   79.84921
#> 13  123.33395
#> 14  156.53576
#> 15  158.90377
#> 16  205.39300
#> 17  141.59178
#> 18  191.18565
#> 19  208.04096
#> 20  112.19586
#> 21  119.04049
#> 22  136.75523
#> 23  131.56695
#> 24  130.46769
#> 25  109.05949
#> 26  132.93206
#> 27  162.85587
#> 28  194.40428
#> 29  161.93930
#> 30  169.98706
#> 31  136.68243
#> 32  163.23409
#> 33  135.04336
#> 34  153.04719
#> 35  142.47898
#> 36  131.89888
#> 37  176.11638
#> 38  111.49928
#> 39  119.63663
#> 40  170.96963
#> 41  101.15520
#> 42  105.16614
#> 43  212.91414
#> 44  137.45737
#> 45  134.26039
#> 46  184.23900
#> 47  168.94830
#> 48  136.83316
#> 49  159.00548
#> 50  170.06304
#> 51  176.31622
#> 52  191.05253
#> 53  161.30480
#> 54  138.89441
#> 55  169.65211
#> 56  145.14067
#> 57  157.52877
#> 58  161.69903
#> 59  159.11484
#> 60  129.45883
#> 61  120.54375
#> 62  226.51814
#> 63  125.19382
#> 64  173.63106
#> 65  150.95956
#> 66  158.31896
#> 67  155.08702
#> 68  135.62074
#> 69  186.15818
#> 70  218.21491
#> 71  116.24518
#> 72  200.24201
#> 73  182.30559
#> 74  129.92959
#> 75  161.94015
#> 76  145.78890
#> 77  143.11031
#> 78  136.91021
#> 79  156.38745
#> 80  152.02844
#> 81  161.29998
#> 82  108.13729
#> 83  144.14245
#> 84  135.05280
#> 85  152.30357
#> 86  150.91198
#> 87  134.86830
#> 88  131.32105
#> 89  122.45291
#> 90  115.34674
#> 91  145.83662
#> 92  151.57205
#> 93  173.34844
#> 94  134.12057
#> 95  156.65114
#> 96  154.37352
#> 97  151.02290
#> 98  120.66953
#> 99  176.29729
#> 100 162.49003