Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10, 10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(
  subset = rep(c("A", "B"), 50),
  level = rep(1:4, each = 25)
)
rownames(meta) <- 1:100
scRNA <- Seurat::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA,
  by = c("subset", "level"),
  min.cells = 10, max.cells = 20
)
#>         A.1_1     B.1_1     A.2_1     B.2_1     A.3_1    B.3_1     A.4_1
#> 1   150.48697 145.79162 170.07761 125.79557 219.47925 156.3765 130.97231
#> 2   182.01159 116.46332 173.51700 166.54582 159.50148 142.6302 111.03053
#> 3   142.83034 198.97297  96.59735 129.23291 172.34359 161.8914 123.41657
#> 4   162.74730 183.62597 163.49043 152.47215 131.44627 154.0107 140.22516
#> 5   120.89409 135.42631 147.83354 235.91431 144.41354 132.7299 132.73400
#> 6   171.77351 155.60027 128.59881 161.84625 157.63278 164.6563 145.16767
#> 7   121.69188 180.23613 131.44650 134.76449 119.62181 133.9744 190.77265
#> 8    92.95642  84.16403 178.67352 134.52484 183.25947 162.2854 102.90901
#> 9   119.50621 139.09268 197.18632 111.75267 113.68655 183.6291 121.21832
#> 10  241.31385 190.13327 107.99547 180.20929 170.94598 172.0486 141.30122
#> 11  172.15946 127.19155 176.56948 152.22799 160.01981 132.4462 121.20386
#> 12  158.46324 151.88708 141.81768 190.76045 149.65374 175.0941 149.15721
#> 13  126.55363 117.28096 106.60076 155.64685 175.15127 171.9755 107.07556
#> 14  190.09975 137.53371 132.26813  99.30740 161.62029 148.2539 123.44929
#> 15  155.60504 129.54905 128.17220 163.16440 142.89707 198.5619 126.26380
#> 16  132.81488 113.11799 161.75989 146.19783 210.44167 104.1753 141.49159
#> 17  182.51135 131.98149 116.85826 129.57918 176.33500 182.6234 101.28679
#> 18  121.37168 154.60817 166.45999 149.15132 142.37061 239.9354 181.15997
#> 19  133.09104 112.11706  95.70158 131.29822 152.51452 101.9581 166.19550
#> 20  139.03074 144.58488 139.11778 214.29677 117.32828 142.8186 166.96488
#> 21  152.61423 159.26755 120.31423 140.13312 182.10048 146.2636  93.62638
#> 22  124.26403 144.67994 132.73534 135.24328 118.56828 102.0381 105.44568
#> 23  111.31380 158.87688 182.75582 150.90036 109.64422 167.5246 194.06572
#> 24  152.40552  96.22908 173.95473 133.08660 156.93971 122.7331 154.21613
#> 25  170.27003 120.24880 103.65731 149.39831 153.23596 115.6886 119.27206
#> 26  179.23710 121.40875 116.92169 165.03444 127.92162 159.4163 159.44826
#> 27  140.20995 110.19627 119.54215 179.05594 153.10719 157.2069 127.46360
#> 28  147.34842 194.52214 109.12380 129.93284 164.06968 224.3784 170.46021
#> 29  175.49202 132.50284 144.54016  83.70444 180.24557 114.0801 150.66273
#> 30  160.72612 143.10636 164.39794 160.81894 157.21362 184.9094 157.55509
#> 31  160.33238 100.81708 121.61745 194.93218 176.75748 128.4157 104.84580
#> 32  134.85763 155.17550 121.91908 166.72753 143.16938 105.9430 139.61816
#> 33  150.94656 131.01730 157.33380 184.30232 154.34695 130.6531 191.25952
#> 34  112.25048 154.05212 136.84268 160.48284 116.93275 124.3836 169.92907
#> 35  161.94145 157.66022  92.82454 162.08420 148.29071 113.8570 121.38810
#> 36  157.86924 169.30053 129.83126 124.18245 202.76345 121.5127 181.53726
#> 37  166.40848 152.47616 139.66982 171.77411 122.89313 154.1202 136.46873
#> 38  163.00503 143.64037 165.61636 132.83353 157.11682 176.4895 146.37528
#> 39  133.12278 124.20005 126.03079 146.27341 134.84656 178.0948 116.05530
#> 40  133.57369 167.73604 159.65552 173.93739 182.39565 123.7240 122.57713
#> 41  159.62296 111.75694 170.16709 144.22816 122.57612 112.4003 103.27798
#> 42  170.30708 149.06126 154.72241 148.20933 163.09359 137.5580 176.67716
#> 43  169.68750 156.31727 133.79342 102.39150 108.63369 163.8174 132.32988
#> 44  154.36917 143.21202 124.67702 136.72842 159.45720 175.4093 107.52738
#> 45  102.30580 140.29144 143.01168 170.19298 136.64197 114.1128 149.66270
#> 46  123.69413 102.02303 164.18230 132.93048 162.80050 153.3915 145.91036
#> 47  111.56374 173.29741 124.21304 170.03092 204.12186  91.5001 133.37297
#> 48  134.39073 148.03345 140.42622 166.96791 135.55952 155.6177 119.23922
#> 49  146.58630 120.76777 176.01305 108.57399 138.08476 184.1058 134.19229
#> 50  135.37756 216.31148 186.88007 168.51508  94.05111 153.8256 169.72324
#> 51  205.07164 124.06344 131.35683  93.47961 143.58239 181.4091 142.03650
#> 52  163.81282 180.03802 118.26651 153.40068 176.87090 104.7101 125.35812
#> 53  149.31243  67.03871 135.78200 121.93376 142.73389 107.0966 137.42212
#> 54  133.02315 182.06465  76.19257 165.19072 177.07137 109.1920 131.50191
#> 55  165.76331 106.61382 228.72077 160.33996 158.17008 108.3982 120.77538
#> 56  190.39408 111.71629 159.76823 114.36501 151.22524 115.9678 142.91095
#> 57  158.70038 116.61770 159.94249 190.23380 138.83517 101.4517 129.24246
#> 58  111.79649 117.08973  96.61842 117.62655 117.83197 158.0228 115.08043
#> 59  129.30414 179.34416 178.59622 178.49834 141.61344 145.7631 148.67959
#> 60  165.94543 179.60638 170.05933 139.99149 158.36648 124.8108 132.90685
#> 61  172.14871 168.96072 127.47424 167.10335 153.98070 167.3859 149.68450
#> 62  154.29925  93.71749 134.56800 113.34027 127.79217 102.0472 188.32008
#> 63  154.84812 170.45283 168.31366 177.96436 187.63563 145.9065 150.80368
#> 64  179.26264 130.82879 131.11407 114.06126 121.30440 138.2717 145.06337
#> 65  128.23483 107.19054 141.51351 160.51013 205.68791 162.5320 155.07516
#> 66  133.77120 130.41573 117.80659 159.73220 143.53319 157.4193 131.51677
#> 67  182.08707 152.12205 159.15742 137.39154 178.96840 120.1368 123.32538
#> 68  116.55704 135.94355 145.53073 173.57718 146.21699 133.8139 143.95342
#> 69  149.28772 140.71761 100.22190 140.30798 103.79269 104.8867 147.78232
#> 70  170.28255 149.83086 169.03482 167.29944 139.46228 176.6696 126.04150
#> 71  203.05638 107.42995 125.74843 173.30900  85.55407 112.5689 145.07244
#> 72  136.99656 141.28285 191.65988 126.35726 146.57273 155.3073 118.26892
#> 73  183.33102 108.71215 144.82728 155.83869 142.99384 137.6525 176.36567
#> 74  141.36147  92.39713 112.79016 135.50093 125.75508 170.8110 134.76421
#> 75  129.93783 137.44968 148.10601 129.45297 140.20957 189.2621 154.81214
#> 76  119.31491 100.67520 168.62921 137.17036 124.33571 114.0989 205.12755
#> 77  152.31890 127.60755 161.61656 175.90863 149.61864 113.8872 139.40589
#> 78  118.20308 173.52922 114.73671 158.12139 171.06375 160.1115 183.20613
#> 79  162.45925 144.43162 131.33734 143.50765 151.14012 138.7882 135.31469
#> 80  132.35128  93.08537 149.94238 170.06568 122.71134 117.3210 169.37600
#> 81  187.17844 159.96346 142.65646 137.53059 146.81992 139.5665 145.93327
#> 82  192.32670  96.31261 131.67298 161.50850 111.34872 151.7326 178.26751
#> 83  148.84915 185.56681 108.70961 103.57142 139.69821 141.7629 133.07987
#> 84  164.44432 178.49768 127.46383 139.34993 134.07324 139.6532 122.44671
#> 85  160.73955 117.66181 152.01400 172.41931 141.40115 141.6269 127.66481
#> 86  183.33596 116.03386 181.70690 195.13690 171.44101 121.0405 142.25578
#> 87  146.80051 140.72570 137.93144 178.15058 198.58270 147.8067 145.14428
#> 88  173.89841 201.32813 119.70485 132.91919 169.45247 133.4117 124.13145
#> 89  160.20638 140.80327 147.04974 167.93216 178.21365 157.8395 136.82137
#> 90  151.26944 119.11706 102.11640 117.53494 182.87264 146.0862 138.37975
#> 91   94.29526 117.12249 169.55968 153.75319 131.86549 129.4488 139.33442
#> 92  225.00582 111.17541 135.73412 106.96860 127.95863 143.4157 114.87264
#> 93  144.01642 119.98874 104.80658 118.13769 128.70165 138.9930 183.51144
#> 94  146.80017  84.25489 162.31290 126.01415 227.83742 156.7703 123.25820
#> 95  207.47144 176.14261 156.46741 122.50142 178.21203 167.5872 177.15610
#> 96   99.89824 152.20224  97.33950 158.31978  84.51107 143.6612 148.67021
#> 97  224.60060  97.04792 157.16385 138.46779 109.32495 157.0212 104.80418
#> 98  156.60190 112.18029 124.82730 160.33587 116.71022  78.9658 102.27194
#> 99  159.01320 129.65952 104.42525 135.56308 140.22817 135.2344 129.21835
#> 100 143.89149  86.40919 139.02487 103.65110 153.42671 144.5698 127.88234
#>         B.4_1
#> 1    99.62305
#> 2   187.34823
#> 3   136.36196
#> 4   161.39976
#> 5   142.84102
#> 6   123.13866
#> 7   135.01413
#> 8   179.46229
#> 9   150.95817
#> 10  134.60047
#> 11  139.00977
#> 12  262.73999
#> 13  155.58517
#> 14  182.32327
#> 15  165.20042
#> 16  137.88342
#> 17   79.39395
#> 18  153.70621
#> 19  148.30244
#> 20  119.04347
#> 21  115.40992
#> 22  188.89646
#> 23  141.96110
#> 24  183.84432
#> 25  139.40640
#> 26  125.73651
#> 27  175.93187
#> 28  144.90185
#> 29  125.67864
#> 30  105.05858
#> 31  142.67404
#> 32  168.14338
#> 33  172.08336
#> 34  172.95507
#> 35  166.77100
#> 36  182.00910
#> 37  201.53058
#> 38  150.72283
#> 39  135.31234
#> 40  121.14375
#> 41  126.99011
#> 42  139.23665
#> 43  178.11277
#> 44  151.43321
#> 45  128.04435
#> 46  149.18858
#> 47  114.34622
#> 48  194.00785
#> 49  110.79961
#> 50  154.71540
#> 51  132.85602
#> 52  156.78332
#> 53  182.69640
#> 54  114.25777
#> 55  149.97782
#> 56  113.13422
#> 57  133.64007
#> 58  172.80105
#> 59  176.91394
#> 60  183.95081
#> 61  159.62725
#> 62  171.77227
#> 63  182.41697
#> 64  188.77853
#> 65  135.74201
#> 66  134.37017
#> 67  174.87879
#> 68  171.11210
#> 69  146.34104
#> 70  167.16306
#> 71  160.28980
#> 72  174.15496
#> 73  135.61200
#> 74  143.73992
#> 75  119.11888
#> 76  127.32101
#> 77  144.45122
#> 78   92.60857
#> 79  145.61488
#> 80  110.49108
#> 81  173.91006
#> 82  142.91563
#> 83  124.51452
#> 84  146.47413
#> 85  160.44960
#> 86  149.89163
#> 87  160.87707
#> 88  144.19827
#> 89  176.13583
#> 90  146.53952
#> 91  221.05675
#> 92  148.62582
#> 93  138.60872
#> 94  158.51179
#> 95  165.57029
#> 96  172.66411
#> 97  174.62763
#> 98  161.41989
#> 99  137.90066
#> 100 187.81212