Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10,10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(subset = rep(c("A","B"),50),
                   level = rep(1:4, each = 25))
rownames(meta) <- 1:100
scRNA <- Seurat::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA, by = c("subset","level"),
               min.cells = 10, max.cells = 20)
#>         A.1_1     B.1_1     A.2_1     B.2_1    A.3_1     B.3_1     A.4_1
#> 1   140.93518 160.93727 170.18492 138.55469 162.9897 147.73245 206.56906
#> 2   128.55157 141.47220 157.32908 157.79335 172.4588 166.92068 132.79402
#> 3   196.59976 146.85275 136.36741 140.09041 174.5660 113.29570 142.69632
#> 4   192.28978 168.36569 142.34685 152.13659 214.6833 171.18908 101.90864
#> 5   159.16416 141.07642 161.56865 129.72902 125.8440 158.52617 190.16249
#> 6   160.55703 132.25054 133.64481 132.20610 137.5061 101.00940 135.18892
#> 7   166.53461 179.46875 155.41593 127.54225 163.6973 148.41388 170.84329
#> 8   137.16675 103.66314 132.57521 166.86194 123.6396 110.63614 168.42396
#> 9   182.60715 171.71170 118.31403 146.02552 118.7485 119.87837 171.37997
#> 10  128.48264 126.67945 114.89329 110.19122 142.1996 145.23211 127.91790
#> 11  118.73290 157.12494 118.97973 134.35595 188.1129 123.82002 134.72351
#> 12  173.63737 145.81048 179.21294 102.65047 186.1233 194.74343 170.73523
#> 13  113.94560 155.12455 154.35738 104.47162 225.7342 121.90344 160.29897
#> 14  119.05351 105.60679 157.28286 122.91237 159.7416 188.88444 109.32871
#> 15  159.46266 126.66496 143.92246 127.89651 110.2019 140.51143  82.30297
#> 16  174.63420 123.68348  99.24459 155.23317 147.9122 121.24188 168.94177
#> 17  127.66187 128.98496  44.85836 182.84221 129.3727 156.67443 140.11781
#> 18  168.38357 142.00694 125.08209 214.00174 178.9675 111.15392 157.83970
#> 19  137.11513 108.95451 145.34829 159.20510 131.2037 108.43881 150.64594
#> 20  145.40202 207.52641 194.04735 188.90148 141.8601 137.48064 151.41095
#> 21  180.99940 141.82429 103.76870 157.15864 119.7453 163.34479 197.85474
#> 22  149.44385 100.90310 174.99185 140.66390 154.6693 125.81551 136.68929
#> 23  183.68558 153.47662 171.87096 150.51424 157.4472 126.80197 168.19806
#> 24  168.80626  96.52731 145.80517 127.70456 207.4920 137.13010 110.03556
#> 25  141.90229 128.91196 114.50271 181.57822 196.8873 153.14853 159.19911
#> 26  213.32165  79.72054 102.18390 174.77565 141.3835 147.52613 149.16628
#> 27  139.39174 120.90011 151.51396 137.89924 181.9406 104.59137 130.65449
#> 28  161.41371 164.65606 128.21743 136.64161 157.3426 101.80104 123.01540
#> 29  152.37400 192.94546 150.81768 181.30825 176.0101 149.55652 121.60405
#> 30  144.26728 126.33056 129.38203 176.38416 128.9074 145.01081 117.42849
#> 31  149.82246 140.35337 108.35034 164.31480 174.4799  98.66454 177.82265
#> 32  129.08974 189.05799 181.26425 191.59694 114.7891 204.49620  96.36109
#> 33  148.17847 166.76168 150.13132 167.68711 137.9877 160.68113 140.90323
#> 34  129.45083 191.60507 112.64539 103.04456 198.9098 170.84944 169.11033
#> 35  177.38983 134.34799 146.32328 169.29928 136.1124 114.77129 193.41107
#> 36  154.69447 146.84179 142.56483 156.27910 163.9971 106.38079 142.06990
#> 37  129.01813 179.26928 172.62183 126.28342 166.4158 197.30437  90.14569
#> 38  106.98611 138.63095 151.74410 162.12440 155.9216 149.39188 136.60422
#> 39   89.91450 185.52660 165.53764 189.95164 159.0130 125.72798 132.68213
#> 40  150.46537 152.29959 169.64438 214.09063 164.8726 115.01274 133.61993
#> 41  195.82242 111.11326 128.94501 123.06764 174.8177  90.94532 138.82193
#> 42  168.44074 114.03479 165.99037 215.29331 116.6227 179.60946 129.07909
#> 43  167.01047 161.38540 142.02935 150.44093 138.4868 153.58158 129.05558
#> 44  156.39064 110.79922 151.76834 129.98619 151.0960 144.82975 115.98878
#> 45  120.52012 193.13067  89.02800 155.71978 173.5651 113.53556 142.15980
#> 46  143.97608 108.49558 125.58321 120.64323 155.4940 188.81542 135.44514
#> 47  103.93901 143.94392 161.02275 147.80541 114.0191 129.18324 139.66472
#> 48  104.16221 222.43448 152.53564 186.80494 155.3317  85.80390 200.05557
#> 49  161.83915  98.65157 132.49375 192.95358 117.4469  89.60782 129.52812
#> 50  160.43716 142.86631 169.18680 157.55766 185.9667 117.68678 129.67193
#> 51  204.66977 128.21549 184.81295  99.35717 123.6266 152.61764 117.03061
#> 52  183.32523 101.61132 137.70558 181.20982 169.7560 121.72945 160.00463
#> 53  160.52285 154.04237 191.62458 192.84020 126.3115 122.93183 126.45694
#> 54  145.13506 147.26645 128.69868 123.53375 131.4804 173.72759  90.01583
#> 55  136.78199 165.62383 186.54726 205.55256 152.1025 139.47479 189.35536
#> 56  164.16895 123.46039 121.72092 137.55039 170.0814 166.40033  89.32864
#> 57  171.39636 151.21383  89.97128 159.56885 111.2287 144.05370 179.07101
#> 58  152.51729 123.57833 139.81724 179.12467 158.0216 193.43401 128.73702
#> 59  150.17263 165.41372 109.39129 189.04284 107.8804 149.96526 115.92284
#> 60  171.97169 174.34485 119.17504 179.67584 153.3004 165.07751 145.65132
#> 61  211.55480 117.56535 112.53723 162.88290 138.9378 126.26579 137.87579
#> 62  140.63369 205.07362 129.65462 114.39161 201.7296 153.56239 136.74524
#> 63  164.95303 154.61826 143.07844 217.87653 126.9131 136.60901 142.73162
#> 64  111.22310 127.91666 153.39769 157.08901 163.8264  79.03113 149.90578
#> 65  222.64516 133.01202  84.48807 166.41774 186.6948  90.28074 148.10898
#> 66  101.70951 125.45741 105.16334 201.90101 151.0466 174.49831 176.13513
#> 67  173.18590 155.77178 103.03856 142.75575 183.2042  83.64094 120.15211
#> 68  191.18607 144.96963 127.71778 224.64506 177.9093 210.02569 129.43248
#> 69  161.41936 173.10935 172.72912 157.97375 179.3132 139.48450 137.97416
#> 70  151.64090 131.33428 139.34832 165.43550 147.5758 118.62426 140.37030
#> 71  144.87419 144.57561 208.32987 108.41052 106.5956 141.86806 127.29362
#> 72  150.39388 159.78001 138.17176 147.44976 105.5410 139.50101 132.52132
#> 73  127.12351 190.51065 154.17402 160.12330 136.5294 137.65545 109.88668
#> 74  184.66011 125.17121 104.92195 154.81938 201.2322 116.31061 121.66137
#> 75  144.94625 135.63340 174.16384 158.61003 187.9593 144.38279 159.44382
#> 76  141.84464 165.72741 112.44077 115.09712 169.3982 199.52858 137.90159
#> 77  136.31562 122.81297 146.00660 107.08813 181.1379 199.58271 157.55701
#> 78  207.09394 180.28713 172.52869 154.91219 156.1864 152.75352 156.33975
#> 79  153.24652 170.09942  96.76488 150.34713 206.2996 180.96153 134.98033
#> 80  153.84104 131.40797 188.57542 170.10232 155.8548 182.78677 176.78375
#> 81  134.40227 122.86515 125.17556 174.55875 160.8562 145.96765 178.73843
#> 82  184.73742 165.01626 154.08919 180.74967 171.8693 205.40970  94.33309
#> 83  132.51953 133.30817 171.46588 184.69188 172.4483 161.40247 155.40942
#> 84  133.93630 154.81370 132.84599 152.57870 127.9109 165.23217 139.63313
#> 85  123.34275  87.43307  86.76677 129.42598 110.8541 156.67059 111.55923
#> 86  158.86753  88.44998 174.08673 177.03222 151.6544 173.55135 174.78202
#> 87  198.69305 150.36763 154.49733 156.77130 165.2672 138.31036 156.44601
#> 88  121.21505 142.15593 146.03211 153.03113 159.4132 119.48039 167.24924
#> 89  149.61471 102.86751 165.61361 104.89182 183.1263 107.48512 181.39574
#> 90  132.79537  95.41210 196.15106 115.09992 117.6873 192.25260 104.66895
#> 91   91.90369 134.32867 133.11871 218.40475 137.9304  95.47104 158.45108
#> 92  157.16198 108.91135 164.25762 130.96410 189.2798 169.56448 197.64566
#> 93  159.21329 103.38616  91.63730 160.43588 160.8369 112.97617 160.77756
#> 94  142.28085 174.97213 118.84225 153.17137 179.7618 178.37907 182.50638
#> 95   86.63706 113.57206 195.71069  73.47459 103.4153 180.15010 155.46723
#> 96  183.31504 123.21363 147.27197 176.54686 154.6017  89.97790 133.68198
#> 97  148.79265 168.80332 161.16855 148.95553 163.8579 102.91823 146.75336
#> 98  160.14928 106.99122  87.48074 180.16936 123.0700 117.59181 108.14701
#> 99  108.23741 131.56531 135.83267 164.99947 131.0873 182.55183 210.35545
#> 100 191.32882 172.44183 108.47203 140.59590 135.5209  70.16291  84.39980
#>         B.4_1
#> 1   131.15090
#> 2   152.62270
#> 3   140.25795
#> 4   121.57907
#> 5   180.39397
#> 6   173.29137
#> 7   142.68860
#> 8   131.27089
#> 9   180.44874
#> 10  145.60718
#> 11   83.88089
#> 12  119.11359
#> 13  149.14446
#> 14  148.83438
#> 15  197.47905
#> 16  120.50336
#> 17  166.98189
#> 18  183.82911
#> 19  162.66766
#> 20  159.89808
#> 21  159.53211
#> 22  144.98254
#> 23  140.91082
#> 24  169.58237
#> 25  193.85556
#> 26  166.39208
#> 27  127.83180
#> 28  113.55279
#> 29  130.76351
#> 30  151.53158
#> 31  151.13378
#> 32  158.34833
#> 33  133.53024
#> 34  162.53992
#> 35  146.51165
#> 36  122.32465
#> 37  204.60926
#> 38  114.40647
#> 39  176.07603
#> 40  132.73414
#> 41  177.08854
#> 42  165.29586
#> 43   89.56918
#> 44  134.89172
#> 45  157.62294
#> 46  153.78978
#> 47  194.65223
#> 48  173.90808
#> 49  116.62168
#> 50  198.25383
#> 51  127.91404
#> 52  184.44292
#> 53  163.44186
#> 54  170.78923
#> 55  165.99480
#> 56  142.58195
#> 57  169.68374
#> 58  142.50613
#> 59  183.75354
#> 60  167.24448
#> 61  139.52952
#> 62  201.75386
#> 63  158.82759
#> 64  166.76781
#> 65  124.63026
#> 66  128.34207
#> 67  137.56227
#> 68  163.75432
#> 69  179.99377
#> 70  120.38931
#> 71  134.64145
#> 72  108.46541
#> 73  166.55736
#> 74  149.88086
#> 75  208.95450
#> 76  224.35609
#> 77  168.74981
#> 78  107.45184
#> 79  157.97753
#> 80  130.70761
#> 81  159.54036
#> 82  111.47204
#> 83  175.90020
#> 84  138.29628
#> 85  129.39343
#> 86  129.95640
#> 87  109.45819
#> 88  146.54280
#> 89  144.64856
#> 90  123.47081
#> 91  150.97502
#> 92  146.42515
#> 93  145.24352
#> 94  158.64475
#> 95  203.30079
#> 96  155.09844
#> 97  153.04944
#> 98  110.86745
#> 99  122.89028
#> 100 194.88025