Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10,10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(subset = rep(c("A","B"),50),
                   level = rep(1:4, each = 25))
rownames(meta) <- 1:100
scRNA <- Seurat::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA, by = c("subset","level"),
               min.cells = 10, max.cells = 20)
#>         A.1_1     B.1_1     A.2_1     B.2_1     A.3_1     B.3_1     A.4_1
#> 1   154.96732 173.67891 132.81625 104.72841 156.71496 163.51226 112.67650
#> 2   139.49189 153.17907 126.30751 200.59383 140.52977 110.52271 109.28161
#> 3   128.35502 190.26225 110.36953 143.22685 164.53166 162.09401 114.30803
#> 4   194.45186 132.62919 125.13338 114.85536 121.04805 129.25197 180.08046
#> 5   162.90295 121.48260  87.47729  88.26265 182.42499 108.63695 169.01381
#> 6   126.34844 121.89641 133.05829 127.20235 183.49380 151.61631 139.18724
#> 7   191.71141 132.03091 147.18076 151.14836 183.78128 173.13695 134.62912
#> 8   113.62894 118.15091 113.31526 159.82702  96.98324 127.30637  96.50438
#> 9    94.55261 140.70824 193.29962 149.33900 152.62657 147.13182 233.76083
#> 10  205.32010 192.44278 156.55733 125.84234 124.52502 186.41691 145.38195
#> 11  160.05188 127.67705 140.56536 143.99311 159.20799 134.35206 196.20335
#> 12  176.33368 127.60686 139.85934 231.58993 141.59034  88.96342 135.66028
#> 13  131.07372 154.73206 120.87212 157.32091 220.89714 153.69990 113.10200
#> 14  145.47893 114.86246 137.61176 156.84944 195.74002 171.04393 131.55662
#> 15  212.56070 125.01570 126.79044 142.88982 144.66292 134.17542 165.80138
#> 16  129.14848 132.57684 170.75691 126.59616 150.74564 134.70850 141.70384
#> 17  178.21590 146.59642 173.18653 183.82695 159.47222 146.39668 118.33485
#> 18  132.56326  94.21507 105.71661 188.96051 164.30942 126.30688 125.75724
#> 19   98.54616 131.93397 128.93016 142.44271 160.33015 148.96102  91.11653
#> 20  121.25857 152.99308 146.35819 182.13407 184.33899 125.46566 185.21771
#> 21  135.39223 138.36818  93.12951 116.24419 195.64029 152.33624 121.00792
#> 22  156.48503 194.03446 148.30247 158.80501 130.99857  76.11924 134.51987
#> 23  114.96766  99.78875 148.00917 167.26804 149.97274 125.83397 177.18718
#> 24  183.81326 145.06845 202.87874 140.11566 178.28372 192.58480 126.16462
#> 25  215.52237 137.36994 136.18121 152.36352 172.06260 114.07450 138.13272
#> 26  128.70969 164.69833 164.32745 109.03184 150.98398 111.92643 116.83984
#> 27  217.73566 152.22381 110.84500 189.27100 164.09832 167.41590 136.74665
#> 28  140.83403 188.74581 133.13200 110.31132 109.12782 121.29433 111.50814
#> 29  152.37616 116.79593 107.85467 186.59601 173.00763 130.78412 162.79463
#> 30  179.32566 187.20217 208.53871 171.01833 181.97165 171.48858 107.41382
#> 31  146.48699 140.76389 183.63994 132.09080 162.46158 115.61794 104.76122
#> 32  201.85341 112.05259 129.75264 164.88754 133.47599 163.86714 165.54549
#> 33  128.89830 134.46758 192.30781  80.11710 158.26014  97.40767 161.98165
#> 34  180.67419 153.10152 162.77020 150.11737 174.18762 136.25877 127.76568
#> 35  144.02748 106.33034 177.65927 155.69209 168.71740 155.05857 140.27004
#> 36  139.41616 136.82332 130.37376 114.75685 123.68819 183.39093 154.37578
#> 37  140.94509 152.91840 165.47535 129.38109  62.06567 111.49662 107.54040
#> 38  145.58548 202.64503 125.23156 129.49446 126.81955 142.43791  89.68328
#> 39   92.85121 179.75668 153.19439 203.45701 145.11064 150.86872 152.84415
#> 40  108.21663 132.58816 136.09452 142.41832 164.78417 135.15847 132.94079
#> 41  136.58241 161.00352 159.98864 177.66418 109.84745 175.52018 170.59194
#> 42  161.98784 134.61224 132.63294 179.33776 146.10246  96.37898 114.58429
#> 43  137.21433 158.28980 122.27922 180.31389 143.42247 156.77857 126.92677
#> 44  128.02921 141.22191 127.89927 138.82812 123.70963 150.36444 184.40484
#> 45  122.25199 122.58474 123.13999 138.79070 139.18935 161.12414 113.80729
#> 46  111.17143 147.73835 126.89614 178.10466 147.84301 108.56796 207.87296
#> 47  152.25271 144.07038  61.66525 168.45874 140.38450 124.01611 102.98593
#> 48  174.44074 159.95378 102.81320 169.81614 168.49299 152.21911 135.20237
#> 49  133.55817 132.66789 128.56806 176.41797 136.23301 115.76427 136.69918
#> 50  149.56549 113.87858 108.07303 128.16376 159.03902 138.40824 122.31298
#> 51  104.05162 192.51310 123.90916 132.37695 141.45733 141.96710 105.98713
#> 52  140.45611 117.33507  95.14703 113.01657 133.38096 101.34898 124.75035
#> 53  150.83808 100.79989 132.72675 173.87714 152.39725 146.64121 144.60051
#> 54  155.41776 120.75710 115.54866 157.16179 147.73848 157.57208 124.57492
#> 55  140.34589  99.71910 134.14432 130.76634  77.57234  97.45158 145.78848
#> 56  136.72380 182.98015 187.04026 168.33972 141.71549 145.19090 104.98793
#> 57  120.54200 120.10080 154.75792 146.54244 108.86583 124.73509 104.00748
#> 58  116.35442 116.98771 108.49238 181.09797 178.48248 129.68045 108.99251
#> 59  128.84767 140.42039 111.11938 172.77877 155.76019 164.54649 148.27541
#> 60  105.25112 123.37005 135.04407 198.44252 140.76115 171.89303 146.76662
#> 61  146.62888 117.62887 138.26245 199.88864 189.06697 126.69838 156.75416
#> 62  193.31005 132.40045 129.98955 128.19841 130.22130 161.38349 188.44392
#> 63  142.68458 157.76895 176.67988 153.72904 184.78274 158.50487 195.66066
#> 64  165.71929 175.24169 118.38925 152.25101 188.01902 197.12428 136.57592
#> 65  139.08155  98.82897 177.05296 151.26551 151.71037 196.34054 141.33063
#> 66  208.22058 165.80631 153.35706 144.45216 146.45099 142.72905 158.12852
#> 67  208.95593 132.99741 132.28266 172.55462 163.73641 126.37600 171.52965
#> 68  149.65101 161.70918 145.28094 126.36819 216.59766 146.80207 108.45819
#> 69  139.99184  82.54773 115.83992 136.78650 158.72757 150.12170 126.58712
#> 70   85.20892 158.17080 108.43438 117.22156 153.76829 156.29591 159.77145
#> 71  110.15608 113.28311 169.09369 133.29477 171.72269 137.28367 140.13032
#> 72  191.17660 130.76553 142.53612 131.63299 168.83592 131.69996 109.01205
#> 73  171.19580 166.73174 121.49809 184.91273 186.40502 109.38741 113.53689
#> 74  152.38835 104.05054 139.84722 112.39075 172.64205 142.50990 134.35451
#> 75  180.72539 117.30787 185.93227 170.25277 172.40277 144.13274 121.89080
#> 76  122.98069 200.10724 118.71202 186.36286 101.32710 142.16241 156.00573
#> 77  104.74883  86.87823 122.28946 132.01238 135.29378 147.16605 127.98146
#> 78  153.63114 154.84605 103.22020 154.23874 150.85718 163.38786 147.70060
#> 79  152.49398 156.08262 156.13444 180.12248 157.86764 126.08904 142.22821
#> 80  183.30854 174.59021 193.45669 178.01883 156.62600 123.34792 122.69133
#> 81  151.96488 103.05719 155.08952 173.45264 164.63484 128.83235 148.90499
#> 82  167.29053 131.89857 173.49457 112.41435 137.41908 169.99447 148.30640
#> 83  134.43648 137.14348 169.18435 155.91817 184.11198 136.99715 151.70571
#> 84  159.76567 119.63029 117.69965 142.05380 163.24138 187.07775 116.51953
#> 85  153.97284 109.19956 133.06195 110.86839 169.78214 145.93943 145.22355
#> 86  150.12850 127.21143 148.95414 196.82066 150.15623 135.33276  86.15654
#> 87  153.96948 147.65612 157.87682 137.74524 144.35886 212.97181 121.14971
#> 88  178.21549 190.26403 125.27625 138.10218 172.70307 107.53621 150.91250
#> 89  150.23407  84.38549 184.85278 120.48777 161.04342 133.70121 139.38294
#> 90  103.06204 111.16189 140.42341 138.74266 179.55445 128.83416 157.59974
#> 91  181.65601 120.00964 147.20697 183.50115 164.51826 183.12860 133.13921
#> 92  156.24664 138.29586 159.09905 186.73495 190.67375 141.77878 120.15731
#> 93  202.13778 159.70606 180.47340 141.34118 160.00079 145.59229 153.29381
#> 94  142.61276 106.52622 139.73616 139.30597 142.66695 168.32579 160.02983
#> 95  190.69605 137.95961 167.99398 164.49859 146.95191  92.17471 136.49409
#> 96  133.94455 187.39231 170.91724 127.35666 155.75353 136.56648 169.68414
#> 97  155.24039 139.86537 186.98631 152.63400 177.08700 109.52282 141.66388
#> 98  138.44980 105.97724 139.58748 141.47679 141.50720 107.65934 101.48006
#> 99  146.34707 156.28370 152.56976 147.03343 145.42838 110.03235 160.70275
#> 100 150.76102 157.96509 145.93802 184.96011 207.59946 112.48142 167.72832
#>         B.4_1
#> 1   130.51972
#> 2   164.27693
#> 3   156.05594
#> 4   102.87916
#> 5   155.66798
#> 6   123.08484
#> 7   154.96800
#> 8   161.62322
#> 9   112.23868
#> 10  218.22598
#> 11  148.36012
#> 12   91.67830
#> 13  115.21400
#> 14  157.92717
#> 15  213.27959
#> 16  179.11059
#> 17  159.68133
#> 18  167.49184
#> 19  195.36585
#> 20  157.95610
#> 21  155.23780
#> 22  186.48690
#> 23  132.43361
#> 24  166.44512
#> 25  158.79183
#> 26  202.17868
#> 27  157.71441
#> 28  136.05280
#> 29   93.54701
#> 30  151.24405
#> 31  171.39681
#> 32  135.00815
#> 33  184.11924
#> 34  116.33834
#> 35  208.03936
#> 36  132.35740
#> 37  157.32318
#> 38  169.45758
#> 39  168.72029
#> 40  170.79113
#> 41  178.24215
#> 42  161.38116
#> 43  146.23185
#> 44  182.17678
#> 45  144.04154
#> 46  188.29180
#> 47  211.77304
#> 48  101.25164
#> 49  224.02425
#> 50  119.68956
#> 51  139.09252
#> 52  151.83013
#> 53  202.94871
#> 54  140.30186
#> 55  158.67661
#> 56  152.21822
#> 57  190.63119
#> 58  128.27597
#> 59   96.45036
#> 60  145.04177
#> 61  162.17699
#> 62  164.06270
#> 63  159.36007
#> 64  123.93048
#> 65  139.24708
#> 66  166.90713
#> 67  170.76830
#> 68  159.92879
#> 69  194.08294
#> 70  151.37502
#> 71  189.32385
#> 72  145.02355
#> 73  158.32637
#> 74   98.61697
#> 75  155.17317
#> 76  138.89719
#> 77  133.33090
#> 78  125.51412
#> 79  158.45504
#> 80  149.04542
#> 81  134.43768
#> 82  130.25997
#> 83  128.22664
#> 84  133.54971
#> 85   87.83687
#> 86  162.79358
#> 87  101.73982
#> 88  164.71787
#> 89  142.12429
#> 90  129.57215
#> 91  123.66480
#> 92   85.79342
#> 93  156.45376
#> 94  142.96289
#> 95  155.96840
#> 96  139.52236
#> 97  166.86027
#> 98  141.17455
#> 99  155.56655
#> 100 145.80375