Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10,10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(subset = rep(c("A","B"),50),
                   level = rep(1:4, each = 25))
rownames(meta) <- 1:100
scRNA <- Seurat::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA, by = c("subset","level"),
               min.cells = 10, max.cells = 20)
#>         A.1_1     B.1_1     A.2_1     B.2_1     A.3_1     B.3_1     A.4_1
#> 1   119.02905 129.53115 160.55544 141.60562 140.11742 114.47208 161.35879
#> 2   173.72411 156.34410 192.65732 151.94120 125.80035 135.57586 133.56436
#> 3   152.34488 174.74162 144.83518  73.82184 158.64929  95.73646 125.53002
#> 4   163.01364 181.83446 148.74400 117.94288  98.38824 127.13166 161.53626
#> 5   156.47008 139.41238 176.16661 161.47272 159.35528 130.16846 110.29146
#> 6   179.21703 161.39020 121.72975 149.41096 129.20139 109.89771 157.52332
#> 7   154.55073 129.83980 159.80869 128.65800 119.69710 122.85833 126.68683
#> 8   121.53348 168.50504 131.75350 152.62728 219.01231 132.65362  80.39084
#> 9   156.34195 159.17479 160.32288 167.81306 146.65952  94.67285 108.80168
#> 10  119.74255 119.70870  99.79058 118.79154 147.00674 117.90131 113.28580
#> 11  146.10978 151.05216 161.83691 136.73475 143.84909 202.92180 124.83703
#> 12  175.58748 139.57971 144.80268 179.93726 127.73940 145.71093 112.06905
#> 13  159.26051 105.96695 150.15568 152.09028 149.57804 142.81475 152.04109
#> 14  118.11925 172.06695 170.77011 155.42021 112.69177 107.97359 117.37104
#> 15  189.39313 179.98449 165.27380 182.00568 162.74556 135.38262 147.09282
#> 16  155.50136  99.45149 129.83213 167.19637 117.90649 164.00120 164.59756
#> 17  149.42696 127.84932 157.69202 152.17279 126.87634 108.25180 154.09539
#> 18  172.96210 124.51597 129.92231 131.18175 171.15458 128.92068 126.14209
#> 19  196.31581 150.61335 161.27993 130.07663 199.80012 144.78640 159.58234
#> 20  152.35540 138.73710 126.79598 176.26717 122.89010 184.64802 125.93467
#> 21  110.39169 135.51661 133.06238 148.10486 126.15965 141.79707 148.96033
#> 22  185.87814 127.07064 131.75347 157.80337 108.78606 159.92240 102.97682
#> 23  107.38100 161.56256 129.34611 112.44711 135.89038 125.60030 101.96486
#> 24  116.50974 166.41120 165.86453 111.99609 142.36912 116.77137 128.60995
#> 25  171.82077 146.14157 168.68256 146.53802 178.38149 183.74598 132.10811
#> 26  167.24147 126.90795 187.03822 171.90210 144.60761 128.32778 145.97050
#> 27  230.65937 162.86494 148.66393 201.64506 133.49824 160.84765 139.64151
#> 28  170.00575 144.94097 150.11041 165.89729 173.70021  73.96210 127.72090
#> 29  149.32686 137.56320 140.21759 157.71165 174.06152 124.14066 185.79917
#> 30  122.77299 138.90930 134.96662 137.37324 187.60642 115.23089 107.54727
#> 31  134.88523 181.27879 111.70178 108.38080 201.23998 109.67901 156.45836
#> 32  187.29804  90.54452 182.05683 164.92734 162.55086 131.00151 141.84571
#> 33  197.82447 137.69502 176.92820 113.64526 174.51840 123.11439 131.27638
#> 34  108.07014 118.67138 131.16759 145.71172 133.50713 107.96809 120.34827
#> 35  186.50930 105.23127 119.72248 152.06972 172.49846 207.23269 145.26104
#> 36  206.64832 166.52819 174.21978 162.72656 120.24618 156.56622 149.07664
#> 37  125.70810 148.81126  96.47592 100.66790 109.75350 167.94870 140.47037
#> 38  143.20435  85.93003  76.74818 169.28773 171.36798 179.30641 134.03171
#> 39  182.03929 127.12938 148.72915 124.32884 160.57642  99.57993 177.76535
#> 40  218.85155 112.69253 146.08368 166.68367 136.74458 157.57563  82.58353
#> 41  111.79203 172.68156 173.90244 162.55405 190.53104 132.96706 198.73783
#> 42  157.02288 214.92210 130.54997 178.70850 143.15954 110.26419 156.37736
#> 43  126.27593 112.72276 131.40511 138.68232 130.13415 102.39935  95.60010
#> 44  159.98523 168.93795 175.20959 125.85865 155.12776 134.41420 128.61329
#> 45  120.51676 150.81703 137.35634 140.89597  99.78914 123.54582 165.93873
#> 46   92.96002 127.16410 118.90545 182.42138 202.60048 147.24614 159.99080
#> 47  128.15978 120.42732 122.96949 161.27148 163.39395 123.54454 160.42256
#> 48  165.70608 102.79904 147.50352 164.01716 147.99657 138.91990 133.30860
#> 49  127.33106 156.56898 152.53136 159.69895 116.06472 125.03455 130.24945
#> 50  144.79753 136.44218 134.74116 181.15407 144.03728 112.36680 150.86400
#> 51  131.83978 112.39037 123.81117 186.10716 114.00766 119.99705 145.82797
#> 52  139.70098 142.35405 124.93156 129.81467 128.94671 169.89118 125.64247
#> 53  189.09005 111.43693 139.20059 129.63747 140.69365 155.09196 117.79353
#> 54  166.96367 111.63822 160.31363 135.70190 158.83973 100.63804 136.05412
#> 55  158.41120 135.10938 142.32419 153.68453 132.43307 188.49806 136.15516
#> 56  139.00575 148.43470 144.21505 207.71180 168.06745 119.68386 178.83696
#> 57  137.69651  83.91071 169.06385 177.00682 127.81548 152.62510 106.58702
#> 58  139.05142 160.57269 115.29680 144.24096 170.32443 161.40183 172.47238
#> 59  158.13468 130.01086 158.85335 157.12683 181.35384 167.18179 136.64354
#> 60  133.17506 138.11519 138.96534 111.58482 132.03960 107.34390 131.18892
#> 61  120.02201 191.48602 111.43317 184.99652 125.05098 113.22566 144.55438
#> 62  147.15412 107.12221 103.11216 155.65970 116.87277 155.76109 108.35048
#> 63  155.23647 108.56347  75.55353 158.97070 161.05959  99.04165 126.61019
#> 64  125.56819 167.62383 147.07160 139.33567 145.48597 134.34118 139.24530
#> 65  164.42039 185.93450 116.48893 160.65260 128.25728 130.77454 133.20206
#> 66  139.42365 176.09575 163.35848 181.03544 170.38456 158.76750 143.29324
#> 67  110.58012 155.28754 105.64072 157.08786 102.69585 156.70304 124.33225
#> 68  149.83513 125.12982 115.27909 197.17330 108.77146 134.64202 170.19152
#> 69  183.58588 123.70601 132.93814 159.51851 174.47693 196.31206 141.29292
#> 70  205.12306 142.70323 104.76710 146.99651 161.95996 106.64998 122.92039
#> 71  180.45967 158.71531  90.09616 186.30908 146.39003 145.66980 136.21903
#> 72  136.08725 104.73603 185.25534 179.13642 145.90690 149.27580 118.15393
#> 73  135.37517 151.94853 111.67967 161.91937  84.52191 127.39190 191.44559
#> 74  158.89119 145.52638 112.63291 212.97956 138.93392 103.42247 146.08679
#> 75  171.39646 164.83690 117.37797 172.96204 201.09202 154.57361  93.47180
#> 76  163.65309 147.56911 169.15598 139.00241 183.37674 120.51694 102.32906
#> 77  167.53102 170.09533 112.09592 178.73245 138.44844 166.85374 159.05130
#> 78  179.61971 144.36624 133.31440 127.50256 169.24651 153.64075 162.69145
#> 79  168.66499 124.32686 137.61956 132.47364 134.23875 163.56738 129.73584
#> 80  157.98767 184.37192 135.56305 171.25735 135.88553 111.19512 160.34785
#> 81  120.94264 131.51551 164.54375 118.22460 169.71046 133.01309 102.34763
#> 82  115.31106 170.76647  94.99124 155.86332 176.60581 148.01102 131.14728
#> 83  124.38380 159.50748 154.73543 119.32280 165.19959 117.61717 143.55184
#> 84  140.31105 168.57731 129.26319 217.07099 124.66148 159.91893 119.99625
#> 85   94.79741 157.95079 162.19864 156.60891 187.18670 178.15386 152.91269
#> 86  199.21617  82.68761 148.14141 142.36855 156.66383  98.35664 110.51216
#> 87  153.50748 107.31057 125.70803 135.36957 120.38412 102.48854 129.90095
#> 88  170.16466 140.10021 109.10523 123.72846 222.74635 146.62734  92.86521
#> 89  122.89623 156.78333  98.10397 145.22761 186.18701 150.64189 141.12826
#> 90  185.91263 107.84630 182.11404 163.06282 169.16120 103.81187 122.60449
#> 91  175.78002 140.85833 154.09749 140.35376 138.44792 113.50990 136.41577
#> 92  113.47986 163.60632 119.44365 149.16878 139.05841 125.99201 129.72358
#> 93  126.22092 169.79069 169.00281 122.02856 160.90272 120.67382 101.13594
#> 94  192.28883 108.52296 164.42482 152.91503 128.10662 145.78127 176.03115
#> 95  125.10600 129.66069 140.48533 200.00995 218.58040 144.35657 172.97561
#> 96  125.78839 167.19665 102.86218 132.07186 160.55359 103.80545 126.65215
#> 97  176.30102 138.85529 139.19673 109.21567 183.66026 132.75301 146.71027
#> 98  146.85779 190.64353 152.35669 139.46781 168.68038 127.83724 122.59539
#> 99  182.14152 136.38874 131.77296 159.52154 128.02386 171.11852 160.53747
#> 100 175.16221 116.89373 150.59255 214.11098 128.77444 150.37063 156.30593
#>         B.4_1
#> 1   118.55318
#> 2   180.95398
#> 3   132.07651
#> 4   171.74714
#> 5   224.99644
#> 6   142.29565
#> 7   191.97274
#> 8   125.15432
#> 9   126.60147
#> 10  145.31595
#> 11  116.05606
#> 12   91.62449
#> 13  125.68167
#> 14  122.17386
#> 15  138.73493
#> 16  200.37247
#> 17  165.82695
#> 18  139.21303
#> 19  136.43607
#> 20  178.27989
#> 21  196.85246
#> 22  195.25706
#> 23  215.61696
#> 24  154.40924
#> 25  146.61038
#> 26  174.55385
#> 27  103.61193
#> 28  121.97471
#> 29  177.78586
#> 30  145.48205
#> 31  111.93176
#> 32  152.44219
#> 33  187.70423
#> 34  165.22270
#> 35  164.08752
#> 36  145.49846
#> 37  181.88957
#> 38  131.38431
#> 39  158.31831
#> 40  158.10741
#> 41  179.83907
#> 42  129.86919
#> 43  150.31378
#> 44  126.25909
#> 45  156.28024
#> 46  126.69033
#> 47  162.83177
#> 48  155.81592
#> 49  157.87329
#> 50  104.09258
#> 51  180.29753
#> 52  171.12849
#> 53  189.55745
#> 54  149.83410
#> 55  134.19907
#> 56  166.65888
#> 57  162.63183
#> 58  125.56562
#> 59  150.83869
#> 60  136.47211
#> 61   84.35296
#> 62  125.18558
#> 63  163.58340
#> 64  163.36349
#> 65  141.39408
#> 66  178.67358
#> 67  119.75539
#> 68  123.23296
#> 69  177.34900
#> 70  138.87310
#> 71  141.29057
#> 72  173.18734
#> 73  137.63261
#> 74  115.32844
#> 75  144.35098
#> 76  148.10458
#> 77  133.58823
#> 78  185.52229
#> 79  176.00559
#> 80   96.27297
#> 81  187.85134
#> 82  193.94379
#> 83  176.69412
#> 84  122.31437
#> 85  194.37143
#> 86  197.85737
#> 87  133.82022
#> 88  154.12723
#> 89  152.78560
#> 90  176.74268
#> 91  162.92331
#> 92  130.97633
#> 93  165.21838
#> 94  120.06556
#> 95  129.66327
#> 96  163.18998
#> 97  140.47449
#> 98  140.84211
#> 99  171.76533
#> 100 131.82812