Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10, 10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(
  subset = rep(c("A", "B"), 50),
  level = rep(1:4, each = 25)
)
rownames(meta) <- 1:100
scRNA <- Seurat::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA,
  by = c("subset", "level"),
  min.cells = 10, max.cells = 20
)
#>         A.1_1     B.1_1     A.2_1     B.2_1     A.3_1     B.3_1     A.4_1
#> 1   195.44419 133.75035 158.73843 127.95338 152.18049 168.83187 105.30232
#> 2   201.62563 148.70206 110.18563 125.55258 128.90969 144.96071 159.12192
#> 3   134.65528 143.13530 139.89954 166.90136 119.07908 143.60396 148.52328
#> 4   110.40152 121.16385 122.51715 104.33875 175.84585 146.24814  82.26987
#> 5   134.27220 121.20938 140.32480 162.49317 232.75439  90.71567 180.61512
#> 6   182.42260  73.75125 171.45863 153.12400 197.47080 122.93745 139.19479
#> 7   112.02620  83.97243 116.84845 170.24234 161.58175 158.74641 168.25103
#> 8   154.05922 108.37108 123.29872 155.12177 154.64705  91.71863 137.83060
#> 9   191.43962 155.11425 124.66809 164.97026 175.25017  89.13084 103.90860
#> 10  162.51317 136.54385 202.77952 152.98009 161.70422 184.47534 112.49699
#> 11  160.70385 113.96405 138.07063 170.87213 121.60274  90.20396  94.94302
#> 12  130.71145 130.59126 166.66976 189.65220 135.09432 119.54243 110.49040
#> 13  157.25270 126.74600 150.67526 146.28074 116.00481 107.39075 113.94753
#> 14  129.16825 140.15867 131.89880 148.62322 125.31665  90.32327 158.75599
#> 15  130.04776 142.90195 126.34573 163.86605 164.24713 160.70472 172.82729
#> 16  110.83407 192.63039 131.82004 209.04779 185.15874 157.82791 150.91566
#> 17  143.81910 103.45999 104.86227 133.85601 167.96522 144.03129 149.05130
#> 18  161.16800 147.99949 173.96789 161.11501 145.31646 121.26840 180.01364
#> 19  170.26443 151.06303 134.44610 119.89553 186.23272 110.88122 144.71628
#> 20  202.64531 183.43188 122.83267 112.58609 161.38800 138.71399 139.55516
#> 21  177.17969 148.13804 111.10134 115.67044 183.03007 124.21640  98.97486
#> 22  197.70536 113.76764 139.37836 173.78481 225.94228 129.17917 115.43184
#> 23  159.35649 136.62816 128.86760 131.97367 153.69343 125.38474 129.49459
#> 24  170.35213 139.59572 122.80756 135.74598 195.35992 109.82741 112.79746
#> 25  126.12485  88.38284 113.81119 105.91441 130.27952 140.03041 146.95528
#> 26  172.82206 109.83869 137.22096 161.02180 149.41982 189.27406 113.78296
#> 27  113.90522 137.88149 147.81455 172.40287 136.35147 141.65972 145.23973
#> 28   92.54529 155.53262 151.34441 133.73888 178.63781 117.42798  96.05809
#> 29  155.75544 164.15586 138.09410 142.94732 131.98186 178.82262 140.54236
#> 30  182.15870 135.23553 146.22230 126.46799 128.50716 129.89901 136.77903
#> 31   80.79017 198.14068 133.74434 198.94514 152.21855 132.93702 132.73219
#> 32  164.70196  89.67926 134.06798 189.10938 121.39354 161.78451 143.30947
#> 33  193.88940 143.53339 163.30162 166.63682 133.45366 145.69226 133.69556
#> 34  188.01294 124.31965 164.71173 152.19064 180.65307 110.69799 196.40477
#> 35  124.54850 120.13265 129.99497 138.90195 188.09495 102.77272 145.51181
#> 36  152.81554 124.80072 161.08358 186.00368 162.49718 144.85117  96.50732
#> 37  144.21897 178.60283 160.79633 146.15502 165.32963 135.03887 238.05714
#> 38  173.78022 141.93200 127.19736 138.48766 117.01815 154.42486  98.44136
#> 39  154.63241 102.23942 162.83943 162.96248 173.57206 171.69349 177.63332
#> 40  169.40568 151.69224 128.51392 107.13966 192.06941 141.82525 102.26880
#> 41  202.24385 134.01011 116.13533 206.48719 134.20365 172.13814 116.15910
#> 42  221.39701 140.53620 126.16369 141.80803 166.70106 171.60845 190.03878
#> 43  153.69228 116.84513 143.62401 154.39323 122.29399  94.03727 119.19699
#> 44  165.02802 157.03554 126.55816 152.04049 107.19849 112.65185 194.71386
#> 45  140.23151 171.77472 136.18817 116.02459 141.40224 160.10428 160.52670
#> 46  127.83916 182.88936 157.00393 134.18376 108.53755 107.70695  79.04285
#> 47  172.01755 154.10410 159.40959 174.47891 155.89924 131.60287 122.20742
#> 48  167.11380 200.02662 111.47618 156.68586 178.75562 113.49992 119.47929
#> 49   94.30864 152.25257 175.38602 201.36047 156.46884 127.49559 112.68009
#> 50  156.80471 173.13859 106.88439  73.54766 152.12752 106.33056 170.12810
#> 51  207.25444 146.67329 123.81705 135.64533 106.21630 151.81394 140.70770
#> 52  176.17148 182.60171 164.09532 213.11873 148.12124 124.94101 164.18924
#> 53  166.39945 171.25630  88.89612 158.39727  89.54296 152.03112 112.50143
#> 54  178.69330 120.33342  90.65724  85.58621 190.25075 106.85204 128.35388
#> 55  166.47058 123.72520 136.47457 182.04811 166.06863 137.97521 138.39965
#> 56  158.64746 124.67992 154.77873 125.57457 192.79054 212.41453 132.24942
#> 57  161.30755 113.50966  88.44551 215.99595 209.44161 155.98833 145.20583
#> 58  196.09169 175.67402 104.84155 227.53325 180.11415 172.52459  96.45258
#> 59  133.80191 199.00747 140.48838 157.88182  99.03458 123.94689 184.65737
#> 60  141.60060 177.90545 157.66201 143.14340 134.18044 148.22020 143.22954
#> 61  110.51024 138.44538 192.58795 166.20364 144.92892 129.99379 111.17115
#> 62  174.79750 149.14117  88.87483 172.20172 146.28877 153.51808 127.46187
#> 63  129.09747 165.82676 199.20367 136.85224 148.90538 123.04572 148.44662
#> 64  146.80820 201.41179 125.29750 161.30969 161.37866 196.38253 139.64575
#> 65  170.37716 143.48625 119.73583 158.98226 144.25506 137.42602 173.90591
#> 66  239.92689 202.36230 146.74127 106.20821 148.47944 141.30435 170.16416
#> 67  186.74261 113.02927  95.53557 109.23493 143.33577 143.38535 166.18750
#> 68  161.20835 111.20644 150.27808 171.76714 139.23630 110.95085 150.88659
#> 69  144.87155 130.85407 176.71493 199.07015 119.29936 131.08190 151.81529
#> 70  178.89870 138.03452 148.15931 135.10934 188.26316 121.34538 155.44467
#> 71  174.69675 169.65935 176.35174 157.53636 125.43215 115.48742 186.38704
#> 72  105.95570 138.68521 105.03840 146.34806 213.63919 143.76377 147.85918
#> 73  157.16575 144.15970 121.71280 176.67725 106.56444 172.62942 145.73916
#> 74  172.55197 157.63419 141.19682 181.80591 121.27196 137.80176 156.25474
#> 75  175.71689 146.98627 172.42892  89.53476 206.86986 160.15696 175.73288
#> 76  131.21193 135.52374 124.37917 107.21604 154.10236 109.54510  84.57205
#> 77  176.93366 155.18111 184.57701 134.40458 168.33219 145.15914 169.50505
#> 78  110.16951 133.89229 129.93471 146.50556 124.92645 113.40888 109.67218
#> 79  121.08050 141.28086  68.65860 199.67217 140.84658 119.89258 141.57853
#> 80  112.94190 107.05751 123.73060 171.63056 155.37817 126.71076 139.03520
#> 81  228.60841 117.61173 153.06489 167.29361 159.22816 131.07218 129.51350
#> 82  169.50960 145.02387 136.38417 228.40822 113.01138 176.27532 129.77870
#> 83  110.65768 115.98231 107.94958 182.26756 108.80210 118.65554 148.96948
#> 84  221.84865 147.35863 183.17179 192.31157 176.39295 148.93081 123.74943
#> 85  141.88625 153.43256 164.20313 159.54795 215.14282 101.42562 164.79256
#> 86  165.06966 146.19184 188.13005 144.35822 165.34711 119.29953 137.45020
#> 87  150.73136 122.81374 181.15280 157.58003 161.98000 122.41200 113.28028
#> 88  111.05393 102.35380 152.03232 198.73829 147.03022 180.51453 125.41826
#> 89  167.77103 156.63373 138.17793 123.05215 170.40764 150.08757 157.56165
#> 90  202.08609 163.68744 138.81040 184.93791 181.93121 136.46047 143.44139
#> 91  148.07251 152.76739 155.96501 120.24141 166.98746 122.56610 158.32633
#> 92  162.98774 133.55104 120.08693  98.08998 191.25429 178.28306 121.65411
#> 93  155.02544 140.66299 169.28660 211.62555 127.64391 129.94145 132.02661
#> 94  154.23625 172.52139 163.56734 168.16118 104.19747 123.25783 122.97629
#> 95  181.33308 157.12706 156.65687 206.39869 110.58124 156.78254 173.37537
#> 96  164.78095 128.21571 145.36552 125.09450 162.52675 136.02554 121.76397
#> 97  149.78405 114.44338 125.40604 149.82407 138.40279 127.64524 118.67558
#> 98  133.50741 143.10244 172.05540 143.22203 130.48388 188.77493 147.19972
#> 99  184.80850 181.30380 164.81562 118.94959 190.30209 157.15875  95.35157
#> 100 188.25806 175.14843 160.01817 115.58181 170.09376 154.11117 163.25055
#>         B.4_1
#> 1   243.98822
#> 2   130.80900
#> 3   170.57259
#> 4   109.50170
#> 5   154.76980
#> 6   103.29263
#> 7   127.11617
#> 8   123.29851
#> 9   168.29731
#> 10  184.08646
#> 11  131.57791
#> 12  147.81937
#> 13  134.32445
#> 14  157.02216
#> 15  179.07395
#> 16  217.12155
#> 17  126.21631
#> 18  116.06203
#> 19   97.46196
#> 20  100.39872
#> 21  143.69583
#> 22  147.40569
#> 23  146.90771
#> 24  164.72427
#> 25  159.45989
#> 26  185.69517
#> 27  204.45009
#> 28  159.52512
#> 29  155.33977
#> 30   93.31028
#> 31  116.04544
#> 32  127.09795
#> 33  120.89568
#> 34  158.51687
#> 35  198.47558
#> 36  135.30539
#> 37  150.00395
#> 38  135.26639
#> 39  171.50514
#> 40  112.29771
#> 41  165.70602
#> 42  108.66645
#> 43  150.60209
#> 44  139.33160
#> 45  152.54503
#> 46  115.91328
#> 47  158.81110
#> 48  149.54920
#> 49  147.81083
#> 50  184.36517
#> 51  154.07954
#> 52  182.76390
#> 53  174.03127
#> 54  105.87214
#> 55  160.80816
#> 56  183.38408
#> 57  124.86541
#> 58  176.88111
#> 59  166.76115
#> 60  119.26342
#> 61  202.82898
#> 62  145.96206
#> 63  178.02670
#> 64  183.03242
#> 65  153.32314
#> 66  146.50415
#> 67  139.99637
#> 68  165.91048
#> 69  122.23424
#> 70  136.93310
#> 71  148.88373
#> 72  144.23128
#> 73  151.75096
#> 74  125.68395
#> 75  178.62230
#> 76  159.64250
#> 77  179.74603
#> 78  162.16261
#> 79  161.25523
#> 80  152.08817
#> 81  144.50540
#> 82  175.97653
#> 83  149.34548
#> 84  142.58100
#> 85  122.03532
#> 86  171.78509
#> 87  121.92289
#> 88  183.67195
#> 89  121.45964
#> 90  127.88783
#> 91  171.55861
#> 92  181.65444
#> 93  188.01225
#> 94  136.15038
#> 95  147.48654
#> 96  168.95773
#> 97  175.78808
#> 98  130.47330
#> 99  164.13864
#> 100 152.93319