Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10, 10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(
  subset = rep(c("A", "B"), 50),
  level = rep(1:4, each = 25)
)
rownames(meta) <- 1:100
scRNA <- Seurat::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA,
  by = c("subset", "level"),
  min.cells = 10, max.cells = 20
)
#>         A.1_1     B.1_1     A.2_1     B.2_1     A.3_1     B.3_1     A.4_1
#> 1   182.43831 162.52661 154.17853 154.96597 150.02795 148.45759 149.15091
#> 2   128.89734 174.26654 107.09954 161.28758 138.50091 123.14072 139.67936
#> 3   139.96907 167.81827 127.94564 179.71801 135.84537 182.36887 166.29358
#> 4   148.90388 115.06305 158.00133 214.05156 165.39047 165.12940 155.89968
#> 5   159.97903 101.31581 137.74506 232.30058 132.85247 128.84512 132.21654
#> 6   174.33281 100.29130 117.44479 222.24639 122.97614 126.97562 142.63736
#> 7   146.89404 158.33753 142.93191 170.21269 163.25029 152.42375 188.46747
#> 8   132.05605 123.27887 138.67488 133.37200 190.05832 154.19851 170.79392
#> 9   153.89160 107.79903 144.14666 206.77833 180.61920 158.48165 143.22691
#> 10  119.64363 114.95025 167.93739 137.64922 178.26437 132.42853 127.60931
#> 11  108.85153 137.35100 183.02043 135.78086 128.23517 181.89133 121.54601
#> 12  172.04847 136.91444 137.82719 117.48502 122.55042 160.14496 167.84246
#> 13  166.30989 123.40712 132.54622 159.54326 135.86699 116.54564 129.50335
#> 14  146.86246 179.94135 112.36695 121.48180 146.41378 108.97307 125.50018
#> 15  181.51979 124.21311 132.97475 187.45725 172.74075 146.26580 117.85934
#> 16  129.73932 113.72823 150.80243 134.67587 140.24745 135.56292 122.04245
#> 17  150.31652 120.69357 135.41601 154.97292 108.03614 156.88441 133.60877
#> 18  160.15122 170.24153 129.66042 177.15941 140.37195 147.90518 131.43117
#> 19  168.53002 135.51386 109.89085 150.11849 162.16057 123.99452 128.86287
#> 20  176.71297  67.32789 155.29966 121.16070 203.91114 137.62005 190.51687
#> 21  138.97064 160.14767 152.37536 176.70877 210.90003 144.62240 178.05801
#> 22  173.75225 158.30802 122.91287 133.73960 129.05261 124.00190 169.59005
#> 23  172.70234 165.62520 132.65936 117.63253 137.91548 129.49690 134.18058
#> 24  195.72322 112.28326 121.71241 161.56420 127.88937 145.29588 175.96213
#> 25  164.23072 139.64594 101.11140 196.32775 176.39595 180.91937 136.99839
#> 26  191.92298 109.75687 127.62621 132.87315 153.22872 143.33040 144.26966
#> 27   89.48172 145.35111 153.02547 138.42790 187.54965 176.76657 168.44817
#> 28  172.16527 133.29454 155.90575 157.03237 154.45015 166.37718 124.94498
#> 29  177.78754 159.27775 105.02676 142.66287 166.60787 150.93496 173.71438
#> 30  181.60901 126.53731 179.11076 113.31393 121.34731 175.31397 173.41283
#> 31  165.91587 152.73251 120.30894 118.58165 159.40927 210.20224 132.94865
#> 32  143.29072 157.03442 129.24927 116.80552  77.51416 139.93661 122.99068
#> 33  112.94789 104.38724 143.42500 102.09064 129.69989 137.28121 134.45637
#> 34  148.01907 195.02129 169.40913 148.93189 117.74344 141.91586 164.18038
#> 35  128.89708 150.12864 139.50636 125.05756 185.93249 164.30091 144.07241
#> 36  228.83280 132.93343 112.64076 152.91165 168.81344 100.30231  94.97529
#> 37  120.31544 170.21761 134.65652 126.72525 191.45698 141.15686 178.42057
#> 38  125.27715 143.74966 142.03677 140.14528 176.33227 131.16750 141.40580
#> 39  126.62908 133.59457 123.81667 132.79587 222.93192 114.41389 172.56021
#> 40  176.34595 149.58308 149.83650 169.00916 104.39956 184.36194 120.12904
#> 41  153.33918 125.71779 146.17580 188.67469 161.83993 177.84091 179.84363
#> 42  130.68449 144.92915 161.81969 131.65081 147.87928 125.58849 129.13427
#> 43  180.80140 196.13557 132.90323 158.09513 177.00874 101.73177  77.24349
#> 44  141.53106 118.08305 156.50205 144.01277 205.12834 202.23358 208.81422
#> 45  160.43945 124.59714 130.30859 130.34974 142.55946 178.22846 108.55800
#> 46  174.24590 142.99627 114.66307 124.30008 160.91335 119.75471 158.70134
#> 47  185.18647 124.16466 169.30412 127.52728 164.37492 147.08206 120.53084
#> 48  176.79936 146.16960 106.88216 166.86736 183.84280 155.54124 109.34897
#> 49  149.23175 113.11939 161.27368 124.68250 186.07527 108.38706 170.41311
#> 50  121.31564  94.78537  95.79413 225.29316 125.18383  97.99105 133.53133
#> 51  178.29457 123.76721 119.72072 165.49326 159.22627 153.86003 210.05569
#> 52  150.14994 138.63172  94.68818 122.15359 145.20725 121.10166 145.84870
#> 53  129.22640 169.08406 141.20515 158.74788 136.87896 114.19580  96.95427
#> 54  176.15772 111.91438 169.86954 137.14230 198.93752 105.25027 149.76848
#> 55  132.58062 183.01317 166.85381 144.38966 151.22879 152.70341 139.96726
#> 56  159.67419 158.17028 122.05996 188.94995 160.00940 161.93540 117.09657
#> 57  125.15703 127.15263 136.26218 148.76738 156.06213 145.98583 192.37735
#> 58  150.51709 169.06034 137.12703 135.48218 137.38242 174.00648 148.07919
#> 59  179.59111 129.84664 159.07679 110.16308 165.55116 125.31109 191.34045
#> 60  168.81277 136.68341 151.62082 185.70756  87.95481 133.18777 129.49938
#> 61  232.29151 164.65349 105.01019 148.58682 123.84767 162.40585 132.55508
#> 62  180.89823 110.68452 138.37940 127.01687 195.05973 172.41446 125.75536
#> 63  161.46524 146.94807 135.05484 147.12518  96.31824 143.26593 109.95745
#> 64  133.25607 198.62533 167.58168 170.03114 169.81731 139.67261 154.64030
#> 65  186.44913 164.76096 136.10008 193.90258 163.09242 181.24934 125.82407
#> 66  193.34888 142.49030 136.40080 154.98862 178.68844 112.70304 154.85421
#> 67  185.31633 213.52593 116.87653 143.46400 168.29906 111.51728  84.15541
#> 68  171.22492 110.56259 145.90405 177.83508 154.38436 117.46901 166.27236
#> 69  169.76802 188.49648 140.47286 145.52805 185.49845 168.70236 144.05351
#> 70  142.87033 164.44249 177.66547 174.29460 142.43904 138.12825 160.95613
#> 71  194.06858 178.51161 105.65788 167.96117 159.65469 145.04301 141.65389
#> 72  152.51890 165.00511 205.37866 118.52577 155.66278 115.82992 154.06111
#> 73  148.90905 114.15351 113.20967  98.56970 157.82410 156.50129 126.68202
#> 74  129.31737 110.50523 129.19815 118.84988 173.36412 110.55193 145.01565
#> 75  149.55402 119.87496  87.14139 204.16572 123.46939 167.72262 206.22661
#> 76  152.96058 205.40008 191.01344 164.81406 167.99809 125.87799 168.51813
#> 77  181.31691 168.53230 141.93658 169.06236 173.79489 153.60606 124.56771
#> 78   90.40946 141.66768 136.99792 160.70475 171.12396 155.84303 151.56141
#> 79  156.54532 166.26832 131.13935 182.97312 214.55086 159.45639 109.58181
#> 80  157.00128 100.67444 167.21229 159.87899 180.31840 181.26311 139.45541
#> 81  107.55120 173.44073 135.90413 181.66131 175.63312 180.83509 118.67683
#> 82  124.17917 168.00888 199.12215 188.68967 101.77991 168.87732 144.96836
#> 83  175.50526 145.92471 125.21635 161.86731 124.91472 165.97607 185.24677
#> 84  245.21036 138.69812 196.30925 175.29805 195.74814 137.03591 142.87378
#> 85  160.76262 162.85218 108.45548 131.55308 142.65327 138.22964 121.94933
#> 86  146.91645 139.41832 128.89679 147.03676 159.77373 162.07536 131.36835
#> 87  141.54776 124.78600  69.50503 150.01340 119.94735 135.65368 140.54575
#> 88  178.61809 157.74586 119.87786 148.69336 190.63342 162.72755 128.76336
#> 89  184.88107 143.08828 157.83027 205.48161 145.05659 130.99710 126.59278
#> 90  143.44377 163.82754 104.90233  86.07452 183.28483 169.70053 159.93905
#> 91  158.37376 172.10001 115.04347 224.01152 197.29680 138.74926 146.23270
#> 92  114.66577 105.55405  97.69176 144.59320 150.15648 185.25165 107.06711
#> 93  117.19287 128.57714 163.66473 148.27228 181.12696 136.38692 177.84370
#> 94  184.68588 164.63004 141.00722 145.01847 186.37204 142.19573 163.27844
#> 95  142.38877  90.52766 151.63261 113.74750 166.03465 133.57139 123.98100
#> 96  162.01827 127.07372 154.09225 130.72715 132.63128 147.16360 109.51090
#> 97  166.29700 108.13166 120.56496 151.57084 137.45091 134.84356  96.38399
#> 98  159.01157 129.24977 129.81173 174.10446 197.07889 116.38334 147.42339
#> 99  185.15542 170.55374 145.77020 160.65281 160.55285 157.05310 140.71408
#> 100 131.63682 154.37898 139.01918 129.28842 167.10339  94.57228 165.23521
#>         B.4_1
#> 1   155.38315
#> 2   162.12650
#> 3   174.54078
#> 4   178.91108
#> 5   170.18749
#> 6   185.11696
#> 7   160.25056
#> 8   172.13364
#> 9   169.26771
#> 10  131.09969
#> 11  150.69667
#> 12  199.17441
#> 13  162.86837
#> 14  126.17372
#> 15  151.57141
#> 16  214.93241
#> 17  171.26686
#> 18  115.96740
#> 19  141.75249
#> 20  130.80154
#> 21  188.15648
#> 22  188.32944
#> 23  148.49976
#> 24  128.71577
#> 25  145.21363
#> 26  145.46508
#> 27  187.51957
#> 28  131.27465
#> 29  140.76531
#> 30  176.21275
#> 31  120.26330
#> 32  150.92699
#> 33  152.50294
#> 34  132.72472
#> 35  169.77224
#> 36  122.69171
#> 37  129.93860
#> 38  181.12902
#> 39  166.73250
#> 40  138.53606
#> 41  182.76694
#> 42  122.46806
#> 43  127.65157
#> 44  128.28620
#> 45  145.95051
#> 46  216.02731
#> 47  141.36833
#> 48  142.21997
#> 49  162.73859
#> 50  163.20809
#> 51  139.31932
#> 52  130.41913
#> 53  142.29398
#> 54  196.53325
#> 55  199.20303
#> 56  180.49976
#> 57  170.98726
#> 58  150.90350
#> 59  157.99821
#> 60  126.88996
#> 61  122.54766
#> 62  168.63303
#> 63  189.71645
#> 64  209.92213
#> 65  148.26704
#> 66  184.37090
#> 67  118.51559
#> 68  102.55062
#> 69  121.25008
#> 70  151.21303
#> 71  148.30331
#> 72  141.57317
#> 73  181.22858
#> 74  128.71880
#> 75  144.03576
#> 76  182.39334
#> 77  168.12604
#> 78  193.40503
#> 79   97.01139
#> 80  130.86695
#> 81  211.78734
#> 82  202.72657
#> 83  108.76422
#> 84  171.09772
#> 85  158.22647
#> 86  146.86739
#> 87   90.56526
#> 88  176.73041
#> 89  182.95669
#> 90  162.06515
#> 91  145.06219
#> 92   79.47590
#> 93  163.09718
#> 94  129.60050
#> 95  164.09047
#> 96  154.24853
#> 97  204.35791
#> 98  121.32181
#> 99   80.92936
#> 100 142.36191