Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10, 10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(
  subset = rep(c("A", "B"), 50),
  level = rep(1:4, each = 25)
)
rownames(meta) <- 1:100
scRNA <- Seurat::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA,
  by = c("subset", "level"),
  min.cells = 10, max.cells = 20
)
#>         A.1_1     B.1_1     A.2_1     B.2_1     A.3_1     B.3_1     A.4_1
#> 1   169.83170 202.05337 123.67196 135.97192 137.04708 133.12109 139.98446
#> 2   285.20559 149.97106 137.98633 205.42888 141.87582 137.71911 128.02595
#> 3   146.11853 127.83879 110.29031 154.23819 153.49874 144.66914 143.22288
#> 4   182.48922 145.18052 115.88898 136.54824 131.18827 111.93695 123.48946
#> 5   141.16358 119.17325 153.68369 149.35009 154.61126 157.51603 130.63469
#> 6   149.46084 129.64080 135.86175 130.88999 141.46590 137.60799 125.17462
#> 7   141.81469 118.33862 119.15444 130.09224 183.23326 164.21223 151.24002
#> 8   153.35591 161.00695 147.05379 128.78436 111.26629 175.29670 119.48691
#> 9   140.35245 108.93230 110.82279 164.22500 141.90976  96.14303 126.43790
#> 10  100.41816 135.16731 137.20307 163.69217 154.81128 186.59540 131.05829
#> 11  124.23570 151.86522 125.74260 180.56120 144.66339 146.73359 151.73819
#> 12  193.53040 144.00014 180.64137 174.59895 136.35617 119.13702 119.91636
#> 13  240.64879 134.43885 121.90817 161.07737  99.54734 121.26127 116.15207
#> 14  159.98565 161.14614 185.76284 183.70633 110.96293 156.91062 153.36796
#> 15  152.66753 140.13807 134.35602 127.43567 151.39928 139.19329 175.09814
#> 16  133.66666 141.67791 171.19228 143.53883 165.94829 143.08432 178.09446
#> 17  122.76049 131.18101 119.14028 157.97425 143.93637 110.01847 135.53144
#> 18  148.42321 119.64462 183.32486 146.26833 119.88202 146.37184 173.47265
#> 19  183.54458 183.69495 109.47888 134.58390 158.43513 158.81705 164.24416
#> 20  187.06632 111.28002 150.26051 193.32189 172.14017 169.16448 161.59851
#> 21  113.09201 129.27180 138.49731 141.65775 217.10890 175.03987 185.03267
#> 22  201.12370 162.46867 123.62052 122.32148 121.72670 150.95541 185.31510
#> 23  137.58652 113.82645 108.97471 143.99437 137.10839  82.70143 145.69971
#> 24  131.57637 203.68960 101.50875 161.59571 128.92398 189.93620 168.02010
#> 25  195.49488  94.49108 139.11358 125.79880 111.69661 169.63841 125.83292
#> 26  157.53235 136.91827 152.54822 194.99142 172.84579 119.59760 117.39309
#> 27  195.75149 124.58363 122.71365 167.57531 150.93746 176.81385 176.45076
#> 28  135.00755 126.36920 118.81010 131.68714 129.66681 156.26981 187.34180
#> 29  132.30014 147.30736 194.12584 127.50788 172.33773 141.98658  93.54530
#> 30  175.69666 169.89755 114.74440 133.09548 149.46302 205.80719  87.37006
#> 31  145.96674 174.96381 123.12149 152.94179  98.68889 146.11182 163.14767
#> 32  120.79085  83.40776 128.28246 163.14572 154.73141 149.80012 161.44365
#> 33  174.42796 157.61609 247.65805 142.74417 194.59534  85.35984 114.14682
#> 34  148.63834 155.05459 169.40324 148.64515 160.42017 120.88272 103.33645
#> 35  179.52763 150.01731 158.59252 135.11070 179.18055 134.23995 154.58391
#> 36  131.93597 167.09129 118.48881 161.57356 147.79335 150.76480 127.18731
#> 37  117.88046  86.09920 159.17234 182.30327 143.23722 128.42076 145.46251
#> 38  185.28223 154.47965 146.97648  92.42366 141.89716 146.70436 110.28656
#> 39  112.81543 137.33919 134.94680 145.55833 132.19917 110.92562 127.32349
#> 40  106.96513 111.51756 119.15650 166.68929 134.63836 137.82687 140.70452
#> 41  139.80637 199.11285 167.81882 138.65231 146.14083 146.93230 140.35263
#> 42  176.11022  82.35731 111.55615 128.13729 136.72547 186.58143 168.92036
#> 43  129.76165 160.59169 190.71282 147.06142 144.33249  97.27162 111.49388
#> 44  162.18295 126.77328  92.23375 137.41468 175.08751 145.36649 132.57106
#> 45  173.17028 143.38958 116.92034 129.11418 134.82450 131.61753 115.62037
#> 46  107.68062 108.00087 160.27279 144.33700 204.06395 133.50630 143.81936
#> 47  162.20698 112.02125 133.72118 153.57309 221.37946 177.45426 104.42347
#> 48  192.91663 134.31141 155.83846 146.27667 173.75232 137.59192  98.77643
#> 49  202.28516 171.95454 128.51198 171.02643 131.80635  98.45877 136.42093
#> 50  158.23347 132.91881  81.41951 146.66363 123.06809 158.55267 107.75663
#> 51  137.50782 172.18156  97.15081 118.74096 149.89102 114.54869  91.99114
#> 52  101.45064  91.58334 132.19368 115.45761 156.60093 135.69728 137.62559
#> 53  158.05541 178.61810 160.65872 139.17233 174.44602 128.07000 181.44948
#> 54  160.34856 152.06492 183.33374 169.67941 168.73816 127.23718 143.36339
#> 55  168.01078 110.14148 197.08101 171.19163 113.27918 121.91635 162.24483
#> 56  142.14372 143.07308 172.07238 138.54186 134.29370 132.11073 158.24053
#> 57  162.12993 154.87534 169.24875 164.78860 166.80339  83.53420 154.76926
#> 58  139.33850 144.64904 141.15753 212.63500 145.02200 165.96739 169.39394
#> 59  118.34709 187.95020 143.00322 161.13105 135.67252 121.06703 189.06295
#> 60  194.85060 146.47264 169.43000 152.43219 132.31825 129.84424 136.46402
#> 61  142.46678 198.59917 170.96384 165.63329 145.80223 125.90154 128.62828
#> 62  153.10019  98.67938 127.83852 130.11842 198.97098 155.48863 178.14181
#> 63  143.66799 146.56918 110.51015 130.01672 122.88554 155.14519 171.73345
#> 64  198.95724 166.06684 132.76011 126.16741  94.07290 116.34575 133.80355
#> 65  154.40845 133.96900 138.51794 138.56793 138.21209 157.99239 154.13803
#> 66  192.52840 165.01650 107.84276 195.23522 139.67411 100.47122 138.92627
#> 67  152.23056 143.71268 185.00355 142.34777  97.14608 182.66915 132.58222
#> 68  134.30258 134.61702 211.01278 191.38011 157.04459 127.44359 156.35064
#> 69  188.94315 158.76075 153.38291 149.54773 150.28978 147.84961 153.99147
#> 70  158.94873 131.70309 166.01365 158.29932 136.15617 151.39786 128.42795
#> 71  145.54784 197.67073 139.85463 182.54123 195.72120 176.20158 160.35949
#> 72   94.75153 152.18667 159.86228 191.51020 159.86364 137.72010 121.52549
#> 73  113.63924 154.81871 208.34524 141.32539 138.80689 137.25864 132.28689
#> 74  167.79675 129.19266 113.65969 190.88245 120.93278 151.49531 118.53463
#> 75   85.25464 120.95199 112.99672 141.09050 152.69111 167.84497 153.53516
#> 76  149.68127 130.44776 129.22130 182.31212 176.07004 165.35557 133.07522
#> 77  153.69965 126.46530 145.41191 145.40365  87.38329 124.92794 174.90971
#> 78  136.71178 130.89802 182.27408 176.07621 104.04324  94.83387 118.24978
#> 79  149.91376 123.46639 148.82793 146.48568 140.12002 187.03423 154.48788
#> 80  147.12826 140.16207 128.13514 140.71582 143.58394 177.20416 181.35972
#> 81  103.59168 165.46792 101.42613 150.99246 131.39735 153.64311 141.88146
#> 82  178.73935 145.27925 149.99526 113.73643 119.36117 158.73849 124.83176
#> 83  148.60364 154.23724 173.92026 185.16508 155.94892 172.60091 146.50918
#> 84  126.47614 157.89786 161.96653 139.63148 142.51852 199.75062 147.30157
#> 85  143.52596 157.90981 119.13843 190.96635 144.57103 151.56871 113.65739
#> 86  179.41140 129.73613 110.09784 180.29510 174.61128 170.54006 139.57363
#> 87  123.34591  66.12182 156.38726 199.37120 138.99292 174.96283 145.30490
#> 88  169.38290 112.48400 173.60620 148.05294 154.18487 178.01719 163.80998
#> 89  189.20959 151.70109 169.47850 152.19533 209.99983 124.01319 186.77108
#> 90  145.58262 157.00290 131.84354 164.83773 148.39531 126.12723 132.46597
#> 91  187.80085 193.34292 126.56698 121.68437 121.53558 132.59306 212.07093
#> 92  141.92150 183.61005 124.14791 152.02604 201.07767 118.21084 114.53143
#> 93  194.70308 183.58477  90.76850 175.12415 161.98134 133.12367 169.74183
#> 94  159.04707 184.53858 120.43732 141.31880 170.27057 116.05084 159.68713
#> 95  182.73283 130.44758  91.66977 179.60271 109.99104 180.57691 132.26804
#> 96  157.70129 134.38899 152.78706 144.91726 174.98564 108.09408 152.65779
#> 97  114.37900 104.05491 143.45992 121.99371 142.11182 180.75488 109.67846
#> 98  153.56614 134.07593 177.98397 155.11534 104.90662 149.88721 115.35294
#> 99  231.14324 136.95059 112.21583 147.18993 179.82211 149.07156 177.63022
#> 100 125.69420 135.46760  93.85458 151.61976 110.09981 202.83390 139.00583
#>         B.4_1
#> 1   157.20407
#> 2   138.30461
#> 3   143.85659
#> 4   180.78314
#> 5   110.62496
#> 6    94.82386
#> 7   178.76889
#> 8   183.36357
#> 9   165.19732
#> 10  137.78496
#> 11  139.97520
#> 12  117.17469
#> 13  184.98775
#> 14  103.96398
#> 15  160.51371
#> 16  140.21682
#> 17  227.20999
#> 18  186.19938
#> 19  171.22179
#> 20  109.13067
#> 21  100.70020
#> 22  127.24602
#> 23  165.17071
#> 24  125.74642
#> 25  137.73592
#> 26  213.78472
#> 27  135.63416
#> 28  129.11945
#> 29  177.37729
#> 30  145.45005
#> 31  164.36800
#> 32  185.34282
#> 33  169.36173
#> 34   99.41724
#> 35  154.04617
#> 36  138.46896
#> 37  139.20307
#> 38  102.57720
#> 39  135.66737
#> 40  182.63186
#> 41  173.86910
#> 42  127.89519
#> 43  152.25531
#> 44  176.50841
#> 45  133.86671
#> 46  100.31442
#> 47  152.78653
#> 48  155.11846
#> 49  153.66746
#> 50  161.79975
#> 51  162.10376
#> 52  151.04215
#> 53  144.19488
#> 54  120.74818
#> 55  157.38398
#> 56  173.32721
#> 57  134.56751
#> 58  120.65223
#> 59  163.73918
#> 60  145.21903
#> 61  167.99804
#> 62  208.36741
#> 63  151.60934
#> 64   98.55537
#> 65  133.83344
#> 66  146.87581
#> 67  195.77155
#> 68  213.48584
#> 69  211.53165
#> 70  141.60817
#> 71  119.75864
#> 72  148.49883
#> 73  108.36102
#> 74  174.55507
#> 75  132.22339
#> 76  106.25092
#> 77  205.00211
#> 78  135.61338
#> 79  187.90559
#> 80  181.51935
#> 81  159.36941
#> 82  169.17005
#> 83  196.71106
#> 84  167.62442
#> 85   93.70239
#> 86  144.60551
#> 87  130.42506
#> 88  217.24007
#> 89  214.16536
#> 90  148.22525
#> 91  161.34247
#> 92  138.38214
#> 93  106.73559
#> 94  160.37508
#> 95  176.13521
#> 96  130.37761
#> 97  162.33827
#> 98  176.21448
#> 99  125.81270
#> 100 151.48973