Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10, 10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(
  subset = rep(c("A", "B"), 50),
  level = rep(1:4, each = 25)
)
rownames(meta) <- 1:100
scRNA <- SeuratObject::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA,
  by = c("subset", "level"),
  min.cells = 10, max.cells = 20
)
#>         A.1_1     B.1_1     A.2_1    B.2_1    A.3_1     B.3_1     A.4_1
#> 1   181.14746 105.07530 126.61875 119.3512 180.3382 171.51794 145.07821
#> 2   122.47230 173.67045 136.90259 177.0181 147.4167  90.47809 203.19001
#> 3   115.55930 153.01611 124.08099 170.6420 124.0805 146.89157 141.00397
#> 4   160.40834 117.70071 162.10841 172.9255 139.0211 136.44689 152.83763
#> 5   120.70950 168.36952 130.88356 140.5478 141.5294 154.30075 204.19108
#> 6   186.43387 169.62835 178.02382 146.8219 191.2936 131.60686 112.24016
#> 7   152.35513 162.58465 135.07713 169.8512 208.9635 141.76008 131.19852
#> 8   161.97602 122.21584 103.50650 137.4910 119.9967 104.61761 172.72050
#> 9    80.42474 136.07640 148.28787 137.5032 163.6800 140.51269 129.04427
#> 10  168.67593 102.72665 149.63542 165.7253 162.4765 151.03236 136.94076
#> 11  219.06798 138.58683 102.28540 177.2818 133.8256 121.85332 118.30827
#> 12  249.06443 156.93271 120.19379 102.2078 160.1250 121.87677 171.41495
#> 13  188.41154 116.14537 129.57210 135.3419 118.1957 179.70867 138.83708
#> 14  146.38994 129.18399 135.07534 170.8667 157.4211 230.71744  93.20550
#> 15  177.93293 137.32619 180.21599 142.7211 149.1658 174.84670 175.43866
#> 16  136.53552 127.94939 122.15267 151.6516 173.2103 146.80026 127.05021
#> 17  154.66087  99.72864 154.10392 142.6085 106.5474 144.20885 146.74123
#> 18  175.74042 144.85443 158.79183 157.2291 126.7316 130.18455 160.93842
#> 19  122.47398 149.93449 121.82117 114.0221 173.8808 111.91811 149.11152
#> 20  142.85591 192.67434 165.13853 128.4569 165.5542 113.49820 181.05722
#> 21  142.05447 101.83176 104.21740 130.9421 113.7110 128.66293 147.28934
#> 22  175.06478  95.47981 117.77421 161.7438 165.5732 127.32046 157.24653
#> 23  120.21916 154.36074 200.97659 226.7038 176.9725 134.97668 131.52180
#> 24  158.22818 154.48072 150.78326 136.2622 159.5603 162.17265 194.69152
#> 25  180.96686 141.69856 163.58403 147.6680 148.3820 143.32597 176.27751
#> 26  215.99021 156.23390 166.96281 151.8504 148.3088 144.24137 128.37413
#> 27  122.30037 144.43872 128.34666 143.9118 159.1765 146.27567 150.76337
#> 28  143.27588 131.98212 125.21802 206.8542 143.6937 130.87079 123.37631
#> 29  131.21184 136.56804 110.96442 100.3668 141.6105 135.21072 129.46478
#> 30  147.74521 127.18147 110.41222 189.2805 171.3259 148.67067 109.94472
#> 31  128.47970 180.10631 126.06031 179.0670 134.9108 155.31412 131.17684
#> 32  184.02081 145.60687 155.48352 138.7140 142.3103 170.91345 138.21599
#> 33  151.81852 145.08981 132.87860 122.8805 154.4344 127.44554  80.95817
#> 34  171.62362 104.40718 124.05460 176.0130 175.9057 132.76319 146.84896
#> 35  166.20779 152.00524 166.74531 226.0207 141.0428 124.32653 156.64030
#> 36  132.23370 146.07659 167.53948 135.5901 127.7218 119.79300 138.06440
#> 37  120.70328 164.65267 161.91177 145.2270 161.0318 199.89121 165.11870
#> 38  171.33139 132.62788 142.00022 202.3638 158.6923 178.66189 201.22377
#> 39  134.41243 184.52687 156.42439 184.1869 120.2818 160.24223 154.79818
#> 40  152.52310  97.34848 143.10393 142.6165 176.3245 116.08697 126.49490
#> 41  175.59047 126.66789 138.05419 154.8294 150.9492 114.81324 112.75493
#> 42  128.32919 141.89426 113.11786 175.1426 145.4062 182.98110 149.59544
#> 43  190.99916 166.12095 157.87148 157.9924 157.2308 168.76987 217.09201
#> 44  182.31178 155.38772 171.28036 130.3541 151.4247 146.23311 111.51631
#> 45  183.88499 166.11739 159.13302 153.4731 146.9303 173.04026 124.40183
#> 46  152.67424 168.15417 111.67965 166.4287 158.4964  85.71103 117.39734
#> 47  179.69438 145.09524 113.82819 137.0747 134.0436 120.58965 182.97605
#> 48  149.08381  71.54665 196.34286 194.2250 155.6183 192.44769 150.06018
#> 49  142.44179  92.91134 138.40221 146.8233 158.0926 172.91750 144.87742
#> 50  139.89810 175.93537 122.99354 153.1657 153.7498 143.35574 192.32627
#> 51  130.10890 155.47788 122.74545 208.1084 140.6892 125.62955 128.46341
#> 52  106.75968 124.39697 121.04203 129.5148 108.7929 160.96142 118.31083
#> 53  157.83929 162.59867 139.89784 118.0368 123.2412 135.40374  90.17036
#> 54  146.21865 190.79833 163.77307 137.2084 167.5379 130.47943 133.38043
#> 55  194.35312 118.56571 183.82945 131.7355 117.8127 118.99624 196.86068
#> 56  130.72195 150.42332 196.10911 155.1485 159.6847 101.98376 165.45726
#> 57  128.73050 135.95575 135.31367 137.0301 153.1569 105.03098 173.07044
#> 58  137.88153 164.90125 116.88005 157.2044 183.0770 127.75713 148.50298
#> 59  160.68185 120.94648  95.00419 119.1093 161.9849 148.07507 173.82822
#> 60  142.84208 145.64688 143.26116 138.9764 109.8096 155.68751  86.62672
#> 61  118.10605 136.58942 143.74546 205.3641 127.1136 146.30610 116.56859
#> 62  162.81598 115.92968 127.84265 220.0789 134.4923 152.99674 129.53268
#> 63  174.82044 111.53343 103.28672 144.8867 172.4770 168.27092 119.29571
#> 64  139.64747 114.54525 113.21230 134.5393 132.2608 153.80589 185.73307
#> 65  153.24594 117.93821 145.24029 108.5831 132.1402 148.23738 161.88723
#> 66  221.01836 117.76880 160.83649 162.5125 123.8022 144.67210 150.58779
#> 67  120.27410 157.26482 162.11762 147.2522 153.3191 125.36403  96.83576
#> 68  129.02327  90.73897 137.46712 142.3544 161.0482 146.60558  66.70320
#> 69  203.68609  85.45364 111.00745 137.1974 156.4813 126.45190 147.89342
#> 70  180.78865 124.46896 112.28472 161.2842 126.9140 123.80732 149.41465
#> 71  170.13841 158.32624 146.54677 122.0747 154.2886 204.32336 106.36254
#> 72  125.59903 117.91844 114.49939 121.5053 115.3641 145.44545 173.72891
#> 73  173.73836  77.38775 116.52526 136.2417 182.3149 167.99505 156.35706
#> 74  193.12746 164.38574 157.41612 178.6032 111.7019 141.60652 146.44563
#> 75  201.78920 121.63291 150.68971 137.2925 161.1100 112.58551 147.88422
#> 76  153.57615 192.23152 148.74635 163.8129 150.3750 158.49227 133.97677
#> 77  116.80883 154.88287 136.88939 192.9735 141.1276 146.86057 174.30292
#> 78  111.44942 133.80822 125.74238 141.7110 160.2075 163.08639 118.14476
#> 79  128.87880 130.64042 164.83234 150.8117 113.2529 125.37535 140.00292
#> 80  156.21797 123.37586 169.97935 194.4803 151.6825 137.87251 123.94132
#> 81  128.10956 109.66842 186.02993 180.7234 158.2216 115.60604 115.02688
#> 82  127.56095  97.36418 120.48579 152.3245 192.0845 114.02755 163.16460
#> 83  141.71641 180.74587  74.52776 148.3064 121.1847 190.61397 128.19311
#> 84  175.37414 127.87092 157.21837 131.5514 154.5239 131.12751 117.52852
#> 85  231.14315 151.86418 135.17099 131.5388 148.8799 134.28181 152.71382
#> 86  120.96058 153.82781 154.12905 183.8225 193.3381 108.63320 180.20590
#> 87  157.08073 131.35698 136.34653 124.9639 129.2040 176.12919 243.77944
#> 88   87.08359 123.18982 148.76904 158.3480  84.7561 129.28973 124.13103
#> 89  169.65798 131.79250 125.60083 141.9965 130.8801 172.80610 104.73489
#> 90  140.47852 133.37608 153.94240 128.7291 220.6473 159.23805 104.88647
#> 91  120.69462 156.52474 129.76000 141.9108 148.8587 113.11277 180.14599
#> 92  188.98148 195.39213 152.27857 179.0003 154.7709 114.54212 137.73462
#> 93   87.64934 176.69355 124.55712 134.0977 142.5377 194.95369 122.29151
#> 94  145.63430 178.19550 133.62532 129.6545 141.4159 103.26453 184.46206
#> 95  144.94862 102.75841  98.25647 154.8082 137.1393 187.90042 145.02598
#> 96  154.22452 122.88197 141.23438 159.0769 153.4987 168.72600 149.26590
#> 97  135.92230 127.60899 135.26513 179.4727 181.9581 122.42294 126.26071
#> 98   78.44540 141.28557 149.20567 177.7194 124.9270 142.99157 177.94004
#> 99  112.00482 119.42751 164.53926 137.6404 174.6346 142.24863 131.04995
#> 100 159.88239 131.28521 151.18256 160.1292 145.0716 151.40699 118.11954
#>         B.4_1
#> 1   150.68377
#> 2   153.86463
#> 3   203.44052
#> 4   128.62444
#> 5   138.08287
#> 6   168.13374
#> 7   150.87381
#> 8   141.60709
#> 9   231.88248
#> 10  189.86504
#> 11  209.54156
#> 12  148.78752
#> 13   86.49721
#> 14  138.20424
#> 15  134.74835
#> 16  124.29723
#> 17  209.00166
#> 18  149.10297
#> 19  137.83071
#> 20  154.53384
#> 21  176.71131
#> 22  219.77429
#> 23  122.13786
#> 24  131.03804
#> 25  152.60436
#> 26  241.54491
#> 27  113.34945
#> 28  228.38418
#> 29  189.39078
#> 30  159.53495
#> 31  144.97074
#> 32  168.37332
#> 33  154.54346
#> 34  163.11809
#> 35  182.00609
#> 36  169.17432
#> 37  115.10803
#> 38  132.53996
#> 39  167.14072
#> 40  159.46064
#> 41  174.59616
#> 42  127.78686
#> 43  156.35492
#> 44  202.15418
#> 45  239.53751
#> 46  188.88569
#> 47  126.13121
#> 48  169.20752
#> 49  171.40759
#> 50  133.44076
#> 51  218.59349
#> 52  152.29144
#> 53   99.53327
#> 54  164.53671
#> 55  121.13916
#> 56  202.08040
#> 57  150.58125
#> 58   92.09995
#> 59  174.89335
#> 60   83.02980
#> 61  155.89952
#> 62  145.29050
#> 63  120.38848
#> 64  175.05194
#> 65  170.80900
#> 66   89.79286
#> 67  105.08080
#> 68  103.25053
#> 69  169.80566
#> 70  116.78908
#> 71  135.21670
#> 72  147.01250
#> 73  144.09576
#> 74  170.95071
#> 75  176.30139
#> 76  157.78179
#> 77  122.19907
#> 78  158.06033
#> 79  161.20923
#> 80  153.37834
#> 81  164.50238
#> 82  171.68885
#> 83  218.76264
#> 84  175.97790
#> 85  137.61178
#> 86  161.78633
#> 87  187.30698
#> 88  120.71954
#> 89  158.40686
#> 90  141.24758
#> 91  155.10673
#> 92  159.62990
#> 93  146.33308
#> 94  165.51885
#> 95  165.09917
#> 96  134.73263
#> 97   70.96066
#> 98  167.78036
#> 99  112.28570
#> 100 118.73269