Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10, 10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(
  subset = rep(c("A", "B"), 50),
  level = rep(1:4, each = 25)
)
rownames(meta) <- 1:100
scRNA <- SeuratObject::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA,
  by = c("subset", "level"),
  min.cells = 10, max.cells = 20
)
#>         A.1_1     B.1_1     A.2_1     B.2_1     A.3_1     B.3_1     A.4_1
#> 1   154.92690 160.36871 147.72722  85.35806 124.48767 152.10422 139.17236
#> 2   151.00254 101.80123 172.39395 159.60907 129.98954 127.05772 183.35059
#> 3   112.10255 167.91411 154.15277 137.98857 196.88253 141.08554 144.77842
#> 4   183.21008 168.22837 139.70838 160.75529 138.50292 162.92033 179.45897
#> 5   126.80809 168.69941 165.12047 101.67218 159.86256 112.19256 168.07665
#> 6   112.71099 128.90249 147.57687 181.99072 144.37292 147.08317 187.93430
#> 7   143.77722 164.61864 129.42922 110.26267 144.20422 161.45337 135.55009
#> 8   136.07800  93.56794 139.87708 170.69203 138.78445 148.22044 151.35538
#> 9   168.10585 131.40704  93.74876 135.06120 158.25185 176.64358 147.29494
#> 10  189.74134 159.56976 181.85126 151.60956 172.70370 157.37848 167.58454
#> 11  183.24940 161.51515 148.81729 140.90695 189.78651 100.21325 156.35665
#> 12  181.03923 133.49542 126.04321 138.59305 181.21252 117.25534 128.03435
#> 13  135.54595  71.73506 109.93429 152.08529 195.90114 145.50363 249.26789
#> 14  135.99888 151.32704  93.66635  96.55966 130.86673 119.99667 145.29475
#> 15  158.36056 152.02366 105.46077 121.70664 132.39104 138.25242 137.01459
#> 16  194.17256 166.47409 122.16153 196.82510 136.35151 157.01622 140.89264
#> 17  122.03217 147.50352 119.57741 169.86820 194.41241 169.19300 145.18906
#> 18  159.13080 138.76096 138.32870 142.08503 112.58933  67.29347 169.45213
#> 19  144.81719 157.28507 134.84081 146.59582 144.09294 183.89790 130.04731
#> 20  149.55564 130.04640 163.02287 172.13034 129.85186 142.68559 151.22554
#> 21  160.82167 128.76232 120.26890  85.80767 144.27224 117.47642 175.15557
#> 22  108.73923 164.94822 163.97696  97.97569 148.59783 101.46721 130.62446
#> 23  182.69619 165.57343 116.76795 149.96305 155.82098 161.26657 139.16146
#> 24  156.70426 137.99425 165.76897 159.92393 146.53037 119.22233 139.85685
#> 25  140.56732 181.15459 120.55211 145.32742 213.19282 155.24494 139.09270
#> 26  161.07689 126.63582  97.18772 129.92828 155.56183 127.61950 115.15327
#> 27  187.53828 147.23091 125.77956 165.29441 160.89202 161.83887 166.81507
#> 28  130.25033 150.30356 168.10575 156.75623 145.00909 156.69236 139.54895
#> 29  153.48273 148.57992 132.04009 185.40616 147.73214 114.35539 159.73822
#> 30  130.92543 152.28686 177.95105 131.11111 133.96961 117.56410 187.39634
#> 31  188.32829 173.55333 127.38457 224.95056 207.63819 127.30315 158.03301
#> 32  175.09176 196.91935 130.34605 205.16857 152.71638 145.11877 118.18796
#> 33  169.63961 181.47194 117.59401 136.79077 102.11579 127.18676 133.80054
#> 34  165.74456 107.52911 106.84637 181.02351 163.42861 161.71763 148.00910
#> 35  144.64842 155.37407 159.65510 124.68721 117.13936 120.90383 152.48369
#> 36  163.23215 164.05728 183.82223 140.14233 144.80401 159.29370 128.02525
#> 37  137.71298 164.27961 132.28404 184.43648 188.99344 128.75598 161.51240
#> 38  103.84708 188.30355 133.84314 197.41943 190.67722  89.26952  95.68111
#> 39  146.70749 152.41765 209.80515 129.32134 124.00869 158.43523 157.50059
#> 40  205.03096 110.05482 124.34473 159.83071 189.34126 132.17398 133.23436
#> 41  127.02948 168.53443 142.43525 169.04955 163.16240 150.00599 211.94140
#> 42  144.81166 131.78637 146.74360 155.52811 136.16233 139.20207 119.12201
#> 43  159.72134 138.42129 129.00148 167.67870 115.63441 185.36731 116.25086
#> 44  157.34937 138.58046 157.72760 135.33647 155.67825 130.56178 114.97763
#> 45  153.91726 140.48989 129.75432 179.59066 202.66139 107.11554 147.04158
#> 46   97.86381 138.97129 162.00379 141.36900 173.73135 135.47129 151.10662
#> 47  167.20335 128.92051 145.17043 186.55966 151.73232  98.10622 160.69811
#> 48  149.17827 120.00368 118.29237 167.07445 161.72860  93.20663 128.96756
#> 49  177.13164 141.57371 131.25779 152.04765 155.86744  98.82659 170.36947
#> 50  122.44228 126.54817 122.77628 136.07710 207.44063 124.05894 133.02402
#> 51  133.10391 167.30130 123.98337 134.60640 156.33636 130.66674 159.26567
#> 52  130.95578 113.20520 115.85175 128.38658 148.68292 141.30866 114.73562
#> 53  172.37452 140.83509 159.75241 170.85084 180.38965 102.27361 172.20257
#> 54  194.17185 171.29020 129.46261 123.45741  97.42076 135.32850 109.27506
#> 55  190.92239 165.19891 127.77255 119.09947 126.52614 130.18306 136.53156
#> 56  170.88797 107.35960  93.58596 158.49763 183.52388 140.91010 172.10556
#> 57  135.07311 136.65069 151.40004 127.02391 168.58500 207.80005 122.76645
#> 58  127.17510  99.88100 161.38648 139.55367 132.79561 142.92322 137.87787
#> 59  179.55734 126.70345 145.91180 209.50890 114.65262 118.46673 156.00406
#> 60  153.45201 107.90826 192.88077 129.04855 169.97864 161.09516 120.57089
#> 61  135.12306 138.44260 157.64838 182.40557 108.11017 112.35454 148.91728
#> 62  196.43415 117.90586 120.33827 137.15657 121.86406 132.67459 175.80092
#> 63  181.77206 162.21544 139.20578 127.80355 123.46715 161.46691 103.65413
#> 64  130.56510 114.42243 116.48365 211.07823 100.27233 123.25556 136.41656
#> 65  164.36444 128.90201 112.42703 198.01767 143.03505 189.66670 129.01844
#> 66  116.14630 164.24134 131.46493 139.92625 147.71992 135.97700 123.99391
#> 67  173.05245 161.08432 146.08937 141.34675 140.49669 173.76920 219.08660
#> 68  157.52983 147.11848 119.59340 141.56228 247.32275 114.62366 147.02061
#> 69  197.86262 145.53381 136.82983 167.15294 165.01314 192.99827 146.94917
#> 70  135.70113 129.04754 136.35903 164.07487 111.71078  90.74235  95.51926
#> 71  157.98918 105.05236 152.00917 159.26136 165.97628 151.65984 168.61238
#> 72  161.84677 141.51469 185.79535 130.21296 104.53824 150.05290 149.30797
#> 73  218.99138 119.12043 154.85582 153.73413 130.75301 129.68547  96.35341
#> 74  162.82144 132.19325 124.65046 170.25071 157.83514 129.06156 152.06676
#> 75  136.99409 128.27794 179.23794 143.59805 200.88003 149.44538 188.56592
#> 76  126.63004 116.35475 184.04841 136.43117 174.38751 129.78802 180.89574
#> 77  176.80918  93.82109 142.27966 169.72620 119.30608 163.21608 124.64335
#> 78  204.65372 168.84886 108.58886 190.19922 250.76245 113.25028 186.20253
#> 79  141.62353 140.39756 146.68152 133.26423 152.14036 184.36227 118.95737
#> 80   84.07822 154.07965 159.05469 117.79869 153.66490 158.94613 143.94477
#> 81  132.92106 159.48114 122.98281 171.41324 118.33561 173.15665 118.17349
#> 82  187.88868 139.30538 146.55595 227.94109 151.58067 104.35550 155.68942
#> 83   99.25229 143.94267 129.45228 160.07165 145.89387 212.72147 135.91087
#> 84  132.60779 129.83226 107.75538 118.60872 150.25276 192.71128 139.72809
#> 85  145.63143 154.89065 151.79058 210.84964 107.53310 167.88397 114.29898
#> 86  161.05357 155.75026 115.81618 158.76246 126.98730 131.70066 157.61759
#> 87  186.17783 116.00402 150.36196 190.79895 169.00319 128.37103 147.27204
#> 88  187.94846 156.84200 148.52425 136.53727 195.30679 107.44712 111.64442
#> 89  180.11565 131.83544 159.66692 124.35567 158.61011 146.77897 115.65717
#> 90  179.79399 150.89710 181.62718 118.18812  91.59022 178.73091 119.07619
#> 91  161.37058 150.97467 163.60568 147.80154 142.78027 154.48130 134.19472
#> 92  167.95319 133.29811 113.36441 150.51821 169.88101 170.39853  91.84830
#> 93  181.67139 153.21534 134.89189 108.25085 178.14267 174.58843 170.22127
#> 94  115.99039 123.32305 172.66469 148.66453 131.19108 131.73726 147.92370
#> 95  122.98167 119.03030 174.76507 137.36287 202.70452 157.39376 149.79456
#> 96  163.01998 149.16693 133.95456 131.96490 191.81222 156.84165  92.22109
#> 97  149.02000 144.23581 143.31841 134.21882 133.53910 151.11425 131.39168
#> 98  111.44959 132.64026 208.71257 145.13196 134.22130 126.35893 121.79068
#> 99  160.87461 155.41261 134.18185 183.93974 173.39947 134.74924 178.86200
#> 100 125.94562 122.66995 114.66412 141.34191 184.32383 112.88784  89.22679
#>         B.4_1
#> 1   206.25270
#> 2   151.75228
#> 3   145.81021
#> 4   150.23975
#> 5   147.10389
#> 6   135.79157
#> 7   172.11441
#> 8   165.91349
#> 9   145.36597
#> 10  133.58684
#> 11  173.00059
#> 12  125.80496
#> 13  177.57530
#> 14  155.04701
#> 15  104.74347
#> 16  141.43591
#> 17  117.27463
#> 18  141.57519
#> 19  166.71493
#> 20  183.92311
#> 21  144.35936
#> 22  149.05161
#> 23  112.00283
#> 24  157.77663
#> 25  160.30243
#> 26  148.04796
#> 27  141.47857
#> 28  136.25526
#> 29  151.76056
#> 30  157.40937
#> 31  138.44366
#> 32  128.28353
#> 33  170.03611
#> 34  212.75001
#> 35  131.72659
#> 36  182.57223
#> 37  160.88628
#> 38  183.42011
#> 39  137.56446
#> 40  164.09554
#> 41  110.15251
#> 42  156.00945
#> 43  200.72903
#> 44  136.80629
#> 45  131.57553
#> 46  160.48416
#> 47  151.67683
#> 48  159.99504
#> 49  115.87315
#> 50  148.25922
#> 51  124.50181
#> 52  112.71358
#> 53  190.99247
#> 54  108.73664
#> 55  133.23048
#> 56  155.78900
#> 57  127.65253
#> 58  139.12893
#> 59  103.95694
#> 60  143.30273
#> 61  147.48484
#> 62  177.13233
#> 63  215.64148
#> 64  141.90224
#> 65  150.25401
#> 66  183.22063
#> 67  202.89859
#> 68  153.54149
#> 69  196.76191
#> 70  103.03593
#> 71  159.32756
#> 72   97.14699
#> 73  167.39284
#> 74  173.63041
#> 75  132.96492
#> 76  186.49172
#> 77  172.09584
#> 78  154.21974
#> 79  175.18966
#> 80  134.96189
#> 81  158.82584
#> 82  166.49191
#> 83  150.59105
#> 84  186.73237
#> 85  137.29769
#> 86  150.76100
#> 87  137.19260
#> 88  164.95440
#> 89  110.13991
#> 90  114.88267
#> 91  145.69133
#> 92  107.80543
#> 93  197.68411
#> 94  126.77900
#> 95  191.45000
#> 96   97.17917
#> 97  122.32513
#> 98  164.56186
#> 99  173.89029
#> 100 136.65622