Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10,10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(subset = rep(c("A","B"),50),
                   level = rep(1:4, each = 25))
rownames(meta) <- 1:100
scRNA <- Seurat::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA, by = c("subset","level"),
               min.cells = 10, max.cells = 20)
#>         A.1_1     B.1_1     A.2_1     B.2_1     A.3_1     B.3_1     A.4_1
#> 1   136.28369 192.53778 162.14879 140.77620 139.36564 122.95625 135.04054
#> 2   152.96188  75.87015 159.48761 150.21270 150.15016 195.89694 148.66250
#> 3   126.39605 159.59422 183.70874 199.39368 147.11309 124.00379 146.90349
#> 4   160.31342 138.79392 124.55243  92.32974 118.09405 123.37670 183.37441
#> 5   114.90193 177.40575 197.53612 208.26743 207.67330 106.61503 150.88341
#> 6   163.95241 160.33750 119.27116 177.88796 129.25305 179.49399  90.31547
#> 7   137.47533 174.43969 160.74053 155.04834 152.97065 157.07128 158.30925
#> 8   148.64660 136.34289 147.47291 158.86259 101.76826 154.88702 185.49115
#> 9   178.39100 101.23061 129.65548 130.94253 139.57431 148.93260  98.75251
#> 10  126.89760 133.89953 116.25346 149.01011 133.20167 126.25406 186.34518
#> 11  162.19798 193.36991 125.45405 123.03136 137.72732 122.65300  88.73242
#> 12  158.64455 116.26853 166.86253 183.38242 202.19024 103.35495 117.73676
#> 13  122.98021 141.83123 149.67320 143.10713  80.92142 110.37090 133.11899
#> 14  162.99091 162.51370 130.92292 136.50589 148.39800 101.32997 136.40477
#> 15  129.06648 145.83957 168.47221 161.08962 195.92848 108.70829 146.04137
#> 16  126.63856 146.73942 138.26625 119.23201 143.80354 137.97819 134.91833
#> 17  138.15681 160.61553 192.66405 208.55229 119.50265 192.37586 165.72725
#> 18  147.11465 186.51802 183.48469 186.84836 150.79866 166.66048 182.71560
#> 19   56.68901 162.00253 106.02841 192.32620 115.62262 114.72907 146.76157
#> 20  140.63111  77.94138 151.92642 184.49231 167.82077 167.13672 133.31866
#> 21  125.22272 120.15137 153.24648  90.95901 154.30878 163.58854 120.76943
#> 22  125.04858 202.84725 178.07542 131.42153 174.17280 115.46202 151.60708
#> 23  163.51839 132.80586 145.91331 139.91540 155.01410 114.53931 195.73706
#> 24  174.93725 131.69244 135.55439 112.97019 114.40736  99.07750 124.79149
#> 25  183.69535 146.39480 169.41346 100.72329 162.45263 119.64460 184.05540
#> 26  168.02406 121.53729 115.86667 163.72423 148.61441 179.64285 124.04808
#> 27  163.19078 182.75100 166.11379 183.00257 139.76954 111.00450 133.50055
#> 28  113.03040 136.92477 121.27443 168.20292 131.72599 134.74583 163.28831
#> 29  132.40722 102.86010 156.56396 146.80455 124.08144 107.69964 127.17337
#> 30  137.58162 128.83739 120.78662 113.14611 154.52222 109.12799 148.38356
#> 31  200.76953 123.23090 120.13523 195.26269 211.80083 146.75664 124.10111
#> 32  144.61917 133.44178 157.29514 155.20055 155.86625 180.13415 172.98217
#> 33  151.44244 170.18440 173.79323 146.50076 138.50103 176.39557 139.16574
#> 34  128.41377 138.79574 203.85017 118.39875 186.74417 121.09048 123.82460
#> 35  167.54717 148.45300 156.95334 164.05033 126.33736  85.25181 126.88945
#> 36  162.52551 204.31643 150.66540 197.56695 116.66129 125.13028 181.06124
#> 37  128.54349 183.20043 129.06151 170.62102 143.23373 131.49008 156.75543
#> 38  159.82815 181.64347 139.48291 175.27681 160.79188 112.51497 149.45704
#> 39  139.91294 142.81342 132.12830 157.66159  84.96861 156.01403 166.28907
#> 40  160.21984 129.16175 142.05775 140.64053 152.87845 185.79276 197.49318
#> 41  181.31573 197.18020 138.63619 176.78578 139.73287 115.55102 170.29521
#> 42  152.89516 130.56218 152.24435 140.63684 130.17176 156.92294  98.75079
#> 43  147.60720  88.58050 119.94017 110.97735  90.34417 112.46499 103.46238
#> 44  167.37268 114.89328 141.91127 165.82112 129.93549 176.45726 147.81169
#> 45  125.20246 132.72416 134.98212 145.30362 131.21957 148.49048 111.37282
#> 46  152.03557 127.45896 146.06290 150.41351 205.61009  95.60173 148.20474
#> 47  188.12054 173.15541 148.44014 172.25871 126.02926 116.18067 118.31719
#> 48  123.47562 158.23964 131.85736 138.54593 228.63090 171.77485 125.39118
#> 49  142.55516 180.95735 164.54089 133.50696 129.35229 191.50935 154.56271
#> 50  133.17864 164.42750 112.36008 187.46760 193.14237 174.41168 135.20634
#> 51  134.19593 116.47136 169.31486 188.46371 164.03843 155.62366 130.88802
#> 52  151.75569 120.93217 179.94209 171.65468 138.12572 154.60569 149.21344
#> 53  120.90920 149.78117 182.18201 125.88356 157.45368 132.86554 154.83383
#> 54  140.15588 117.46834 168.28484 119.18536 119.16405 165.55336 171.73738
#> 55  147.10040 130.18640 169.23956 190.65373 161.52514 111.88851 131.37085
#> 56  127.96744 149.50462  72.25018 156.52337 179.82378 143.54351 116.37913
#> 57  128.81911 147.89169 170.89733 110.65200 143.78109 108.46434 156.06796
#> 58  141.28393 128.37797 158.38037 146.95496 126.32238 163.03847 144.44095
#> 59  138.24066 105.65664 113.64470 143.38855 204.97571 111.01495 190.12673
#> 60  136.68804 114.38966 158.72555 194.37352 147.62790 144.23322 163.16606
#> 61   92.78913 125.60609 111.57135 184.26373 195.76190 112.52892 146.62076
#> 62  121.14260 136.14869 125.63460 133.54489 135.38938 127.44355 189.60753
#> 63   82.87381 204.67803 140.03503 146.48560 183.28041 154.10167 129.89158
#> 64  100.81831 192.52717 157.36837 123.92556 191.97086 144.96097 100.02959
#> 65  181.56180 214.11612 136.91238 173.47433 154.70588 125.49249 156.08673
#> 66  224.15123 132.82136 131.13786 163.90036 179.30425 121.17595 153.52922
#> 67  159.68300 127.10327 148.93717 162.99620 164.35346 115.51110 160.82162
#> 68  118.88498 109.71162 167.42491 144.62410 121.87094 157.69952 127.06042
#> 69  144.37683 135.72071 114.14928 108.44097 188.72343 156.94065 134.91715
#> 70  156.39934 103.02897 181.89873 170.39392 118.42611 168.83561 118.72852
#> 71  189.27598 120.61771 159.23849 128.64323 149.32583 137.59413 200.91452
#> 72  140.90925 109.99501 140.46335 119.17071 133.41852 130.24648 177.40961
#> 73  135.25400 163.59320 135.38814 160.78210 166.17911 109.25841 121.51414
#> 74  148.86307 176.68251 156.88106 154.71894 163.17997 187.23594 135.44997
#> 75  119.23235 131.89755 191.93239 148.65378 140.56525 147.97397  93.49683
#> 76  169.72568 166.70263  98.37478 136.23141 218.83842 146.84320 130.28475
#> 77  132.53921 146.58084 137.54591 159.48297 218.32230 133.75325 229.20709
#> 78  121.46056 156.95722 121.27163 179.97444 178.60992 144.57747 153.50899
#> 79  246.15975 168.21234 124.07008 149.53972 125.25792 222.19679 112.91892
#> 80  143.75902 103.39253 109.62158 190.63259 185.11183 163.10724 163.32577
#> 81  189.93320 189.03305 137.89875 104.05267 115.49650 118.00790 119.92316
#> 82  196.75679 105.50107 132.41646 198.72442 170.82330 115.97403 151.02001
#> 83  184.57950 142.79389 191.73460 148.99090 162.21449 121.92081 114.20087
#> 84  115.40856 171.67072 170.71915 113.78488 155.43056 157.20023 119.51541
#> 85  172.50308 108.62689 141.08365 200.00702 135.92189 191.34697 133.70742
#> 86  158.02308 138.24964  98.53840 160.99385 189.49210 119.93282 148.93229
#> 87  201.51773  91.73438 138.06364 154.25598 153.03077 139.48862 128.04301
#> 88  183.17877 153.46103 161.17566 119.10898 149.67250 139.62329 143.87781
#> 89  158.20782 169.46595 100.36597 165.35450 178.44754 131.39002  98.59530
#> 90  132.07892 157.09084 144.25364 115.91794 168.79088 139.98982 125.64102
#> 91  181.77895 155.31950 134.31886 111.01770 156.65883 160.28822 125.40449
#> 92  193.61211 141.14227 145.84112 138.19536 159.41862 175.17284 141.14171
#> 93  129.17776 136.81836 146.80600 152.86543  89.46983 188.13178  85.46142
#> 94  201.53040 168.55775 107.49458 182.11185 113.38151 160.38836 132.25265
#> 95  133.64064  91.74672 195.51804 160.92567 152.80888 139.40393 194.72894
#> 96  115.90647 115.17385 100.47460 165.17166 138.03991 121.79590 154.41200
#> 97  153.02602 154.32176 161.63179 176.94439 178.09418 126.48460 117.33896
#> 98  143.36912 134.86808 125.67012 132.88301 188.06512 137.88087 175.40707
#> 99  168.14036 128.39560 151.65156 139.99363 164.90576  95.76612 178.41737
#> 100 106.46611 149.75708 183.83656 135.04146 129.94175 124.19577 130.94443
#>         B.4_1
#> 1   205.70621
#> 2   158.31548
#> 3   145.21014
#> 4   171.44048
#> 5   174.14548
#> 6   200.55691
#> 7   165.33002
#> 8   164.42589
#> 9   158.97242
#> 10  180.59858
#> 11  126.11617
#> 12  175.02672
#> 13  166.21203
#> 14  147.67777
#> 15  148.02028
#> 16  204.55333
#> 17  112.01313
#> 18  148.17992
#> 19  183.91264
#> 20  179.22043
#> 21  124.92474
#> 22  127.94276
#> 23  168.30275
#> 24  108.20101
#> 25  147.64728
#> 26  134.94842
#> 27  161.33922
#> 28  124.62477
#> 29  157.12999
#> 30  198.81702
#> 31  160.30549
#> 32  125.11192
#> 33  207.94261
#> 34  136.25510
#> 35  143.22318
#> 36  211.18269
#> 37  142.39871
#> 38  176.43292
#> 39  154.83261
#> 40  146.25698
#> 41  149.00417
#> 42  189.07860
#> 43   96.46566
#> 44  178.20984
#> 45  133.92002
#> 46  166.26889
#> 47  157.79953
#> 48  163.55581
#> 49  120.11693
#> 50  171.28690
#> 51  156.38642
#> 52  126.19616
#> 53  156.77422
#> 54  111.80471
#> 55  176.88721
#> 56  118.11884
#> 57  109.74905
#> 58  178.86859
#> 59  162.72928
#> 60  111.46897
#> 61  120.12783
#> 62  164.75751
#> 63  176.49328
#> 64  127.33806
#> 65  147.87571
#> 66  178.84376
#> 67  159.34982
#> 68  140.32614
#> 69  148.25971
#> 70  171.10350
#> 71  123.60695
#> 72  182.48166
#> 73  148.44004
#> 74  123.23232
#> 75  123.71032
#> 76  162.37992
#> 77  138.89198
#> 78  178.78850
#> 79  161.20158
#> 80  135.03272
#> 81  133.75323
#> 82  197.41459
#> 83  143.05882
#> 84  190.24145
#> 85  168.80662
#> 86  116.78519
#> 87  150.21945
#> 88  140.40543
#> 89  152.55240
#> 90  148.98306
#> 91  208.18309
#> 92  156.71241
#> 93  209.41455
#> 94   98.41931
#> 95  186.20655
#> 96  177.98787
#> 97  154.15598
#> 98  172.42366
#> 99  204.92424
#> 100 162.06080