Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10, 10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(
  subset = rep(c("A", "B"), 50),
  level = rep(1:4, each = 25)
)
rownames(meta) <- 1:100
scRNA <- Seurat::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA,
  by = c("subset", "level"),
  min.cells = 10, max.cells = 20
)
#>         A.1_1     B.1_1     A.2_1     B.2_1     A.3_1     B.3_1     A.4_1
#> 1   190.38164  67.22350 153.64447 177.56362 132.35223 137.60404 116.66857
#> 2   146.32705 118.87385  99.60004 176.47859 144.61058 125.22023 151.75566
#> 3   145.83672 139.95294 130.77315 171.44415 187.91820 158.36442 122.70789
#> 4   177.30459 169.25993  75.85986 188.48323 151.51816  79.19918 145.65357
#> 5   119.92653 155.91161 117.71614 113.83410 148.31100 155.90638 136.39148
#> 6   204.22153  91.15326 107.98720 147.16186 155.22044 156.79373 150.71774
#> 7   159.91740  95.91325 143.33320 159.95538 148.95428 117.42108 168.38403
#> 8   141.61971 149.44675  65.20997 160.95176 169.75495 144.75436 102.25822
#> 9   160.37367 126.69050 163.72610 183.65604 136.00976 147.13128 114.46802
#> 10  149.81117 169.03295 165.12174 111.74560 157.57132  82.39332 120.16643
#> 11  114.25755 200.43937 152.99392 136.13949 214.36101  94.97328 119.20352
#> 12  140.11078 118.64052 133.95842 170.45034 122.80708 140.70214 162.32673
#> 13  161.96479 176.83392 134.29628 125.07409 161.09767 140.44854 164.39428
#> 14  141.13341 112.48880 118.92945 189.50475 172.58344 126.88086 129.49588
#> 15  107.17693  97.97224 164.45204 177.50847 150.00659 118.84342 176.20152
#> 16  144.79011 155.54224 208.77824 129.03336 114.71774 169.46640 163.49062
#> 17  156.72500 162.55487 122.05661 238.07904 146.34740 112.88607 137.92127
#> 18  196.92166 142.24612 128.92480 143.18446 156.03014 210.62636 175.33285
#> 19   73.35191  82.91715 126.56171 145.25065 145.35107 140.24066 123.86986
#> 20  142.58880 145.18988 137.79278 157.10888 154.18615 123.94630 144.77282
#> 21  162.35819 148.96049 121.48090 179.72117  74.64688 124.32237 105.28132
#> 22  179.04640 152.87756 149.09695 157.27498 127.00541 180.10558 148.60123
#> 23  182.34224 144.19162 120.98573 109.40510 120.29671 102.49456 167.36661
#> 24  140.25337 168.70589 156.26454 192.66258 153.08442 108.24771 163.04130
#> 25  179.91958 187.04402 115.27869 162.04129 139.75776 149.54396 114.77854
#> 26  162.58259 149.75428 102.56888 160.55404 152.67403 145.07202 140.02822
#> 27  165.21893 138.64163 188.22677 174.01245 167.30001 108.14279 191.14837
#> 28  190.52758 119.57365 167.43904 165.46364 128.03008 132.55857 167.07151
#> 29  145.02728 157.45035 107.59472 139.51754 112.48115 109.99117 117.74061
#> 30  159.01734 201.95221 144.03222 154.26669 150.87259 111.30164 180.16020
#> 31  150.62430 183.53897 132.51070 161.41734 182.14718 112.36903  92.70091
#> 32  153.31164 163.96027 160.71125 159.65832 104.05730 142.95975 149.84327
#> 33  121.79309 156.12060 122.44448 133.95168 168.90424 129.40926  97.74664
#> 34  117.31943 165.49128 156.87494 175.79821 152.00159 163.79641 137.92656
#> 35  131.82531 133.29516 110.98866 115.68566 164.28941 168.29621 144.58155
#> 36  193.52936 154.85305  96.67567 157.74417  99.88336 141.77084  82.03337
#> 37  159.07704 122.67116 140.41233 161.22934 138.83811  91.86958 178.44576
#> 38  147.80482 132.29217 128.52681 184.14474 199.09708 127.57491  87.27701
#> 39  114.06816  95.00057 109.03816 141.46331 151.84975 152.68607 191.35941
#> 40  138.10233 213.30283 126.53363 136.89005 116.78568 159.41429 181.50920
#> 41  176.82975 101.89961 155.21544 149.91521 148.81690 160.43864 147.06958
#> 42  211.33601 137.01665 136.90842  82.54766 136.52935 123.82746 166.50330
#> 43  148.35666 145.91648 132.63408 197.94867 157.13641 126.56187 108.67548
#> 44  122.85981  94.51812 109.66489 131.87156 105.96832 144.38519 173.80844
#> 45  163.85789 150.03493 122.96561 141.24880 127.53238 187.37284 131.51315
#> 46  175.40301 105.88113  96.67892 194.22192 157.37818 178.01064  95.44942
#> 47  136.70950 115.34139 133.55778 125.07415 180.68880 122.16492 136.58155
#> 48  167.31785 118.41900 106.87495 210.78281 113.82671 108.31719 113.22468
#> 49   83.82078 186.04609 156.94422 149.87465 138.85457 128.76031 168.43360
#> 50  143.04553 204.15715 117.22633 150.46925 189.98504 153.67673 107.56155
#> 51  126.58742 146.58270 109.38626 144.93063 116.29016 195.49889 142.96002
#> 52  210.12290 116.46539 138.55141 178.22188 140.65388 117.25494 112.48291
#> 53  192.96120 184.53096 136.05398 139.10796 200.33405 155.33527 147.86398
#> 54  184.01235 108.32847 125.26405 173.83827 158.43551 176.99863 134.19374
#> 55  168.23188 157.16814 156.54433 101.22930 144.42131 114.39996 162.12286
#> 56  181.46929 100.38711 151.57526 123.44919 141.09368 128.10203 113.70721
#> 57  114.42948 178.66059 113.52545 118.95884 178.36857 147.49884 138.43516
#> 58  171.89936 126.21814 139.71487 108.49098 191.10341 107.56146 136.12698
#> 59  176.18615  99.07787 127.35314 111.12637 134.09265 145.29999 148.27666
#> 60  131.21979  91.55996 128.16715 149.79240  97.98189 140.58692 140.70669
#> 61  160.32162 208.79355 128.07843 163.26335 136.31944 122.73455 127.47097
#> 62  234.01690 141.03420  93.46358 189.95094 166.40207 120.23159 158.83614
#> 63  181.14179 111.34129 159.64096 165.93142 179.42890 130.95820  80.00335
#> 64  156.08510 129.18741 149.26424 156.85146 144.32463 144.89248 142.72729
#> 65  136.86106 133.87885 145.28353 173.60002 166.50140  87.67837 176.38606
#> 66  191.34119 120.61239 123.16350 225.57733 202.95396 149.72036 155.93054
#> 67  110.61069 165.31163 165.38333 197.10801 169.46387 128.71258 124.64403
#> 68  187.23622 113.79520 112.35246 138.93124 145.53335 180.30730 148.29397
#> 69  164.19337 138.09333  97.59798 124.60245 192.61528 152.76314 205.05590
#> 70  122.10729 114.93566 130.76274 142.49353 149.81958 155.89387 135.89412
#> 71  164.13285 138.52138 158.45367 123.18507 150.76378 221.21457 182.84993
#> 72  151.61451 144.12103 110.91248 153.16993 121.74792 118.66685 117.45650
#> 73  134.92555  91.23416 120.51034 193.54136 151.90376 146.42602 126.06918
#> 74  143.76622 150.66508 150.18584 159.02292 196.98032 137.46916 156.30942
#> 75  126.78641 136.16804 190.12987 135.69243 147.51217 123.81218  88.63941
#> 76  116.37234 127.46302 101.12428 122.35322 113.61152 139.71625 131.71787
#> 77  215.94166 135.49390 166.38190 199.19032 145.64200 205.90711 183.68591
#> 78  116.67738 121.65065 188.56529 155.58283 104.65419 112.14137 147.42747
#> 79  163.53355 137.98932 167.62560 135.54681 125.45121 160.61556 146.72883
#> 80  147.26982 108.48420 148.80303 140.80590 149.76801  88.32277 148.23979
#> 81  177.81369  91.42535 104.57872 122.12331 117.31801 112.77998 112.39595
#> 82  173.32813  96.31054 158.80055 151.52574 130.80602 124.86214 101.67674
#> 83  141.97714 107.45021 115.49872 134.00115 170.40698 174.54000 136.50966
#> 84  162.73357 139.71146 137.38641 135.46093 158.54033  82.06079 146.05632
#> 85  164.12475 117.03034 130.69689 133.10421 193.66076 162.93591 138.92871
#> 86  122.06264 140.20019 164.07757 176.22215 188.10599 160.39243 130.23793
#> 87  168.24619 115.18305 172.91288 131.67356 163.24109 166.25026 124.86761
#> 88  141.60083 150.83835 119.42884 177.49201 128.18282 101.29940 165.22492
#> 89   98.23287 167.71539 147.64850 142.11400 128.69226 141.18512 137.58593
#> 90  155.61559  83.54032 137.28229 144.68147 160.49621 122.42676 133.74116
#> 91  102.72892 141.85321 150.78516 143.98262 162.60398 173.85065 144.37050
#> 92  132.42680  94.75429 149.22814 180.38210 118.21298 118.85317 145.97897
#> 93  166.02001 118.19967 116.51189 161.25589 177.28173 164.12018 145.16477
#> 94  143.86049 110.33189  85.23727 154.16278 162.13004 149.86795 102.83290
#> 95  131.20170  87.21950 156.73783 161.06030 109.43320 138.35509 114.19041
#> 96  139.04125 189.93264 146.80678 168.03822 147.40602 180.98162 153.11376
#> 97  139.11712 138.82809 119.14108 100.38636 147.97020 128.62480 164.66125
#> 98  100.95361 133.34413 174.18792 129.23356 200.28977 113.97105 116.75811
#> 99  188.21398 148.52744 114.24656 128.85746 207.90506 170.01171 163.32106
#> 100 140.92696 135.91641 143.17617 120.14613 162.84091 154.76761  78.90908
#>         B.4_1
#> 1   207.83585
#> 2    99.62738
#> 3   118.83584
#> 4   150.71163
#> 5   199.24877
#> 6   161.47242
#> 7   151.40800
#> 8   103.94970
#> 9   132.30651
#> 10  147.75002
#> 11  169.24484
#> 12  179.67449
#> 13  144.30720
#> 14  147.44484
#> 15  109.97499
#> 16   99.48245
#> 17  121.20687
#> 18  150.33838
#> 19  132.39001
#> 20  201.84471
#> 21   77.12999
#> 22  166.02528
#> 23  209.41101
#> 24  144.96239
#> 25  219.13471
#> 26  160.13334
#> 27  184.59062
#> 28  131.96159
#> 29  126.38180
#> 30  136.80371
#> 31  126.12744
#> 32  157.85553
#> 33  139.47666
#> 34  150.13373
#> 35  148.56574
#> 36  176.89135
#> 37  153.54146
#> 38  158.78294
#> 39  169.38016
#> 40  149.31402
#> 41  199.94161
#> 42  143.32861
#> 43  163.07173
#> 44  162.84684
#> 45  113.06278
#> 46  140.52339
#> 47  139.65047
#> 48  134.76780
#> 49  128.70734
#> 50  127.45105
#> 51  118.81022
#> 52  172.75638
#> 53  141.48424
#> 54  121.28367
#> 55  140.57702
#> 56  173.14580
#> 57  161.90587
#> 58  145.52984
#> 59  140.93216
#> 60  167.18031
#> 61  148.09236
#> 62  136.98604
#> 63  100.28724
#> 64  117.91204
#> 65  198.29503
#> 66  151.30240
#> 67  185.19924
#> 68  131.28120
#> 69  103.88371
#> 70   90.95056
#> 71  102.38781
#> 72  152.14702
#> 73   94.82807
#> 74  155.00974
#> 75  142.10916
#> 76  163.52826
#> 77  145.69598
#> 78  175.70928
#> 79  175.40069
#> 80  118.43639
#> 81  213.97234
#> 82  151.03772
#> 83  182.44581
#> 84  171.72373
#> 85  209.04998
#> 86  142.89007
#> 87  179.38880
#> 88  165.96018
#> 89  161.97946
#> 90  136.30439
#> 91  148.19713
#> 92  122.26840
#> 93  160.11325
#> 94  145.00189
#> 95  190.87850
#> 96  172.87199
#> 97  156.47574
#> 98  203.01104
#> 99  138.03048
#> 100 194.32651