Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10,10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(subset = rep(c("A","B"),50),
                   level = rep(1:4, each = 25))
rownames(meta) <- 1:100
scRNA <- Seurat::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA, by = c("subset","level"),
               min.cells = 10, max.cells = 20)
#>         A.1_1     B.1_1     A.2_1     B.2_1     A.3_1     B.3_1     A.4_1
#> 1   100.66141 160.77867 123.85177 169.02898 185.54209 125.24817 116.37883
#> 2   129.72906 158.89630 174.71389 130.40652 152.93481 140.05124 112.89234
#> 3   114.84211 130.25259 140.37767 133.20029 177.52476 136.83805 194.54117
#> 4   158.72157 148.52298 124.25818 125.53834 208.73582 114.29981 114.45197
#> 5   125.40006 138.37432 127.68279 165.94569 137.94907 122.74268 138.63615
#> 6   144.74580 202.93946 141.58728 208.69006 150.17991 179.70979 171.24800
#> 7   102.51959 118.02516  97.08758 132.87796  97.91536 123.96700 155.86957
#> 8   117.46133 142.23747 160.89523 182.53060 149.60407 143.12873 116.50160
#> 9   133.69863 123.61082 147.84438 121.85340 119.70709 120.95226 154.93039
#> 10  158.14902  94.44517 135.69520  93.09483 135.51147 153.81989 133.69799
#> 11  182.50347  92.46629 100.75975 147.19861 132.82666 108.24024 133.48430
#> 12  189.32813 111.54677 114.26410 179.87039 140.93122 129.44449 148.04823
#> 13  121.31604 192.13429 145.66700 135.30605 149.76748 124.94149  94.82672
#> 14  130.91288 140.90971 136.07932 171.49042  85.27929 136.41564 146.98625
#> 15  128.79121 144.13717 150.57047 144.19628 145.63833 160.92967 116.15595
#> 16  141.74399 159.11658 124.65282 141.82856 126.11958 102.70743 167.67257
#> 17  221.44963 221.71561 123.53301 112.49016 138.73605 124.54819 167.67632
#> 18  134.06583 208.30356 159.78964 167.31638 107.74002 108.45484 151.23346
#> 19   98.26644 157.54389 145.05713 200.50661 190.93112 124.40919 144.88208
#> 20  153.93695 166.28889 189.13534  81.33443 131.98580 155.17992 166.86840
#> 21  161.00909 103.16671 121.07638 112.58376 171.20916 120.24041 164.76721
#> 22  171.16483 103.23779 122.41445 168.52845 160.77896 181.48813 121.74272
#> 23  183.86955 169.40760 161.35485 145.81531 153.33338 154.47571 134.65204
#> 24  149.87941 122.80554 115.12040 169.42454 154.55702  77.68214 153.73062
#> 25  204.89381 204.19835 127.61536 152.65118 161.06046  97.07701 149.36759
#> 26  193.32585  71.09244 122.66011 155.36047 166.54091 146.68723 102.47035
#> 27  179.79765 142.90304 145.43026 167.63114 159.39881 150.42881 144.87017
#> 28  201.34468 134.60087 102.23289 128.90075 134.46794  76.25359 106.07441
#> 29  181.34038 160.09644 180.75493 174.67222 136.46455 129.57196 137.95109
#> 30  158.53667 162.27004 180.69120 172.41040 129.96839 161.57816 144.62941
#> 31  133.65765 148.39473 110.75977 151.39796 160.19055 133.90103 107.38022
#> 32  120.16893 142.06255 144.05263 174.41953 128.67554 179.73287 174.77840
#> 33  128.91404 139.13356 135.44247 154.31405 121.58526 124.44021 130.11915
#> 34   88.57273 104.27678 110.01150 139.12282 109.47622 161.14504 117.49679
#> 35  190.29127 153.29629 130.07433 153.22234 120.90771 201.26374 171.15553
#> 36  169.95723 118.71027 120.47964  96.52360 168.00342 135.36505 117.04047
#> 37  153.50321 135.69401 142.83972 180.95377 181.82601 130.21417 101.11123
#> 38  109.34602 134.20910 134.13522 136.98216 159.46550 155.06297 130.05291
#> 39  145.91194 169.32220 127.37618 149.34695 129.73218 122.65339 139.35297
#> 40  176.86111 135.25495 194.18096 197.88097 165.36408 139.31166 153.23959
#> 41  127.34701 188.97430 112.58489 181.66163 118.82756 123.90202  73.29152
#> 42  141.75965 126.65027 150.49171 139.93624 149.14234 160.38254 142.01461
#> 43  154.15056 149.04854 146.78689 177.13768 169.05302 138.66230 133.20444
#> 44  126.67830 102.39149 150.86375 158.80009 145.34990 136.03780 139.23631
#> 45   89.36136  69.90253 106.07077 158.65091 122.35670 136.07681 116.01304
#> 46  143.20733 120.76380 106.56169 138.74129 155.04034 160.12838 168.27155
#> 47  100.43022 147.25551 130.18854 142.41633 142.65809 163.97620 148.57544
#> 48  184.32102 149.90500 146.97825 187.01086 200.40250 147.26682 163.66837
#> 49  171.60364 161.03334 107.15506 123.78821 148.04813  85.60568 134.67847
#> 50   76.04946 147.10127 149.90319 142.84737 114.58620 200.13368 141.42341
#> 51  236.15098 141.95820 129.14692 124.10690 113.92443 112.30588  70.73152
#> 52  127.82289 170.95749 153.16741 153.53099 185.86026 119.20094 152.13338
#> 53  170.02441  90.42522 136.63486 107.40165 178.27434 162.09584 110.56535
#> 54  152.36150 123.91695 122.36543 183.91462 162.68603 132.03819 141.39431
#> 55  157.38363 136.43152 125.87901 178.42128 110.16673 150.61509 172.57258
#> 56  109.95663 141.15215 137.73063 180.64159 130.18175 163.10818 125.67317
#> 57  114.00267 131.90685 145.64981 158.59900 148.70163 184.29637 140.64205
#> 58  197.95506 146.47044 168.28091 164.42851 124.75066 128.24082 182.42500
#> 59  178.54992 142.67266 155.23677 134.60343 171.46199 196.43002 104.25974
#> 60  141.60093 146.99761 104.31858 132.68582 145.68042  68.30637 124.23006
#> 61  158.15991 153.19345 147.46207  98.56797 171.80704 167.49053 174.30045
#> 62  169.06325 120.42906 135.07209 160.86780 116.03639 141.08479 142.15852
#> 63  117.39988 167.27909 128.97941 151.87021 151.28402 122.81481 127.08463
#> 64  140.42201 164.89300 181.36110 100.87365 180.80987 112.39101 172.77697
#> 65  132.69308 135.99096 110.72338 179.67713 130.24382 133.77100 112.75387
#> 66  174.27565  96.46595 170.68467 118.93386 122.83515 158.34820 145.71694
#> 67  105.81602 121.93236 178.24329 152.90519 124.04712 158.20874 137.51935
#> 68  162.08484 135.65199 103.51874 146.65555 127.68752 113.61866 178.29327
#> 69   67.58376 117.21286  99.65930 131.84936 134.32620 124.15977 116.21343
#> 70  145.40436 188.19568 213.95648 134.20430 208.18573 223.37399 188.01943
#> 71  158.79590 137.20112 155.37112 128.69994  91.89889 105.05628 148.21802
#> 72  127.53942 188.67865 101.31238 205.78925 179.86066 145.49196 166.43573
#> 73  146.37837 178.35439 134.81651 188.00309 141.54205 109.90962 176.76563
#> 74  158.37258 195.10911 146.00552 154.26640 116.26372  97.03553 137.12519
#> 75  105.97467 111.94610 106.44922 145.07428 183.11457 160.76212 155.02350
#> 76  125.07446 163.83412 126.97201 145.18227 139.96417  91.41281 100.07721
#> 77  115.39531 155.36198 161.70352 183.37016 138.22076 121.27844  86.61260
#> 78  117.15806  87.03800 139.78202 119.26274 140.83949 156.98694 179.20339
#> 79  149.14497 121.92229 146.14444 105.99238 165.75110 151.01242 119.97372
#> 80  188.35091 175.75792 156.82948  87.40050 131.50997 115.68814 101.94310
#> 81  121.19369 111.80370 155.72483 145.98336 141.42598 167.95955  87.90168
#> 82  162.56670 161.46264 188.61411 152.43010 185.89527 127.16721 126.76576
#> 83  183.74754 136.18640  89.23969 167.95926 171.65257 140.24915 151.41852
#> 84  150.30271 163.99635 125.01039 133.27605 140.84532 136.71386 125.17374
#> 85  171.38010 164.08393 169.83718 188.94084  92.21754 188.25944 172.79438
#> 86  132.27497 139.77148 151.57408 189.48283 152.90977 106.85639 200.42192
#> 87  152.64246 115.63255 109.13661 177.51125 188.34972 115.25123 137.32429
#> 88  148.76858 123.17866 149.26317 119.49356 163.24401 181.51438 158.71676
#> 89   97.06306 128.71536 146.12589 181.09298 143.19742  92.38474 115.82074
#> 90  128.95310  76.63167 149.40201 156.53668 150.08660 114.26739 146.69136
#> 91  168.77866 220.44984 138.13723 171.65734 172.27871 119.79591 150.90318
#> 92  162.84878 145.93292 103.16122 128.23682 162.92861 153.13156 155.69756
#> 93  115.60331 174.95858  73.43249 133.22299 163.57998  96.83622 151.60917
#> 94  175.21080 175.17302 105.39721 168.50591 137.86227 130.34393 159.56635
#> 95  131.54847 148.93470 156.87941 123.82774 168.81310 108.94796 165.97980
#> 96  141.66922 110.33272 111.54153 140.05125 185.51629 130.67390 168.71124
#> 97  176.87499 125.67412 155.03196 175.58364 182.64981 170.94683 116.63088
#> 98  174.59586 181.48909 162.12115 146.57957 186.06289 199.12390 139.98048
#> 99  170.89518 115.06561 146.70868 142.69883 129.03232 146.25235 110.31100
#> 100 185.25828 127.75458 129.83725 218.92880 149.22715 179.83412 150.46077
#>         B.4_1
#> 1   146.47383
#> 2   190.00799
#> 3   125.52788
#> 4   203.10566
#> 5   175.78271
#> 6   161.62783
#> 7   201.50466
#> 8   111.31493
#> 9   164.92831
#> 10  159.69666
#> 11  180.01673
#> 12  184.30499
#> 13  150.10490
#> 14  159.81690
#> 15  159.70573
#> 16  195.15585
#> 17  219.34779
#> 18   92.59412
#> 19   96.68408
#> 20  140.68181
#> 21  124.65476
#> 22  144.98132
#> 23  115.94546
#> 24  154.42557
#> 25  134.55126
#> 26  137.69218
#> 27  153.34552
#> 28  150.40726
#> 29  163.00692
#> 30  125.39026
#> 31  118.67395
#> 32  175.36020
#> 33  159.23726
#> 34  153.07421
#> 35  149.42328
#> 36  117.75949
#> 37  161.93905
#> 38  144.13171
#> 39  164.58945
#> 40  183.51115
#> 41  104.25814
#> 42  140.30444
#> 43  123.70261
#> 44  162.41795
#> 45  133.25033
#> 46  175.29907
#> 47  167.93390
#> 48  124.47716
#> 49  168.24315
#> 50  152.97777
#> 51  175.09370
#> 52  187.68894
#> 53  133.33476
#> 54  204.02882
#> 55  203.27864
#> 56  129.33313
#> 57  147.59940
#> 58  165.17338
#> 59  192.15287
#> 60  152.36877
#> 61  163.28836
#> 62  193.82514
#> 63  134.97837
#> 64  142.47785
#> 65  142.07984
#> 66  128.39323
#> 67  139.81076
#> 68  123.09910
#> 69  155.28082
#> 70  120.29299
#> 71  162.80537
#> 72  163.27477
#> 73  113.88786
#> 74  121.94420
#> 75  160.23150
#> 76  221.48749
#> 77  135.69239
#> 78  132.41673
#> 79  147.32851
#> 80  144.30044
#> 81  129.71943
#> 82  131.02938
#> 83  129.20804
#> 84  159.13258
#> 85  105.73425
#> 86  140.58991
#> 87  148.06400
#> 88  180.19113
#> 89  196.79835
#> 90  177.87271
#> 91  199.58987
#> 92  179.46600
#> 93  163.29070
#> 94  132.15761
#> 95  114.30024
#> 96  170.11612
#> 97  121.38384
#> 98  156.51052
#> 99  144.88214
#> 100 113.03639