Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10,10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(subset = rep(c("A","B"),50),
                   level = rep(1:4, each = 25))
rownames(meta) <- 1:100
scRNA <- Seurat::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA, by = c("subset","level"),
               min.cells = 10, max.cells = 20)
#>         A.1_1     B.1_1     A.2_1     B.2_1     A.3_1     B.3_1     A.4_1
#> 1   136.16954 156.41634 117.77681 120.42663 127.08684  86.52262 126.42707
#> 2   146.51124 193.52593 110.17368 135.69154 161.23967 107.61363 106.57967
#> 3   166.32805 118.48374 148.84021 128.36028 172.89145 104.24934 116.43548
#> 4   108.40436 157.64573 147.15797 156.34126 189.44914 142.20444 114.04564
#> 5   151.11895 135.64341 133.61451 179.97252 175.88381 145.82827 114.07858
#> 6   160.02233 158.15017 185.73708 108.82399 182.07509 105.89277 168.21209
#> 7    96.63455 143.64772 180.32297 217.84617 180.92325 140.16723 152.94618
#> 8   121.64451 111.77254 115.61998 113.75554 131.02782 163.93004 116.38709
#> 9   120.46948 121.79326 153.45252 150.02207 105.66956 120.55375 111.00952
#> 10  164.51977 133.54995  99.95312 149.25471  91.45082 149.43849 133.16297
#> 11  127.50820 157.76924 131.05588 163.23588 187.73734 189.04064 147.52519
#> 12  193.16207 152.35359 209.34429 182.26954 105.83912 218.30860 151.46311
#> 13  161.35058 181.33747 167.73788 148.71035 163.30699 152.27272 150.54250
#> 14  147.16065 137.08662 129.76533 184.44721 150.38376 135.92111 108.46005
#> 15  114.92582 152.99332 147.97644 152.01044 145.06124 124.04987  92.91741
#> 16  211.41566 167.53501 153.16931 165.42108 124.10837 207.40424 118.65189
#> 17   67.12175  89.58983 171.43879 134.80703 183.33792 129.12125 140.93734
#> 18  176.74577 135.53352 145.95326 125.86387 229.92418 130.36236 164.24847
#> 19  156.23110  93.63785 128.84728 122.78042 108.93096 173.05232 113.29630
#> 20  137.32538 155.87381 168.46348 174.21713 139.09958 119.37363 151.27503
#> 21  103.74265 136.26886 134.48050 214.33018 104.29799 152.74833  87.92736
#> 22  162.42778 129.96324 133.03042 133.28577 137.50722 184.41624 184.35051
#> 23  151.92943 136.68277 101.54438 150.75010 145.79069 119.94033 150.13300
#> 24  154.13471 170.36231 126.75426 160.85227 124.36405 165.22862 175.11387
#> 25  185.61207 156.68440 155.52835 116.34070 148.21196 164.37528 116.60193
#> 26  160.32593 125.09770 155.52055 195.91077 158.75398 136.04335 178.22558
#> 27  141.37210 192.53374 121.33890 151.02533 199.10337 129.40636  98.10343
#> 28  128.43520 129.47509 129.73535 187.80283 132.43715 118.98329 165.08804
#> 29  190.37349 164.02612 154.31648 114.21027 136.08799 139.13584 104.03811
#> 30  143.79237 124.74061 144.85427 192.57297 163.08292 132.22224 118.47400
#> 31  106.98453 121.96849 162.58600 182.61501 188.86527 195.53799 193.22304
#> 32  153.04674 184.50982 111.33337 161.88224 190.06803 122.64318 151.46768
#> 33  146.06608 194.08824  87.75509 107.34026 150.91888 139.78085 153.85223
#> 34  151.36757 103.44346 156.46835 167.65029 161.58704  86.00716 141.86170
#> 35  187.86555 148.30938 140.64689 135.71042 164.08030 157.36862 170.28749
#> 36  168.59416 142.34275 143.74616 143.25389 136.42605 120.05598 122.13084
#> 37  139.47729  93.37295 152.08000 166.24314 190.33679 171.17400 110.80988
#> 38  188.86777 147.26497 128.23933 175.21350 139.23125 126.31641 130.40503
#> 39  120.67377 130.49093 115.70655 149.78025 148.68759 135.01242 159.01706
#> 40  183.25773 156.32049  82.88235 169.89882 188.49598 192.25182 163.86040
#> 41  138.25645  95.47366 147.81671 138.76125 142.94285 174.53368 106.97027
#> 42  154.83831 149.66176 169.44243 106.48161 132.80790 107.09664 137.18694
#> 43  134.45608 100.05360 120.79703 132.69355 141.64050 117.77115 132.34767
#> 44  156.05543 135.64425 158.84985 160.32758 192.63748 172.88673 144.84653
#> 45  174.01130 154.07893 136.08414 168.46120 151.99057  87.18150 140.13951
#> 46  149.12595 153.76998  96.52880 214.82167 137.25948 139.51713 147.36185
#> 47  166.04324 127.57961 126.91919 137.86286 119.51941 130.96821 179.32158
#> 48  174.07482 172.41151 162.42807 203.16860 151.77878 159.24335 182.79207
#> 49  166.09894 174.48224 121.78882 121.16271 172.83100 159.14889 142.27647
#> 50  138.00443  98.68423 148.53673 179.16567 171.20979 147.17398 173.56883
#> 51  151.99697 172.53835 123.13521 162.25274 180.60904 162.18565 112.17098
#> 52  112.38729 202.53339 139.98111 132.56505 141.64377 158.21144 132.98067
#> 53  141.52030 140.48870 151.96563 148.20696 123.81981 189.06634 109.94448
#> 54  163.11789 111.71178 140.94266 158.15105 149.32874 129.55083 152.22058
#> 55  123.37523 111.62826 121.55550 202.11225 142.83804 128.43117 178.32675
#> 56  140.14414 109.95791 133.52075 158.34217 129.50843 128.87990 113.30355
#> 57  111.95795 135.21742 114.08604 172.69098 146.72085 115.97115 146.98625
#> 58  125.82913 131.18321 114.70889 187.48697 149.70731 164.66847 120.88850
#> 59  169.53162 124.21868 169.19784 152.69215 174.15991 172.52094 154.21560
#> 60   78.16324 208.42864 152.48029 151.53730 168.97480 156.61823 135.85548
#> 61  149.53459 119.74667 141.16348 156.93685 136.30189 110.01890 170.78178
#> 62  110.14067 108.41155 157.51348 162.81992 181.13415 196.79637 168.70042
#> 63  158.22367 164.67492 141.59508 132.37861 164.19015 176.07978 151.51393
#> 64  144.75355 144.54856 148.04393 184.58409 153.52516 186.25803 172.04942
#> 65   86.62794 192.85394 176.30632 129.71714 145.16444 102.04747 161.66638
#> 66  123.53668 114.95713 106.53487 128.56040 181.75153 135.91146 102.72364
#> 67  216.26488  73.07365 115.16763 174.05567 178.54332 159.04016 143.32075
#> 68  107.93586 108.53546 117.41241 209.48107 160.32286 128.40565 171.43991
#> 69  198.13137 132.39171 158.47510 148.68911 142.11587 166.50497  90.18186
#> 70  136.62126 131.44983 115.83478 154.44159 156.12402 134.60958 198.70641
#> 71  120.60416 106.67517 139.83709 119.24269 108.64519 134.62005 131.06779
#> 72  129.30899 125.52531 183.39813 108.70074 152.56244 123.14488 118.31793
#> 73  127.41737 102.95371 145.32141 149.18371 165.36401 194.25734 125.80743
#> 74  137.16960 135.54599  68.92429 181.54902 118.82753 139.97552 130.60833
#> 75  223.04804 115.73457 114.47164 164.51613 100.17209 167.73037 131.27494
#> 76  152.18050 131.72827 111.22528 169.35366 168.44821 162.60747  77.58910
#> 77  128.25799 146.82158 143.13808  97.75612 175.44033 131.00718 144.03922
#> 78  162.87613 108.74957 196.05601 116.58381 118.52309 146.98269 142.05573
#> 79  172.39051 144.41076 148.98196 113.34982 168.30051  94.52411 146.64207
#> 80  150.16704  98.74984 164.36498 180.01253 173.42423  96.61347 119.30386
#> 81  138.91223 136.85270 178.16684 216.75845 203.09295 130.56303 161.85756
#> 82  146.39334 122.18259 118.34686 152.03185 167.34802 133.23946 128.56839
#> 83  119.87711 163.39826 171.98326 154.67977 141.43647 171.87690 150.24553
#> 84  156.48760 199.89085 164.66175 133.69984 132.99705 135.26967 136.45025
#> 85  138.86588 121.06041 215.18719 105.56605 129.99742 138.82508 134.98885
#> 86  172.81105 123.47306 137.85711 140.00881 118.26419 121.69505 164.28310
#> 87  148.33862 136.72599 138.45243 162.47432 152.07066 138.61366 132.36711
#> 88  205.13236 154.04538 148.90056 147.62565 151.95290 148.32201 158.95139
#> 89   83.26839 131.10641 153.15751 145.22390 131.93266 183.46904 129.32305
#> 90  152.83720 142.17837 151.44264 123.67113 153.95447 101.58315 143.66103
#> 91  197.03987 133.29818 128.77875 154.31855 146.98630 226.60390 175.14747
#> 92  193.83524 127.11349 111.35672 120.67248 236.66622 116.00435 139.88100
#> 93  125.72494 150.91592  98.44238 107.09696 143.48053 154.32071 142.23405
#> 94  153.43830 142.52370 119.75502 137.97325 160.65654  93.09015 158.69504
#> 95  146.61384 125.08271 142.44609 154.45212 117.48371 119.77743 158.92419
#> 96  112.85288 119.37676 144.33637 175.42894 162.27820 151.51039 178.00106
#> 97  127.25110 122.60674 169.70888 174.57938 118.13204 122.67062 131.43147
#> 98  146.49918 129.14762 128.14370 122.09500 138.90965 143.90766 100.03173
#> 99  124.39698 134.56614 166.94060 137.18588  91.57182 137.13182 138.72722
#> 100 123.71587 158.93609 122.47059 144.65338 186.94516 112.30129 123.84731
#>         B.4_1
#> 1   217.49423
#> 2   157.35012
#> 3   141.72645
#> 4   186.18826
#> 5   153.08609
#> 6   156.95863
#> 7   160.28643
#> 8   162.44895
#> 9   162.69431
#> 10  111.17249
#> 11  178.02471
#> 12  156.88417
#> 13  128.87094
#> 14  154.54317
#> 15  162.78920
#> 16   90.71507
#> 17  125.20242
#> 18  107.98024
#> 19  135.85654
#> 20  138.72182
#> 21  116.38018
#> 22  105.08946
#> 23  157.27343
#> 24  114.72528
#> 25  153.80528
#> 26  201.31403
#> 27  155.67243
#> 28  144.29744
#> 29  173.04511
#> 30  224.14921
#> 31  127.11736
#> 32  149.32924
#> 33  127.71632
#> 34  139.78451
#> 35  127.25084
#> 36  123.83424
#> 37  134.88267
#> 38  137.62581
#> 39  115.49975
#> 40  172.26128
#> 41  156.82655
#> 42  163.56183
#> 43  138.70979
#> 44  144.22139
#> 45  132.26452
#> 46  145.54962
#> 47  152.04125
#> 48  168.60369
#> 49  109.45263
#> 50  119.17514
#> 51  120.74405
#> 52  167.64558
#> 53  218.03196
#> 54  137.68987
#> 55  155.95512
#> 56   92.64480
#> 57  160.59771
#> 58  184.25862
#> 59  172.30567
#> 60  161.78509
#> 61  172.57403
#> 62  197.35969
#> 63  157.15212
#> 64  182.58579
#> 65  131.81121
#> 66  165.39486
#> 67  125.13641
#> 68  187.03880
#> 69  176.67131
#> 70  173.06728
#> 71  158.19880
#> 72  118.16368
#> 73   96.40918
#> 74  162.21981
#> 75  201.97298
#> 76  155.93782
#> 77  147.33223
#> 78  214.70715
#> 79  210.02477
#> 80  165.63358
#> 81  132.80890
#> 82  156.86612
#> 83  164.30299
#> 84  132.41369
#> 85  140.52540
#> 86  162.67917
#> 87  149.84701
#> 88  137.30591
#> 89  146.14537
#> 90  183.14646
#> 91  166.82989
#> 92  117.45477
#> 93  145.59413
#> 94  158.48124
#> 95  222.62681
#> 96  125.68513
#> 97  143.45555
#> 98  111.11247
#> 99  159.28253
#> 100 126.53164