Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10,10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(subset = rep(c("A","B"),50),
                   level = rep(1:4, each = 25))
rownames(meta) <- 1:100
scRNA <- Seurat::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA, by = c("subset","level"),
               min.cells = 10, max.cells = 20)
#>        A.1_1     B.1_1     A.2_1     B.2_1     A.3_1     B.3_1     A.4_1
#> 1   152.5793 154.70898 161.64308 155.73501 116.67698  86.69242 165.49154
#> 2   170.9509 134.25023 139.93874 109.52877 123.41906 142.99626 116.59687
#> 3   165.4551 191.48198 100.93216 158.71410 152.28871 140.13548 141.24324
#> 4   177.6789 141.71850 143.78117 135.79730 139.00176  87.63355 115.76151
#> 5   123.2645 126.12897  87.17963 142.58954 155.63759 102.28278 111.99304
#> 6   166.9194 143.68708 182.19785 138.89025 149.42839 238.38224 127.15185
#> 7   160.6259 141.48948 147.97911 203.42505 131.99739 155.39445 132.69588
#> 8   132.9810 140.18869 128.16852 132.25185 192.56630 140.24531 165.78166
#> 9   171.7710 129.52831  97.80143 125.80446 224.51392 127.79578 147.13013
#> 10  171.7816 114.86770  83.56148 143.99083 131.52616 157.28439 170.24523
#> 11  141.6311 108.09559 144.58169 179.65936 177.09096 106.66361  92.66716
#> 12  123.7084 180.19643 100.03777 117.15161 173.94436 119.53601 186.24690
#> 13  184.5701 126.78898 122.91555 116.53683 204.46838 148.32634  87.41472
#> 14  158.0311 195.52704 176.14524 168.26636 152.42645 138.43642 110.27542
#> 15  163.8752 160.75637  96.74198 135.62485 174.94515 196.95159 130.36182
#> 16  158.2010  63.80887  89.35030 163.69701 180.40183 124.25417 163.98506
#> 17  160.6970 162.36685 125.71431 149.74897 189.85921 103.96912 134.91798
#> 18  121.3010 185.06035 121.76572 162.91386 147.54639  88.48832 125.61524
#> 19  132.9281 102.76205 137.23688 119.22975 155.35287 135.25277 167.32832
#> 20  174.9550 127.67797 100.99713 162.85002 137.43591 118.69655 125.66728
#> 21  156.6167 128.05101 146.12918 177.90117 157.80850 114.87415 174.86814
#> 22  126.8638 156.52997 103.03584  76.05362 170.90622 148.69983 142.34092
#> 23  155.7220 197.29179 107.06986 196.81976 206.51819 152.99540 121.68943
#> 24  162.2192 127.25363 107.69140 130.65684 163.56716 123.10804 139.98079
#> 25  158.2280 136.56048 150.64867 195.52303 165.78763 114.11980 144.59693
#> 26  137.8865 174.04246 202.00569 191.85343 173.71580 126.58371 194.28362
#> 27  120.8399 110.78515 165.53803 137.95490 113.72822 129.37078 182.72025
#> 28  169.4586 170.63319 134.40658 153.18766 193.00683 119.28075 116.54720
#> 29  130.7772 124.17319 129.61569 145.15095 202.33447 159.08950 133.09443
#> 30  149.7939  95.27320 146.50639  98.50768 185.74247 150.57649 183.63777
#> 31  171.5680  98.45456 102.36248 154.96709 149.94561 110.60301 153.36401
#> 32  160.2329 179.80427 139.66141 116.37772 176.87817 113.26813 181.04250
#> 33  166.9601 146.56743 113.29035 105.51467 136.39936 144.96660 128.29987
#> 34  146.3602  95.33530 161.35153 109.15007 137.52583 163.38817 131.34305
#> 35  171.1417 140.86035 159.84902 155.47589 162.77317 145.49020 113.24341
#> 36  114.3873  97.82461 172.81638 167.56754 137.28446 138.89122 121.07513
#> 37  143.7132 143.80245  96.60136 158.03833 145.02061 154.89538 171.25225
#> 38  159.5156 103.05503 131.00488 133.15729 115.56873 164.68537  84.44654
#> 39  143.1842 118.82287 162.98873 132.48974 150.44429 145.40715 137.22362
#> 40  141.2162 162.56712 101.53267 164.67526 153.24700 111.94374 148.40407
#> 41  149.8756 138.94773 103.94718 167.53898  78.06410 165.96470 126.96476
#> 42  130.2302 138.09132 126.59557 164.56979 137.98822 170.94248 122.18606
#> 43  166.8109 188.85427 156.81697 164.51992 160.48026 117.04176 190.30696
#> 44  140.9245 136.04937 170.97554 171.90054 176.57251 192.94414 133.15899
#> 45  156.8488 148.63832 162.93336 146.26779 121.21172 142.39940 158.24957
#> 46  164.7557 146.33064 118.96472 157.57796 163.13336 154.08625 175.67007
#> 47  178.2832 208.79945 146.36655 167.39067 146.03528 111.50435 129.58976
#> 48  122.0733 141.18966 110.96915 118.79513 175.92053 120.95639 161.95144
#> 49   84.1651  94.93985 105.03838 102.46563 151.62540 174.29296 195.78320
#> 50  155.5298 143.83841 113.42088 149.47904 161.80109 103.63598 138.43102
#> 51  137.4885 134.48977 149.18389  89.65558 136.46400 119.48565 173.84321
#> 52  117.1410 106.00935 145.58923 151.74343 171.49750 176.89304 130.71896
#> 53  157.0153 192.36116 140.95883 179.18063 162.46204  83.86311 168.30337
#> 54  184.7407 111.20678 150.15066 174.46926 179.62920 108.65778 150.91691
#> 55  147.9528 151.09388 131.60248 169.69798 115.33823 209.34265 118.76838
#> 56  176.0580 128.65055 156.80902 143.22210  95.88061 137.47553 137.26097
#> 57  201.0561 202.36607 134.84054 240.93179  90.28200 150.30842 163.50158
#> 58  167.2862 142.98242 199.19271 164.86582 139.52890 183.60340 183.06518
#> 59  152.5663 170.85506 141.76990 159.48169 151.19817 149.33099 114.25168
#> 60  129.9811 163.89747 110.95878 124.83102  97.50335 162.81642 111.30146
#> 61  127.8431 160.14047 181.00793 164.37168 210.45076 108.04422 192.94546
#> 62  162.4820 136.93425 120.14134 152.96906 122.23538 149.26116 105.66932
#> 63  176.2703 176.78982 135.74075 158.99804 174.77752 134.36287 130.34599
#> 64  137.5122 182.96719 173.06997 192.06442 115.76972 153.81982 168.28719
#> 65  135.7464 131.09863 127.88265 110.79210 120.05628 114.94836 148.81181
#> 66  164.9056 123.84808 123.91376 154.63868 131.47233 151.50204 129.97771
#> 67  157.5228 149.74300 189.01030 182.87422 165.38275 131.50615 155.41343
#> 68  160.6848 120.34415 115.58079 129.45426 160.22654 126.62020 144.70935
#> 69  127.2384 128.10946 122.53916 123.49908 201.56318 194.20995 134.16817
#> 70  189.2782 174.35009 123.46868 114.16912 141.67156 161.21700 127.42950
#> 71  165.6118 182.40339 165.91969 124.80180 169.11456 168.51348 163.78832
#> 72  137.1345 151.66453 194.71787 191.30439 137.18130  93.36053  87.78035
#> 73  129.5664 120.73890  96.36386 142.17850 150.65317 170.01766 154.50298
#> 74  125.1851 148.44646 135.58505 136.00738 139.90174 163.73030 133.72354
#> 75  159.5269 141.84621 145.80058 131.74829 134.55846 164.43055 107.57538
#> 76  146.6648 118.39028 112.11495 140.87371 148.75101 140.55703 220.61503
#> 77  136.4818 209.58264 144.65498 198.60111 150.68682  81.31708 155.93307
#> 78  152.2086 170.47105 155.60257 154.45103 179.65953 119.83088 145.64970
#> 79  152.2676 147.18792 171.70655 184.00838 135.88496 169.99212 104.15732
#> 80  130.6570 196.59218 139.72174  99.63702 164.72192 121.00053  87.66934
#> 81  112.8406 137.88308 138.53008 173.01332 152.08632 125.20510 213.05357
#> 82  109.4323 167.94919 108.52585 172.51150 145.61454 123.59776 101.99283
#> 83  141.6979 189.06154 105.98103 154.77966 158.01453 154.35066 189.61140
#> 84  199.1630 201.39781 171.79144 146.19015 135.16826 107.02128 116.41239
#> 85  170.6984 100.72679 128.48603 186.99700 157.06784 170.84556 116.09460
#> 86  140.5785 119.43450 134.81105 157.33785 143.96763 129.17405 133.63257
#> 87  164.2186  91.86604 131.07333 161.07364 152.95872 108.87739 159.07058
#> 88  151.3007 147.69337 164.60467 111.09754 174.88651 169.44760 169.21811
#> 89  173.2754 160.34913 150.48391 170.16211 163.72491 146.84531 136.45101
#> 90  177.4110 197.26542 132.07352 179.70371 165.26911 156.67766 131.57461
#> 91  148.8910 138.22677 125.40530 145.81367 132.35742 177.27652 139.61724
#> 92  126.2028 153.48895 131.02799 127.81169 172.85734 179.01421 111.29179
#> 93  147.1750 157.86260 151.10244 151.62690 147.21284 192.76412 192.47745
#> 94  167.1451 139.59208 136.51123 136.17537 170.42249 162.39370 151.33499
#> 95  173.6869 143.34351 137.57265 147.47830 149.50579 158.86832 114.66454
#> 96  111.8803 128.95183 151.05519 180.17508 119.06791 137.72174 163.78505
#> 97  124.5424 156.47898 123.36097 135.36155 174.01226 176.93232 119.65779
#> 98  178.1691 118.63490 150.15745 152.35296 134.46192 117.64904 138.64763
#> 99  182.9061 143.29906 186.34309 169.34887 129.89848 105.78488 130.31159
#> 100 146.4519 153.66194 119.08463 134.90713 172.71705 147.60471 118.32087
#>        B.4_1
#> 1   137.8377
#> 2   170.7786
#> 3   117.9324
#> 4   189.9936
#> 5   158.4183
#> 6   123.6673
#> 7   173.9122
#> 8   205.7506
#> 9   108.4422
#> 10  112.4950
#> 11  128.5355
#> 12  172.3940
#> 13  179.4709
#> 14  186.6910
#> 15  137.8376
#> 16  111.3853
#> 17  143.2317
#> 18  167.9485
#> 19  138.9267
#> 20  127.5411
#> 21  164.2493
#> 22  113.1634
#> 23  173.8049
#> 24  187.8205
#> 25  147.1208
#> 26  144.7776
#> 27  221.8123
#> 28  129.8259
#> 29  156.3993
#> 30  157.8086
#> 31  182.4907
#> 32  145.4386
#> 33  183.4409
#> 34  171.0381
#> 35  136.3773
#> 36  178.0158
#> 37  165.3912
#> 38  106.1631
#> 39  149.5612
#> 40  119.0741
#> 41  180.8279
#> 42  146.4534
#> 43  195.3117
#> 44  146.8614
#> 45  142.2103
#> 46  129.8177
#> 47  183.7897
#> 48  179.4189
#> 49  121.5982
#> 50  145.0956
#> 51  132.2515
#> 52  128.5047
#> 53  159.8130
#> 54  146.0616
#> 55  161.4795
#> 56  184.6030
#> 57  158.6254
#> 58  141.8878
#> 59  199.8165
#> 60  131.7745
#> 61  147.1302
#> 62  123.4200
#> 63  147.2937
#> 64  121.5794
#> 65  166.9953
#> 66  190.6992
#> 67  147.2165
#> 68  162.2627
#> 69  153.9360
#> 70  171.1826
#> 71  153.2866
#> 72  137.2692
#> 73  153.6421
#> 74  141.2801
#> 75  137.9217
#> 76  177.4449
#> 77  161.6924
#> 78  139.0900
#> 79  188.5816
#> 80  133.1942
#> 81  129.8116
#> 82  117.0339
#> 83  165.1387
#> 84  135.5100
#> 85  185.2662
#> 86  140.7992
#> 87  195.3287
#> 88  155.3808
#> 89  114.8426
#> 90  165.3843
#> 91  154.8274
#> 92  141.7043
#> 93  138.2286
#> 94  129.1065
#> 95  127.2823
#> 96  163.8962
#> 97  164.1413
#> 98  175.3106
#> 99  146.8886
#> 100 146.0345