Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10, 10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(
  subset = rep(c("A", "B"), 50),
  level = rep(1:4, each = 25)
)
rownames(meta) <- 1:100
scRNA <- SeuratObject::CreateSeuratObject(counts = counts, meta.data = meta)
#> Warning: Data is of class matrix. Coercing to dgCMatrix.
pseudo_samples(scRNA,
  by = c("subset", "level"),
  min.cells = 10, max.cells = 20
)
#>         A.1_1     B.1_1     A.2_1     B.2_1    A.3_1     B.3_1     A.4_1
#> 1   191.44304 114.10249 153.37909 146.57290 174.5301 159.90505  93.34917
#> 2   109.79898 124.51993 115.18523 150.29539 122.2098 114.52627 117.24675
#> 3   147.27763 113.66921  96.16048 184.72314 142.8078 145.77499  84.49412
#> 4   134.03342 117.06103 108.00184 203.77652 136.0761 102.60595 121.10066
#> 5   129.79755 156.93089 176.87443 127.22327 213.1971 147.46292 151.27154
#> 6   102.47260 148.89167 159.08827 162.62347 160.6238 162.72362 131.15681
#> 7   125.72613 162.37743 115.03600 153.17001 130.2934 165.88121 150.70401
#> 8   177.50225 148.56633 142.50768 147.27963 165.6890 126.23008 146.94523
#> 9   180.58189 131.35018 151.32681 168.50140 107.6140 126.18498 127.56602
#> 10  150.25711 178.63941  90.30696 149.20170 168.0749 182.97755 133.93723
#> 11  107.88950 153.14873 135.08302 171.50005 176.5765 142.83650 128.21048
#> 12  167.40777 110.53503 114.23605 102.37910 111.9253 150.75786 168.10419
#> 13  130.24888 154.27349 119.58585 153.22251 143.4066 156.18402 180.60048
#> 14  146.43182 132.04300 148.86695 132.30180 156.3449 143.30817  90.83645
#> 15  183.40519 110.16323 171.63956 156.00709 151.4153 117.03110 123.95312
#> 16  137.38584 139.75418  92.40959 155.80533 137.1738 112.71985 134.50341
#> 17  154.34760 118.04330 157.84136 170.01438 144.7467 115.05188 154.42147
#> 18  165.78977 101.40317 136.41598 175.32213 159.6539 105.33066 133.62554
#> 19  123.55215 173.47952 133.50688 205.05783 147.9443 196.71818 179.62018
#> 20  141.89509 149.21513 128.18404 174.18379 177.7161 142.62272 124.78429
#> 21  161.89422 146.76197 156.51057 149.28564 154.3234  94.58553 186.84083
#> 22  109.19229 104.65942 131.22112  72.02998 164.3827 163.11762 139.30901
#> 23  168.77850 150.27734 161.09107 163.88891 153.8307 107.43463 117.04532
#> 24  169.78388 125.04028 141.73676 121.54066 157.7752 152.62036 133.65522
#> 25  131.45854 185.07265 132.65864 138.20941 159.8433 144.59759 106.75442
#> 26  206.01796 126.47139 174.12073 148.05196  94.3650 166.59889 120.56604
#> 27  211.79352  79.23535 167.10948 118.66381 159.0140 151.38181 151.49937
#> 28  179.36420 141.06780 137.12690 188.08913 157.5910  90.72566 173.51233
#> 29  144.08700 156.85545  86.71435 107.46827 185.4207 129.77925 140.75842
#> 30  114.69045 145.27746 136.40936 123.41750 125.4356 101.12883 137.91793
#> 31  123.68177 172.78934 147.33825 176.42603 170.0844 131.37250 114.38870
#> 32   95.12202 140.79330 143.04850 128.04329 130.1133 147.25248 100.00736
#> 33  135.99805 172.57316 163.79647 109.65874 138.1937 101.86223 120.21768
#> 34  166.66220 125.58842 147.09842 142.21378 156.6715 113.27252 126.89527
#> 35  103.82072  97.40614 109.13121 108.17475 151.8634 197.14408 159.12530
#> 36  134.07895 155.56714 134.62450 120.95729 151.8708 183.63378 124.67438
#> 37  157.16875 115.75447 151.56032 112.11562 124.1311 167.59472 167.81142
#> 38  132.45432 118.15251 172.00990 113.00108 188.2037 103.77932  96.21958
#> 39  201.66595 128.38933 136.96378 130.82255 120.6101 105.52627 186.05044
#> 40  196.63031 157.83563 148.82027 193.20722 128.4300 166.62689 104.48288
#> 41  125.21922  97.61492 119.82983  99.24984 204.4109 134.20765 130.74018
#> 42  110.20218 114.99832 126.96788 189.85016 156.9217 145.54993 128.90274
#> 43   92.52837 133.57546 168.04470 134.85660 156.8341 158.18128 181.53687
#> 44  144.72800 148.60912 116.81239 101.09141 144.3243 137.35119 104.86631
#> 45  176.24210 126.98844 128.75056 106.20841 195.9534 147.20343  94.57859
#> 46  144.03638  98.18225 135.47259 135.72103 167.1565 142.83420 134.07110
#> 47  103.67581  99.82198 181.44560 172.85292 147.5702 151.12196 154.34693
#> 48  161.63389 118.21837 154.56085 164.22324 112.9868 138.16023 157.17812
#> 49  116.27745 179.98486 157.95461 235.62547 184.1973 161.41847 143.28614
#> 50  103.49470 121.27268 100.72994 201.57193 212.4526 187.44733 192.84481
#> 51  117.06864 154.09742 111.51397 187.60661 145.7043 146.54692 147.64225
#> 52  169.10419 121.37784 130.73316 173.65169 128.8841 153.02867 172.82933
#> 53  146.96943 135.76972 110.39911  94.27134 157.6530 121.83977 138.12599
#> 54  136.68204 155.96360 177.58456 152.49392 240.1635 133.04329 121.43127
#> 55  181.43863 160.37939  77.89033 170.70710 129.4716 140.25717 127.86257
#> 56  153.41504 155.93021 121.64167 151.39386 165.1731 123.58725 144.21397
#> 57  149.01063 133.01315  96.05622 160.75611 204.9709 167.37172 169.07030
#> 58  207.21793 139.16170 121.57179 150.59095 108.4895 103.26812  90.55439
#> 59  111.29769 106.87583 130.52868 152.96360 158.2077 119.69147 115.49723
#> 60  181.18040 113.60799 165.43254 143.33198 187.0781  95.58694 150.13215
#> 61  100.52766 103.40296 159.78414 181.54783 161.2223 173.78159  98.38163
#> 62  151.63028 138.25333 141.39735 138.91354 215.7509 131.75137 103.58076
#> 63  102.12816 154.68542 138.72690 149.66420 129.2445 173.15897 122.29159
#> 64  127.80158 107.47057 160.40868 122.15889 148.3884 126.21084  93.78899
#> 65  141.36039 135.77057 152.41337 113.47856 139.8644 124.50726 172.05934
#> 66  143.29567 126.65225 138.24519 158.62760 129.0881 121.69142 123.68314
#> 67   97.61386 128.54093 159.04057 225.38278 136.2571 131.03045 203.05796
#> 68  190.77691  66.43889 159.02675 147.79575 140.7040 190.79108 149.50314
#> 69  116.70810 149.82553 127.63221 160.80290 189.6694 162.57382  80.64927
#> 70  182.69875 139.31196 149.24013 148.77462 137.9483 114.58495 142.48682
#> 71  185.34379  87.14711 156.26145 138.62351 129.8019 121.50143 148.11682
#> 72  123.07016 168.90414 118.73125 203.24321 108.1045 117.56821 190.77383
#> 73  179.14788 174.06600 120.89130 110.69100 166.4921 137.57839 144.88396
#> 74  128.46537 144.99448 148.68093 187.84214 143.7425 147.80957 149.11068
#> 75  152.41671 171.90600 147.13687 183.38332 164.0133 131.03496 108.91340
#> 76  120.44559 113.68977 195.73552 114.20401 186.8475 143.43043  72.63823
#> 77  112.33095 151.06611 149.19175 156.81686 172.1196 124.91104 132.40402
#> 78  174.01794 115.38247 166.71968 140.86180 150.9009 167.03244 155.59414
#> 79  128.84652  96.55443 103.75597 131.17261 155.3527 149.72465 100.27502
#> 80  184.95370 123.12095 107.67415 173.28105 156.8663 159.18418 125.52324
#> 81  175.93199 151.96321 179.97964 164.85580 154.9957 196.84023 118.70510
#> 82  195.78370 119.37191 113.79923 169.60728 142.1339 119.53175 133.75364
#> 83  157.29683  97.56617 109.17456 146.77282 134.4595 153.56267 151.61021
#> 84  182.79771 169.28896 145.31173 133.91981 139.2534 107.39061 167.04902
#> 85  189.44828 150.47906 158.36662 125.20028 125.2247 111.39216 143.80948
#> 86  201.75169 153.14302 123.84266 181.02722 126.9603 142.05314 142.71403
#> 87  149.61254 135.38403  93.78493  96.17314 112.0617 257.70371 134.21203
#> 88  179.45752 160.13843 150.10167 145.93523 193.9516 152.32311 127.48785
#> 89  206.72879 143.62876 108.18936 137.34696 170.9874 147.50884 144.84954
#> 90  181.92511 112.68916 162.83486 128.54230 129.1425 119.99550 138.20907
#> 91  131.24203 189.77345 155.81880 131.63109 160.1085 122.72607 164.75590
#> 92  181.19532 142.02844 112.83429 204.06894 221.8077 116.51171 143.43877
#> 93  136.10394 135.03387 123.83827 161.75317 173.1507 137.90202 130.06392
#> 94  139.86056 151.00356 112.02401 144.72152 199.1752 162.01555 180.82041
#> 95  150.56044 140.84017 162.83859 154.48571 161.6847 177.06364 134.19304
#> 96  147.16670 120.04959 145.06142 215.64955 170.5077 125.57297 158.94685
#> 97  210.56268 125.98447 178.14785 198.55915 125.4472 142.84067 120.89458
#> 98  187.92320  92.08809 117.81817 143.41757 122.0555 169.24688 147.52183
#> 99  149.10921 155.12860 223.54541 148.44580 187.4752 134.89962 202.99619
#> 100 156.78869 128.62491 104.36452 169.97442 109.2426  79.25579 139.52791
#>         B.4_1
#> 1   105.82421
#> 2   177.37824
#> 3   161.13589
#> 4   189.73173
#> 5   110.96501
#> 6   178.35269
#> 7   117.77957
#> 8   199.10824
#> 9   156.99089
#> 10  160.53105
#> 11  111.67397
#> 12   94.00559
#> 13  165.89597
#> 14  142.93828
#> 15   93.40791
#> 16  163.60504
#> 17  183.92346
#> 18  105.30665
#> 19  157.12897
#> 20  156.21246
#> 21  130.83430
#> 22  107.31571
#> 23  142.34581
#> 24  168.50394
#> 25  168.82950
#> 26  142.24006
#> 27  193.92784
#> 28  103.75949
#> 29  180.15132
#> 30  147.57920
#> 31  144.30266
#> 32  136.50165
#> 33  117.40274
#> 34  204.44043
#> 35  139.82475
#> 36  136.74319
#> 37  120.38839
#> 38  181.79673
#> 39  138.97349
#> 40  191.33966
#> 41  150.97981
#> 42  183.28218
#> 43  122.60017
#> 44  183.09100
#> 45  161.03919
#> 46  162.97337
#> 47  140.86585
#> 48  148.96677
#> 49  165.12373
#> 50  158.96806
#> 51  105.32286
#> 52  224.68020
#> 53  147.11270
#> 54  163.75031
#> 55  120.99774
#> 56  168.10549
#> 57  156.01966
#> 58  142.81159
#> 59  113.52186
#> 60  170.32799
#> 61  198.23814
#> 62  140.23070
#> 63  164.10552
#> 64  189.82269
#> 65  144.71176
#> 66  115.15564
#> 67  181.55439
#> 68  143.42053
#> 69  174.02937
#> 70  193.09833
#> 71  133.83115
#> 72  108.27502
#> 73  162.62284
#> 74  166.74800
#> 75  192.80740
#> 76  139.05948
#> 77  155.55973
#> 78  150.05431
#> 79  159.20245
#> 80  132.94499
#> 81  176.35198
#> 82  159.71687
#> 83  159.01248
#> 84  178.11447
#> 85  171.40103
#> 86  110.59247
#> 87   79.24923
#> 88  163.11062
#> 89  194.55072
#> 90   82.08710
#> 91  146.54832
#> 92  176.74902
#> 93  105.28140
#> 94  153.62085
#> 95  150.32718
#> 96  167.02859
#> 97  162.81641
#> 98  155.35806
#> 99  130.18838
#> 100 174.86036