Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10, 10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(
  subset = rep(c("A", "B"), 50),
  level = rep(1:4, each = 25)
)
rownames(meta) <- 1:100
scRNA <- SeuratObject::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA,
  by = c("subset", "level"),
  min.cells = 10, max.cells = 20
)
#>         A.1_1     B.1_1     A.2_1     B.2_1     A.3_1     B.3_1     A.4_1
#> 1   197.58030 153.43386 139.41709 124.30041 153.74016 143.70021 136.38483
#> 2   154.50603 126.63339 105.12937 140.41964 140.72246  81.39627 162.63623
#> 3   173.68042 138.22548 212.36880 129.75402 157.04603 110.88553  82.53463
#> 4   150.60530 167.39174 134.30506 140.78145 156.93597 132.10262 147.72604
#> 5   115.51014 115.22003 129.90295 141.00492 173.35803  87.24945 145.88190
#> 6   146.36667  81.43772 194.36954 158.36770 169.32287 114.28687 165.12895
#> 7   127.63207 146.76655 153.60491 133.12454 158.47195 124.57491 128.76140
#> 8   152.32544 119.95059 158.74256 198.41699 133.91809 137.12103 130.61249
#> 9    86.04143 123.12682  69.63551  88.67178 176.90222 123.69886 158.44537
#> 10  194.37597  91.27201 155.84955 143.92493 136.32662 162.03481 151.65006
#> 11  109.40309  89.81698 132.00363 187.13876 121.23611  98.20033 158.00547
#> 12  188.92401 113.10566 125.10121 164.15748 141.63844 157.00248 158.42072
#> 13  170.35381 146.84307 151.08571 143.18014 137.27708 146.44807 140.77133
#> 14  118.46532 140.62283 120.06290 150.65881 150.41091 157.67739 180.68925
#> 15  134.61148 111.67164 119.39681 201.54350 164.78323 131.48263 171.84990
#> 16  100.39442 168.56497 126.52225  98.63952 214.87155  87.40812 159.40104
#> 17  165.36083 130.95058 137.46357 132.38237 239.66270 135.77537 129.87011
#> 18  136.59315 113.59769 144.67125 140.47172 155.83266 113.02884 161.86161
#> 19  179.77974 176.32784 150.39991 121.28812  65.81394 144.76306 135.33647
#> 20  184.96812 164.04567 138.01678 157.83092 146.28336 216.20838 151.21624
#> 21  137.76844 160.96966 182.41654 158.64554 190.92999 155.39972 134.03437
#> 22  110.68444 110.94612 146.14467 150.82703 156.52486 165.98653 151.94531
#> 23  132.18965 189.57383 178.04717 147.97894 129.98944 134.61684 188.98479
#> 24  114.69002 112.34114 190.41184 109.45921 171.01289 116.76833 125.33659
#> 25  181.51162 114.95352 164.91992 161.25023 166.58820 120.77290 168.27701
#> 26  159.57757 137.32184 159.86506 181.00644 160.97271 123.58993 121.94445
#> 27  211.26602 127.64782 127.33581  68.84096 152.25731 169.73352 159.28639
#> 28  141.75925 163.11809 113.33236 126.03748 116.29263 145.06794  81.33809
#> 29  170.66261 112.70449 106.42581 171.00097 160.05296 117.25207 120.80519
#> 30  123.93540 147.45445 148.86846 138.26440 136.66068 132.22411 120.50757
#> 31  111.41571 148.68387 154.75922 100.21503 154.03988 188.39202 104.09985
#> 32  190.11754 110.16377 161.23914 129.48629 162.53948 132.59630 150.02797
#> 33  200.98356 131.84283 127.04848 109.93938 180.30215 117.75388 145.95303
#> 34  156.13927 132.53419 142.45654 191.04622 126.19091  91.15258 140.19466
#> 35  123.75113 150.63605 172.36950 150.50899 149.12643 129.71799 153.13750
#> 36  136.27794 170.34970 149.64765 150.23261 130.45193 177.50325 145.48182
#> 37   94.60103 177.38134 102.61336 112.71391 133.29533 108.17778 139.62657
#> 38  154.32311 123.77342 127.25890 158.44717 186.25208 158.60915 119.44168
#> 39  157.85551 138.58919 121.32547 182.65758 129.32632 141.45023 147.78111
#> 40  124.71912 220.95547 124.81586 199.23060 150.77094 168.77275 131.79863
#> 41  156.86540  93.90797 117.45258 166.89710 130.58761 122.02567 119.00939
#> 42  106.98406 123.93125 132.26821 176.95645 172.80062  81.61448 152.70172
#> 43  215.12856 144.21615 135.15908 144.33907 179.53507 155.18119 134.86112
#> 44  173.12761 118.33191 104.80971 153.09017 121.24788 105.25819 165.88794
#> 45  167.03954  85.70702 161.30051 178.99641 123.07552 137.62987 159.07490
#> 46  128.36089 144.62287 145.33302 216.66922 206.87767 136.40931 183.20786
#> 47  120.78182 146.99124 136.24253 155.68018 202.48415 157.07868 126.39210
#> 48  191.67267 194.95732 146.46684 131.20230 157.01327 152.84522 185.54514
#> 49  147.17424  99.11917 131.66313 148.54458 135.80530 146.33240 131.50573
#> 50  147.09586 146.20258 148.96887 142.88287  93.02347 101.49638  96.17223
#> 51  153.33951 186.03849 150.44725 135.74152 123.08318 132.19002 116.77155
#> 52  129.76183 144.93025 133.69555 129.28949 133.72872 171.75833 106.65868
#> 53  109.52244 178.97689 136.49418 129.28910 164.08032 138.74810 198.19147
#> 54  127.45051 106.37207 138.78720 124.59262 126.21102 158.61976 100.33261
#> 55  195.03201 175.07419 102.74378 170.76482 106.25596 139.35388 148.32220
#> 56  203.08732 160.85434 131.25342 162.37853 146.72575 141.14126 123.30929
#> 57  142.09935 188.03901 121.72230 160.07924 145.24900 133.82630 122.74268
#> 58  159.70875 186.10780 136.81874 145.23136 113.27233 129.71095 118.70692
#> 59  158.81278 116.20003 124.67658 164.40925 116.28426 151.18691 147.21998
#> 60  154.28908 150.22273 157.84675  85.67803 215.22488 110.10488 124.53555
#> 61  160.86577 115.94300 112.17828 148.82903 164.16433 148.36389 167.22540
#> 62  193.83175 146.20285 147.74458 169.11036 190.91021 164.25671 107.59546
#> 63  136.12348  90.09602 154.28291 182.61981 148.22557 190.59521 148.81002
#> 64  157.20182 139.89343 125.55384 137.14680 137.35588 117.50684  95.44558
#> 65  150.59521 124.89725 179.84158 165.08596 129.53079 152.05325 161.05566
#> 66  125.78708 137.20663 169.05503 133.30538 131.82689 137.41517 112.04228
#> 67   76.19386  89.57076  84.14026 113.59213 145.54267 103.95406 128.35121
#> 68  148.46585 161.88947 134.97267 171.85305 139.30483 147.68055 141.11527
#> 69  161.17568 177.37758  84.10323 161.27224 183.44451 100.72228 160.87163
#> 70  117.48900 172.61571 100.95440 169.61133 154.67814 178.99396 180.36137
#> 71  118.95930 117.27087 149.28167 129.56065 121.88087 139.33128 140.94702
#> 72  143.10114 162.44876 128.67663 164.07403 151.21628 128.96765 120.38641
#> 73  139.86703 119.55060 111.33447 183.43227 141.61429 146.63868 164.07967
#> 74  148.35769 108.64261 111.79618 157.33011 151.27569 108.71394 117.33284
#> 75  174.86394 117.48020 131.00354 147.77834 127.22628 142.90153 160.72389
#> 76  147.41586 125.91968 144.73735 121.73558 116.96099 150.36586 106.81888
#> 77  145.52816 134.79935 104.80850 149.71610 147.90943 121.46389 119.35215
#> 78  140.11284 140.08558 175.21538 129.97675 206.26720 106.44983 153.63808
#> 79  171.85738 173.34070  93.81711 128.56789 138.25451 190.94058 102.86597
#> 80  194.72280 137.17157 204.47909 136.25732 186.72347 132.98880 157.64081
#> 81  177.61912 178.42520 115.24595 157.62064 165.52387 149.46010 185.94760
#> 82  159.63527 151.82747 118.15759 163.25075 151.76800 100.47006 136.93163
#> 83  185.08267 134.23242 150.03986 143.55581  89.07886 126.33647 162.99996
#> 84  188.04413 112.17002 204.25713 138.74906 140.65635 146.23808 187.26348
#> 85  158.23064 150.53492 220.61337 151.38311 146.26416 164.65466 160.04095
#> 86  174.70822 131.02855 192.33465 145.87338 151.82613 122.42395 146.44591
#> 87  126.98407 148.62740 122.66137 149.97494 195.22549 168.19765 113.85717
#> 88  138.86130 157.80736 178.36476 133.93151 182.24219 175.47957  86.55256
#> 89  153.98528 147.66105 192.28063 157.05960 201.35030 183.89466 178.54239
#> 90  172.78561 136.81380 119.06679 131.17066 143.16445 118.53739 175.32915
#> 91   96.79533 115.52788  97.16317 109.54122 148.84208 130.23299 108.22194
#> 92  186.90708 106.28526 136.44876 160.96510 155.54415 147.50928 169.81629
#> 93  123.15308 144.35776 158.55691  83.99098 225.02181 157.18090 120.63687
#> 94  186.32043 123.14952 157.90806 113.15666 101.71309 176.79783 141.05291
#> 95  131.61360 208.84192 120.03610 158.84743 178.16573 144.81470  94.12867
#> 96  131.49690 113.00731 118.93928 104.78564 160.54858 104.89413 109.76015
#> 97  209.88983  69.32918 147.25407 143.00014 177.12749 128.54688 147.36419
#> 98  142.54448 158.40677 114.30866 151.03628 104.71537 132.76089 182.53186
#> 99  151.30540 129.00962 151.42114 159.28631 169.31098 105.70312 122.01655
#> 100 147.76654 103.51699  86.37677 180.55076 100.88660 140.17871 107.35538
#>         B.4_1
#> 1   147.67269
#> 2   113.13603
#> 3   130.93513
#> 4   131.98773
#> 5   199.79958
#> 6   146.79303
#> 7   148.27663
#> 8   176.36719
#> 9   126.92313
#> 10  123.08892
#> 11  193.82474
#> 12  167.59225
#> 13  105.90657
#> 14  131.12224
#> 15  162.18430
#> 16  137.82838
#> 17  160.25478
#> 18  165.88531
#> 19  159.31569
#> 20  172.07883
#> 21  142.28220
#> 22  132.04662
#> 23  175.49075
#> 24  193.13560
#> 25  158.10317
#> 26  137.80367
#> 27  147.63964
#> 28  149.80210
#> 29  109.01289
#> 30  151.20543
#> 31  123.75877
#> 32  126.70730
#> 33  144.36570
#> 34  124.10860
#> 35  159.12639
#> 36  116.18261
#> 37  118.55781
#> 38  126.77649
#> 39  135.94572
#> 40  183.98942
#> 41  154.38322
#> 42  168.98264
#> 43  205.14167
#> 44  232.71765
#> 45   94.46652
#> 46  139.58745
#> 47  120.40174
#> 48  148.03314
#> 49  158.54014
#> 50  158.82312
#> 51  198.24861
#> 52  200.11009
#> 53  143.27372
#> 54  151.04821
#> 55  191.27541
#> 56  160.86486
#> 57  159.49188
#> 58  135.88705
#> 59   83.46368
#> 60  164.04460
#> 61  217.92607
#> 62  149.44155
#> 63  199.43217
#> 64  133.56261
#> 65  195.30835
#> 66  128.37758
#> 67  110.93484
#> 68  120.86362
#> 69  138.03710
#> 70  130.60362
#> 71  132.81256
#> 72  181.02886
#> 73  119.61537
#> 74  122.11917
#> 75  148.90368
#> 76  132.58456
#> 77  153.89999
#> 78  133.84070
#> 79  105.88406
#> 80  135.17347
#> 81  134.02555
#> 82  127.58178
#> 83  176.01537
#> 84   90.74559
#> 85  144.55275
#> 86  164.82999
#> 87  109.75163
#> 88  173.91027
#> 89  171.86996
#> 90  158.64797
#> 91  158.76564
#> 92  129.62184
#> 93  123.80993
#> 94  227.77432
#> 95  156.51808
#> 96  178.18304
#> 97  198.99978
#> 98  135.78133
#> 99  160.53462
#> 100 178.90202