Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10, 10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(
  subset = rep(c("A", "B"), 50),
  level = rep(1:4, each = 25)
)
rownames(meta) <- 1:100
scRNA <- Seurat::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA,
  by = c("subset", "level"),
  min.cells = 10, max.cells = 20
)
#>         A.1_1     B.1_1     A.2_1     B.2_1     A.3_1     B.3_1     A.4_1
#> 1   124.71742 115.32071 151.29867 129.62820 157.97576 116.84862 126.81054
#> 2   124.02741 140.44938 144.15651 129.80836 120.25018 140.76977 127.38131
#> 3   156.91142 112.82033 148.04560 158.05650 104.33308 143.20736 165.25537
#> 4   102.29224 150.37877 133.03287 108.08735 164.73996 118.51555 113.38954
#> 5   138.83316  92.81574 131.25331 101.27059 130.84765 110.43408  96.11309
#> 6   171.18021 139.77150 112.56453 136.44566 152.40095 119.83221 137.92668
#> 7   135.04536 115.07391  89.15178 120.44877 101.24279 132.90091 153.21248
#> 8   133.79043 131.12752 125.97888 162.17667 123.57095 133.99058 138.60898
#> 9   134.66345 113.96456 167.97744 123.10913 200.94007 202.55508 130.55657
#> 10  126.97102 151.48239 115.20496 124.84054 153.19178 120.35845 156.97476
#> 11  156.18674 128.14177 124.64381 154.71418 112.31863 164.54337 164.11154
#> 12  143.78549 197.08110 118.76508 155.11466 116.61950 136.07573 203.93214
#> 13  184.66597 149.31435 163.47202 103.37031 123.37974 166.84647 156.87853
#> 14  153.87620 125.42636 156.74182 163.36936 122.82805 127.08347 128.38130
#> 15  166.26259 151.75303  95.51639 182.42320 186.01211 202.88418 180.61990
#> 16  185.66275 136.23865 160.42236 160.11934 184.75064 110.17135 128.70240
#> 17  194.48134 133.39273 165.79620 133.25850 124.55777  75.56945 189.44058
#> 18  181.05421 100.71061 149.50137 132.65078 169.88904 146.75336 127.24274
#> 19  156.59691  61.50924 159.06404 178.00269 124.44243 143.61943 147.56564
#> 20  181.25657 169.00701 113.07651 143.46992 131.77158 116.44216 145.30270
#> 21  156.60642 137.77187 169.51403 178.47257  98.80444 105.37250  70.87519
#> 22  118.72923 114.37490 175.18397 125.00851 138.66463 144.00788 144.16660
#> 23  159.53079 179.86709 124.61368 193.49432 198.16663 136.60239 136.32488
#> 24  150.93668 145.04478 154.28461 129.68114 141.23806 157.07886  88.98495
#> 25  149.52128 164.72032 191.98834 155.13275 152.49195 144.45437 157.34942
#> 26  173.06564 137.44187 107.91478 158.22577 199.12019 154.44327 135.61286
#> 27  197.95233 160.16357 203.08600  78.19667 169.82097 113.95513 126.34243
#> 28   95.93264 199.24424 124.80028 121.11025 115.63099 202.09823 133.59915
#> 29   78.72659 148.82302 105.52662 156.77258 190.88562 106.50594 134.09314
#> 30  144.87265 170.67775 103.66413 153.40336 128.16539 104.23611 163.70607
#> 31  156.09039 109.49820 113.23878 181.40148 162.72921 153.51137  65.49269
#> 32  142.45447 133.61874  96.12674 123.69172 164.51859 101.99238 156.87267
#> 33  131.93108  98.59761 167.16044 128.94716 178.30749 147.69698 144.24182
#> 34  127.89402 147.76787 214.90048 178.05871 118.10903 192.86635 130.15898
#> 35  189.39403 137.95630 148.23938 103.64948 137.59006 106.52777 181.41309
#> 36  148.38032 103.12353 142.60290 174.08506 126.32064 129.03371 161.93335
#> 37  140.69463 136.51505 149.53549 164.18702 120.63625  81.91878 172.92392
#> 38  136.26276 137.86047 127.05860 178.08229 140.73131 126.59395 183.03119
#> 39  108.18729 109.73302 156.07199 164.47196 134.27968 123.51426 163.43335
#> 40  162.56513 175.60810 138.16272 120.71437 169.76248  99.35106 152.52714
#> 41  182.50343 122.95698 114.05138 121.00918 140.34080 151.21693 152.39014
#> 42  146.65299 137.01523 136.93086 180.33261 132.91020 119.48310 120.03594
#> 43  171.79044 116.82810 168.62506 147.46724 111.01104 172.32892 158.32107
#> 44  111.16977 123.08265 155.10115 148.77768 170.76231 134.67160 102.19352
#> 45  178.25598 104.41112 139.40279 115.06501 144.95977 161.43651 171.91680
#> 46  131.60141 153.56356 184.30893 158.44379 122.74272 138.90472 165.49818
#> 47  191.37365 110.27472 183.84340 184.80904 176.09004 185.30955 157.63005
#> 48  124.31487 129.05176 145.11302 167.89355 126.15581 137.85320 121.71696
#> 49  154.72772 151.22106 161.04559 152.87060 163.19566 105.84461 177.57782
#> 50  165.57363 161.41299 148.68117 213.87647 162.25952 150.45837 106.32397
#> 51  161.56995 165.66273 149.11979  76.54716 175.49648 168.28735 103.13869
#> 52  151.64560 142.84537 110.77755 121.92204 160.28177 116.39238 142.63091
#> 53  187.36331 125.31025 140.87820 146.79864 192.42521 159.02931 197.17279
#> 54  165.50869 112.97898 156.86811 178.35767  79.71730 154.07350 110.59643
#> 55  157.28001 149.13109 185.32103 124.76715 145.86457 139.43007 116.93956
#> 56  110.14814 105.32770  76.39587 135.00488 149.92700 157.84548 137.08072
#> 57  122.77336 168.19091 127.45092 127.25790  96.75274  97.91596 120.16165
#> 58  149.69819 113.89732 182.52169 149.23402 141.99084 152.51944 142.52136
#> 59  168.06218 159.97202 167.09875 145.40252 133.43224 104.67887 153.58696
#> 60  182.16339 152.01112 119.79253 135.11163 153.69356 127.80537 103.41305
#> 61  169.68732  92.46939 129.36562 183.95032 154.43285 158.31254 125.15462
#> 62  131.76687 126.59250 153.44997 123.13921 136.43468 141.52853 119.75569
#> 63  117.82977 148.39168 112.69140 152.92332  92.46799 167.83593  85.56659
#> 64  188.85996  79.55594 132.58242 160.60352 126.33261 157.49327 124.25170
#> 65  190.89704 151.01384 129.32991  92.89898 154.78430 157.84242 163.07209
#> 66  195.56152 144.22136 119.69988 161.38395 129.53078 128.44461 136.67413
#> 67  119.63664 122.05802 114.39906 189.33367 136.72092 145.97416 149.99176
#> 68  161.57719 133.56508 128.31015 110.25661 134.44633 145.75256 101.80605
#> 69  122.20844 168.92899 128.85873 178.69099 170.52913 105.43161 136.13533
#> 70  179.13197 142.82658 126.76951 198.31851 151.74367 172.61542 147.13348
#> 71   73.25739 170.05058 107.74924 139.72850 129.49606 149.25304 134.99391
#> 72  173.27083 155.57110 150.04438 133.91345 175.41795 159.81194 220.03068
#> 73  120.54966 188.41192 153.50958 186.38056 134.66452 103.83801 140.41389
#> 74  156.82499 132.40119 132.15164 170.62816 161.25107 133.29798 149.48750
#> 75  142.75039 124.39099 194.45706 159.93952 131.59895 144.88592 123.10482
#> 76  142.98685 161.68478 121.90209 153.02263 120.47571 152.23415 113.98705
#> 77  138.96548 113.54905 197.96694 142.01363 162.97472 157.90336 135.70237
#> 78  147.59104 123.65085 120.59474 133.98651 133.63331 170.57810  89.20378
#> 79  201.45680 117.96754 132.39312 181.41037 137.59329 200.17677 112.88550
#> 80  131.29907 183.90589 119.13013 131.14382  88.22079 168.30581 207.61257
#> 81  140.45882 112.39609 158.78379 177.36373 143.84658 103.66473 145.58117
#> 82  124.96800 107.75628 120.78146 131.78569 164.65402 174.86974 110.17934
#> 83  133.02145  88.40776 174.58046 141.95431 141.13057 133.32961 127.50871
#> 84  149.83934 141.58443 122.39776 174.80858 175.72251 103.85202 116.05794
#> 85  158.19510 146.63471 179.68625 150.96698 172.97136  99.52935  97.10164
#> 86  140.43770 154.77100 160.43183 154.94173 145.35312 148.14092 125.52135
#> 87  193.49478 128.00423 179.38032 130.74706 216.69592 141.61472 128.53190
#> 88  131.55233 141.47214 108.84200 192.59462 178.95493 155.07996 191.97579
#> 89  127.66099 155.52194 127.95113  94.53110  96.66380 140.06131 140.50111
#> 90  191.29911 188.50818 144.21884 102.83774 130.84859 133.07238 162.79221
#> 91  172.32822 116.22409 159.14509 179.02204 180.67105 164.33156 149.62989
#> 92  139.39284 158.07803 199.80662 138.98906 151.96876 132.58763 100.59210
#> 93  157.79448 119.58411 140.97339 146.05587 167.50644 145.26611  87.01330
#> 94  112.94177 117.55322 174.10480 130.81888 160.14410 132.08168 147.56208
#> 95  146.80083 157.16769 127.44978 141.70538 152.36766 104.64940 146.77840
#> 96  135.69593 152.89807 161.07326 148.49064 149.01738 142.00613 115.60016
#> 97  144.76415 143.85058 105.55318 190.16092 173.63409 167.24763 130.84499
#> 98  133.99355 142.03709 143.59186 129.24450 174.41294 205.19948 149.52976
#> 99  172.20978 146.27295 184.80861 138.61689 193.07920 151.69833 193.42418
#> 100 118.77856 141.79061 125.91478 146.43603 131.27301 216.20495 121.10706
#>         B.4_1
#> 1   117.55251
#> 2   134.85695
#> 3   109.21884
#> 4   171.10922
#> 5   128.94240
#> 6   121.11235
#> 7   162.75675
#> 8   141.03690
#> 9   127.66240
#> 10  172.21480
#> 11  169.06261
#> 12  166.49520
#> 13  157.37765
#> 14  179.44501
#> 15  137.60225
#> 16  217.97695
#> 17  177.46591
#> 18  106.99058
#> 19  183.23230
#> 20  127.60689
#> 21   93.15949
#> 22  122.38872
#> 23  143.84460
#> 24  183.46532
#> 25  122.16841
#> 26  138.52290
#> 27  207.20931
#> 28  124.37536
#> 29  218.95633
#> 30  176.33637
#> 31  145.34437
#> 32  110.35228
#> 33  161.86877
#> 34  111.60230
#> 35  191.30279
#> 36  177.52420
#> 37  147.09533
#> 38  149.25123
#> 39  178.28328
#> 40  134.32310
#> 41  135.45186
#> 42  168.43892
#> 43  103.11922
#> 44  167.47684
#> 45  175.23494
#> 46  125.90894
#> 47  168.26120
#> 48  211.75632
#> 49  124.85858
#> 50  157.78748
#> 51  130.84469
#> 52  168.26377
#> 53  168.45752
#> 54  175.27509
#> 55  149.23052
#> 56  165.09687
#> 57  214.27818
#> 58  107.45648
#> 59  106.43908
#> 60  123.12598
#> 61  131.32242
#> 62  178.43306
#> 63  149.60429
#> 64  122.59857
#> 65  157.82680
#> 66  153.32361
#> 67  153.77370
#> 68  152.34946
#> 69  106.79953
#> 70  216.03549
#> 71  152.39960
#> 72  139.69553
#> 73  193.28494
#> 74  148.38063
#> 75  139.47937
#> 76  143.11117
#> 77  113.45557
#> 78  190.71180
#> 79  146.70256
#> 80  130.64075
#> 81  153.40447
#> 82  165.20469
#> 83  156.84780
#> 84  169.77912
#> 85  165.70492
#> 86  142.42946
#> 87  157.13775
#> 88  177.13834
#> 89  186.42259
#> 90  199.00181
#> 91  150.95735
#> 92  160.49026
#> 93  166.22228
#> 94  172.97505
#> 95  209.17577
#> 96  115.19221
#> 97  160.53324
#> 98  134.55611
#> 99  184.19050
#> 100 183.80277