Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10, 10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(
  subset = rep(c("A", "B"), 50),
  level = rep(1:4, each = 25)
)
rownames(meta) <- 1:100
scRNA <- Seurat::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA,
  by = c("subset", "level"),
  min.cells = 10, max.cells = 20
)
#>        A.1_1     B.1_1     A.2_1     B.2_1     A.3_1     B.3_1     A.4_1
#> 1   166.8767 106.75023 155.65972 178.48133 128.83151 164.60667 127.57823
#> 2   135.3204 174.16753 135.48697 157.36421 164.49747 120.56625  98.32575
#> 3   184.9898 156.05037 180.61094 168.76495 174.40944 146.33755 111.35990
#> 4   167.2642 119.37599 125.97101 160.90396 172.78975 171.35883 169.35158
#> 5   157.4227 122.16547 188.42308 180.14191 150.15256 133.82190  89.94540
#> 6   147.0981 183.73853 153.69981 118.30281 111.36554 178.34261 154.40204
#> 7   152.1667 207.16284  83.93076 200.90670 153.58375 200.68338 147.77508
#> 8   140.1872 193.49067 156.00105 165.12836 118.49635 112.10444 118.44972
#> 9   166.2639 126.68320 182.77790 115.04659 157.82520 117.00515 139.21783
#> 10  119.5994 127.81793 170.02368 173.53862 141.73272 130.12299 235.16140
#> 11  145.4192 130.86752 135.94142 154.37048 179.31230  91.69319 143.52217
#> 12  115.6712 182.87673 161.69205 184.75643 129.74114 122.59220 159.83734
#> 13  192.8874 173.59670 148.54792 175.10393 128.48255 132.66626 154.78370
#> 14  185.3317 167.07120 147.72458 143.47759 173.54189 118.53820 137.36129
#> 15  163.6608 124.43002 126.56255 132.25215  87.26085 139.37595 142.54763
#> 16  124.3747 139.68055 113.86869 177.73061 128.00766 109.85678 167.14264
#> 17  168.4398 163.37277 100.57514 148.11324 201.02620 123.90323 158.11290
#> 18  176.7542 209.87083 160.52528 134.52333 147.53115 105.45158  81.67908
#> 19  117.8623 153.89570 167.39365 148.70272 148.83238 133.42838  99.75436
#> 20  129.5433 115.83424 149.75790 178.06453 117.83776 158.30687 167.58742
#> 21  163.1957 143.42496 160.97898 147.33544 182.93286 134.22584 184.20930
#> 22  228.6313  87.69571 141.64187 152.57151 154.53429 144.93419 148.00775
#> 23  188.2433 121.53083 122.08187 173.81430 143.21417 139.79923 153.36344
#> 24  144.9519 174.63591 126.04086 127.42211 117.36917 133.17237 116.09116
#> 25  131.5932 161.41525 134.00314 179.87053 156.23423 199.13911 113.86261
#> 26  219.5129 177.56153 114.51066 172.62484 151.60596  96.23309 165.51558
#> 27  130.2409 115.18846  97.29473 104.17410 145.44879 144.66897 157.96250
#> 28  165.2232 155.81272 118.41302 134.61070 108.36297 160.02743 151.00196
#> 29  188.9772 111.09386 173.04360 152.88985 112.27764  97.36728 123.26690
#> 30  135.9686 177.42730 156.63737 120.35803 139.97691  92.65559 106.66623
#> 31  144.0605  99.70229 160.85895 153.84919 186.38814 118.54146 137.16833
#> 32  142.6004 116.51933 114.79310 153.74745 119.94425 167.81795 122.43661
#> 33  155.2259 169.13426 121.03156 175.83836 153.98829 159.04411  92.19302
#> 34  158.0558 151.51422  93.47463 157.91237 148.23514  90.19218 189.14051
#> 35  153.4287 127.43085 201.13205 149.77972 164.20820 124.29023 187.30283
#> 36  144.8111 175.40188 123.67451 190.82619 142.40778  93.91061 131.92152
#> 37  120.8330 133.07378 155.88226 120.76655 150.51678 108.46222 138.32594
#> 38  153.0729 186.17772 152.35862 130.75397  93.55287 164.26245 152.75084
#> 39  166.6093 109.73896  93.73833 180.11889 154.76586 145.78440 130.50088
#> 40  148.6795 123.32578 161.36133 159.10486 183.63057 132.54572 124.33470
#> 41  133.2596 155.28689 117.16713 135.33126 132.22589 203.40639 148.40680
#> 42  193.5753 123.93871 112.72629 124.88058 155.45158 141.59167 137.58176
#> 43  193.3594 133.43579 133.15058 155.31798 171.00883 148.41779 104.79880
#> 44  140.2924 106.67756 153.61200 173.55972 163.14110 109.96481 112.66018
#> 45  145.4649 123.73587 144.77262 139.31258 156.05790 122.80839 100.32225
#> 46  145.7186 129.71530 197.15164 121.82360 152.93648 137.72185 128.30690
#> 47  180.3765 187.01434 179.07003 157.79207 184.81941 148.82423 156.25217
#> 48  194.1763 129.90717 140.13920 117.19913 111.89024 168.88566 162.73895
#> 49  112.5097  99.21381 135.33716 114.91932 135.40709  95.01898 124.68806
#> 50  195.3528 166.88404  90.90672 128.50000 146.83074 123.00461 175.58674
#> 51  137.8834 132.93427 153.08736 157.12129 121.35170 140.10184 126.38784
#> 52  134.7955 136.17213 140.04795 154.98492 155.51183 104.08806 142.84650
#> 53  105.4633 146.36224 125.30168 197.74893 165.94676 115.86019 138.25735
#> 54  102.9474 169.75993 186.97606 193.80712 107.99730 140.44366 138.53430
#> 55  131.1274 178.86412 132.39149 174.30335 175.91693 162.41160 131.91409
#> 56  134.0737  98.60696 137.34390 130.54162 116.50687 108.73514  81.80572
#> 57  112.9193 164.24406 198.85061  90.91216 152.00332 144.93191 179.09666
#> 58  130.1368 131.88710 121.13685 128.80929 233.12481 160.00339  81.50527
#> 59  129.1856 174.47569 101.99612 141.78522 213.09645 158.19887 127.36934
#> 60  102.1299 111.21650 165.35788 205.80262 210.98017 185.09351 108.79811
#> 61  150.7400 144.16948 106.56636 158.11270 173.87509  81.79331 164.54605
#> 62  158.8054 166.42370 145.73967 151.17065 170.83049 130.63116 140.98782
#> 63  132.0840 186.99457 121.24641 109.78568 132.62988 129.12789 133.52095
#> 64  187.8196 205.23467 105.55742 174.32669 145.47945 162.09090 149.62785
#> 65  104.0434  90.56771 161.08502 169.18841 139.80332 141.66628  93.54807
#> 66  163.0014 166.36545 180.20349 174.96206 212.70164 195.64326 147.36881
#> 67  135.8057 160.76370 153.27078 100.76505 168.98013  93.25424 170.06293
#> 68  147.6502 200.27188 110.35206 201.29845 177.86317 145.35527 184.73609
#> 69  152.3274 160.38702 178.64568 137.76864  94.45704 109.05564  97.18731
#> 70  133.2374 186.54881 162.03867 132.07694 114.76475 124.02067 162.43686
#> 71  121.3664 121.33960 101.13382 200.83275 159.32409 136.70040 115.17673
#> 72  154.9301 130.51325 142.81672 148.54801 142.28154  98.64400 143.25075
#> 73  129.4037 175.58790 111.84001 175.28437 108.09804  98.68236 101.78634
#> 74  140.7792 139.59005 185.45541 181.45872 215.87619 140.65774 174.47393
#> 75  154.6836 156.79531 160.15730 183.38982 201.31009 143.20943 140.70509
#> 76  135.9720 127.05053 151.83003 219.28649 138.17125 190.29637 118.79406
#> 77  155.1293 102.62055  99.62332 172.89801 157.72687  94.48981 146.92484
#> 78  119.6677 125.03925 131.82697 142.88617 157.68751 143.68621 117.28236
#> 79  183.4486 145.21401 110.25541 149.50680 165.79666  87.13966  92.43990
#> 80  194.3933 151.84353 163.70850 147.84845 157.79031 168.18795 123.44207
#> 81  118.3272 182.46450 119.29240 152.51825 131.31130 140.61705 165.07310
#> 82  176.9719 125.74626 150.61582 127.20302 153.96919 166.57710 150.71251
#> 83  230.7068 179.14462 100.79067 133.02185 185.20039 131.35915 133.68678
#> 84  225.2536 158.77136 144.80310 154.40390 176.97290 147.57221 137.12925
#> 85  157.1492 109.19126 157.51886 184.79902 133.01818 142.57556  95.69180
#> 86  125.7550 136.63449 127.17690 190.99206 193.56855  84.62379  88.60761
#> 87  164.8100 154.68139 125.00913 110.70216 138.56844 132.52031  94.36430
#> 88  174.1647 182.78457 119.72138 141.97302 216.70158 159.58862 125.25348
#> 89  176.3674 109.70443 143.79921 129.88684 169.11923 117.80397 104.24916
#> 90  163.4828 154.98929 171.54412 206.67843 178.33877 167.03449 176.29153
#> 91  121.3643 135.00155 143.44497 100.22599 136.97183 156.08849 125.51097
#> 92  140.6892 135.77497 148.33050 212.91962 149.76598 151.49641 124.06016
#> 93  143.1141 107.36127 180.79922 184.75069 143.87958 103.67052 124.69359
#> 94  188.4804 153.81227  87.26498 190.31939 124.67553 157.51616 114.66990
#> 95  123.6330 137.02441 144.25855 167.02655 225.22949 121.06205 147.24402
#> 96  124.4889 114.44153 146.89312 141.24004 148.07254 130.04673 138.00236
#> 97  200.8715 152.08989 156.30678 141.29338 143.36724 149.85053 127.91646
#> 98  155.9149 166.29706 155.54791 146.12020 172.93114 159.34778  84.56148
#> 99  141.6397 175.13695 166.91682 152.41262 193.82673 172.57403 141.85803
#> 100 131.2505  99.66846 131.85363 191.97746 142.90334 165.87967 141.36227
#>         B.4_1
#> 1   157.67820
#> 2   128.05943
#> 3   183.85409
#> 4   167.22242
#> 5   147.35933
#> 6   170.40344
#> 7   139.00091
#> 8   121.46751
#> 9   121.17892
#> 10  142.31551
#> 11  161.01096
#> 12  220.85326
#> 13  166.38295
#> 14  153.27005
#> 15  141.31370
#> 16  162.17704
#> 17  155.05055
#> 18  154.50949
#> 19  171.39421
#> 20  119.45929
#> 21  130.83263
#> 22  148.13535
#> 23   93.71452
#> 24  157.20134
#> 25  127.76488
#> 26  112.08606
#> 27  173.41404
#> 28  162.29428
#> 29  184.94603
#> 30  142.33355
#> 31  129.15413
#> 32  113.81266
#> 33  173.26136
#> 34  134.01199
#> 35  130.53944
#> 36  122.79346
#> 37  130.81364
#> 38  129.26429
#> 39  139.45015
#> 40  141.73544
#> 41  154.06658
#> 42  132.35645
#> 43  163.75471
#> 44  136.37981
#> 45  172.06038
#> 46  163.63842
#> 47  157.04310
#> 48   95.71654
#> 49  156.58145
#> 50   93.51190
#> 51  163.32507
#> 52  225.25846
#> 53  136.92894
#> 54  184.87004
#> 55  192.95842
#> 56  153.62189
#> 57  126.78489
#> 58  111.28589
#> 59  245.32890
#> 60  148.42425
#> 61  172.89546
#> 62  185.76652
#> 63  117.94669
#> 64  138.12366
#> 65  127.11784
#> 66  150.68978
#> 67  201.45618
#> 68  154.36258
#> 69  161.98821
#> 70  179.78696
#> 71  170.13307
#> 72  160.62841
#> 73  163.37795
#> 74  172.61727
#> 75  126.58548
#> 76  178.84258
#> 77   99.98975
#> 78  115.31987
#> 79  185.09332
#> 80  134.06555
#> 81  127.17408
#> 82  148.70050
#> 83   96.48573
#> 84  155.39103
#> 85  178.44365
#> 86  192.87268
#> 87  160.85278
#> 88  158.31737
#> 89  141.96201
#> 90  165.45512
#> 91  112.22091
#> 92  137.33570
#> 93  139.15558
#> 94  157.64544
#> 95  128.24539
#> 96  117.65195
#> 97  119.34582
#> 98  155.28519
#> 99  150.72189
#> 100 207.50382