Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10,10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(subset = rep(c("A","B"),50),
                   level = rep(1:4, each = 25))
rownames(meta) <- 1:100
scRNA <- Seurat::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA, by = c("subset","level"),
               min.cells = 10, max.cells = 20)
#>         A.1_1     B.1_1     A.2_1     B.2_1     A.3_1     B.3_1     A.4_1
#> 1   226.42963 115.49174 107.92533 124.68058 123.15770 141.80300 145.79066
#> 2   153.46428 109.43192 107.43862 181.14672 213.35288 103.93851 156.11580
#> 3    87.14953 128.72301 129.17350 132.74043 125.62788 145.67084 180.25471
#> 4   168.21213 153.05240 243.14871  91.87269 152.25659 141.87357 120.21373
#> 5   119.52036 133.14027 153.19849 187.28250 114.43889 171.00986 134.11353
#> 6   185.58891  65.51716 111.78452 135.65610 142.38670 112.17752  87.91395
#> 7   149.00355 217.21013 179.68862 139.10157 141.79962 169.89374 141.18264
#> 8   147.32091 206.94782 115.18656 150.89585 123.43764 107.27184 161.91521
#> 9   173.94838  93.23476 147.55357 148.34187 143.23450 106.50592 130.64795
#> 10  134.81804  91.70351  70.68448 167.83180 173.59550 110.87882 150.06542
#> 11  140.10687 184.98088 122.12645 143.91869 174.09111 121.62236 116.78327
#> 12  109.40830 164.36742 127.84644 157.67755 142.02290 169.02048 137.86852
#> 13  173.47399 164.92682 122.12449 135.77922 192.58178 212.85176 111.87610
#> 14  166.62568 128.06304 129.64246 175.78640 135.74142 136.78879 137.43634
#> 15  111.57355 217.08157 171.65381 194.51674 159.67921 137.91078 206.35836
#> 16  159.43511 145.83737 137.09436 207.71833 113.54150 142.25451 164.22624
#> 17  140.87389 174.88199 113.70270 187.91068 148.23126 164.73085 120.52482
#> 18  187.09323  76.44624 106.16223 167.83136 155.52799  87.69847 113.45571
#> 19  178.02730 119.89940 102.51903 124.37322 137.00750 123.17726 164.26718
#> 20  156.62674 101.46750 130.91816 165.66964 165.68781 158.70588 105.49671
#> 21  192.30969 112.70367 109.82531 157.84715 189.96658 153.42859 184.47271
#> 22  161.52003 156.79696 179.81469 118.02057 106.97673 108.53241 117.47113
#> 23  170.34648  85.60872 170.25590 183.13756 100.57000 119.86642 149.03536
#> 24  124.59977 178.72669  85.53154 130.99106 192.11301 101.13031 140.96780
#> 25  164.81296 123.12345  80.95368 134.84213 164.09094 215.88938 147.19886
#> 26  131.20881 151.57690 147.46155 159.65877 142.73132 124.11074 156.11810
#> 27  115.77021 142.30587 130.27467 206.45118 161.40276 155.96067 125.12736
#> 28  129.48615 128.09975 178.58866 139.52473 144.99986 155.11868 125.39596
#> 29  197.10156 141.77342 177.48776 162.54675 176.99211 180.87110 122.66126
#> 30  110.94759 123.64106 133.97637 224.45714 172.17422  81.05879 176.66927
#> 31  170.73181 146.53072 110.98752 120.16656 210.50196 102.62676 134.40882
#> 32  153.13184 165.84439 157.79072 168.84763 145.06588 148.15229 136.02717
#> 33  189.60373 136.08470 123.07335 142.07592 115.74793 136.58520 168.78151
#> 34  124.35440 160.35301 130.45463 139.45545 192.22542 115.95782 130.32572
#> 35   88.17755 128.65256 203.62156 126.22891 171.39297 131.83072 145.21422
#> 36  207.96180 101.39055 213.74850 160.26908 177.88170 158.23898 124.69099
#> 37  142.04547 188.14255 141.11846 145.40026 177.29051 155.43235 154.12667
#> 38  166.32045 151.43570 149.13026 185.44161 155.70004 164.88807 125.35876
#> 39  138.96836 158.86306 117.37360 168.59591 173.58827 175.47432 167.91728
#> 40  136.66353 135.12124 149.21684 173.45907 135.34744 130.37553 127.93363
#> 41  173.64977 142.03183 109.02112 150.25431 185.98236 118.52626 128.78442
#> 42  165.26598 124.55815 125.04572 198.29047 145.94698 143.76263 141.07734
#> 43  131.88839  96.11242 137.61171 132.72901 109.21086 116.04963 138.79441
#> 44  185.00008 144.05093 183.99877 151.51897 101.18734 134.33729 104.04837
#> 45  151.24644 162.23778 188.39777 162.56697 129.71375  88.21252 112.99182
#> 46  162.45326 166.48975 120.25548 206.69485 110.94553 130.40790 113.02073
#> 47  197.87596 130.88135 121.66618 219.35818 137.80487 149.25933 124.53322
#> 48  156.86880 129.23053 126.70288 216.95416 124.85479 101.19095 132.29634
#> 49  156.23140 159.12202 182.51183 133.92331 143.52129 121.30457 137.92990
#> 50  161.27387 127.43787 108.71739 160.99780 139.53591 137.22009 152.14498
#> 51  151.47229 110.05758 137.54591 192.07115 122.38092 184.41530 127.38427
#> 52  206.55632  90.62093 161.40591 176.35554 108.33720 168.58774  90.10822
#> 53  165.50360 130.08400 137.42595 173.00349 124.21791 125.17746 136.38842
#> 54  191.97570 147.13858 124.11684 158.25909 185.76137 140.67854 141.07723
#> 55  109.83260 205.97991 143.67486 153.08453 141.31320 136.11316  95.93753
#> 56  176.89647 176.11183 148.01664 195.68503 148.67275 115.10016 102.79795
#> 57  148.64161 186.71400 135.94087 160.86178 170.64139 203.34500 104.74409
#> 58  211.97646 103.01096 113.45072 131.37188 158.96978 134.90822 177.00358
#> 59  165.05229 103.09722 102.56468 106.36093 137.07885 118.20907 135.27873
#> 60  176.58494 150.84700 128.91137 173.87338 173.32104 145.68081 153.18187
#> 61  233.10408 202.69967 176.90414 139.67493 204.94961 141.62548 162.40394
#> 62  147.44041 116.08449 196.16586 132.83814 139.24106 180.04219 138.31410
#> 63  128.07232 153.14298 112.06688 185.44757 160.38582 143.34024 136.20656
#> 64  236.02254  96.71562 104.94953 128.63954 153.21494 143.29416 115.03117
#> 65  156.92525 160.82218  78.14767 157.89566 154.72299 129.83446 141.72988
#> 66  179.06262 135.93717 129.98547 152.36615 147.78223 116.72305 154.52613
#> 67  126.43053 144.95326 106.18785 166.08825 160.33859 121.63229 144.79315
#> 68   93.53127 120.70188 113.64730 166.23853 103.79726 157.29212 111.11540
#> 69  111.25948 136.70652 149.77352 144.30641 174.59241 125.09655 128.85577
#> 70  120.23811 134.65702 146.05465 129.07106 150.59383 155.85045 145.59342
#> 71  195.42095 111.21706 166.05785 174.00346 105.79347 135.74923 120.01871
#> 72  175.78155 120.66950 138.46678  80.10411 151.45027 111.43351 184.72753
#> 73  166.61951 150.62224 132.18330 218.35684 160.53914 112.09940 166.69060
#> 74  181.91064 155.67397 107.32712 132.79338 157.37761 138.75787 168.48204
#> 75  195.46100 116.19459 196.47406  88.50944 167.46950 152.16762 182.58950
#> 76  138.41574 159.02291 143.34683 164.37315 166.99314 182.19249 134.02927
#> 77  121.88950 135.67751 161.61912 151.95656 145.49824 118.69150 123.94295
#> 78  154.12509 148.17437 158.08697 139.71629 185.24720 150.28875 111.13180
#> 79  132.41205 188.82651 144.90314 179.47354 153.56642 145.01292 103.13165
#> 80  152.47085 149.62142 134.66394 115.81984 118.23217 135.45995 164.31895
#> 81  148.35328 136.99307 110.65118 127.94479 190.91228 109.51262 166.09509
#> 82  186.87829 114.88635  81.68406 139.54744 158.77498 104.27420 144.19562
#> 83   98.37597 148.22255 124.84954 152.81914 174.57148 193.00240 148.69811
#> 84  138.63842 119.20510 143.60691 153.69062 191.57183 158.83615 151.41601
#> 85   81.69909 133.80772 162.20806 148.95918 122.49476 134.25115 145.18080
#> 86  129.96199 126.44612 157.10207 152.27690 124.68612 142.73008 146.85683
#> 87  168.17241 149.85577 125.36878  95.08306 150.74723 140.00015 170.82134
#> 88  114.82305 157.61213 173.11708 164.54238  71.94799 167.45208 142.51378
#> 89  191.36236 115.04431 122.64745 148.09833 109.55973 111.13007 180.61898
#> 90  142.68700 128.52597 154.25163 177.26188 118.04139 185.10293 148.72264
#> 91  128.58913 144.28901 142.76523 122.68017 134.70994 165.19221 122.16774
#> 92  144.20906 110.87953 127.27858 124.34387 112.28808 141.77359 165.23818
#> 93  108.91056  78.29866 125.92668 164.16230 115.06304 208.18880 167.56039
#> 94  198.45520 160.02729 148.97565 158.74216 145.40885 185.18930 140.96663
#> 95  100.68499 154.52185 219.90548 251.28998 142.61359 141.64336 122.61358
#> 96   94.79691 140.88203 108.69953 162.05936 144.56369 150.10258 136.60556
#> 97  165.40182 164.62182 125.46764 200.56903 133.82067 121.43632 202.15126
#> 98  173.65694 155.51495 139.16927 125.11425 169.31346 129.24415  95.34086
#> 99  227.22833 144.66558 180.25824 192.98696 128.88005 147.30483 171.00638
#> 100 172.08309 161.69703 134.04923 169.55314 102.32972 131.06711 141.46936
#>         B.4_1
#> 1   178.53223
#> 2   156.93704
#> 3   180.62009
#> 4   194.77226
#> 5   122.31814
#> 6   146.59698
#> 7   153.01912
#> 8   119.58438
#> 9   143.88297
#> 10  149.39566
#> 11  163.42710
#> 12  181.51875
#> 13  106.05565
#> 14  135.33706
#> 15  146.57577
#> 16  169.01623
#> 17  111.77986
#> 18  159.25529
#> 19  138.43636
#> 20  118.35888
#> 21  162.88457
#> 22  158.22364
#> 23  175.94062
#> 24  129.76730
#> 25  152.31861
#> 26  162.67072
#> 27  191.78399
#> 28  131.89059
#> 29  145.64460
#> 30  104.34856
#> 31  153.52159
#> 32  176.29471
#> 33  170.43785
#> 34  149.84679
#> 35  139.13092
#> 36  153.61335
#> 37  164.65583
#> 38  184.76736
#> 39  171.88180
#> 40  151.26166
#> 41  205.83851
#> 42  131.79182
#> 43  154.89712
#> 44  118.11321
#> 45  148.35995
#> 46  160.74639
#> 47  150.81837
#> 48   88.76949
#> 49  162.49635
#> 50  177.36421
#> 51  143.53479
#> 52  158.69693
#> 53  151.50022
#> 54  114.51679
#> 55  150.87702
#> 56  183.16429
#> 57  124.10425
#> 58  164.23289
#> 59  116.28889
#> 60  143.46471
#> 61  154.61094
#> 62  140.91755
#> 63  152.21374
#> 64  144.50866
#> 65  112.22517
#> 66  150.15139
#> 67  162.60091
#> 68  123.27919
#> 69  157.43512
#> 70  174.28543
#> 71  215.33601
#> 72  145.68685
#> 73  188.36195
#> 74  123.40077
#> 75  198.45424
#> 76  204.21721
#> 77  131.14372
#> 78  186.11776
#> 79   87.83060
#> 80  137.54752
#> 81  156.42327
#> 82  190.38899
#> 83  168.59158
#> 84  147.95739
#> 85  152.39696
#> 86  203.92662
#> 87  125.71926
#> 88  194.68932
#> 89  138.85870
#> 90  212.10532
#> 91  149.45808
#> 92  120.69727
#> 93  162.16382
#> 94  144.48988
#> 95  126.39048
#> 96  180.32453
#> 97  128.54106
#> 98  149.50661
#> 99  122.20001
#> 100 167.23423