Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10, 10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(
  subset = rep(c("A", "B"), 50),
  level = rep(1:4, each = 25)
)
rownames(meta) <- 1:100
scRNA <- SeuratObject::CreateSeuratObject(counts = counts, meta.data = meta)
#> Warning: Data is of class matrix. Coercing to dgCMatrix.
pseudo_samples(scRNA,
  by = c("subset", "level"),
  min.cells = 10, max.cells = 20
)
#>        A.1_1     B.1_1     A.2_1     B.2_1     A.3_1     B.3_1     A.4_1
#> 1   133.4718 116.63118 104.03120 206.97555 241.36028  98.96757 123.37676
#> 2   182.3319 187.78385 125.88171 100.00500 184.97449 155.86870 145.84370
#> 3   110.1766 175.21765 134.66147 163.56761 124.90500 126.50698 149.93699
#> 4   106.5022 142.06961 157.22021 127.17893 111.99822 181.47010 117.27401
#> 5   170.6425 124.50876 156.04242 168.17741 167.77236 118.36936 174.40738
#> 6   147.1655 183.35078 120.12560 129.03935 135.34024 131.36327 119.74137
#> 7   197.9369  92.29946 152.84558 127.34375 110.61056 108.34112 109.64430
#> 8   134.2275 127.50479 157.94606 165.74399 136.58957 162.34337 172.76792
#> 9   166.7330 155.64163 117.58660 139.04924 127.74442 147.89020 161.77252
#> 10  139.1982 174.00008 158.36471 128.07160 124.54719 242.86962 156.19739
#> 11  140.6150 142.00538 168.54153 168.45329  80.08528 124.77066 162.84459
#> 12  130.7340 125.29367  72.60881 129.03360 128.56484 145.14680 143.53381
#> 13  182.4087 167.81064 111.97005 183.00229 207.40076 155.27916 163.93355
#> 14  134.4779 148.10730 180.04040 126.98743 173.86655 152.61467 130.97881
#> 15  128.1338 190.54541 149.36808 147.19475 150.14232  98.61885 202.13858
#> 16  127.2369 138.48301 141.77726 121.53138 136.69428 152.39969 189.41383
#> 17  140.8559 112.54574 122.57005 133.45171 131.93556 131.38267 154.08619
#> 18  156.9232 129.13106 141.11772 171.56221  97.51278 138.00802 145.37963
#> 19  156.4097 154.87703 137.91044 116.86009 126.09812 146.64286 141.14541
#> 20  131.1254 206.73578 144.55788 182.98107 117.00879 128.24574 176.70648
#> 21  176.6178 148.08696 155.54598 110.61881 156.66169 193.47047 195.14257
#> 22  129.9864 128.55792 147.29181 185.55826 128.40836  95.86348 154.81060
#> 23  108.5068 146.43622 171.29348 164.65745 174.31413 189.88778 185.94857
#> 24  117.8669 168.01298 195.38170 147.44636 106.83828 130.61426  90.57220
#> 25  163.2083 141.11697  93.14332 121.71924 181.50295 126.65110 103.94961
#> 26  143.5013 143.63846 165.47590 153.34878  96.92796 154.28961 183.89316
#> 27  131.6470 157.57075 118.89211 156.03182 176.62026 142.66144 164.46007
#> 28  109.3561 203.06610 103.58762 170.80542  99.69954 139.79728 169.69956
#> 29  126.3739 136.17892 135.24410 148.89495 171.09675 158.13171 175.62291
#> 30  182.1398 125.96843 135.65628 207.74521 173.12972 110.30652 131.56693
#> 31  227.8230 123.01219 130.36336 148.30063 160.98123 147.47046 104.75822
#> 32  157.5173 150.19732  84.28694 169.65642 152.44900  96.10631 169.10421
#> 33  155.2870 122.35903 127.55467 129.84239 108.41194  95.63392  86.56483
#> 34  120.6953 113.68373 127.22934 177.78193 158.67339 100.38262 169.92397
#> 35  166.0319 115.67317 116.13024 110.87385 136.13061 134.21570  98.24095
#> 36  153.7177 155.28907 185.70576 135.83644 140.19066 128.71151 136.01674
#> 37  157.4679 139.86817 137.50101 159.84290 196.65404 134.66963 140.99938
#> 38  145.8044 108.57741 181.11805 175.33093 132.67755 126.54508 201.32009
#> 39  141.2264 109.96958 111.61698 124.36083 120.20555 139.07270 107.19089
#> 40  109.6406 114.63801 181.74561  98.40943 159.11942 145.89379 115.34311
#> 41  132.7286 158.67950 167.86163 154.58127 169.63332 160.64254 146.63957
#> 42  125.7867 143.43033 107.08921 110.92159 123.34808 150.20130 163.95763
#> 43  172.4766  77.42885 157.22476 206.42718  99.25001 166.43182 131.75724
#> 44  158.5686 106.15350  88.66303 153.14206 132.41788 140.09343 138.57005
#> 45  181.3330 167.75495 171.06118 140.35575 208.43208 146.20245 173.76960
#> 46  153.8431 148.02434 108.35859 128.57235 190.35559 163.53375  97.48431
#> 47  204.6228 123.52927 195.15522 166.32417 180.73917 122.10408 170.67663
#> 48  111.8829 191.63872 128.28951 239.00305 180.39639 116.90278 115.42853
#> 49  168.2329 151.67961 118.88740 122.79101 208.33372 163.38380 203.10013
#> 50  186.2878 127.69761 126.69666 162.71272 145.28390 171.95036 105.80714
#> 51  113.8989 122.72129 118.57259 138.01936 144.19898 157.06021 151.28702
#> 52  148.2763 209.11449 155.71854 202.80666  92.43353 153.72378 143.84650
#> 53  163.7214 120.80729  91.93687 135.57055 184.55014 147.67956 139.23960
#> 54  162.5418 146.05273 115.37915 158.37068 161.89180 114.08265  78.19680
#> 55  152.1224 142.25707 151.68062 110.27242 144.69836  95.62520 141.90120
#> 56  182.8846 183.92696 116.85136 117.52407 166.06843  69.76185  95.91002
#> 57  131.3532 155.67409 218.52457 145.03491 113.38870 138.31763 109.83499
#> 58  172.4448 100.19741  84.20220 141.57052 182.34426 125.62021 165.46187
#> 59  169.6616  65.95888 129.91837 190.54305 172.94338 156.35780 148.96755
#> 60  126.8920 109.80145 171.46146 119.55329 142.58974 154.93767 146.03333
#> 61  133.4446 170.71572 115.56612 155.25324 169.63558 137.04501 161.55198
#> 62  185.7086 125.83332 162.90379 198.47202 177.71496 154.41107  98.71031
#> 63  113.4393 114.85534 134.77950 165.93494 178.91217 117.43760 105.70830
#> 64  167.1679 130.55027 139.15359 143.15233 148.73971 132.87723 148.28175
#> 65  129.2211 118.91860 130.42906 130.37375 129.79147 163.88814 165.56776
#> 66  182.9015 136.42720 141.98763 171.22185 144.68853 115.77392 179.13348
#> 67  133.8361 181.93105 161.44097 114.31997 200.64729 161.48533 156.16965
#> 68  138.2432 173.11402 154.36199 200.92530 140.38299 139.98654 125.83645
#> 69  121.4954 114.44610 108.71554 145.45897 131.03930 101.22943 139.91682
#> 70  125.7210 126.05418 190.34916 139.55427 189.45982 141.72145 133.69002
#> 71  195.1627 188.88402 219.58223 147.84050 171.61533 187.12044 170.40444
#> 72  164.5459 111.79904 145.37785 176.07573 167.16914 130.53170 168.55335
#> 73  170.8174 128.37420 109.74342 135.02279 136.72959 110.09747 135.70103
#> 74  115.5342 173.02859 146.60727 133.70648 167.40511 132.00006 162.70888
#> 75  141.9232 129.90471 146.68745 147.76013 228.50620 144.48718 164.08967
#> 76  130.5752 181.53019 152.04241 154.46982 175.72003 117.04860 115.13838
#> 77  171.6828 121.62735 163.40206 138.47189 135.18673 132.52232 100.76710
#> 78  153.7791 135.49315 164.55606 206.01001 167.11924 162.72330 129.21099
#> 79  155.3022 153.48449 210.06181 163.45570  85.83703 115.82143 142.91916
#> 80  152.0757 192.89444 129.14937 104.03223 123.53101 145.70075 134.25342
#> 81  141.3458 140.43470  94.84377 188.67536  98.67115 122.48054 141.87814
#> 82  161.3938 169.21146 148.25981 131.85361 172.56376 127.40131 102.58749
#> 83  166.2365 174.23600 158.89292 186.74056 194.19885 130.35408 125.02393
#> 84  144.1385 123.46055 145.76387 135.17028 138.29650 148.93835 113.67396
#> 85  172.5362 114.63970 131.71410 174.78629 155.88719 197.07675 160.66998
#> 86  161.8932 140.13073 139.73710 136.41310 140.49537 128.45293 149.07151
#> 87  171.2167 150.08633 160.84035 152.82916 123.01499 158.24699 174.16567
#> 88  105.9949 115.57779 118.93013  98.39294 139.44511 130.53911 163.75865
#> 89  144.0558 115.55007 172.07660 227.26851 138.85178 151.16676 161.52352
#> 90  126.7659 145.92262 163.65465 121.70747 121.46435  94.35441 154.13752
#> 91  101.8843 128.43230 164.13001 129.17934 110.79454 131.73371 157.27248
#> 92  208.2204 139.51955 171.16989 134.20751 123.53372  96.97620 122.13348
#> 93  142.8270 145.47195 148.48215 189.66029 170.69034 119.46346 183.76112
#> 94  134.4120 155.51973 124.76619 149.05925 180.63197 127.90320 127.28728
#> 95  123.7581 120.47149 143.57431 115.16636 169.20723 128.08120 150.47346
#> 96  142.1145 105.00005 129.02579 182.78050 147.53044 141.29505 178.21230
#> 97  144.6444 125.14379 126.97139 187.40278 150.15002 104.61511 148.34793
#> 98  147.1957 198.11755 125.23476 156.68220 130.00416 192.56552 171.77018
#> 99  154.0700 139.18712 181.34151 122.98965 166.35203 147.55609 176.73428
#> 100 141.6787 116.63778  71.73803 156.85518 167.19277 148.73914 165.68117
#>         B.4_1
#> 1   205.64631
#> 2   126.19201
#> 3   212.33989
#> 4   123.69270
#> 5   142.33938
#> 6   177.40200
#> 7   184.65879
#> 8   185.24065
#> 9   147.71631
#> 10  116.46837
#> 11  121.11857
#> 12  140.58986
#> 13  115.90970
#> 14  153.69657
#> 15  118.28176
#> 16  174.25510
#> 17  158.34453
#> 18  157.28935
#> 19  155.77299
#> 20  152.50494
#> 21   92.65146
#> 22  149.37162
#> 23  118.90000
#> 24  178.19516
#> 25  128.73680
#> 26  132.48202
#> 27  169.45961
#> 28  125.52077
#> 29  214.75382
#> 30  110.54760
#> 31  125.83790
#> 32  149.07306
#> 33  221.37521
#> 34  109.77443
#> 35  181.01928
#> 36  155.03058
#> 37  127.95551
#> 38  162.70901
#> 39  123.68516
#> 40  145.31731
#> 41  186.90018
#> 42  111.73749
#> 43  166.54819
#> 44  159.77993
#> 45  168.56484
#> 46  146.50542
#> 47  187.64452
#> 48  143.63543
#> 49  182.62060
#> 50  160.03119
#> 51  204.50935
#> 52  129.21283
#> 53  105.72701
#> 54  204.77771
#> 55  147.15996
#> 56  173.52116
#> 57  206.61720
#> 58  158.48255
#> 59  105.24725
#> 60   78.80104
#> 61  202.45956
#> 62  187.08330
#> 63  148.01185
#> 64  123.58838
#> 65  146.18359
#> 66  158.84922
#> 67  154.99406
#> 68  168.17891
#> 69  157.08811
#> 70  116.34903
#> 71  132.14971
#> 72  103.15920
#> 73  154.83258
#> 74  161.32696
#> 75  112.60895
#> 76  189.02167
#> 77  101.50099
#> 78  155.14329
#> 79  134.10727
#> 80   90.11567
#> 81  131.80605
#> 82  165.69391
#> 83  128.98378
#> 84  133.95245
#> 85  137.57036
#> 86  173.91407
#> 87  141.59547
#> 88  157.56274
#> 89  191.61532
#> 90  147.92896
#> 91  146.64761
#> 92  182.01087
#> 93  135.68360
#> 94  119.50107
#> 95  141.09699
#> 96  161.49376
#> 97  142.94919
#> 98  157.31427
#> 99  125.79268
#> 100 139.41511