Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10,10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(subset = rep(c("A","B"),50),
                   level = rep(1:4, each = 25))
rownames(meta) <- 1:100
scRNA <- Seurat::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA, by = c("subset","level"),
               min.cells = 10, max.cells = 20)
#>         A.1_1     B.1_1     A.2_1     B.2_1     A.3_1     B.3_1     A.4_1
#> 1   167.99541 113.11568  86.14900 130.35130 143.65267 155.49248 160.75318
#> 2   150.32683 127.82220 189.07838 159.78374 204.71979 125.40675 136.29879
#> 3   126.50965 127.07675 120.68324 144.86568 137.11558 127.31932 128.13233
#> 4   128.21021 150.68724 121.42518 150.69496 221.93519 198.93610 138.87488
#> 5   135.49555 162.16403 125.21401 150.51557 187.53761 119.14358 128.37854
#> 6   192.09179 170.64487 133.51576 126.57873 187.99975 168.68511 168.71743
#> 7   150.20393 143.21872  99.82867 126.73313 143.89000 139.61126 177.13238
#> 8    98.47118 129.21810 124.28690 137.66698 142.96268 135.97680 133.20977
#> 9   160.85690 129.42952 158.23287 236.40593 154.64652 124.32609 156.14685
#> 10  197.52157 137.29675 127.51082 160.85821 119.36406 135.35951 158.38140
#> 11  201.70600 125.08576 180.23879 123.52086 196.64738 103.28883 110.09980
#> 12  170.13691 197.58781 127.73946 141.31785 134.59798 196.52672 143.93514
#> 13  137.54823 137.40301 136.33017 146.52670 130.37713 147.37937 184.77892
#> 14  107.27557 144.50436 143.67652 156.43918 130.81521 109.81409 115.40124
#> 15  160.03024 114.14764 182.59559 183.54385 164.88083 114.58296 118.53189
#> 16  133.50588 173.29622 133.12623 174.31469 130.63778 107.70495 124.93698
#> 17  132.80120 140.91296 170.24918  67.30372 147.83845 124.60622 144.44575
#> 18  139.27971 144.64023 111.70914  98.25542 201.03343 144.25778 222.90643
#> 19  178.90594 161.89364 110.95582 180.72116 185.57179 139.34622 150.27590
#> 20  149.93680 140.31768 172.03695 141.80472 208.58330  91.14220  93.35672
#> 21  105.38750 168.71637 136.80189 131.60403 130.99982 137.08683 167.72294
#> 22  151.04973 116.73981 120.65917 150.78067 139.14081 169.52946 110.13586
#> 23  158.26204 216.29465 173.23444 135.78276 175.28456 161.31506 187.00175
#> 24  192.75739 168.45255 155.00348 160.46872 132.18539 150.63594 139.25018
#> 25  177.76969  87.76316 102.89062 109.21843 158.19886 155.31033 120.86172
#> 26  156.84785 107.32202 180.73354 197.24756 201.74274 153.35395 158.74314
#> 27  156.35868 174.44093 140.42021 135.92274 161.70476 200.49814 102.12860
#> 28  120.19839 142.48088 210.43551 106.37487 157.65697 108.60900 116.72680
#> 29  150.68564 115.02349 131.33262 169.23456 127.38407 102.22366 157.12030
#> 30  182.85023 178.85325 145.79686 197.31767 120.81503 111.24620 179.89392
#> 31  167.82133  95.90166 119.01528 106.95033 133.53192 126.58015 162.81631
#> 32  172.77010  85.75363 128.50235 193.45780 128.87854 170.42768 114.13823
#> 33  168.97390 135.41112 121.59891 182.97265 124.91109 151.34845  80.20965
#> 34  140.40096 183.41128 114.84269 148.23559 125.27112 157.82325 122.24460
#> 35  177.15867 134.72831 168.29646 147.86188 151.45896 172.61878 118.36771
#> 36  110.96756 157.70442 107.90714 158.55060 142.17949 188.04320 141.71680
#> 37  134.76966 107.20017 119.18458 107.75124 184.34140 119.66052  82.68741
#> 38  192.29933 138.22142 139.25145 229.17870 205.37049 160.34238 145.75670
#> 39  152.06680 135.81069 169.76658 100.96849 157.73198 205.90040 158.01266
#> 40  127.22070 118.77353 118.77501 137.23546 194.46348  72.48743 130.46829
#> 41  120.20145 116.57630 119.90061 116.51215 129.02501 135.56847 198.75483
#> 42  169.40006 140.50039 163.27624 177.14321 230.84401 131.88962 117.08817
#> 43  152.67559 141.63974 138.48114 139.68668 163.81957  96.42464 170.81295
#> 44  153.63045 112.85367 127.83540 146.44048 117.56687 146.27201 143.64398
#> 45  166.49589 127.07834 130.58725 143.22086 131.68379 142.39297 129.76619
#> 46  131.20716 129.08928 127.71996 154.52168 160.83249 180.27158 133.98873
#> 47   95.45888 165.63653 189.18627 163.12042 171.34303 189.53418 147.60646
#> 48  149.00247 138.58146 136.85984 144.72103 170.16997 145.08150 158.00198
#> 49  135.66201 188.20021 126.74146 111.65629 142.89441 193.86128  88.88854
#> 50  129.08913 142.38995 144.03789 132.14284 191.75697  97.51825 156.45755
#> 51  131.80436 142.56276 124.76465 187.78806 161.18313 120.93777  97.23242
#> 52  158.45684 122.96140 115.56155 170.88684 134.55120 131.72279 114.86821
#> 53  136.19506  84.09637 152.71778 123.18729 138.09777 154.39172 145.08327
#> 54  204.13763 154.00320 197.41591 105.34338 151.58369 113.92322 149.39901
#> 55  132.91737  91.47055 160.68623 149.21833 199.80160 178.79524 141.17534
#> 56  138.46288 122.67254 202.19574 120.06149 139.97279 114.09659 135.75827
#> 57  146.59585 132.96666 146.52293  83.43645 183.20058 188.65211 108.87755
#> 58  133.62521 164.66115 128.36012 134.71352  95.26383 133.16554 187.03479
#> 59  121.05146 145.93481 110.04837 137.82658 121.95472 123.94801 114.63143
#> 60  157.13632  93.44109 146.78396 159.46231 148.15448  94.82135 167.04965
#> 61  138.44826 134.73143 105.93023  95.66391 184.73589 142.56220 106.30790
#> 62  176.67378 136.34294 123.94689 118.09675 135.96494 117.30717 142.39922
#> 63  117.88088 178.46752 138.38902 145.81697 147.46208 102.30250 138.19904
#> 64  192.93934 155.72588 159.86753 167.59635 130.72097 123.72408 113.06749
#> 65  159.97325 133.89522 153.77327 155.50527 177.89677 156.89263 153.50652
#> 66  174.05719 175.50068 148.10014 176.46989 185.84748 167.22448 158.06738
#> 67  170.55186 150.22552 142.60897 108.18503 160.54504 114.45318 156.05240
#> 68  162.19278 104.39804 153.27085 107.66501 113.74515 170.54826  95.92975
#> 69  177.21243  94.95889 102.50684 192.88921 162.89535 139.34287 180.32096
#> 70  138.43155 153.40504 169.59507 162.35320 101.13328 187.10441 171.23028
#> 71  207.53739 105.95995 153.39660 134.83120 165.20867  96.21294 134.42511
#> 72  126.51988 161.86777 110.27791 141.76579 172.59524 145.12767 133.64051
#> 73  105.72811  88.58920 156.52183 143.92310  98.79512 136.05111 185.33694
#> 74  145.39780 141.59574 172.02264 149.15897 184.00823 129.58441  91.83662
#> 75  151.10844 166.16832 127.84581 153.55678 134.32301 117.13682 131.82420
#> 76  121.75321  97.24834 140.49940 153.82226 220.06283 151.41013 162.10131
#> 77  167.16405 133.65492 150.26949 178.52017 128.89619 122.49990 123.93804
#> 78  135.36744 135.59844 148.18685 160.74507 158.06175 150.25473 168.26255
#> 79  153.41089 162.69446 126.55027 161.13751 184.60510 127.76456 128.31504
#> 80  146.13830 130.63173 174.63950 146.47134 203.10004 177.65999 118.29635
#> 81  147.09623 142.88348 137.85652 143.91792 199.78362 143.54960 124.40282
#> 82  158.64714 151.08502 146.93857 171.62377 197.03265 167.37433 185.58278
#> 83  131.13910 128.36702 186.85306 156.07785 129.78029 145.59424 110.97719
#> 84  191.06407 138.42043 113.95787 123.04661 200.56653 109.76904 164.53557
#> 85  114.66178 123.93958 160.56983 143.79324 237.48978 164.57627 134.04133
#> 86  131.30862 118.17434 143.99966 124.41696 171.37903 146.13757 138.91265
#> 87  184.59448 107.38451 158.12816 153.37863 155.97901 146.97039 148.88712
#> 88  171.62364 119.02495 127.58902 158.03672  70.56991 112.21453 194.15485
#> 89  102.56255 178.09853 110.88362 199.28163 173.24465 124.21603 118.47209
#> 90  164.92648 132.36094 121.28841 150.71801 126.22942 127.70259 114.65062
#> 91  136.98803 113.50923 160.72995 173.73139 148.68887 130.80435 120.57174
#> 92  122.23582 143.23145 156.67362 175.48765 157.76185 137.51037 142.99035
#> 93  159.33486 144.07551 109.47036 164.30674 131.42422 133.09046 149.44648
#> 94  160.23956 144.74196 131.84403 106.49263 116.92621 141.59665 133.00146
#> 95  149.25424 162.54025 147.46429 209.21951 161.91888 172.35921 214.53926
#> 96  190.23632  98.98518  91.92392 136.06756 118.23818 127.69153 180.59072
#> 97  162.47781 117.78913 149.75119 179.43856 162.39572 115.05696 121.70622
#> 98  143.22025  99.25099  99.31968 170.03551 140.18210 118.91057 145.87660
#> 99  178.45599 126.54005 135.98166 192.02457 113.15785 195.43977 126.16946
#> 100 132.59207 147.17592 144.64167 125.45203 148.44433 218.23517 147.84359
#>         B.4_1
#> 1   177.04424
#> 2   166.74190
#> 3   121.38699
#> 4   115.46946
#> 5   172.01441
#> 6   178.76922
#> 7   110.45363
#> 8   217.47617
#> 9   165.00574
#> 10  170.55863
#> 11  179.00138
#> 12  118.98079
#> 13  155.82459
#> 14  138.93360
#> 15  103.75523
#> 16  103.50234
#> 17  156.29667
#> 18  171.07739
#> 19  100.56671
#> 20  126.77097
#> 21  150.55412
#> 22  163.27442
#> 23  142.80783
#> 24  109.89259
#> 25  165.36531
#> 26  131.82673
#> 27  143.46297
#> 28  167.81827
#> 29  164.58764
#> 30  188.55617
#> 31  231.24502
#> 32  212.22859
#> 33  130.22261
#> 34  144.78365
#> 35  168.98457
#> 36  153.30084
#> 37  212.46314
#> 38  135.51735
#> 39  180.21613
#> 40  147.38966
#> 41  144.47899
#> 42  200.78841
#> 43  115.43776
#> 44  123.81060
#> 45  205.75910
#> 46  127.04032
#> 47  136.54532
#> 48  160.68270
#> 49  148.08946
#> 50  186.06292
#> 51  129.66367
#> 52  147.09040
#> 53  182.08882
#> 54  187.64003
#> 55  129.98972
#> 56  140.95028
#> 57  209.91116
#> 58   94.37623
#> 59  177.43855
#> 60  159.51280
#> 61  181.63119
#> 62  129.59170
#> 63  141.20335
#> 64   96.90859
#> 65  114.45839
#> 66  137.38463
#> 67  136.01092
#> 68  140.09978
#> 69  131.36342
#> 70  125.65284
#> 71  113.69961
#> 72  177.66989
#> 73  168.53160
#> 74  143.55124
#> 75  115.08159
#> 76  164.78698
#> 77  141.26721
#> 78  140.05346
#> 79  166.95438
#> 80  138.68961
#> 81  156.96542
#> 82  137.08165
#> 83  182.53824
#> 84  148.25249
#> 85  137.45986
#> 86  198.89948
#> 87  144.48137
#> 88  139.17290
#> 89  195.29052
#> 90  101.24414
#> 91  146.20639
#> 92  100.22452
#> 93  148.92900
#> 94  153.20235
#> 95  211.02598
#> 96  167.21043
#> 97  144.15692
#> 98  147.11895
#> 99  160.60465
#> 100 155.27690