Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10, 10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(
  subset = rep(c("A", "B"), 50),
  level = rep(1:4, each = 25)
)
rownames(meta) <- 1:100
scRNA <- Seurat::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA,
  by = c("subset", "level"),
  min.cells = 10, max.cells = 20
)
#>         A.1_1     B.1_1     A.2_1     B.2_1     A.3_1     B.3_1     A.4_1
#> 1   156.30755 125.38586 177.18421 176.62582 100.45373 163.02464 103.61809
#> 2   155.26534 144.56808 147.49328 140.44107 130.03110 119.02949 147.86462
#> 3   110.41045 106.46587 139.41569 167.64356 173.42888 176.46419 114.77394
#> 4   140.06032 123.96424 146.76037 158.09890 185.73931 141.66855 171.60528
#> 5   162.12159 153.11163 146.96302 159.01062 111.29029 182.26915 164.21637
#> 6   165.32240 121.03438 156.65202 126.20832 121.10130 167.01873 174.38469
#> 7   186.44614 127.30021 103.27093 145.31104  87.60189 154.18916 156.10785
#> 8   153.39034 139.15027 107.68502 179.14765 148.73086 146.05730 125.81207
#> 9   146.56942 115.41961 140.40455 138.49768 158.22764 181.65925 121.19865
#> 10  155.96594  99.69813 169.74796 183.53843 129.03833 152.05772 156.01800
#> 11  184.53200  99.27508 145.54797 138.75201 123.10225 134.46271 161.45031
#> 12  148.16455  98.47421 152.84752 176.42519 123.24649 148.66517 114.35156
#> 13  124.35802 119.44404 177.14026 143.99827 105.64561 163.65407 147.63979
#> 14  196.69480 151.80815 137.79732 161.72247 126.43757 134.51417 219.01402
#> 15  121.21592 176.25966 139.17416 170.94020 129.42442 176.28895 148.54574
#> 16  155.12139 169.80947 143.96833 164.70522 129.21667 145.32704 145.66239
#> 17  147.06267 137.54563 152.36633 162.20524 176.89247 143.37826 134.44878
#> 18  160.07769 137.15664 102.72206 158.71413 141.79734 132.20545 195.91090
#> 19  124.47842 112.97423 126.92716 173.81332 156.80129 133.23662 121.54639
#> 20  139.06963 179.70510 147.38363 178.21610 124.50870 142.83592 160.36928
#> 21  116.35261 139.11986 139.96859 160.42745 177.60343 146.15926 168.15828
#> 22  119.33903 172.28065 140.92065  98.83297 179.83888 197.83537 125.72001
#> 23  133.17598 142.80679 124.28563 165.10067 164.16182 129.97801 123.10544
#> 24  155.30911 149.99035 136.58524 140.46142 126.74858 154.53355 138.70064
#> 25  111.38157 137.53421 108.20186 145.95965 106.02265 144.08832 116.99996
#> 26  141.47033 105.90852 134.37471 166.21928 176.36872 135.80222 184.50341
#> 27  134.95791 125.74089 129.35554 186.37085 152.25060 121.07394 153.99287
#> 28  107.62582 132.59437 115.81813 165.64005 130.18360 131.48493 135.25132
#> 29  111.72942 139.78902 121.52400 185.61213 158.40182 132.80424 121.48981
#> 30  143.41891 137.56869 131.87029 108.59634 106.14497 153.29722  87.22215
#> 31  166.17391 179.83385 126.92601 136.76334 150.06231 136.63109  98.17604
#> 32  196.51159 159.96590 168.27384 179.29363 123.74842 135.27815 147.51372
#> 33  174.76937 126.45108 153.10394 140.84044 128.09030 157.51573 122.67902
#> 34  150.04666 110.51957 181.32236 143.78203 156.54155 131.19183 109.98262
#> 35  145.74589 149.51319 144.67775 141.44709 160.04088 178.37623 108.25202
#> 36  172.73046 153.01129 149.10526 169.77197 126.22900 119.93194 120.48890
#> 37  197.97857 116.97782 160.63614 155.08545 155.38317 131.48216 165.57359
#> 38   99.93434 139.95286 131.93398 168.03250 150.61721 126.07954 213.32607
#> 39  146.53674 110.98413 187.29923 166.77447 150.15293 154.54819 126.12923
#> 40  165.93755 105.02930 150.68925 131.22890 115.91073 160.20528 162.75310
#> 41  187.19204 141.19603 158.75321 152.46210 147.67255 118.66248 127.70975
#> 42  164.44517 141.06263 181.08743 122.03851 145.17707 119.33153 138.51724
#> 43  189.04852 108.85328 152.59845 122.39433 167.17207 182.38246 165.26073
#> 44  129.93353 121.85871 140.28318 211.14801 124.33644 143.58473 109.93730
#> 45  142.47218 177.72532 119.78677 158.27062 176.72426 134.73469 138.51963
#> 46  174.66032 117.51119 121.09590 117.83379 176.30757 216.42054 109.10038
#> 47  161.14574 141.87548 153.72976 145.13921 140.46246 152.75558 185.84169
#> 48  181.24368 123.56660 119.54605 154.67954 173.45535 153.84877 116.59329
#> 49  184.15101 141.96307 164.85146  84.78462  94.14069 139.63773 171.49135
#> 50   72.37538  62.23261 144.12002 177.07576 181.67874 170.59803 187.21636
#> 51  191.23805 146.98901 123.61780 157.24671 165.61861 127.54521 180.04835
#> 52  148.51768 163.09916  90.99750 128.64959 172.43888  84.99837 121.87092
#> 53  103.77482 105.23730 120.22615 142.76496 142.14769 102.81720 147.13939
#> 54  163.66203 146.37141 124.51711 163.23839 129.89497 141.86818 156.15874
#> 55  170.33660 146.17033 152.98431 130.07112 119.38707 167.05714 161.18459
#> 56  141.78595 115.03255 158.08796 145.16853 156.45437 197.58186 115.12859
#> 57  204.36480 135.37292  88.55215 120.33611 151.39126 110.88383 140.02778
#> 58  199.64861  77.38210 172.81206 233.11992 142.21985 133.03706 118.38754
#> 59  155.26754 127.94513 151.28770 157.87645 142.60185 188.26792 136.88964
#> 60  142.73076 135.39958 138.41709 146.68274 174.45446 110.02031 121.60196
#> 61  184.33384 131.36453 153.72696 137.54211 132.64131 115.38398 170.25176
#> 62  146.51259 115.99434 172.99071 169.41188 183.59511 197.22271 162.42635
#> 63  150.98231 130.10693 142.04682 190.51280 131.63591 183.21501 107.31010
#> 64  144.88624 144.19457 109.50121 165.72380 140.72739 148.43819 132.73047
#> 65  180.89436 171.23984 136.03719 157.97271 148.60583 125.15945 147.99129
#> 66  150.47959 123.24118  90.74123 143.54154 169.08716 183.75085 112.94731
#> 67  161.61316  57.18088 160.93345 123.74552 123.29379 145.14962 158.13338
#> 68  166.57116 161.05956 145.69237 196.42377 138.08481 138.50810 111.50112
#> 69  105.16725 126.34378 137.63608 148.12864 149.94978 132.63216 193.90590
#> 70  148.79189 134.73173 162.72717 114.10040 156.81170 130.18071 148.05249
#> 71  205.58668 178.52800 166.56067 171.71555 150.45175 123.79188 132.41452
#> 72  181.48405 131.48757 133.08691 105.39178 241.21306 148.07494 130.99524
#> 73  107.36250 127.84138  95.05038 131.79065 114.81162 118.78996 114.47882
#> 74  191.66658 135.94649 145.87328 142.53118 142.92239 151.91392 143.66764
#> 75  146.79129 162.57985 114.11706 128.94429 106.49007 140.71257 158.29416
#> 76  137.04091 121.43437 174.37595 172.59086 179.12433  91.12575 116.24569
#> 77  116.79796 108.21107 160.93621 124.21262 148.38228 205.84593 131.68996
#> 78  134.33209 110.99743 139.84453 165.87965 156.44282 141.90791 125.79223
#> 79  127.03233 151.08754 115.92468 188.55529 207.12896  91.69661 154.36087
#> 80  127.95656 113.39769 182.59208 142.32241 125.38855 165.01885 166.63426
#> 81  114.57365 158.57874 119.17850 139.22026 118.11195 107.04690 181.58291
#> 82  112.45339 131.36576 136.26047 161.85253 122.45510 110.19954 118.96612
#> 83  177.11438 120.37229 158.87615 115.44689 105.17295 192.10830 139.62860
#> 84  137.90646  97.96683 150.41195 120.31243 134.01488 168.44214 161.77688
#> 85  134.82674 130.22919 131.02506 172.06331 149.59137 125.89074 130.87992
#> 86  134.03179 113.05820 159.43633 133.36429 161.14931 130.56265 163.90734
#> 87  192.89596 125.90437 168.26823 150.86099 117.42402 162.12935 139.14920
#> 88   99.31518 133.48121 172.36334 209.42340 169.51272 166.37055 164.37677
#> 89  169.62230 136.37560 120.85074 140.97659 129.55682 159.80200 190.45562
#> 90  135.23837 137.39920 128.83743 147.09784 144.58405  88.08094 148.50779
#> 91  116.41350 150.75013 140.91068 102.36433 133.73153 101.36194 163.51305
#> 92  137.54131 145.95998 159.78145 100.62292 120.46624 165.15423 132.31247
#> 93  174.32121  98.68091 120.99830 155.39581 160.37182 177.54776  95.66849
#> 94  187.17232 124.95587 113.34279 204.43102 198.32356 118.41427 156.73900
#> 95  130.44591 188.14963 196.67123 162.94079 108.36710 162.46853 107.14352
#> 96  127.48409 100.80117  97.25140 178.44594 118.46471 217.94966 125.86923
#> 97  113.73183 148.09304 117.67362 120.13109 120.96152 167.67377 108.59708
#> 98  115.19435 168.67466 115.45977 155.05395 182.07988 118.74240 149.56672
#> 99  125.77859 151.30415 120.06289 171.56218 118.64259 110.54580 171.15305
#> 100 158.99008 135.44068 148.30784 136.95803 126.14597 125.61188 117.61713
#>         B.4_1
#> 1   149.62930
#> 2   138.00925
#> 3   167.82624
#> 4   184.73571
#> 5   175.56826
#> 6   160.29616
#> 7   180.90193
#> 8   142.62616
#> 9   163.92310
#> 10  120.23102
#> 11   76.79413
#> 12  206.02918
#> 13  135.06995
#> 14  173.43516
#> 15  131.25728
#> 16  144.61880
#> 17  135.94425
#> 18  161.11881
#> 19  138.05171
#> 20  165.81803
#> 21  115.15701
#> 22  149.87910
#> 23  172.50885
#> 24  179.36435
#> 25  131.04680
#> 26  158.19614
#> 27  112.71120
#> 28  119.33898
#> 29  148.86039
#> 30  166.99820
#> 31  191.85264
#> 32  120.54224
#> 33  126.24077
#> 34  165.35302
#> 35  164.60753
#> 36  144.84494
#> 37  121.95097
#> 38  196.79680
#> 39  112.50637
#> 40  198.93767
#> 41  185.15635
#> 42  138.81525
#> 43  136.97774
#> 44  137.07729
#> 45  209.44849
#> 46  158.41541
#> 47  203.07274
#> 48  141.90759
#> 49  149.45738
#> 50  173.41268
#> 51  148.17791
#> 52  113.06267
#> 53  154.06219
#> 54  168.03899
#> 55  183.18343
#> 56  191.73061
#> 57  133.46099
#> 58  150.20012
#> 59  174.44234
#> 60  129.82419
#> 61  111.06419
#> 62  168.16427
#> 63  106.76295
#> 64  193.30014
#> 65  149.73198
#> 66  178.08936
#> 67  179.60818
#> 68  151.98637
#> 69  187.20051
#> 70  137.99490
#> 71  148.82542
#> 72  125.49337
#> 73  126.70288
#> 74  140.37116
#> 75  175.28276
#> 76  161.59711
#> 77  148.18943
#> 78  130.19054
#> 79  121.51065
#> 80  156.24569
#> 81  133.17135
#> 82  116.82122
#> 83  149.90914
#> 84  200.67846
#> 85  154.17125
#> 86  137.38762
#> 87  181.10542
#> 88  131.36930
#> 89  116.28782
#> 90  131.02080
#> 91  147.58511
#> 92  144.52851
#> 93  134.84754
#> 94  101.72525
#> 95  141.84818
#> 96   97.30501
#> 97  157.33914
#> 98  132.75364
#> 99  125.59684
#> 100 128.32817