Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10, 10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(
  subset = rep(c("A", "B"), 50),
  level = rep(1:4, each = 25)
)
rownames(meta) <- 1:100
scRNA <- SeuratObject::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA,
  by = c("subset", "level"),
  min.cells = 10, max.cells = 20
)
#>         A.1_1     B.1_1     A.2_1    B.2_1     A.3_1     B.3_1     A.4_1
#> 1   230.09805 166.56067 120.80976 147.8470  94.60386 133.05980 108.24842
#> 2   147.00730 190.23282 121.47773 141.5633 142.51679 204.04682 152.97252
#> 3   186.60533 131.92081 136.16432 109.0583 168.05715 144.64031 153.27226
#> 4   164.72321 103.49850  97.37293 116.1363 109.45937 177.34988 154.29397
#> 5   167.40307 189.88349 103.31777 125.5871 168.34402 159.83814 132.88903
#> 6   151.28207 106.79630  74.70023 189.2718 124.92663 131.43045 133.69201
#> 7   113.95298 109.18893 143.04087 183.7008 125.46293 185.94162 126.15675
#> 8   149.54569 133.81727 130.22405 127.2749 129.56839 114.77904 138.82867
#> 9   189.57079 137.60164 172.53347 189.8022 179.35712 146.22754 138.79867
#> 10  161.74136 152.16275 124.83423 175.1057 116.69615 182.99623  92.43052
#> 11  141.49275 141.03535  75.80347 183.4983 184.20999 137.13267 155.36280
#> 12  145.44783 111.75617 178.70473 148.0150 129.78599 164.15389  93.48762
#> 13  171.17174 172.17289 181.71156 177.4206 152.67963 109.24813 117.23158
#> 14  177.70927 171.67635 125.64508 165.0328 114.72108 127.25738 152.70085
#> 15  220.10010 126.27998 128.65174 165.2398 159.99069 143.38348 177.47875
#> 16  133.05948 141.80024 127.76532 122.1477 153.59293 142.43487 174.00571
#> 17  153.56454 137.88075 119.42914 167.4856 154.99436 164.03011 140.84227
#> 18  141.42564 147.22612 107.89008 191.3995  95.80590 159.06000 136.25392
#> 19  115.61033 114.05804 158.75547 160.6569 176.55000  91.83655 145.23188
#> 20  103.05319 166.23422 214.38264 173.2808 156.15360 189.59582 114.08576
#> 21  161.36773 139.93221 188.95233 108.4936 213.39261 138.60720 148.48448
#> 22  111.46760 131.78601 117.02259 145.1701 196.13358 114.34827 183.03516
#> 23  157.44105 145.27032 158.69633 129.4745  82.09585 162.51286 163.04698
#> 24  110.88562 102.83356 144.64285 121.5260 132.85947 132.30110 133.57425
#> 25  137.73854 110.35212 191.47129 176.1525 199.83633 123.41880 153.68825
#> 26  177.60534 119.48992 161.54951 107.3964 128.15379 123.08752  75.84303
#> 27   99.53387 154.58970 108.85650 167.8378 178.76841 121.40979 146.22069
#> 28  139.53107 119.55095 158.24123 140.2008 167.52385 131.44352 180.34344
#> 29  150.10585 122.97562 140.99782 169.0196 169.84911 167.55656 172.11576
#> 30  190.01901 169.72149 179.61812 151.1841 212.01299 129.36776 108.81395
#> 31  179.03566 147.58990 146.89573 179.8145 156.85008 105.23844 112.35663
#> 32  145.57860 180.12920 137.76173 116.1042 158.41989 157.38633 190.28910
#> 33  193.30809 102.69209 159.01548 117.1041 159.34340 157.99600 120.32919
#> 34  162.10180 135.78414 169.15574 207.9070 125.21487 104.54963 155.26184
#> 35  117.44281 131.86872 137.24103 141.1343 177.08260 121.88048 122.10141
#> 36  203.72470 118.12211 112.16864 126.1109 205.49541 123.73333 171.84996
#> 37  149.02692 173.57879 172.92843 189.7888 119.74644 134.10635 147.00086
#> 38  119.88967 152.10114 109.31733 158.8862 184.76697 164.22391 129.43801
#> 39  176.70163 179.47705  92.89921 168.0233 133.93186 106.32429 136.85716
#> 40  198.59123 149.65043  80.97541 114.4024 147.84951 117.23321 139.24698
#> 41  185.24641 103.27596 195.59239 125.9866 147.93010 168.24078 103.08346
#> 42  179.01660 121.51695  99.02008 126.4761 125.59317 151.59691 227.10225
#> 43  177.37666 169.74559 157.55621 131.9177 118.83472 138.21182 160.70220
#> 44  146.68626 165.32172  92.96042 126.6161 139.92136 159.91377  72.25538
#> 45  175.04082 113.46139 167.15542 140.5542 151.54096 132.27823 122.06573
#> 46  173.86580 186.19125 138.06359 168.3584 158.72576 187.35180 166.20061
#> 47  133.52096 103.24777 162.03567 165.7580 147.95926 123.94740 136.59376
#> 48  173.50802 148.80580 115.62640 164.8268 173.71022 152.21901 154.99160
#> 49  131.19819 128.52925 203.00729 151.4266 161.09104 169.84269 119.50461
#> 50  188.91893 187.19778 112.71434 151.7638  97.37918 126.69174 151.09281
#> 51  119.49263 146.18159 181.26698 150.8009 179.88985 143.12932 164.79805
#> 52  190.08815 149.94254 143.68565 167.4564 178.04108 188.22180 163.84792
#> 53  210.76589 177.46102 194.20739 145.9796 172.35172 137.69008 119.75419
#> 54  180.03103 143.42610 123.89958 169.4507 133.35410 121.40033 160.01788
#> 55  141.61688 145.67893 116.75251 164.6777 155.69584 103.34555  98.63435
#> 56  145.00442 114.29101  97.10625 132.6549 111.55420 147.88100 119.78937
#> 57  136.07999 140.91432 123.37094 165.5137 153.14556 121.35614 102.52615
#> 58  175.18955 112.25201 114.15679 143.9733 141.85577 148.21871 132.36515
#> 59  106.45481 144.16373 155.89666 138.2169 118.21030 123.20045 154.34266
#> 60  153.01941 134.89411 101.79170 146.1637 205.37505 149.52190 140.12572
#> 61  110.48314 113.57365 116.04500 185.0461 166.02588  87.66379 213.26150
#> 62  148.58246 165.11244 138.23313 178.4263  85.80963 114.60160 122.20253
#> 63  146.57347 140.11618 140.32951 142.5424 150.84196 153.25141 136.72588
#> 64  140.55571 135.95949 127.66653 155.9235 132.02344 168.41971 231.78154
#> 65  147.17952 110.84412 125.75775 170.3935 232.96076 112.47228 125.61507
#> 66  129.46764 179.92774 142.00186 169.7379 162.92617 156.51816 102.88648
#> 67  138.49133 147.65011 181.56703 180.1808 173.75888 134.12492 162.40975
#> 68  156.27246 114.21802 143.18787 167.6705 119.96995 111.42786 137.36766
#> 69  184.69274  96.72584 138.26673 182.8101 181.31526 155.93496 153.41898
#> 70  157.96706 112.15935 165.98544 131.1943 158.05940 155.06080 131.88589
#> 71  141.67690 153.50030 128.76714 142.9013 146.70805 106.06543 169.25776
#> 72  125.74731 127.61294 118.54313 140.2794 164.41668 163.13715 112.29604
#> 73  148.85228 141.75299 165.40840 133.5481 142.31214 114.60502 119.91319
#> 74  134.31187 146.82190 130.13353 128.5387 120.09636 178.98932 102.94339
#> 75  160.66826 149.03823 150.38511 146.3491 163.39661 171.74541 102.65747
#> 76  108.33901 134.37165 171.43771 143.2106 153.79083 148.97484 101.71991
#> 77  147.38459 137.58250 143.72604 128.9556 134.50745 161.07155  88.70380
#> 78  134.68749 137.79725 143.01741 191.6209 163.94449 132.36719 140.59977
#> 79  126.05411 144.62537 145.55083 140.0646 134.50897 124.50173 155.49698
#> 80  164.85084 138.18483 179.58307 151.9736 142.78216 151.99193 124.03219
#> 81  135.76559 110.20947 114.76304 233.6957 143.98425 178.62383 155.71121
#> 82  154.89818 177.22645 175.59538 151.6049 110.53931 162.47702 145.90506
#> 83  160.81881 137.18774 103.18993 198.3237 147.31961 161.60071 147.12137
#> 84  114.30870  96.81929 103.39893 121.9300 186.32317 119.49412  82.22067
#> 85  132.60842  97.20582 136.61465 165.4538 135.29615 131.04776 144.33366
#> 86  161.42779 133.29922 123.88133 133.6047 134.91939 149.06753 209.79409
#> 87  178.43321 155.99795 182.21079 162.2073 182.38180 173.71614 115.02094
#> 88  131.15534 176.85800 106.49487 191.0508 107.94606 136.75511 155.01699
#> 89  180.91821 130.86826 117.00532 130.1365 221.11712 162.35454 118.01488
#> 90  180.00141  91.48886 144.15512 168.0970 162.75593 219.18470 138.81712
#> 91  136.33737 134.12053 127.94874 201.8964 157.23727 153.99783 108.81963
#> 92  171.54655 130.86904 114.48636 186.8943 165.06072 175.84214 167.39635
#> 93  195.09574 114.67072 158.98749 172.4274 155.55069 158.36080 154.01085
#> 94  137.66346 134.51340 128.09601 138.7495 167.59596 119.69924 117.74357
#> 95  139.09518 146.08379 174.01632 149.6146 170.82743 157.33217 156.72425
#> 96  137.22904 153.30901 112.94526 192.5956 166.65684 116.54146 141.12045
#> 97  128.38792 139.56624 161.72532 106.7138 132.15033 117.97969 134.67714
#> 98  118.69013 143.74109 105.10829 156.0043 147.32559 172.38946 113.80497
#> 99  150.05223 129.99052 167.39404 139.7409 153.91657 152.83008 108.70669
#> 100 146.95777 106.51510 200.15660 179.4454 161.30836 158.28845 136.53770
#>         B.4_1
#> 1   100.07874
#> 2   154.14142
#> 3   161.39374
#> 4   195.25787
#> 5   170.73505
#> 6   144.56178
#> 7   182.55628
#> 8   178.49208
#> 9   191.65188
#> 10  150.97317
#> 11  200.43698
#> 12  122.26904
#> 13  158.70766
#> 14  136.44741
#> 15  122.22783
#> 16  112.18501
#> 17  126.21419
#> 18  185.24150
#> 19  149.24663
#> 20  135.96786
#> 21  120.97941
#> 22  138.38648
#> 23  180.73807
#> 24  126.49249
#> 25  143.42582
#> 26  147.83014
#> 27  138.33634
#> 28  199.21057
#> 29  140.28695
#> 30  117.63109
#> 31  172.27413
#> 32  154.02089
#> 33  146.14778
#> 34  166.53355
#> 35  153.39893
#> 36  137.41852
#> 37  100.28697
#> 38  125.44585
#> 39  124.56537
#> 40  166.10956
#> 41  152.45754
#> 42  168.93491
#> 43  136.40937
#> 44  156.34434
#> 45  163.05784
#> 46  154.39726
#> 47  176.42707
#> 48  143.32639
#> 49  155.84316
#> 50  173.41244
#> 51  139.72584
#> 52  208.69926
#> 53  172.75323
#> 54  154.83011
#> 55  147.61045
#> 56  203.21151
#> 57  183.94768
#> 58  161.28673
#> 59  109.20868
#> 60  159.94803
#> 61  110.95608
#> 62  156.59211
#> 63  161.56956
#> 64  142.02283
#> 65  150.07731
#> 66  110.31916
#> 67   77.40295
#> 68  170.82990
#> 69  175.71434
#> 70  118.73509
#> 71  127.25155
#> 72  103.09033
#> 73  120.11247
#> 74  170.67932
#> 75  166.93989
#> 76  179.23105
#> 77  145.76074
#> 78  201.50670
#> 79  132.28906
#> 80  117.88649
#> 81  159.04401
#> 82  123.52076
#> 83   96.81929
#> 84  122.81016
#> 85  165.89161
#> 86  150.16719
#> 87  122.94460
#> 88  153.37827
#> 89  141.84849
#> 90  131.82722
#> 91  163.19443
#> 92  167.66919
#> 93  199.99729
#> 94  188.54586
#> 95  139.28943
#> 96  171.39472
#> 97  145.75176
#> 98  165.73388
#> 99  153.40581
#> 100 157.60569