Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10, 10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(
  subset = rep(c("A", "B"), 50),
  level = rep(1:4, each = 25)
)
rownames(meta) <- 1:100
scRNA <- Seurat::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA,
  by = c("subset", "level"),
  min.cells = 10, max.cells = 20
)
#>         A.1_1     B.1_1     A.2_1     B.2_1     A.3_1     B.3_1     A.4_1
#> 1   160.44644 123.65124 121.61460 179.00621 115.50589 142.70358 133.55639
#> 2   145.36405 142.28158 193.32016 138.88401 187.83917 111.91623 132.48370
#> 3   128.59647 126.22871 104.69604 138.36851 138.26565 140.40088 155.12852
#> 4   148.23210 109.96383 156.01373 142.22385 135.96054 166.33907 165.74710
#> 5   158.13632 194.68190  93.38513 138.07278 217.39348 157.60266 173.55284
#> 6   189.51023  97.37917 157.70429 134.30328 216.12062 106.71197 125.47374
#> 7   142.35930 177.09160 150.51700 155.96588 173.01503 134.71847 173.81881
#> 8   161.76525 211.31539 106.32047 209.31310 125.30483 138.78367 137.48345
#> 9   107.07582 122.76637 126.42654 168.72637 119.59787 123.05129  87.41453
#> 10  118.80406 125.18719 173.08300  66.03221 197.03299 145.82514  98.25900
#> 11  151.28918 165.77001 138.04797 182.37858 150.34102 130.62046 180.39110
#> 12  124.22656 162.15660 134.00284 207.40121 178.34038 121.53227 153.75141
#> 13  174.89945 135.24253 120.63813 145.38517 210.76349 139.81773 134.40201
#> 14  145.25466 130.76733 133.87435 171.81475 176.67081 124.22575 120.10860
#> 15  141.47648 143.91637 124.51799 159.92525 168.09760 123.38436 172.72030
#> 16  145.23025 106.25963 161.84665 171.40663 174.65129 174.21335 160.86274
#> 17  195.83881 163.66135 135.91241 127.92832 206.52449 154.68593 139.59818
#> 18  190.77865 149.81566 150.35801 187.58918 168.21866 172.82600 172.37179
#> 19  184.71559 166.52240 100.95180 126.16527 139.02589 150.97624 204.02923
#> 20  150.18278 182.73291 150.59507 224.58934 200.82229 120.05495 135.59985
#> 21  153.66064 132.65322  88.42413 151.41366 132.23282  87.97894 160.31975
#> 22  124.65301  97.60027 180.61049 180.71667 134.36684 130.64361 109.23174
#> 23  162.54283 218.45683 106.01310 118.22052 136.06196 160.82723  99.03808
#> 24  107.31212 190.05599 119.43337 164.27361 181.58136 152.75335 113.88800
#> 25  186.57058 119.72845 154.14704 128.62817 171.50616 102.08668 158.12567
#> 26  198.24527 143.07416 125.69690 119.65912 187.80029 139.28422 119.98078
#> 27   90.42925 143.05779 118.49254 180.95679 135.76523  96.68443 139.42647
#> 28  137.61585 188.21057 123.14897  83.27287 184.97566 135.45421 109.57495
#> 29  185.43334 155.30048 161.56552 140.93859 142.15763 135.92119 136.36318
#> 30  148.62904 117.39310 167.94754 127.44712 113.87509 145.53872 154.62367
#> 31  160.84259 165.03725 155.43725 169.21197  87.85287 139.01218 119.15798
#> 32  129.00595 165.56034 110.67462 154.46384 171.69610 165.92893 147.12001
#> 33  141.47634 173.43167 123.02072 185.94030 146.14296  90.99100  93.18747
#> 34  201.49290 135.20486 140.13635 187.63307 129.17701 134.12779 142.74117
#> 35  158.36640 175.18903 168.51910 162.62053 119.50378 208.06782 147.33115
#> 36  182.33638 112.43736 111.62359 144.98340 153.10986  74.48369 149.06227
#> 37  161.90005 118.32883 127.30292 172.36876 144.90950  99.93689 116.12891
#> 38  137.09878 189.31159 129.35025 168.20359 121.79686 167.75665  76.06130
#> 39  162.44931 128.19839 167.45717 180.92530 131.81061  89.71156 150.11410
#> 40  136.60885 160.56219 150.75767 158.05732 135.77166 160.06187 168.35161
#> 41  123.38019 152.79011 162.70673 158.52213 131.02816 171.98199 139.70530
#> 42  144.41082 124.85532 162.46619 207.34541 211.12102  93.15907 154.33402
#> 43  125.82920  85.47055 117.93069 141.65772 135.68102 131.45109 160.32431
#> 44  140.63905 103.95763 122.76645 223.38152 201.86649 131.36343 176.07864
#> 45  111.01240 122.58250 146.29904 145.26270 133.07481 123.25470 110.69842
#> 46  185.01857  68.76413 107.22955 126.71115 118.74769  88.23868 101.14738
#> 47  118.77447  91.78697 147.61866 176.71025 205.30182 126.08218 135.45893
#> 48  170.57867 131.04243  85.69820 171.28274 147.95869 122.67946 103.31934
#> 49  159.88368  80.49650 139.28919 105.32683 204.50330 111.27093 163.94495
#> 50  141.30943 153.19904 149.81677 107.58361 134.83913 145.39432 148.08035
#> 51  146.80722 189.84114 133.70604 139.20368 122.93153 103.21058 158.19407
#> 52  203.68056 167.22855 149.45558 188.07336 151.67149 117.74076 148.94577
#> 53  142.58067 156.16253 136.01806 173.52953 167.19652 130.59581 168.98313
#> 54  148.61482  99.89541 148.94076 147.21749 143.43108 118.29548 120.81551
#> 55  146.15670 129.32216  91.78454 148.83509 220.04380 165.15405 168.40753
#> 56  125.16641 102.11340 146.36585 169.83506 194.20175 161.16540 167.88495
#> 57  137.02471 162.53805 159.76619 133.41062 132.02799  81.41259 117.48734
#> 58  186.04753 135.84557 199.49751 114.45074 155.20871 140.21421 189.39651
#> 59   99.02762 168.16205  88.06241 145.35997 189.96548 150.49078 143.14638
#> 60  143.65112 168.97043 183.05286 175.94533 143.33015 130.75076 132.90867
#> 61  165.07564 149.84680 194.47366 141.43431 195.31969 112.61054 150.52975
#> 62  105.47809  94.84101 123.21527 182.76145 162.86356 119.08754 214.44783
#> 63  143.32763 145.66731 173.74495 151.92450 142.86315 161.13134 122.69468
#> 64  127.01885 152.42300 142.44031 106.67779 158.05075  84.90238 164.81047
#> 65  149.54289 170.77721 172.97654 147.60674 143.41840 105.22279 106.53492
#> 66  191.86144 162.45988 187.97525 147.51764 135.14718 162.68595 113.19011
#> 67  136.56239 116.74071 153.74409 169.41040 159.89199 147.41822 126.51165
#> 68  146.72124 109.56193 109.81072 155.70896 125.21791 164.48632 154.76378
#> 69  171.05293 151.25802 175.86429 175.87012 138.42101 106.19979 144.44218
#> 70  133.63708 132.97383 103.26516 139.97050 128.74420 157.24182 190.78769
#> 71  125.97436 146.65820 126.18055 141.14405 216.89539 136.79059 185.74994
#> 72  158.26283 107.83069  97.50698  97.25771 180.38673 148.62713 133.82745
#> 73  175.32013 171.59819 106.87457 179.42502 160.07016 114.02246 148.94329
#> 74  149.58900 108.22293 139.90227 109.69331 132.59710 132.67406 179.49746
#> 75  109.97111 106.92963 181.16152 181.62880  83.00930 139.56242 152.49470
#> 76  180.66522 123.61042 144.50208 127.98179 107.82195  99.29595 161.35743
#> 77  154.25893 145.60759  99.82656 172.78812 162.70825  73.51314 145.22257
#> 78  145.85304 154.61662 195.82947 110.70308  95.84643 103.27442 165.62748
#> 79  197.52820 140.06291 107.62494 125.61354 179.46416 163.22292 124.97068
#> 80  108.16585 157.84539 185.11976 110.27352 118.22337 130.89359 162.84812
#> 81  113.93164 148.75200  87.13823 127.38964 167.60743 153.62686 168.94720
#> 82  168.37788 130.19257 105.48161 118.81585  94.59562 123.50758 141.52117
#> 83  185.37629 130.09107 120.09072 177.51064 160.47333 148.42857 158.91788
#> 84  165.19651 105.53449 172.60214 108.46975 187.50434 141.18366 175.05541
#> 85  141.30260 179.41643 129.59235 172.89369 128.12590 169.04549 180.91568
#> 86  134.86751 139.90190 111.56688 133.70549 161.84690 124.24589 146.24006
#> 87  144.08437 139.90739 148.29777 147.70611 120.66991 135.56201 114.03828
#> 88  149.60931 127.07773 144.30803 132.60716 147.52661 147.81618 177.05090
#> 89  204.42861 130.63976 152.52222 187.79297 143.32663 116.32064 137.67639
#> 90  125.95091 168.11817 157.94704 158.46545 151.61328 153.10110 164.66692
#> 91  172.64456  92.67404 159.40581 166.07550 102.38722 145.36040 125.48323
#> 92  117.79192 161.34424 137.76743 118.35585 111.76905 133.79679 170.72265
#> 93  160.94315 147.62256 171.49268 192.74194 151.60185 168.75148 156.61169
#> 94  153.93137 118.15960 146.33650 129.42321  88.66459 129.65677 151.61554
#> 95  172.15037 154.85399 112.03032 132.51435 151.58611 106.87563 110.35182
#> 96  132.16039 194.10624 107.68668  91.13386 161.90744 145.56349 151.53641
#> 97  139.69731 127.01833 185.67263  96.83014 147.83545 122.95901  89.63511
#> 98  138.78476  90.22651 147.86807 213.75440 133.07737 137.11750 135.01593
#> 99  173.84845 146.27712 151.70643 170.38322 146.02687 143.76171 138.01958
#> 100 185.92609 191.36985 169.11445 138.38907 141.39435 171.93947 109.05607
#>         B.4_1
#> 1   166.90922
#> 2   264.92240
#> 3   107.00623
#> 4   155.95204
#> 5   121.86676
#> 6   142.27662
#> 7   119.79995
#> 8   191.83687
#> 9   176.76147
#> 10  165.30023
#> 11  165.31008
#> 12  138.55153
#> 13  104.85952
#> 14  166.49200
#> 15  182.67756
#> 16  141.63043
#> 17  151.29776
#> 18  113.36985
#> 19  177.22607
#> 20  159.68898
#> 21  149.44069
#> 22  131.04926
#> 23  130.08281
#> 24  127.77679
#> 25  184.73197
#> 26  175.49974
#> 27  149.02012
#> 28  125.93874
#> 29  160.30812
#> 30  122.62600
#> 31  129.10461
#> 32  163.96973
#> 33  158.99628
#> 34  120.31946
#> 35  178.29209
#> 36  150.66974
#> 37  179.35773
#> 38  125.21852
#> 39  188.65493
#> 40  133.98505
#> 41  136.68124
#> 42  207.14677
#> 43  173.09162
#> 44  152.89843
#> 45  151.22028
#> 46  133.30457
#> 47  142.22487
#> 48   98.48702
#> 49  185.04891
#> 50  130.44293
#> 51  148.69313
#> 52  193.10713
#> 53  154.45319
#> 54  177.87616
#> 55  163.30470
#> 56  125.36090
#> 57  174.80946
#> 58  201.79715
#> 59  174.14204
#> 60  150.77416
#> 61  197.48633
#> 62  176.09507
#> 63  144.72312
#> 64  207.71987
#> 65   97.70558
#> 66  168.32550
#> 67  104.01878
#> 68  163.71618
#> 69  127.62999
#> 70   89.18805
#> 71  175.11776
#> 72  203.54150
#> 73  139.85356
#> 74  139.75445
#> 75  162.07320
#> 76  172.83590
#> 77  192.57625
#> 78  158.57522
#> 79  141.72884
#> 80  157.44498
#> 81  109.45235
#> 82  149.69873
#> 83  143.29400
#> 84  139.94739
#> 85  121.75358
#> 86  140.43130
#> 87  147.80150
#> 88  143.45398
#> 89   91.50681
#> 90  159.21114
#> 91  209.28443
#> 92  116.57640
#> 93  146.96436
#> 94  108.46509
#> 95  100.30410
#> 96  175.25716
#> 97  168.70531
#> 98   73.24078
#> 99  178.46416
#> 100 187.90553