Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10, 10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(
  subset = rep(c("A", "B"), 50),
  level = rep(1:4, each = 25)
)
rownames(meta) <- 1:100
scRNA <- Seurat::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA,
  by = c("subset", "level"),
  min.cells = 10, max.cells = 20
)
#>         A.1_1     B.1_1     A.2_1     B.2_1     A.3_1     B.3_1     A.4_1
#> 1   143.40652 159.19585 126.29470 105.90950 187.78566 123.07260 163.08783
#> 2   209.13756 148.64442 142.50312 196.67787 113.33348 138.64106  96.97083
#> 3   158.85370 107.40527 146.94933 220.51750 184.27213 153.47447 153.57221
#> 4   158.10336 106.57094 224.50539 111.70211 155.12122 144.82242 129.75028
#> 5   153.50243 149.57900 129.32459 162.56351 182.77530 119.21279 103.33350
#> 6   140.47149 119.00961 176.00415 228.20535 176.13547 172.27365 113.94562
#> 7   132.48528 174.60596 183.28365 159.17118 169.58129 100.63169 127.85420
#> 8   113.66490 137.42989 114.71922 102.66574  98.83114 157.43452 150.24687
#> 9   109.27038 108.25125 138.42611 142.84014 166.01640 171.35558 160.81137
#> 10  179.67195 143.80992 134.51320 110.12775 145.38743 122.24736 139.37801
#> 11  226.86493 130.81638 131.85198 151.60191  91.14148 128.27933 148.45557
#> 12  137.55157 129.42704 135.42791 140.33646 146.95068 120.89287 143.60794
#> 13  163.84252 109.11191 117.52232 148.77451 126.80130 158.27839 165.86254
#> 14  173.86468 113.71588 176.25895 131.73777 115.38865 116.99361 177.22413
#> 15  206.93166 113.97504 153.25287 122.46721 188.46666 122.91764 147.14858
#> 16  119.50489 190.07048 152.66184 187.38867 122.51064 124.87999 124.70280
#> 17  147.09083 176.40159 130.09273 152.34875 154.63641 132.94312 149.04869
#> 18  140.23631 108.58848 161.66605 135.89595 164.35011 197.08521 159.73627
#> 19  163.73451 179.12093 170.76558 163.73638 138.77832  87.71383 159.63692
#> 20  162.15344 110.59606 160.93851 132.00007 176.14644 111.57663 145.85901
#> 21  181.28310 122.39107 114.39336 150.32797 128.68084 141.35763 106.60636
#> 22  187.20771 169.63202  96.71729 135.71571 162.20627 119.89980  94.80720
#> 23  172.85208 115.22178 179.68030 157.91205 188.78593 176.29159 140.56338
#> 24  178.95621 145.29692 101.20528 142.87533 151.98038 104.26970 169.14516
#> 25  147.14667 137.24703 156.82768 177.23713 181.12527 171.36404 102.48087
#> 26  147.34597 134.14022 163.05331 131.90287 149.52655 156.29978  93.85413
#> 27  152.90802 146.44020 141.04257 136.31287 190.61304 151.23797 151.20930
#> 28  122.64459 158.02154 179.88275 183.42480 180.33340 141.46881 176.49277
#> 29  169.27507 157.97140 146.05984 142.03223 187.53567 173.29455 187.17228
#> 30  183.20167 137.62149 132.80841 233.41303 129.29707 145.60545 140.76333
#> 31  205.07273 132.18154  99.61973 108.48317  83.77808 130.04458 154.05556
#> 32  153.14495 169.85843 119.65128 190.10035 177.05808 153.55394 126.26308
#> 33  194.51605 142.52559  94.00471 116.59319 144.85648 109.46530 102.41677
#> 34  154.40182 175.60992 171.33469 130.67705 157.88759  63.67768 102.30205
#> 35  140.68773 196.98943 166.85760  96.02842 140.08974 111.03223 127.58223
#> 36  145.29138 109.26896 113.86982 126.98892 155.35792 174.89209 162.48573
#> 37  136.29800 167.17226 128.23291 143.96162 151.36465 161.83121 142.49691
#> 38  138.34913 139.21405 158.48999 122.02024 153.90369 133.46921 137.58231
#> 39  203.95084 177.02216 121.63181 128.92411 185.71896 190.84917 174.79181
#> 40   92.27296 184.12148 118.35234 126.23725 172.29486 102.02231 144.31396
#> 41  165.62221 118.89003 150.67016 167.72983 116.25621 143.12755 123.26424
#> 42  120.09619 117.58697 137.62403 203.37445 152.86398 133.36488 113.58487
#> 43  154.01298 184.08227 160.86258 136.43813 167.72010 131.03363 164.47512
#> 44  183.84032 146.82819 128.89687 135.63127 130.66426 164.32944 120.30612
#> 45  146.54563 147.88798 115.96455 157.28687 159.74410 151.40129 128.38285
#> 46  141.27954 146.45331 160.99522 185.30688 103.96540  55.45779  83.62427
#> 47  116.33863 190.01271 148.71669 166.01007 152.15226 158.90413 143.35355
#> 48  209.51634 120.52017 104.04472 172.84170 147.99156 139.80114  73.08048
#> 49  201.74409 102.92829 111.71592 129.57773 157.99937 171.55231 137.59224
#> 50  145.92663 157.07831 128.76880 155.85804 153.46084 122.10967 128.87079
#> 51  144.42131 127.25171 184.87934 151.69412 174.08285 133.39079 108.38017
#> 52  178.59416 166.61688 162.00395 139.33027 119.13503 152.44811 167.38440
#> 53  168.28815  88.56617 114.32569 152.09070 140.18294 137.44816 215.70272
#> 54  191.78922 171.47532 137.66901 107.37353 155.38607  94.06961 126.53785
#> 55  151.36068 141.83232 172.98298 104.98409  93.44747 112.86842 114.41871
#> 56  149.45640  81.18699 136.78756 225.08754 116.88361 177.84743  97.15361
#> 57  134.01041 133.70258 151.39688 144.04690 164.62236 116.00353 131.44997
#> 58  142.71839  99.54939 180.73226 198.32469  92.51900 120.68374 165.68247
#> 59  161.53631 115.81430  81.74668 163.06231 179.22364 138.07161 133.38004
#> 60  154.13544 113.31766 161.97270 122.62656 148.50519  98.47586 139.14340
#> 61  133.91343 117.67349 149.99551 148.90195 177.61383 148.72199 116.10315
#> 62  144.18691 101.08970 145.62494 192.11055 123.46835 143.80180 130.74062
#> 63  141.04330 150.65080 142.70402 143.13659 144.29083 164.99212 142.17921
#> 64  152.30264  93.21002 201.58266 170.03180 134.01665 121.69138 118.79642
#> 65  141.60341 121.23864 154.73529 165.27043 156.96272 169.20308 111.79984
#> 66  149.22470 144.95631 146.93603 136.31118 159.04964 120.61650 164.07418
#> 67  165.26689 134.35241 130.34397 166.08009 132.68189 149.39123 159.59509
#> 68   93.70987 136.72605 189.66651 168.36487 165.14250 128.09215 154.42188
#> 69  144.44594 201.85194 127.34098 161.05558 112.60248 160.74880 179.90268
#> 70  138.56266 206.95359 187.78220 123.81745 136.51862 141.47565 179.45647
#> 71  151.84377  93.40661 167.98474 166.31206 137.43458 127.46616 132.44722
#> 72  133.96882  97.06429 109.94430 141.02320 194.07383 103.39738 135.17769
#> 73  140.22884 163.93619 164.97822 181.86986 176.59289 201.95099 189.49240
#> 74  174.53459 151.98333 145.08969 109.00022 143.96829 166.03861 125.40180
#> 75  145.43176 157.40777 138.27963 165.15764 188.99717 148.18887  97.25255
#> 76  142.09490 120.15069 178.03475 125.58067 137.49105 143.21638 192.14689
#> 77  215.64771 123.95952  91.09863 129.97654 121.94234 114.61252 128.73562
#> 78  200.23633 138.57138 143.03072 144.07052 123.12958 125.82163 162.28231
#> 79  159.16939 132.17561 151.51804 217.50027 142.63800 186.81040 170.43471
#> 80  155.03118 149.20595 172.72645 120.53003 111.67834 130.00157 161.73335
#> 81  169.56329  85.83751 117.46484 142.16235 139.26039 107.80004 159.62124
#> 82  138.17478 148.54647 181.08927 169.69090 119.24406 126.89615 168.53589
#> 83  133.17261 118.37413 116.07333 130.82440 126.55619  83.41754 179.41210
#> 84  166.55315 152.47014 142.27170 134.01810 191.42161 123.62483 163.53694
#> 85  114.52974  85.62377 143.91223 156.08183 126.08921  84.47541 127.41401
#> 86  159.58086 208.28660 131.87681 154.09740 176.50067 101.05134 200.40177
#> 87  164.35313 163.62668 126.78444 120.40321 139.03668  98.61702 106.43947
#> 88  141.12515 129.71218 198.33904 132.65153 137.89526 134.46279  97.18129
#> 89  155.26990 136.82135 153.50532 118.49927 140.56268 151.69332 152.40494
#> 90  135.71115 127.24424 141.41553 126.29564 159.08045 102.69184 235.60492
#> 91  152.81987 133.27170 189.10082 163.21296 138.37590 145.37936 156.00490
#> 92  175.65610 160.34345 146.70458 182.79588 167.10377 159.98218 164.77385
#> 93  166.43462 171.92143 114.64200 164.74260 109.61458 124.59176 124.42486
#> 94  155.00910 178.15907 127.55839 139.54011 124.19385 139.15089 141.95260
#> 95  137.86677 139.46385 149.15161 111.37147 167.28271 172.33258 141.27430
#> 96  150.59313 153.23852 148.62512 215.93447 203.26831 133.79098 110.76672
#> 97  134.70135 117.30370 167.94646 141.66179 139.19647 155.28693 141.24996
#> 98  120.60047  77.17827 139.82978 122.31811 153.41985 169.48187 154.16593
#> 99  141.65780 139.06013 176.34947 144.77479 126.65734 147.63398 188.45352
#> 100 138.61298 152.87614 160.57629 135.83110 138.33135 139.13258 122.18574
#>         B.4_1
#> 1   132.10583
#> 2   193.87917
#> 3   131.91025
#> 4   131.10351
#> 5   135.62012
#> 6   132.88321
#> 7   161.27809
#> 8   190.28858
#> 9   182.58103
#> 10  217.45947
#> 11  163.35330
#> 12  148.59067
#> 13  161.19988
#> 14  117.06387
#> 15  151.87545
#> 16  196.51645
#> 17  173.54606
#> 18  121.11716
#> 19  196.77076
#> 20  158.33885
#> 21  196.61024
#> 22  151.89068
#> 23  132.55789
#> 24  152.52599
#> 25  115.43153
#> 26  157.62745
#> 27  164.43924
#> 28  176.94637
#> 29  110.30678
#> 30  172.46978
#> 31  234.28364
#> 32  139.77858
#> 33  168.83903
#> 34  150.01131
#> 35  127.16225
#> 36  180.30160
#> 37  104.99792
#> 38  166.29537
#> 39  152.22259
#> 40  197.84068
#> 41  143.66752
#> 42  163.94620
#> 43  203.39836
#> 44  149.28523
#> 45  229.44371
#> 46  123.40040
#> 47  141.46609
#> 48  137.44381
#> 49  116.39846
#> 50  147.11308
#> 51  159.94676
#> 52  121.77747
#> 53  171.23341
#> 54  136.01087
#> 55   94.17876
#> 56  116.72829
#> 57  173.70894
#> 58  174.94294
#> 59  159.69042
#> 60  148.40575
#> 61  137.35404
#> 62  160.43514
#> 63  138.25945
#> 64  120.43944
#> 65  153.76111
#> 66  178.34311
#> 67  137.61616
#> 68  164.49875
#> 69  179.45673
#> 70  143.80948
#> 71  179.34836
#> 72  148.18808
#> 73  177.02789
#> 74  146.43330
#> 75  110.58303
#> 76  129.61910
#> 77  102.23716
#> 78  156.40942
#> 79  187.54360
#> 80  171.73010
#> 81  118.47271
#> 82  190.92404
#> 83  153.80671
#> 84  209.27932
#> 85  121.35361
#> 86  149.12564
#> 87  150.06387
#> 88  169.62926
#> 89  124.34729
#> 90  109.04972
#> 91  155.64066
#> 92  179.50534
#> 93  110.12417
#> 94  168.16725
#> 95  205.40802
#> 96   92.15110
#> 97  159.31478
#> 98  156.38382
#> 99  152.91713
#> 100 172.32322