Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10, 10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(
  subset = rep(c("A", "B"), 50),
  level = rep(1:4, each = 25)
)
rownames(meta) <- 1:100
scRNA <- SeuratObject::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA,
  by = c("subset", "level"),
  min.cells = 10, max.cells = 20
)
#>         A.1_1     B.1_1     A.2_1     B.2_1     A.3_1     B.3_1     A.4_1
#> 1   151.85851 129.32866 119.12705 234.07095 211.53891 165.37947 133.65546
#> 2   135.98506 124.43642 124.23891 159.56484 156.11300 143.48370 209.31983
#> 3   140.05823 146.41021 169.99050 197.74414 115.71811 114.87678 170.04947
#> 4   147.28887 138.37102 168.56292 159.18502 143.83947 140.31843 140.98844
#> 5   173.91706 168.70002 158.94123 151.38023 162.83527 110.50706 147.00482
#> 6   153.72791 155.57275 159.54987 209.50205 124.50988 132.29883 176.72112
#> 7   177.08839 119.74624  83.23707 170.12898 119.25473 156.40787 121.24138
#> 8   142.04517 125.07181 108.59396 219.00630 119.92358 101.48448 167.55287
#> 9   123.79483 171.21915  94.88767 102.42527 174.58928 130.26716 121.11512
#> 10  135.64620 111.98192 110.31113  99.04334 103.39140 121.85037 152.25216
#> 11  158.73954 158.32048 106.15238 174.06191 118.98743 133.59611 147.14807
#> 12  160.38350  82.90031 106.07851 136.31031 158.05118 148.90359 112.01827
#> 13  191.12099 131.54739  94.02595 160.18088 104.98622 116.08430 156.39457
#> 14  150.37128 239.72177 141.71805 135.54152 136.90635  97.79972 143.70247
#> 15  173.18881 141.38063 142.27053 160.45302 140.76242 117.93315 176.00701
#> 16  150.67925 163.73680 126.79456 142.13623 130.92973  77.14715 163.28908
#> 17  119.52625  88.88855 165.22124 130.32872 134.78696 129.64964 131.96993
#> 18  248.13265 124.18420  88.15014 148.68296 125.69516 136.49270 127.08199
#> 19  170.44371 136.34589 161.37495 148.68355 167.95914  85.89449 142.64490
#> 20  159.40151 160.26304 139.83074 142.77816 115.06299 139.93691 165.88519
#> 21  143.33677 168.31443 120.19197 184.16236 115.94542 140.23190 137.16040
#> 22  147.58840 114.44306 158.48975 147.39224  96.61831 126.14714 148.19701
#> 23  186.04189  94.23673 120.82571 168.87330 119.28318 174.51426 133.12327
#> 24  164.92400 137.31923 129.28520 163.69552 161.63219 108.62270 139.93268
#> 25  112.79368 123.27052 110.87550 207.45082 153.58634 126.09794 138.09728
#> 26  102.41428 158.15757 110.22894 179.03758  99.68767 128.49614 145.19686
#> 27  164.41611 218.87934 122.11660 167.43519 153.67641 144.76459  69.37894
#> 28  205.86369 170.43007 145.19439 170.86451 175.67065 148.62264 116.03079
#> 29  170.34991 154.81904 107.62298 159.32608 155.87805 150.81831 115.46603
#> 30  133.26441 110.03899 176.88155 123.93829 131.31392 166.79356  95.82185
#> 31  183.72171 142.08023 138.20646 149.07919 154.64927 133.76990 111.65679
#> 32  153.12925 124.83960 149.21968 192.87872 143.03775  94.24735 126.88189
#> 33  212.59073 147.85971 134.03520 155.34376 123.05304  98.47616 173.31186
#> 34  144.63741 113.71642  99.66684 153.36986 131.32803 149.70413 172.41676
#> 35  125.94151 116.52211 186.64853 113.13858 125.31413 136.56896 109.39069
#> 36  164.08554 192.59213 146.70021 174.89682 162.14713 172.98620 108.23646
#> 37  163.97779 156.78344 183.69031 111.52898 196.02674 153.81809 166.01638
#> 38  176.23570 199.81547 126.56491 113.36145 154.53510 104.27681 162.74575
#> 39  107.14148  97.51883 122.22025 137.25367 126.62877 106.79387 138.98112
#> 40  170.77348 172.53251 165.99286  79.86555 123.14941 128.34464 168.63959
#> 41  186.87060 106.48137 149.89388 131.47169 159.78089 173.82740 105.24261
#> 42  154.67187  95.14881 109.57716  77.87293 145.06069 176.03566 105.62550
#> 43  157.53286 133.31347 132.17366 193.42247 165.94681 156.03993 160.62374
#> 44  179.20035 157.73368 151.20325 187.58880 131.62993 119.43472 126.29731
#> 45  132.95523 158.67827 153.72457 141.82300 146.49142 135.82046 151.51257
#> 46  137.79891 132.23115 106.90416 135.98878 186.59391 185.22507 140.43820
#> 47  134.64723 164.82198 138.23596 114.83858 218.09023 140.74100 134.21983
#> 48  175.00320 125.86132 160.51962  88.96192 187.54853 110.17173 152.73432
#> 49  247.70398 145.85491 111.62562 142.37519 115.64175 122.98024 134.44938
#> 50  185.69814 186.21819 109.33647  82.31628 151.61994 164.34013 130.00882
#> 51  162.33982 101.46987 141.85616 125.20847 142.80655 180.78118 118.68259
#> 52  170.38283 139.24741 225.98330 127.67771 163.26122 131.03722 138.94624
#> 53  167.78977 129.91394 147.20676 118.54645 178.60334 143.18870 161.94267
#> 54  151.47010 149.63798 166.18216  92.39452 146.08436 108.98394 105.50801
#> 55  165.64171 160.35045 107.40779 117.99663 153.38410 128.27553 100.70525
#> 56  183.53357 148.57915 149.38065 212.65067 127.86338 154.83453 143.73035
#> 57  141.32845 126.72844 119.95015 157.94271 170.30345 103.88900 178.48106
#> 58   95.00959 123.59684 135.72698 195.55666 106.54184 170.89627 151.31217
#> 59  165.33349 138.82236 131.25859 150.14242  91.99837 151.50209 111.15866
#> 60  199.37679  99.17994 148.34215 107.37527 138.40994 151.90797 184.30715
#> 61  196.91480 113.50609 139.29880 227.69739 191.88822 195.18886 191.29107
#> 62  177.70394 197.99680 177.98212 169.94241 156.59752 126.65625 157.16873
#> 63  157.02786 142.74437 192.28492 108.42320 123.54692 193.95376 102.42286
#> 64  170.34261 125.18126  78.53219 184.87720  97.66982 159.78988 105.29832
#> 65  112.15696 118.06160 131.25462 114.87749 160.91500 126.13690 130.75475
#> 66  124.51072 161.28982 130.01561 136.95008 136.75783 123.88291 129.20667
#> 67  108.74387  93.23117 144.50955 187.36911  95.15502 139.34157 126.11615
#> 68  181.36236 183.54790 109.88236 151.03233 217.52999 171.77211 158.50207
#> 69  159.28640  75.50441 159.06825  85.85345 161.30819  94.77747 169.23673
#> 70  166.66329 159.94169 132.21866 161.48861 175.59579 157.33497 139.43314
#> 71  109.88816 165.90667 145.45092 153.98424 139.19376 110.08093 134.72859
#> 72  161.29659 123.62214 151.58721 107.06941 151.42624 126.68452 132.67061
#> 73  110.26524 149.70208 120.15129 162.50488 149.84300 120.38077 146.49418
#> 74  151.16151 137.79163 144.92049 151.35339 133.44235 115.96623 157.80127
#> 75  143.04016 194.28933 127.50870 121.43262 243.96068 115.37309 121.61085
#> 76  168.33757 140.32708 165.16675 195.03306 157.78992 128.94015 103.86268
#> 77  159.96082 140.32107 132.98550 176.13698 197.03607 147.09140 105.19030
#> 78  132.56533 102.38932 120.20024 168.03888 181.42208 139.68037 146.61223
#> 79  138.19369 105.84971 140.65363 193.37316 107.80388  86.59532 184.73841
#> 80  136.02866 121.95110 134.64391 113.37838 125.51384 154.80162  80.78908
#> 81  139.80824 124.32849 138.61371 153.35718 175.26538  88.64275 151.13256
#> 82  198.65941 161.14378 106.65693 115.40917 141.48121 108.68353 164.96884
#> 83  159.44117 160.23601 126.27397 173.91229 128.98668 148.85181  97.79192
#> 84  140.45931 115.67620 164.06418 126.26447 213.15602 134.81490 168.07734
#> 85  207.59411 122.20133  86.71463 157.48935 151.32613 106.88647 142.35241
#> 86  128.10834 215.70941 138.22396 173.45591 117.10122 130.96313 199.29089
#> 87   88.97844 169.95369 119.81350 182.85624 203.61626 131.24998 117.67796
#> 88  184.50403  95.44179 122.20776 178.43147 151.23919 202.13556 130.48543
#> 89  156.39644 179.04826 106.91336 102.48595 129.70507 144.67810  97.68123
#> 90  138.19359 143.75926  82.61368 187.71362 163.12922  96.17376 173.15966
#> 91  136.66135 117.67470 134.54826 189.75078 168.39392 150.56504 149.54110
#> 92  182.88741 169.19108 118.88157 171.33526 142.26496 111.63835 155.77087
#> 93  157.87221 125.05395  94.19577 136.97721 119.82663 130.82713 125.43281
#> 94   99.85745  98.67581 134.02041 135.96771 181.14083 145.71071 147.03572
#> 95  156.80936 125.25770 153.67030 164.83712 152.72942 162.49866 155.45762
#> 96  137.50888 165.15250 163.54746 139.93291 165.87736 126.77545 116.59088
#> 97  108.99969 156.04582  90.97252 142.33997 135.81835 155.32659 110.00038
#> 98  157.75687 163.13114 160.61010 119.52152 143.32283 170.93352 123.90907
#> 99  168.98651 173.01212 100.61654 130.58318 133.76913 131.19003 101.09571
#> 100 136.94216 179.25083 104.27475 136.36970 136.01082 121.97406 155.25192
#>         B.4_1
#> 1   199.76046
#> 2   126.76170
#> 3   179.63188
#> 4   109.96637
#> 5   183.29512
#> 6   152.34980
#> 7   141.42806
#> 8   154.83541
#> 9   181.46272
#> 10  136.03392
#> 11  214.45711
#> 12  171.59880
#> 13  119.16397
#> 14  167.09242
#> 15  159.51982
#> 16  129.12640
#> 17  194.44134
#> 18  117.58687
#> 19  196.10886
#> 20  110.77655
#> 21  164.45866
#> 22  115.61390
#> 23  162.67784
#> 24  127.08220
#> 25  192.52148
#> 26  127.31739
#> 27  127.41254
#> 28  147.97872
#> 29  153.69087
#> 30  154.63372
#> 31  133.64866
#> 32  161.75484
#> 33  122.55331
#> 34  146.56158
#> 35  154.69392
#> 36  126.57938
#> 37  137.49633
#> 38  166.88723
#> 39  198.91757
#> 40  110.01994
#> 41  164.71280
#> 42  150.87960
#> 43  134.18546
#> 44  161.64283
#> 45  156.71127
#> 46  146.86410
#> 47  209.50895
#> 48  139.66801
#> 49  131.40509
#> 50  140.06980
#> 51  128.75631
#> 52  136.83176
#> 53  102.83288
#> 54  155.65895
#> 55  187.00020
#> 56  135.03058
#> 57  142.17525
#> 58  153.14479
#> 59  184.64516
#> 60  180.69315
#> 61  134.42375
#> 62  205.67120
#> 63  165.45439
#> 64  132.08365
#> 65  148.15791
#> 66  161.40073
#> 67   82.70774
#> 68  101.50217
#> 69  162.31547
#> 70  127.16499
#> 71  137.43876
#> 72  117.74963
#> 73  108.09279
#> 74  163.60144
#> 75  165.90934
#> 76  178.72401
#> 77  115.07008
#> 78  159.10884
#> 79  123.37111
#> 80  169.34514
#> 81  125.22665
#> 82  145.45190
#> 83  153.71556
#> 84  152.19334
#> 85  154.66862
#> 86  212.94222
#> 87  130.09051
#> 88  147.52623
#> 89  145.12568
#> 90  171.25253
#> 91  130.06615
#> 92  133.71002
#> 93  192.31611
#> 94  155.11604
#> 95  149.45461
#> 96  119.62282
#> 97  149.99595
#> 98  189.93239
#> 99  123.33583
#> 100 137.92327