Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10, 10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(
  subset = rep(c("A", "B"), 50),
  level = rep(1:4, each = 25)
)
rownames(meta) <- 1:100
scRNA <- SeuratObject::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA,
  by = c("subset", "level"),
  min.cells = 10, max.cells = 20
)
#>        A.1_1     B.1_1     A.2_1     B.2_1     A.3_1     B.3_1     A.4_1
#> 1   171.8952 129.63762 122.49938 164.04706 146.27869  73.52776 169.89775
#> 2   167.3485 201.42306 153.32119 153.84424 159.19431 125.14279 168.61448
#> 3   176.2009 159.89588 188.23349  82.15291 125.98376 105.62394  94.17028
#> 4   158.1543  93.27863 132.17809 173.56757 197.47566 140.07371 120.62443
#> 5   191.9134 100.01310  79.57310 212.05805 167.47398  87.53377 155.38162
#> 6   170.6351  96.16676 142.90564 158.03851 139.21305 163.28290 145.84662
#> 7   160.9183 117.19491 102.86440 188.80246 130.35473 120.55131 178.32554
#> 8   140.0892 144.28821 112.71551 164.99318 151.25878 126.73413 112.05833
#> 9   145.7507 122.56854 123.99980 152.70429 180.70844 117.11373 129.12069
#> 10  155.5118  93.59104 107.24808 122.37171  97.09199 162.00884 143.69379
#> 11  123.3543 152.15127 185.29692 179.15009 142.47101 194.29845 113.11676
#> 12  144.6802 186.49789 122.97359 165.41808 160.60386 128.15470 124.51715
#> 13  152.1608 141.82927 142.01302 175.89175 199.76348 128.12298 121.55873
#> 14  177.1273 135.16431 187.17839 160.77817  95.39358 102.70801 138.98697
#> 15  191.5447 110.05661 149.72683 138.43824 147.68231 135.94909 164.71634
#> 16  138.8561 153.05744  96.98515 169.73119 157.51057 147.01607 184.96984
#> 17  145.9385 111.86196 187.44338 179.62943 129.26179 157.08849 123.17624
#> 18  167.8197 163.30962 233.52875 167.15887 158.00327 103.65981 133.54656
#> 19  139.4777 169.69227 123.82786 129.25260 168.62696 136.27077 157.91327
#> 20  138.5956 193.36987 160.54984 117.65791 191.36589 112.78073 193.95745
#> 21  118.9909 126.79234  96.14350 181.86431 167.92428 165.99960 129.41595
#> 22  195.0682 151.35544 161.66901 166.80629 190.95528 154.23539 156.71531
#> 23  130.4639  59.68839  99.05875 175.93444 124.37928 149.15627 161.64632
#> 24  172.2393 122.14089 143.57283 145.33677 136.29583  84.95543 201.79548
#> 25  156.9393 179.31602 178.19647 139.58226 138.57555 135.41122 160.89826
#> 26  152.0135 172.51636 135.21184 148.45124 189.34093  92.92599 155.29333
#> 27  185.1893 203.15629 132.33745 124.10881 154.15369 148.94650 133.26569
#> 28  113.5736 138.56129 174.97880 172.32895 163.30652 172.53641  93.35123
#> 29  193.2781 156.55969 183.96802 156.63583 133.27636  74.98250 113.52762
#> 30  125.6344 149.90492 101.03008 156.29828 120.60199 122.35552 113.34499
#> 31  148.6728 155.72144  98.38002 159.69829 176.00898 129.19031 177.48297
#> 32  154.4573 146.77079 164.45995 164.31665 153.02588 144.96535 109.47542
#> 33  128.0851  80.72183 156.26326 169.69430 158.10484 126.75383 120.32463
#> 34  178.5446 109.47331 126.29021 160.98736 150.00909 179.32042 173.18431
#> 35  135.0265 130.68965 137.23123 203.23313  99.86421  83.56512 120.20082
#> 36  182.1704 138.92726 184.10947 201.01310 177.33492 142.72331 137.81868
#> 37  143.6029  79.08520 131.57257 144.29931 152.73508 157.91015 107.87117
#> 38  187.8854 148.11271 102.70322 117.16427 217.54656 150.03467  91.51893
#> 39  167.2694 113.44781 127.33207 146.32374 218.69756 102.23367 112.49948
#> 40  138.2259 140.37370 135.80667 147.49071 120.09318 138.09984 130.22967
#> 41  168.2539 119.35235  88.29794 107.59366 157.15517 187.10685 153.35025
#> 42  163.7235 108.34764 162.41053 142.66392 160.37328 160.08085 149.86120
#> 43  156.4376 153.77053 123.67521 156.74785 146.28217 107.88764 161.93560
#> 44  199.6425 148.20697 129.65585 171.47356 184.01173 185.81857 137.61808
#> 45  160.3284 134.92901  87.56482 154.22931 164.09988 128.67180 177.07474
#> 46  106.5771 154.21313 124.80504 170.16337 162.11841 142.35353 140.08677
#> 47  233.1961 144.19922  96.52038 121.10335 121.31033 167.10910 153.26516
#> 48  177.2978 124.99886 188.54302 174.36516 186.83542 134.45075 141.33830
#> 49  150.3251 145.13878 143.17003 174.93100 158.95860 116.74415 186.03484
#> 50  158.7823 164.07049 188.61120 157.50386 210.28630 124.11041 119.97777
#> 51  108.2581 126.93905 120.72680 147.53669 134.62587 163.84782 142.62268
#> 52  194.6064 147.93780 146.52065 164.23567 137.09946 135.74640 128.26238
#> 53  143.0865 132.00008 139.73471 185.12195 145.16898  85.10313  77.64384
#> 54  176.3748 181.43703 134.36506 209.58163 143.40310 166.87500 119.54222
#> 55  145.1040 112.94273 150.41334 152.81375 158.93801 211.13390 137.89790
#> 56  158.9748 146.85418 202.26598 139.87333 161.16155 139.53439 183.93339
#> 57  121.2426 146.15318 146.85482 132.13492 144.85725  95.64430 163.21719
#> 58  112.1850 136.28522 154.94210 148.19837 146.20151 164.57078  81.26082
#> 59  173.7295  78.75616 219.28398 166.42616 133.00883 147.33765 100.17392
#> 60  210.1795 119.38051 124.99036 145.36214 116.99737 153.38983 169.45854
#> 61  116.6562 145.55497 137.18951 222.40744 205.14426 153.91624 135.97165
#> 62  165.1238 117.86248  86.66905 161.93183 198.25079 116.69410 124.34142
#> 63  187.9139 171.68409 122.25759 168.32988 124.95133 159.50583 113.06247
#> 64  137.6864  95.64253 170.97392 143.96275 179.46386 146.48215 136.96839
#> 65  175.8515 125.51092 143.41113 105.48788 176.44438  92.28797 161.10216
#> 66  169.3483 153.65539 117.43257 182.25322 143.89728 150.17365 186.88867
#> 67  168.3778 134.57271 171.17311 155.24132 197.81857 133.54241 123.63598
#> 68  137.0731 158.45263 113.52904 159.11749 178.30053 110.66216 110.10761
#> 69  164.2737 159.34140 157.95530 116.26268 173.52968 151.04380 155.44510
#> 70  131.3655 169.41772 112.25882 176.78120 136.59522 125.11034 160.59979
#> 71  161.9131 141.00597 153.66446 164.08220 143.87425 117.08661 152.62386
#> 72  205.9322 140.01027 151.74819 168.01248 135.03002 163.11940 136.47732
#> 73  141.2687 157.37793  94.53581 182.93534 166.59152 162.14593 166.86863
#> 74  118.0366 121.20241  91.58943 140.37827 173.45831 161.96567 172.87917
#> 75  134.4180 124.45890 110.47043 146.91807 150.29473 122.34877 128.16363
#> 76  155.1468 177.67270 210.09437 119.39210 145.06346 102.43921 171.70762
#> 77  134.3259 135.86950 119.27954 148.24680 145.50497 145.71043 119.17635
#> 78  126.1044 122.48779  97.37167 110.65214 210.86049 176.65600 148.73552
#> 79  115.0373 158.33785  91.44144 141.19959 174.89910 125.30218 146.23835
#> 80  147.3114 168.26265  78.85815 147.36579 150.44798 129.33843 154.87689
#> 81  160.7717 142.48667 106.81980 154.30866 155.69642 127.07989  88.75983
#> 82  168.1439 191.75417 111.23277 112.67686 112.83900 115.38243 170.02206
#> 83  107.9287 111.04445 162.37586 183.66313 153.48663 236.02081 168.45844
#> 84  103.0740 141.14289 148.77428 198.96205 135.40941 160.04566 149.55458
#> 85  130.7071 143.63721 160.47429 139.12805 203.53908 141.48019 120.83678
#> 86  179.3838 159.33644  81.34169 178.11119 154.39189 133.28539 145.72075
#> 87  131.8064 131.42002 184.48752 133.20054 143.18822 125.68414 123.20607
#> 88  146.7534 128.12198 215.11659 159.69733 158.94061 116.78865 129.47278
#> 89  144.0770 118.00434 146.50934 103.98923 128.44787 118.72629 221.52108
#> 90  142.2244 138.87075 121.32261 147.19710 198.29706 112.13629 154.30888
#> 91  133.3230  91.92584 148.15181 167.79396 169.92098 129.85646 122.13533
#> 92  136.1642  91.82749 116.71349 114.91849 166.88246 104.01264  80.00448
#> 93  145.4193 163.40455 139.05221 132.02957 112.45407 187.78549 123.93627
#> 94  165.8972 123.85941 168.25034 121.49204 162.52578 141.71099 119.43796
#> 95  163.8607 116.17589 129.18130  92.29650 190.33293 158.19397 199.00391
#> 96  120.7970 135.42645 140.81888 147.91931 148.87035 160.20707 158.18626
#> 97  156.4680 114.16372 155.82422 112.87993 157.98293 153.55751 132.54930
#> 98  155.7872 174.16869 187.57604  98.26660 151.02050 124.13861 150.00759
#> 99   92.4853 125.82386 180.39028 127.64712 108.74642 105.63450 193.94026
#> 100 211.2354 122.47750 162.35145 120.72575 192.94206  76.61381 127.06343
#>         B.4_1
#> 1   153.33658
#> 2   128.62954
#> 3   138.58466
#> 4   167.79086
#> 5   167.26789
#> 6   144.91419
#> 7   136.23329
#> 8   147.77112
#> 9   171.76037
#> 10  147.17579
#> 11  105.67575
#> 12  151.04137
#> 13  130.02266
#> 14  160.11832
#> 15  165.54282
#> 16  145.93804
#> 17  167.64533
#> 18  137.71534
#> 19  159.96204
#> 20  158.67557
#> 21  148.77125
#> 22  134.16749
#> 23  159.72132
#> 24  236.80145
#> 25  179.76995
#> 26  110.12481
#> 27  166.21362
#> 28  129.07477
#> 29  235.00885
#> 30  222.63685
#> 31  132.10319
#> 32   93.01964
#> 33  147.23350
#> 34  102.35420
#> 35  185.59869
#> 36  207.28760
#> 37  140.86749
#> 38  144.86404
#> 39  136.93049
#> 40  212.24890
#> 41  150.31273
#> 42  147.86169
#> 43  122.91677
#> 44  127.04585
#> 45  138.97866
#> 46  145.06092
#> 47  158.78093
#> 48  152.51018
#> 49  177.32743
#> 50  117.95740
#> 51  112.42632
#> 52  159.11745
#> 53  161.25623
#> 54  145.82746
#> 55  126.30880
#> 56  157.22790
#> 57  133.11544
#> 58  150.33249
#> 59  207.18652
#> 60  126.60134
#> 61  111.89125
#> 62  175.88331
#> 63  178.61363
#> 64  208.45998
#> 65  172.66761
#> 66  155.27084
#> 67  205.36158
#> 68  220.01146
#> 69  149.45874
#> 70  175.82801
#> 71  113.59393
#> 72  202.17997
#> 73  168.91073
#> 74  163.93932
#> 75  124.46271
#> 76  155.60417
#> 77  167.55969
#> 78  151.75962
#> 79  166.52137
#> 80  142.71793
#> 81   98.91487
#> 82  176.74536
#> 83  186.71563
#> 84  133.43473
#> 85  185.85002
#> 86  134.14708
#> 87  127.44813
#> 88  129.74729
#> 89  131.75328
#> 90  122.98843
#> 91  172.42142
#> 92  217.45651
#> 93  149.19416
#> 94  165.31522
#> 95  157.59870
#> 96  143.91783
#> 97  180.65358
#> 98  147.65392
#> 99  117.64681
#> 100 187.15655