Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10,10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(subset = rep(c("A","B"),50),
                   level = rep(1:4, each = 25))
rownames(meta) <- 1:100
scRNA <- Seurat::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA, by = c("subset","level"),
               min.cells = 10, max.cells = 20)
#>         A.1_1     B.1_1     A.2_1     B.2_1     A.3_1     B.3_1     A.4_1
#> 1   195.99380 164.86923 126.71029 176.92498 121.05402  78.75120 189.29582
#> 2   109.51854 149.85623 177.11426 155.70547 138.29881 115.72587 182.77430
#> 3   105.47757 152.58589 172.86978 161.48161 129.91993 127.39762 152.94969
#> 4   177.22193 152.07809 105.13814 159.59748 111.16167 189.04879 127.81111
#> 5   120.87185 138.52427 175.28872 168.16064 142.62071 130.25789 145.85318
#> 6   152.72538 124.12205 129.28968 166.02657 107.76569 190.40421 191.29544
#> 7   171.96893 129.91988 104.93086 155.21257 123.66318 146.20456 125.29056
#> 8   135.56092 146.58457 129.18658 170.29590 160.09887 160.77117 174.69875
#> 9   125.60911 152.82462  94.18983 210.26719 150.60193 156.82717 124.34016
#> 10  129.04303 129.00373 146.04350 146.15936 104.05985 160.74260 108.72110
#> 11  112.68318 211.60846 146.61633 128.35121 191.06202 163.78090 123.44947
#> 12  184.82625 198.07635 123.99060 141.97927 151.33862 151.09262 163.85191
#> 13  137.49836 121.79281 132.63789 161.19659 116.03298 156.76245 113.65411
#> 14  175.61215  94.86221  95.55601 168.94581 203.82846 121.72464 134.77605
#> 15  169.42974  83.13863  88.91068 184.09957 141.45853  94.71788  95.50985
#> 16  115.60919 144.12914 131.69584 134.95907 110.78421  98.69322 127.49139
#> 17  195.62094 106.13252 143.28950 146.46040 151.25459 171.21206 123.92275
#> 18  157.30890 121.17314 242.13026 162.81518 176.42861 163.14853 156.20581
#> 19  195.28317 154.81706 109.96907 188.90232 141.53655 104.05352 124.42467
#> 20  167.91387 121.16358 190.02506 170.33259 156.13860 112.17394  81.88132
#> 21  181.64985 105.33401 143.71236 144.37683 170.69385 141.60228 164.44345
#> 22  168.57787 135.64919 197.72460 183.51525 122.23119 107.09535 133.70530
#> 23  148.34079 141.42094 150.07827 129.04968 147.28289 177.56710 113.66288
#> 24  157.79264 157.92433 113.74279 144.28615 166.64495 177.39223 129.07301
#> 25  150.48327 124.61490 153.90130 124.18640 155.98870 129.30470 150.72222
#> 26  122.49299 172.97220 103.67487 170.93279 144.35065 129.01120 154.90157
#> 27  164.71653 142.50607 148.45387 102.28802 143.21559 133.63736  98.46823
#> 28  122.57257 125.53323 107.73344 111.00933 169.32376 133.11015 129.04262
#> 29  152.16230 107.41772  58.89974 154.06957 117.77604 159.83486 134.06115
#> 30  144.54926 123.01271 110.27034 197.11620 157.66203  91.80657 170.50159
#> 31  158.63828 182.37486 114.26372 152.50986 183.53352 143.18238 131.72406
#> 32  117.27776 139.97165 153.70512 140.02207 166.79668 140.09876 132.90610
#> 33  162.83395 175.31344 145.81087 167.90079 168.20993 189.45016 149.99349
#> 34  178.27201 163.08598 144.32359 161.02537 168.28296 145.98335 118.45849
#> 35  170.85231 119.77522 161.98482 152.31440 157.56776 159.96286 134.63100
#> 36  153.42709 136.64908 133.99420 140.90382 175.86540 150.59991 206.11503
#> 37  153.17352  78.25825 175.81866 176.29302 161.34473 121.47517 111.98514
#> 38  177.44637 171.97932 122.73073 178.16091 138.93540 233.53366 116.65175
#> 39  120.74850 124.91541 140.55861 125.67413 112.54191 111.48030 144.63609
#> 40  169.24155 196.32875 148.32952 119.32267 143.65262 125.18581 121.09660
#> 41  187.95848 191.51652  96.58788 174.92233 200.58412 109.34442 139.61541
#> 42  174.76026 139.22961 150.65211 162.14267 104.82117  89.09906 157.61647
#> 43  196.76259 178.38762 168.95797 144.04358 122.13530 154.06301 122.50697
#> 44  146.15128 121.56260 191.43537 155.18129 166.84956 160.05946  97.04718
#> 45   96.70615 134.65530 165.84244 185.39031 121.57628 105.81238 118.41826
#> 46  151.47333 163.61086 159.11026 164.58151 131.79720 149.94823 132.39955
#> 47  197.02934 161.82463 154.20401 180.12270 120.56285 141.11322 153.60511
#> 48  273.01349 114.56063 177.15226 151.93103 188.10079 112.92222 135.27464
#> 49   92.41110 141.49961 183.97704 131.14878 146.87222 161.63656 135.09750
#> 50  157.65525 101.62551 156.57708 126.46967 165.48534 100.77367 137.45997
#> 51  146.39860 190.78744  96.77354  98.90488 154.86287 116.81796 139.21131
#> 52  170.69681 112.44692 138.20117 192.34955 113.99454 165.33980 146.31604
#> 53  118.92616 107.40077 135.57894 151.56863 157.13786 135.76470 191.71021
#> 54  139.19224 120.23270 145.15680 118.08140 210.41530 190.53820 148.10524
#> 55  113.78548 156.94741 122.55370 168.11643 163.29580 156.42437 176.89534
#> 56  125.70937 109.53517  98.13517 145.47053 161.76168 134.33292 112.03027
#> 57  130.92719 207.37338 161.54862 148.87909  88.68746 128.75833 159.96466
#> 58  142.36203 117.37927 122.18123 174.22261 129.58345 160.67164 143.73333
#> 59  147.66686 133.17268 148.68265 124.54436 100.75221 111.77866 153.43037
#> 60  115.62903 121.08549 129.24681 213.17132 144.67281 113.79070 140.24665
#> 61  159.30949 175.29187 151.90062 155.97473 199.40261 129.70045 113.74263
#> 62  163.58745 150.56896 101.72176 118.36853 148.70412 117.19223 117.96700
#> 63  131.66680 190.97926 137.67300 216.30631 146.50938 142.63436 118.87188
#> 64  109.56143 153.69698 126.16085 112.41129 170.52735 107.57175 147.17265
#> 65  136.58569 168.03136 110.44137 127.82226 167.44800 152.62202 156.39585
#> 66  121.59764 155.90813 104.17816 155.96836 134.78012 132.38084 126.51519
#> 67  122.84571 151.50533 158.94580 100.65315 137.92366  92.02359 175.89020
#> 68  167.66451 148.89964 150.55004 163.10495 126.92572 134.76410 194.35490
#> 69  167.43380 139.61325  86.72822 160.94241 115.40276 170.45286 218.14323
#> 70  134.37702 152.62390 158.05084 113.69165 250.37220 141.20067 112.05357
#> 71  162.72147 127.36972 179.01827 217.45470 120.22414 154.59168 171.86295
#> 72  176.36743 141.78651 171.27284 114.04292 171.14376 157.20157 165.12551
#> 73  208.65959 137.82716 145.27941 106.65618 108.12726 171.65868  92.48222
#> 74  157.51579 130.90779 126.36850 187.81102 109.11429 136.79194 163.27650
#> 75  203.16537 142.47890 162.53122 129.84085 192.62822 197.15135 154.54558
#> 76  164.05898 189.31799 139.95499 190.93671 162.64806 164.19534 117.56737
#> 77  153.35666 160.71609 149.85956 113.38253 164.08625 177.89924 156.01074
#> 78  109.46365 144.15426 163.10293 155.62349 149.83366 160.85044 147.47986
#> 79  139.03601 126.44780 114.70166 151.39334 206.68230 166.53387 121.13378
#> 80  179.36291 157.69004 167.06512 187.94605 117.29432 133.26171 127.33859
#> 81  180.45792 154.78346 132.23658 154.95146 193.12680 150.37184  85.28416
#> 82  143.98421 141.14043 138.49028 165.22994 141.67981  87.65610 191.94604
#> 83  204.76187 133.34219 151.52330 141.42894 118.55431 121.87637 121.14556
#> 84  124.39010 126.55621 127.19336 116.87307 147.28370 123.62898 143.37924
#> 85  143.16230 125.23156 157.29678 170.18710 154.00678 169.86421 135.35771
#> 86  137.26705  97.05566 168.69505 121.58135 159.83535 134.30090 124.61054
#> 87  101.29643 127.78443 158.69280 124.76597 165.06891 183.86582 121.34618
#> 88  132.06582 138.27465 179.96709 155.51814 154.67198 155.64533 175.99882
#> 89  124.43752 163.35656 212.45172 121.91827 205.34943 123.28199 175.36540
#> 90  157.73805 160.92160 193.20192 187.13394 184.36993 126.26242 102.25207
#> 91  189.58856  92.97552 194.82397 199.39714 137.77205 142.59310 150.98602
#> 92  127.22816 133.33547 126.73355 119.99141 146.22945 151.09521  61.01565
#> 93  191.55192 151.62068 135.03156 148.78976 154.37647 144.87746 146.12122
#> 94  124.94104 148.84083 130.74150 151.11209 156.70906 161.40269 153.66071
#> 95  114.01889 178.53923 144.14031 205.75577  94.86564 147.36587 118.46402
#> 96  149.16487 135.11979 171.23264 132.72792 154.86535 125.44584 154.69850
#> 97  132.67744 141.16701 166.72296  97.08168 167.08952 176.64966 100.20626
#> 98  194.62706 118.50939 132.40821 115.72066 164.40652 116.43605 130.27196
#> 99  150.70342 133.10687 177.23506 132.28482 138.45999 150.18966 133.15560
#> 100 189.83806 113.52257 141.92026 181.25357 203.28917 171.54755 146.90905
#>         B.4_1
#> 1   102.54779
#> 2   177.61652
#> 3   156.09377
#> 4   145.47922
#> 5   123.58526
#> 6   161.48081
#> 7    96.39359
#> 8   142.99915
#> 9   183.73229
#> 10  168.14064
#> 11  180.44308
#> 12  212.41616
#> 13  132.85524
#> 14  137.18019
#> 15  203.43658
#> 16  141.51032
#> 17  121.64202
#> 18  180.56742
#> 19  142.75607
#> 20  123.81773
#> 21  144.18102
#> 22  138.55611
#> 23  182.69306
#> 24  147.06197
#> 25  160.74564
#> 26  167.83465
#> 27  187.72461
#> 28  116.91112
#> 29  172.02689
#> 30  153.91125
#> 31  160.11546
#> 32  155.09570
#> 33  125.74696
#> 34  161.42237
#> 35  124.10502
#> 36  133.62019
#> 37  150.11735
#> 38  104.30687
#> 39  163.40371
#> 40  134.79924
#> 41  121.54485
#> 42  144.73466
#> 43  154.40410
#> 44  103.16218
#> 45  139.23331
#> 46  231.35279
#> 47  186.11940
#> 48  165.37712
#> 49  151.62690
#> 50  203.68664
#> 51  150.68914
#> 52  194.23639
#> 53  162.66488
#> 54  232.05043
#> 55  177.45110
#> 56  152.62431
#> 57  138.47775
#> 58  120.31129
#> 59  147.30962
#> 60  128.19643
#> 61  186.03378
#> 62  141.17804
#> 63  148.70996
#> 64  145.08374
#> 65  139.17297
#> 66  136.11592
#> 67  176.51245
#> 68  122.17033
#> 69  139.41062
#> 70  152.79365
#> 71  164.15022
#> 72  100.89406
#> 73  161.64982
#> 74  174.58726
#> 75  182.89541
#> 76  179.56978
#> 77  130.14293
#> 78  134.60082
#> 79  130.21139
#> 80  189.69051
#> 81  152.89558
#> 82  178.82638
#> 83  100.46544
#> 84  137.11146
#> 85  164.71702
#> 86  168.72667
#> 87  189.59308
#> 88  191.55825
#> 89  162.10692
#> 90  173.62771
#> 91  177.37564
#> 92  147.71363
#> 93  175.38945
#> 94  139.22387
#> 95  193.20858
#> 96  137.54220
#> 97  166.13276
#> 98  156.51973
#> 99  164.77252
#> 100 173.36580