Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10, 10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(
  subset = rep(c("A", "B"), 50),
  level = rep(1:4, each = 25)
)
rownames(meta) <- 1:100
scRNA <- Seurat::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA,
  by = c("subset", "level"),
  min.cells = 10, max.cells = 20
)
#>         A.1_1     B.1_1     A.2_1     B.2_1     A.3_1     B.3_1     A.4_1
#> 1   123.69112 168.35565 129.55290 178.16920 140.54498 164.98566 148.54496
#> 2   181.32433 130.66670 128.47298 142.15712 207.36398 211.12283 167.20071
#> 3   118.55447 132.24672 148.40020 119.47213 150.23688 180.80653 104.16724
#> 4   114.67977 146.32302 102.51123 151.85461 116.52941 100.87942 141.55998
#> 5   185.27451  89.32150 145.70129 163.40734 103.00986 147.21816 152.59332
#> 6   195.94829  83.88172 100.98442 152.41432 179.42777 116.79305 135.76428
#> 7   158.16825  97.68827 149.77217 150.62792 149.46321 140.09821 152.57336
#> 8   151.54519 115.40005 106.64078 122.74940 180.00086 109.85078 126.54867
#> 9   175.30997 124.86750  90.05986 108.83700 174.39189 131.70384 136.13194
#> 10  201.49474 163.67534 167.35495 173.40440 224.25670 145.10950 137.42678
#> 11  138.59933 180.91844 160.92619 175.57478 181.85609 177.55444 117.58937
#> 12  177.88463 137.86411 124.53821 134.42297 156.01435 152.05658 158.55217
#> 13  230.12052 143.66556 112.15531 186.49498 138.25115 143.54326 144.97060
#> 14  203.48635 138.81711 158.44467 127.39883 131.17082 148.50991 140.60998
#> 15  157.08010 141.17549 145.53184 123.18776 165.86301 106.38901 151.56465
#> 16  207.59677 136.02096 142.15101 181.42617 186.75061 128.39606 100.97619
#> 17  170.72795 117.07387 111.78185 145.44565 197.70265 145.41585 145.96983
#> 18  125.30082 124.77943 164.68531 141.75573 167.11526 173.75036 163.62096
#> 19  142.75090 110.25099 118.20036 119.03183 165.88346 137.01946 116.26014
#> 20  160.81809 146.23965 141.37099 132.73952 130.77395 128.99270 148.47556
#> 21  203.67084 172.73196 135.62767 135.30034 160.27947 129.91581 138.02091
#> 22  157.02429 157.66058 119.94452 174.60793 187.58475 172.56237  98.69667
#> 23  136.49139 141.17487 125.70111 142.82585 124.61401 153.49949 144.80451
#> 24  179.94656 180.39081 145.10427 133.59683 167.48446 138.63356 128.08752
#> 25  109.34033 102.91794  94.88109 126.14917 126.86159 134.89127 150.67532
#> 26  146.25455 137.82503 162.50160 147.43609 168.44960 149.42983 166.37465
#> 27  126.05192 132.59505 119.51970 176.39882 155.30976 195.77905 126.29118
#> 28  107.48674 145.19364 152.31555 163.37676 133.87316 158.04917 112.48855
#> 29  162.44404 145.78914 114.97354 113.36686 129.88701 189.66281 140.56648
#> 30  102.96476 132.44021 131.19214 138.57753 177.51018 123.62383 110.30653
#> 31  151.90704 166.26860 112.19737 163.91166 151.94299 156.56977 126.32156
#> 32  173.27177 164.62586 115.14725 145.28910 214.84345 126.76492 109.81142
#> 33  193.58163 165.94065  99.97043 186.92751 159.01592 168.67284 185.45660
#> 34  170.90329 131.89589 159.39945 140.29893 161.41981 116.59421 126.92586
#> 35  175.04479 156.55853 115.45732 196.89653 170.31758 140.30252 138.85005
#> 36  124.59146 161.20129 152.81442 146.83635 167.60068 121.30847 127.97508
#> 37  140.99080 184.32958 147.19529 103.03341 142.50236 149.03979 111.73317
#> 38  150.38826 121.44827 158.46187 173.84171 162.37717 113.85474 180.27615
#> 39  120.35494 163.34106 110.52934 116.93383 146.04696 125.49693 150.80741
#> 40  126.22494 145.78438 108.62476 121.28654 224.86676 149.27734 118.04591
#> 41  137.65865 130.94094 126.52940 202.56161 127.44167 169.17661 116.57847
#> 42  112.93997 162.12381 154.49609 169.06423 161.10802 136.94106 149.69129
#> 43   97.32653 156.24853 130.21490 150.12311 216.95224 190.56802 123.42435
#> 44  185.53194 113.20572 154.31829 128.94753 153.72709 138.15219 130.74220
#> 45  158.95787 146.88034 137.88519 144.68775 215.11031 137.42757 166.06088
#> 46  123.18236 214.54863 126.95555 155.36018 161.05759 131.20181 119.67223
#> 47  131.84469 170.30300 175.75661 142.07699 139.94781 116.60898 176.78277
#> 48  178.10254 179.41580 126.11970 146.12602 136.11222 125.60014 138.14517
#> 49  164.89555 188.70291 122.36999 148.91799 136.75057 108.76860 142.93934
#> 50  191.15910 126.27581 123.54681 146.66804 167.88169 169.63617 169.65036
#> 51  137.04689 128.70321 202.66169 152.34927 128.30478 115.98055 130.13811
#> 52  120.75074  99.54872 196.23042 143.97137 163.71865 145.53243 137.27598
#> 53  122.17668 137.53100 118.83196 136.80949 135.55355 132.68701 112.14462
#> 54  161.06592 147.20111 160.14172 157.36008 140.52572 141.27212 130.31783
#> 55   97.61580 173.74503 146.45033 124.63580  98.14455 148.32318 128.50181
#> 56  163.99733 151.81440 138.52122 145.02168 127.21592 138.76984 168.00976
#> 57  148.29428 117.30558 129.35244 124.89644 177.40271 103.57390 151.31813
#> 58  190.95527 188.64044 124.84237 115.35106 143.35016 151.84659 123.40775
#> 59   98.57975 124.29758 142.62086 160.73314 203.79584 142.48225 159.28719
#> 60  156.73059 146.24310 148.99499 170.35075 155.33506 129.27541 160.17383
#> 61  104.09094 124.32554 139.53408 126.98251 107.82113 166.07353 152.00548
#> 62  150.94446 155.77955 183.31774 135.43014 123.06175 132.76678 140.32394
#> 63  164.75164  92.58656 182.16146 183.98064 150.15080 159.20242 106.47752
#> 64  212.75402 147.39955 157.04103 150.15809 148.95381 137.62550 152.17543
#> 65  137.78443 166.35263 143.02495 139.05933 129.63507 151.73731 148.61775
#> 66  112.38258 134.83306 141.50119  96.19796 187.57135 130.26564 123.46626
#> 67  159.56702 106.74301 189.69579 147.23175 182.13170 160.82444 132.42703
#> 68  113.53942 131.07311 134.54879 159.02716 192.87693 127.99584 145.59209
#> 69  130.61835 154.48478 110.61358 136.14209 134.78512  99.88663 172.00871
#> 70  164.99856 170.32860 179.32638  99.56855 128.69988 168.14668 138.63818
#> 71  156.44286 175.11404  98.35461 135.58301 168.58353 160.42569 136.47328
#> 72  178.81312 143.73677 139.08050 139.04965 130.98924  99.43555 125.76530
#> 73  127.00609 145.81817 131.69307 114.01488 124.10339 156.56526 161.55088
#> 74  190.08766 120.44324 156.53696  92.55676 140.48399 161.61655 156.71853
#> 75  117.81190 149.23383 154.36162 169.13189 277.03459 111.29599 114.61394
#> 76  196.35056 168.89070 103.20885 178.57752 140.35520 165.33873 138.67122
#> 77  166.18979 224.51083 142.08111 155.02144 147.52789 116.39493 123.37697
#> 78  179.03960 126.03096 131.78916 120.84407 106.30104 131.14381 157.19276
#> 79  136.15757 131.13005 132.00309 104.05897 132.51889 141.23035 153.90674
#> 80  160.00379 164.51575 121.71583 163.58232  68.43095 108.28682 154.09050
#> 81  189.93586 181.88707 134.88012 102.86465 164.11459 153.54934 131.31465
#> 82  173.85365 142.04625 146.88216 159.48441 115.40823 126.18581 127.03636
#> 83  171.27914 208.91737 121.35731 145.67732 130.39940  83.84475 105.64951
#> 84  117.01837 160.03669 141.21828 160.30034 171.02015 129.09244 131.99718
#> 85  156.32721  98.38699 157.33125 103.33916 200.74199 129.89964 157.13812
#> 86  166.65839 141.53244  78.49511 150.90300 139.76690 106.53465 191.19562
#> 87  157.86382 128.92646 113.89137 160.28271  92.37437 138.66526 159.76345
#> 88  117.81819 122.40461 151.34338 140.03033 141.77674 101.33344 175.54174
#> 89  171.06142  94.71313 174.14569 120.36838 167.27509 164.68482 177.86651
#> 90  168.64123 152.94141 107.22063  81.79206 124.20847 151.45652  95.47158
#> 91  187.74537 135.79061 154.30729 141.85737 142.16963 148.40950 122.19443
#> 92  150.21197 126.98556 144.30389 139.91176 169.82688 123.04999 168.62801
#> 93  160.91827  96.32454 117.62469 220.46051 137.38643 182.52771 120.63358
#> 94  195.47990 114.18552 166.46718 101.61059 174.59040 136.56414  91.11704
#> 95  137.30497 120.01578 148.88755 147.18327 175.11672 147.20448 171.09019
#> 96  164.69378 133.71778 171.64381 154.45588 175.87773  89.12230 147.43772
#> 97  134.60578 118.66984 126.01407  92.07335 196.63158 130.73502 137.85436
#> 98  110.19337 165.68868 110.28791 146.11963 180.48702 140.82572 195.85893
#> 99  151.05966 141.92470  93.49108 149.33389 195.43391 168.38792 140.64911
#> 100 154.02232 131.26862 190.76378 139.75961 114.05174 142.75645 136.00128
#>         B.4_1
#> 1   174.27372
#> 2   180.78081
#> 3   136.63872
#> 4   151.73726
#> 5   115.25991
#> 6   167.29640
#> 7   143.92409
#> 8   198.30197
#> 9   208.39681
#> 10  117.93459
#> 11   98.35073
#> 12  107.34302
#> 13  142.93890
#> 14  126.09733
#> 15  168.24050
#> 16  143.99423
#> 17  153.85658
#> 18  128.61884
#> 19  133.26262
#> 20  155.44761
#> 21  135.13986
#> 22  109.63217
#> 23  146.46328
#> 24  141.08111
#> 25  172.71702
#> 26  171.51829
#> 27  122.57803
#> 28  146.88011
#> 29  210.73437
#> 30  127.70840
#> 31  156.02407
#> 32  136.87146
#> 33  179.22674
#> 34  172.74229
#> 35   93.38553
#> 36  149.91921
#> 37  165.99051
#> 38  161.24429
#> 39  142.30188
#> 40  132.93694
#> 41  121.33776
#> 42  139.42087
#> 43   99.51342
#> 44  173.15000
#> 45  148.82154
#> 46  147.31035
#> 47  141.89231
#> 48  158.13257
#> 49  134.42934
#> 50  128.40735
#> 51  124.55524
#> 52  158.75356
#> 53  146.07317
#> 54  151.59523
#> 55  171.89945
#> 56  150.19322
#> 57  169.87505
#> 58  132.84623
#> 59  145.52260
#> 60  140.49417
#> 61  139.70517
#> 62  177.07756
#> 63  158.78100
#> 64  161.83711
#> 65  152.98550
#> 66  152.70809
#> 67  107.44698
#> 68  147.34724
#> 69  140.24063
#> 70  188.71432
#> 71  148.98551
#> 72  188.52901
#> 73  186.67346
#> 74  118.99129
#> 75  143.92886
#> 76  140.84214
#> 77  155.77251
#> 78  111.19840
#> 79  149.86730
#> 80  137.63032
#> 81  166.68579
#> 82  169.60235
#> 83  202.05255
#> 84  169.06673
#> 85  144.44089
#> 86  115.76603
#> 87  122.24079
#> 88  128.49749
#> 89  183.02883
#> 90  123.63667
#> 91  147.47641
#> 92  134.64094
#> 93  149.34341
#> 94   93.34951
#> 95  162.87512
#> 96  174.72614
#> 97  136.31544
#> 98  150.86094
#> 99  126.45209
#> 100 148.62632