Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10,10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(subset = rep(c("A","B"),50),
                   level = rep(1:4, each = 25))
rownames(meta) <- 1:100
scRNA <- Seurat::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA, by = c("subset","level"),
               min.cells = 10, max.cells = 20)
#>         A.1_1     B.1_1     A.2_1     B.2_1     A.3_1     B.3_1     A.4_1
#> 1   171.42548 121.89100 111.75910 134.32461 125.04722 145.75106  78.68735
#> 2   196.13559 140.73590 120.17068 136.84720 125.53562 118.34937 143.04543
#> 3   179.66648 172.76653 130.02237 165.31167 120.67442 141.26619 154.86903
#> 4   155.26190 143.50091  95.95086 126.29507 160.95618 144.53689 151.20012
#> 5   135.06830 194.14111 138.96338 161.09704 154.22423 106.62603 132.41075
#> 6   122.80611 130.80295 137.95772 175.78794 141.73945 146.98754 114.39395
#> 7   157.82778 151.41361 133.00655 123.26725 161.72522 114.58628 124.74208
#> 8   128.59498 192.79558 142.23010 189.21594 124.12167 133.48611 134.06363
#> 9   187.60391 136.16617 125.31718 188.19854 183.80502 174.14365 143.12188
#> 10  169.89585 157.43710 153.33531 132.72547 178.16465 142.98008  82.31143
#> 11  172.71541 151.15556 154.00967 159.81444 181.09620 127.15007 129.34859
#> 12  126.82846 162.05928 137.11028 143.10351 150.38386 150.90733 160.87376
#> 13  152.33450 162.40964 203.80753 150.45220 147.94628 152.48177 151.25344
#> 14  144.83835 155.01539 166.15467 157.40416 119.25530 126.04946 168.27412
#> 15  147.15530 158.40081 187.25096 161.25623 159.35324 102.36758 137.24878
#> 16  106.22726 124.88815 160.68734 185.30047 105.83844 122.36107 176.10776
#> 17  216.15382 164.22962 117.72576 180.73462 144.98157 131.67794 105.92068
#> 18   94.01453 160.16701 117.49496  99.18080 182.85237 151.56001 133.67833
#> 19  172.45432 132.83277 164.10809 194.75047 138.57683 123.47809 117.52206
#> 20  130.64327 226.49643 105.86442 119.65686 132.16772 172.44329 174.61138
#> 21  133.76946 185.27661 155.88337 180.98294 170.71384 149.09138 159.46297
#> 22  145.01454 124.18543 167.10174 133.24827 150.46025 137.75257 131.22754
#> 23  141.80506 118.03843 173.04670 133.67865 116.41870 103.14187 146.35005
#> 24  159.52484 198.96722 200.14229 144.74515 176.27307 126.63423 193.65818
#> 25  190.33008  80.73336 176.26947 132.93188 113.56221 108.49585 175.62106
#> 26  128.68336 126.34253 148.64876 195.01944 170.96975 130.14008  96.04856
#> 27  133.22964 122.11112 155.92532 156.34839 161.28935 128.75321 137.42906
#> 28  139.02167 136.45455 176.91651 117.99990 170.31395  80.82815 122.10990
#> 29  127.06635 167.36490 121.80576 141.21841 134.20615 177.18998 131.29495
#> 30  141.73705 135.42400 114.23724 156.66675 149.59161 115.10814 186.79331
#> 31  184.64002 166.71419 115.85236 104.19041 154.52415 123.83397 135.78531
#> 32  149.25335 156.69263 201.57476 125.44038 163.32164 112.46763 143.15156
#> 33  126.99790 110.80528 137.94562 120.08167 163.57902 205.73063 148.84328
#> 34  164.24019 160.36712 170.82375 149.24054 132.13532 144.56967 180.88009
#> 35   92.91445 222.04171 162.66451 182.57435 158.65230 202.12733 105.23014
#> 36  110.81334 173.32405 109.69071 151.61354 141.39991 134.82334 121.37673
#> 37  154.11805 141.47466 199.06837 212.81587 134.42014 190.91360 103.22479
#> 38  162.18123 131.90265 124.23006 192.44114 155.17091 130.82798 125.76101
#> 39  149.11151 103.59166 128.88266 177.12763 167.10490 159.15859 145.43333
#> 40  145.97873 116.74851 147.35556 108.00304 124.00575 138.71722 185.17988
#> 41  172.69367  97.00114 242.13719 166.09447 172.52693 131.99150 124.13792
#> 42  150.36433 167.83587  94.16088 180.79171 142.80943 184.35896 137.91518
#> 43  131.80106 129.29655 160.88932 139.64310 152.18528 128.59511 114.97283
#> 44  191.72623  98.36098 109.67071 100.00005 157.34423 141.09942 116.63807
#> 45  138.12163 121.54329 159.28065 159.28354 168.55439  88.46087 155.12806
#> 46  146.02321 100.10027 145.63831 158.74525 133.04439 102.80862 117.35371
#> 47  142.04170 193.05931 119.32877 175.31675 198.05240 161.60797 188.16870
#> 48  125.36630 103.12930  92.15437 166.37588 117.77344 163.67121 196.63182
#> 49  107.37105 163.54766 136.35310 123.77522 114.09903 137.97157 127.26649
#> 50  166.16874 112.20179 172.40016 143.47066 136.71912 146.04728 164.71068
#> 51  154.45787 136.36471 158.93151 119.12966 141.03903 138.60407 130.26748
#> 52  149.95472 169.84547 134.41368  85.96067 129.01433 146.53062 130.91448
#> 53  167.98048 162.18037 128.57758 140.04989 212.15324  74.93503 142.17035
#> 54  134.66277 189.10408  97.95456 141.81396 187.80281  94.75973 140.06086
#> 55  138.49723 115.07005 149.39930 115.30645 183.74079 160.74678 150.83448
#> 56  176.29337 178.89779 117.68444 168.10562 140.42170 154.68208 142.07285
#> 57  131.19030 123.85581 194.34624 128.02366 143.07908 109.17893 137.22740
#> 58  210.73525 131.43991 104.45239 189.68804 175.04738 166.15221 156.71375
#> 59   87.27560 134.09225 117.75812 141.44735 132.92093 157.05062 149.63836
#> 60  145.65174 148.03591 149.37600 208.99597 164.12855 103.87848 107.81216
#> 61  189.45656 128.39070 124.00763 141.26885 137.14731  51.19865 115.87629
#> 62  151.57405 126.03354 158.42563 171.11051 128.97759 120.99566 123.73386
#> 63  170.14813 120.17789 138.55519 116.93900 208.41596 174.56566 148.30965
#> 64  204.86485 165.93217 128.00586 143.26783 115.76838 141.00130 173.58930
#> 65  156.52323 109.26128 144.24661 182.04084 162.42936 137.55994 174.14442
#> 66  176.11353 106.91461 160.31391 139.21485 180.56118 146.31874 161.16610
#> 67  123.37034 153.34399 159.47723 165.59131 164.16588 178.40729 118.85190
#> 68   87.42784 113.29332 181.91597 127.73952 161.19796  87.67320 129.10897
#> 69  177.38955 180.74371 129.74871 208.14894 129.13962 145.06052 129.99980
#> 70  171.14016 162.69707 141.05366 183.01360 197.94512 135.20092 143.78224
#> 71   91.98667 176.31269 138.54240 160.08793 130.08404 113.57184 126.54674
#> 72  143.59795 160.29442 170.16828 167.71736 185.35977 180.26295 124.87940
#> 73  111.26653 101.87248 155.67714 163.32529 145.97889  99.90223 154.11035
#> 74  174.27181 101.21331 137.92333 176.05325  84.36443 117.66265 142.24185
#> 75  173.03608 153.96354 134.12353 215.65852 128.21971 166.91759  90.09699
#> 76  150.48486 158.73431 166.79598 189.50156 176.20300 172.61047 149.94682
#> 77  118.11118 174.99218 128.92945 183.79177 159.59973 117.01868 138.87876
#> 78  148.98983 148.35779 134.88311 186.18213 221.29726 189.80957 165.37626
#> 79  175.52410 166.84571 135.09516 130.94755 175.18633 117.92533 126.87948
#> 80  185.97535 141.56478 155.81836 114.36766 144.52088 164.17664 127.62371
#> 81   88.22464 145.14139 200.84407 163.04946 150.89603 112.20345 111.68343
#> 82  205.01269 165.69423 123.98277 176.71447 153.14354 109.61102 135.64590
#> 83  122.66739 161.82999 103.55201 165.87297 265.67459 130.66803 184.28295
#> 84  154.61893 137.00477 168.20785 157.00971 130.03624 137.58844 168.21096
#> 85  129.25641 138.96477 124.23722 153.00767 139.28778 117.74039 169.84889
#> 86  151.10508 112.98841  97.40283 100.24082 146.56616 160.12728  78.58946
#> 87  186.54077  97.56809 126.71267 165.34051  90.72308 101.40511  81.11849
#> 88  186.37723 152.44368 167.54720 139.86723 111.67009 120.40500 157.83385
#> 89  186.16388 137.27099  93.93887 154.19997 165.15686 119.72204 110.36575
#> 90  119.64114 134.81516 101.64066 144.17062 139.54069 155.29139 129.44918
#> 91  200.87348 158.04669 134.22459 106.13160 172.10769 146.73086 128.09886
#> 92  185.80751  96.81055 131.76743 188.03320 163.61999 138.03597 101.24469
#> 93  213.73875 173.77590 121.86892 127.82229 114.17973 109.01208  98.14046
#> 94  158.90349 125.76053 169.49754 149.50875 157.54936 173.79820 124.96596
#> 95  187.32308 184.81724 106.48180 158.96894 165.61528 125.15424 159.61726
#> 96  178.53318 160.75257 134.84691  81.73484 192.63760 156.72325 110.50883
#> 97  158.14137 161.01250 121.59120 194.21538 140.40575 149.04885 126.02155
#> 98   97.11865  87.07816 131.45146 156.03447 139.34858 133.87851 121.53976
#> 99  168.96654 140.69789 156.64536 176.12728 147.81207  86.91996 157.50332
#> 100 131.24661 149.85264 118.47156 154.90738 140.55879 139.42575 171.51179
#>         B.4_1
#> 1   152.04466
#> 2   185.27330
#> 3   118.24595
#> 4   198.95316
#> 5   120.33161
#> 6   220.40395
#> 7   209.46932
#> 8   147.17742
#> 9   152.49528
#> 10  138.90222
#> 11  182.21033
#> 12  139.04537
#> 13  142.14180
#> 14  132.72463
#> 15  100.90228
#> 16  166.47851
#> 17  195.37946
#> 18  204.60104
#> 19  155.81140
#> 20  142.10988
#> 21  176.13456
#> 22  160.86851
#> 23   94.42543
#> 24  180.96859
#> 25  196.42238
#> 26  210.99943
#> 27  120.92548
#> 28  181.12833
#> 29  169.29460
#> 30  158.94465
#> 31  124.05592
#> 32  121.16719
#> 33  171.36871
#> 34  126.60060
#> 35  179.27527
#> 36  195.63922
#> 37  174.16499
#> 38  145.64059
#> 39  155.50406
#> 40  178.04121
#> 41  158.45075
#> 42  125.70703
#> 43  171.30009
#> 44  175.60444
#> 45  139.59724
#> 46  136.27991
#> 47  136.19280
#> 48  197.21212
#> 49  143.26564
#> 50  122.02489
#> 51  198.05104
#> 52  156.56590
#> 53  152.01489
#> 54  126.03691
#> 55  184.83577
#> 56  208.09818
#> 57  194.99071
#> 58  147.35872
#> 59  163.78360
#> 60  162.87422
#> 61  165.65937
#> 62  165.61065
#> 63  152.56469
#> 64  170.84091
#> 65  152.28037
#> 66  134.27209
#> 67  163.43972
#> 68  177.50624
#> 69  145.76213
#> 70  227.59188
#> 71  152.22189
#> 72  128.70357
#> 73  143.89850
#> 74  175.18812
#> 75  146.83888
#> 76  206.69683
#> 77  141.99219
#> 78  147.60107
#> 79  131.36989
#> 80  117.31089
#> 81  177.37556
#> 82  116.46368
#> 83  117.47231
#> 84  160.16527
#> 85  202.22228
#> 86  184.41535
#> 87  114.93314
#> 88  159.60117
#> 89  208.03256
#> 90  116.92290
#> 91  133.54192
#> 92  123.63700
#> 93  137.69006
#> 94  128.40394
#> 95  120.45831
#> 96  205.12772
#> 97  179.35539
#> 98  185.74051
#> 99  157.18640
#> 100 153.81363