Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10, 10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(
  subset = rep(c("A", "B"), 50),
  level = rep(1:4, each = 25)
)
rownames(meta) <- 1:100
scRNA <- SeuratObject::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA,
  by = c("subset", "level"),
  min.cells = 10, max.cells = 20
)
#>         A.1_1     B.1_1     A.2_1     B.2_1     A.3_1     B.3_1     A.4_1
#> 1   163.51662 169.42274 140.24854 119.14874 135.12344 153.30490 145.07689
#> 2   145.36670 162.04853 135.01454 162.39903 126.85126 155.37976 150.90677
#> 3   222.77910 140.17439 164.35492 126.13634 128.77472 237.50894 113.30791
#> 4   154.83207 148.28409 101.97633 195.72151 124.02807 108.26419 168.74130
#> 5   125.14676 144.32464 168.37583 121.58148 123.11970 151.78358 119.88870
#> 6   177.47737 143.62341 136.51660 163.49297 177.23041 126.00476 158.25267
#> 7   109.61274 146.69564 128.44715 186.78721 171.78332 133.16397 148.73623
#> 8   162.55628 117.77380 131.84964 111.74392 151.06365 148.12864 134.88062
#> 9   119.04361 153.41896 105.61106 122.39092 220.78752 134.19361 146.43164
#> 10  172.65701  81.39470 157.14194 128.43431 144.75191 118.38649 149.28465
#> 11  164.78343 187.52525 120.53082 144.74854 193.25063 205.97004 169.15706
#> 12  140.32536 100.86307 145.64496 145.95046 134.38451 132.90772 124.20920
#> 13  162.39952 138.25635 121.92176 143.87326 141.89883 158.47074 156.89789
#> 14  151.26107 128.44895 105.05634 160.58227 148.89346 176.58067 124.19210
#> 15  117.88922 178.63427 169.64262 131.18234 185.48494 185.28842 176.19682
#> 16  190.83944  89.00742 142.50085 135.42890 122.51011 137.51233 193.87553
#> 17  138.54402 117.78954 160.37031 129.43729 174.56841 135.19033 148.10087
#> 18  138.70532 157.86940 142.62867 182.03882 151.14236  98.84110 136.29068
#> 19  144.54066 148.67806 170.67385 160.89281 132.92951 128.10068 129.93356
#> 20  161.83806 157.65260 144.82581 138.43838 170.53347 135.85121 130.81361
#> 21  104.24737 111.91727 131.97098 172.53476 122.46916 144.18751 123.39082
#> 22  156.28060 123.51889 110.12038 167.52500 157.26312 115.95394 129.67440
#> 23  163.02809 138.21856 107.35989 122.76511 158.48297 118.35773 163.53498
#> 24  178.85337 177.30821 162.73984 137.80446 186.96927 132.32494 110.53316
#> 25  126.93475 161.33758 139.10685 208.74211 130.90251 158.82041  93.67911
#> 26  168.37579 173.22779 113.13223 164.43809 204.18298 133.53015 106.48149
#> 27  164.27379  72.88919  90.21550 143.47869 142.32010 138.24662 171.03155
#> 28  144.55805 113.86055 139.85739 223.18621 132.52339 138.42953 162.00327
#> 29  151.55966 129.46887 143.00981 111.50092 156.27304 129.09419 154.39794
#> 30  125.89504  92.13959 164.33113 159.01966 161.16697 133.47370 155.03947
#> 31   95.62005 174.83684 134.56786 133.63284 172.36192 144.33817 133.96479
#> 32  194.14909 167.78475 132.49446 153.89131 163.94860 106.95429 132.78701
#> 33  145.36458  85.92626 106.82839 165.97687 154.22892 212.47343 145.69017
#> 34  159.88784 138.27086 152.12978 133.99514 174.83460  95.80062 140.84291
#> 35  179.21118 122.15062 185.73788 166.77531 148.49418  94.15406 206.67934
#> 36  159.49186 185.38940 193.76568 138.43136 164.93869 182.15502 140.42225
#> 37  134.64684 107.63733 132.09168 154.61806 168.18846 131.83338 102.01761
#> 38  177.04862 127.24389 110.13909 183.38727 115.09216 129.01723  85.92171
#> 39  197.28595 161.05683 115.27051 119.60588 122.94625 164.33518 179.21186
#> 40  135.83020 128.35073 195.27350 140.94775 180.75710 185.25670  99.89030
#> 41  110.42906 131.00010  97.46658 169.06370 142.38244 136.38743 183.28668
#> 42  124.33632 108.31108 154.86621 119.57486 185.68587 159.42745 129.47525
#> 43  179.31889 153.01905 148.01339 142.36136 103.39992 149.05473 111.53357
#> 44  150.26279 177.16847 134.18310 136.85142 163.03041 109.35316 161.37457
#> 45  115.71246 171.33931 143.60304 200.48948 133.71467 122.16714 121.25874
#> 46  169.42637  82.43145 112.15304 171.67122 140.37106 180.11362  79.64376
#> 47  145.59017 143.40190 145.63949 143.56771 139.25019 149.17859 104.76890
#> 48  202.19552 191.82870 136.89565 173.07017 127.02427 135.35709 121.36793
#> 49  134.77163 157.86858 132.49236 136.23448 132.77012 136.98529 148.62038
#> 50  153.00309 133.09203 182.93328 157.84674 176.60831  73.46255 142.12915
#> 51  167.52630 144.24932 146.36086 166.53551 200.75926 164.38414  95.12529
#> 52  124.52846 117.65041 170.88297 163.95332 122.27844 151.69436 172.67116
#> 53  148.96271 127.44546 131.22984 140.30942 183.25821 123.50441 189.99837
#> 54  148.61471 185.70064 152.81060 204.84702 182.10349 132.17943 116.11310
#> 55  137.92474 184.04324 148.26795 129.99866 140.59046  81.03263 158.23553
#> 56  120.33864 161.70307 183.81917 182.82032 174.70194 122.16245 187.67638
#> 57  159.93179 148.87207 162.38745 146.78332 118.55956 161.49422 115.32519
#> 58  121.97241 131.89641 142.28103 137.41930 167.20550 140.80822 122.21445
#> 59  179.26171 128.22187 124.49337 135.64985 208.96374 178.95789 109.21349
#> 60  161.32055 141.07635 133.32333 136.88467 117.45124 136.66482 138.00688
#> 61  178.88764 137.55278 130.86982 143.71977 163.99600 109.87183 134.47862
#> 62  238.96750  88.08713 122.82350 146.18028 137.68632  77.75871 118.18002
#> 63  180.95132 116.12496 125.76909 152.76770 214.06866 133.40015 178.10653
#> 64  157.65047 133.73900 177.48158 217.42361 151.37199 145.12667 120.70395
#> 65  165.30507 119.99199 123.26761 128.45599 102.51155 151.80716 157.40923
#> 66  111.49877  88.98613 136.21055 160.63745 154.69827 189.99565 149.76574
#> 67  129.15312 156.33287 190.87137 149.95621 112.13144  86.96310 201.17433
#> 68  191.03921 135.98787 151.68565 142.33127 131.11045 122.73907 108.56987
#> 69  134.43533 138.31520 161.07670 135.73054  94.83958 136.26995 102.04270
#> 70  127.62115  70.20928 182.53783 183.60581 181.14493 113.61711 106.62012
#> 71  189.52208 124.85939 178.57794 163.03177 137.05295 104.41407 138.91522
#> 72  157.66266 102.68793 170.28897 151.59097 149.89652 182.30528  94.50817
#> 73  106.66726  88.97174 152.73270 186.35420 109.00215  86.54519 135.70984
#> 74  105.48435 153.80013 182.48429 156.69060 187.50593 107.95505 186.62793
#> 75  174.84535 117.60892 169.92058 137.27979  86.32947 165.54736 140.73255
#> 76  150.49629  87.51425 104.58687 166.06042 141.66386 150.75171 108.31269
#> 77  162.00647 122.75784 115.12196 182.97754 151.66937 127.45415 135.33814
#> 78  126.81502 105.12661 147.02716 153.91995 161.60090 124.47696 166.15628
#> 79  142.20926 124.90191 156.82660 147.73546 181.68571 163.49387 166.66846
#> 80  135.81838 178.09522 126.31553 160.65156 112.58702 122.74430 120.01639
#> 81  182.34526 127.66975 127.00426 213.70157 141.99371  84.44075 142.05039
#> 82  180.94911 122.63280 171.33560 163.47188 188.19696  86.43159 145.21330
#> 83  146.94113 134.26631 138.76933 160.95099 138.04504 160.57509 134.67914
#> 84  163.88166 121.01095 150.13464 101.39453 113.33815 193.89037 170.06700
#> 85  124.75405 163.18877 145.43531 181.25053 140.47967 171.11287 113.12939
#> 86  118.64930 123.17390 139.75548 122.00841 118.72476 141.79854 164.07241
#> 87  166.09345 185.60361  76.91822 177.10543 116.21980 151.09854 122.66118
#> 88  171.29554 120.64347 139.51990 178.23262 108.50217 122.46820 139.27957
#> 89  190.10493 114.20550 130.27910 125.16566 141.16749 211.81442 177.94532
#> 90  153.87426 111.09886 148.81626 108.03878 210.80543 158.33822 148.16688
#> 91  156.62253 141.01222  97.96252 131.39746 127.61531 153.42766 133.77878
#> 92  182.12038 151.72138 137.09548 180.78999 141.03274 162.52799 173.73118
#> 93  130.19465 135.25144 159.84637 144.23228 139.39169 124.54242  91.81259
#> 94  133.51188 161.56813 103.63998  85.01374 106.04301 136.14984 108.90903
#> 95  167.65200 102.03489 120.00906 150.77896 194.14217 165.90228 139.17942
#> 96  114.60034 125.17040 129.70531 207.57766 161.74999 124.92738 172.61822
#> 97  166.89607 171.74985 150.98066 177.81814 204.24657  90.68338 111.27912
#> 98  110.13335 142.98820 130.72430 182.87518 130.61270 104.93244 124.83152
#> 99  181.49798 114.98512 140.16793 164.99806 167.91916 112.42440 120.03036
#> 100 177.32785  86.18254 164.55975 156.99167 167.05524 143.12110 213.87052
#>        B.4_1
#> 1   162.5902
#> 2   132.9347
#> 3   102.9124
#> 4   158.2114
#> 5   155.7241
#> 6   104.8341
#> 7   146.6328
#> 8   156.0217
#> 9   150.1486
#> 10  126.4451
#> 11  183.1886
#> 12  110.0402
#> 13  115.3115
#> 14  117.1967
#> 15  190.0742
#> 16  194.2656
#> 17  191.1335
#> 18  118.8808
#> 19  143.3937
#> 20  180.1011
#> 21  184.9714
#> 22  172.8048
#> 23  169.6696
#> 24  169.7062
#> 25  157.2159
#> 26  179.0987
#> 27  185.4832
#> 28  130.6809
#> 29  140.4003
#> 30  128.1640
#> 31  181.8460
#> 32  159.3019
#> 33  160.9492
#> 34  105.5237
#> 35  122.1930
#> 36  164.4414
#> 37  165.5420
#> 38  127.1293
#> 39  168.7050
#> 40  151.6407
#> 41  148.9163
#> 42  148.5920
#> 43  183.8068
#> 44  169.9238
#> 45  118.5721
#> 46  180.2531
#> 47  156.5026
#> 48  156.4224
#> 49  144.9124
#> 50  172.4627
#> 51  130.3284
#> 52  130.9378
#> 53  170.1707
#> 54  165.6286
#> 55  165.7360
#> 56  151.7574
#> 57  162.0102
#> 58  118.1289
#> 59  174.2474
#> 60  128.3149
#> 61  149.1696
#> 62  110.3622
#> 63  181.6530
#> 64  198.5300
#> 65  158.7901
#> 66  134.7765
#> 67  210.5228
#> 68  192.6580
#> 69  170.7778
#> 70  149.3134
#> 71  185.2384
#> 72  145.9736
#> 73  119.4347
#> 74  123.0108
#> 75  142.3716
#> 76  132.6080
#> 77  143.8823
#> 78  165.7776
#> 79  123.2750
#> 80  175.2494
#> 81  187.3106
#> 82  178.8114
#> 83  147.9773
#> 84  125.6771
#> 85  130.6116
#> 86  166.1880
#> 87  151.5518
#> 88  122.6838
#> 89  162.6709
#> 90  142.9508
#> 91  192.2463
#> 92  152.4151
#> 93  166.1970
#> 94  153.8386
#> 95  132.4222
#> 96  147.2976
#> 97  166.4174
#> 98  103.7004
#> 99  138.6321
#> 100 144.2339