Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10,10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(subset = rep(c("A","B"),50),
                   level = rep(1:4, each = 25))
rownames(meta) <- 1:100
scRNA <- Seurat::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA, by = c("subset","level"),
               min.cells = 10, max.cells = 20)
#>         A.1_1     B.1_1     A.2_1     B.2_1     A.3_1     B.3_1     A.4_1
#> 1   180.78037 161.96355 143.09435 124.71884 184.63865 117.14285 108.65896
#> 2   153.45731 117.33659 112.37069 192.36214 119.40867  98.27168  92.19133
#> 3   132.26543 111.65776 138.79344 128.80636 113.83818  85.54876 149.81680
#> 4   167.14802 149.11787 129.08286 136.30640 139.46940  92.28081 126.27165
#> 5   159.96128 160.40664 143.03955 130.76992  84.09381 165.93495 137.64966
#> 6   123.78273 166.34015 150.21333 140.25246 135.51999  59.22741 106.35799
#> 7   162.80730 153.66382 104.11556 143.77187 181.73101 151.32583 164.70516
#> 8   161.62667 149.30042 102.68863 140.32880 220.23220  70.18087 161.52371
#> 9   139.45941  96.11456 139.72860 110.29932 141.75552 129.39923  93.71687
#> 10  147.59074 198.48004 106.61748 158.04212 109.91078 183.49463 111.35430
#> 11  151.26918  64.17915 132.15321 200.08004  94.25090 158.83905  96.69940
#> 12  144.88488 116.95802 165.12700 116.92600 178.82810 167.14343 107.23990
#> 13  160.75272  92.40734 146.73797 171.41703 131.76760 136.72130 112.60380
#> 14  128.89067 119.52578 133.63548 158.24781 139.20329 157.86445 133.20409
#> 15  123.34197 149.29094 137.71920 154.05236 108.77223 152.15313 143.76298
#> 16  127.44031 133.83139 200.49395 119.61635 149.03152 129.12895 154.83062
#> 17  191.96861 219.07481 203.82676 161.04426 151.42978 119.48922 138.39179
#> 18  141.02875 136.32481 167.02237 169.42819 141.06891 144.29812 154.25410
#> 19  161.98744 116.72478 163.48561 180.28775 117.66931 124.44061 139.56827
#> 20  151.29938 176.71009 123.44779 178.70627 164.76155 120.86052 157.63495
#> 21  115.89588 101.84206 114.79316 131.14806 150.34008 169.33206 106.76705
#> 22  149.80306 144.47134 115.66025 151.42640 156.37836 179.34338 150.64336
#> 23  198.50859 141.29179 119.54662 115.02806 118.99235 121.74300 141.85814
#> 24  191.51530 152.94107 129.98148 131.35082 168.09655 152.40539 142.14296
#> 25  141.13251  94.80780 161.20874  84.89931 175.11753 164.18892 161.70707
#> 26   99.65272 145.50007  85.08540 125.08100 221.38406 173.28698  98.03716
#> 27  130.27446 134.26386 105.18159 186.48323 226.70843  96.07056  77.23802
#> 28  188.46022 135.06678 141.56602 182.97116 188.04654 175.87372 154.18995
#> 29  139.13126 123.67515 129.98881 164.02574 133.13674 119.68853 146.50852
#> 30  129.01350 153.94608 107.47653 119.84749 138.76052 133.65653 127.85564
#> 31  148.79368 144.44704 135.90458 197.20082 142.68121 174.11479 161.61619
#> 32  168.79255 207.91760 164.60145 182.13154 204.41345 134.09774 139.65455
#> 33  174.50820 134.00811 111.67407 123.34080 150.89748 104.73903 136.56793
#> 34  167.10326 125.51388 144.00867 147.75426 173.78703 131.87200 171.67909
#> 35   98.61907 189.67061 106.37355 120.76285 180.31311 175.38165 146.92067
#> 36  173.18651 138.53719 109.03498 143.44367  91.66943 108.01334 130.67852
#> 37  179.87882 157.88033  97.21671 135.79022 180.75458 118.63008 125.68287
#> 38   88.63645 139.88136 152.78929 135.70356 189.86360 165.16419 128.96740
#> 39  166.66639 187.25346 171.23336 165.32275 139.44475 194.90947 161.55708
#> 40  135.49275 131.75193 176.26770 125.26759 161.46277 111.17680 122.48232
#> 41  141.02404 121.81903 144.57702 144.75196 197.06625 164.70104 127.14326
#> 42  123.18428 112.36159 190.82116 221.66010 179.57555 112.14550 169.96345
#> 43  175.26609 164.30142 131.51405 179.85292 163.84807 164.41223  99.08095
#> 44  131.72394 166.28967  94.33513 179.03780 189.03958 123.07345  89.41511
#> 45  147.63909 176.65334 128.84292 135.91422 138.19075 146.07706 130.79428
#> 46  111.03123 129.33989 132.47352 136.95973 129.15874 149.69701 174.05988
#> 47  143.91361 160.69622 138.09731 182.23501 131.39690 169.00455 143.32552
#> 48  125.95334 128.10063 126.72684 140.25505 165.81410 178.22667 168.83128
#> 49  168.48875 176.42689 164.30036 199.87735 182.56644 124.67089 144.05934
#> 50  159.95673 212.24148 121.51576 162.47370 137.45195 126.05032 151.16030
#> 51  128.03500 125.45221 146.05566 144.16921 198.60903 128.68638 115.72657
#> 52  166.71997 186.11704 177.26350 154.90616 116.91945 139.99589 130.13294
#> 53  145.59697 124.11528 131.59544 103.06221 131.03150 194.16588 129.08573
#> 54  195.28336 110.06436 102.88927 173.13208 160.41099 134.61387 159.56459
#> 55  168.00283 119.70396 158.97143 181.53732 163.59771 113.54094 137.12733
#> 56  122.23643 191.62612 129.84323 129.74963 178.17862 173.96813 130.75662
#> 57  151.63915 152.34518 124.09844 148.01942 112.97703 102.34793 133.30252
#> 58  183.79540 182.10157 168.36237 117.46391 151.82652 121.09532 137.08633
#> 59  173.15268 102.01441  68.58050 134.62159 133.59076 140.36770 140.48607
#> 60  118.86094 117.35246 137.72511 149.66392 158.29964 117.22464 110.21458
#> 61  133.57222 141.07947  96.62541 164.74319 127.49294 171.26486 189.45613
#> 62  176.42958  98.11501 126.18418  89.71760 163.03071  99.00267 154.91151
#> 63  143.11122 144.38204 129.25830 159.71941 177.82012 111.23521  97.83784
#> 64  134.04720 137.32106 179.23005 115.65474  97.26732 122.58076 165.52676
#> 65  192.96611 152.76303 176.81003 182.45752 176.07734 111.47106 110.01613
#> 66  102.01390 156.68801 140.56084 200.51980 111.94332 109.69549 178.19395
#> 67   89.18438 138.67891 133.11278 178.86983 186.73592 180.88321 132.54269
#> 68  144.75417 128.41982 148.50009 152.86822 192.46243 160.39637 161.79442
#> 69  180.76534 139.45436 124.27520 154.11915 211.70996 152.53856 151.23645
#> 70  148.81352 137.79991 140.89435 122.95153 129.52177 132.25674 183.87080
#> 71  172.38574  95.55421 156.56819 139.94925 158.66493 148.12725 163.78888
#> 72  183.05074 121.84756 131.88065 160.05653 161.81743 169.30041 110.17445
#> 73  187.71703 131.60732  80.09688 121.60566 160.16010 115.02741 143.66660
#> 74  172.59553 137.04745 160.21898 152.90840 135.41271 183.44646 132.09809
#> 75  125.88564 194.57870 104.89759 157.50673 128.21655 170.86744 146.33898
#> 76  202.62022 108.36532 153.31682 135.48256 135.13835 147.15204 167.27486
#> 77  166.65729 116.75016 103.43036 116.22833 163.66906 181.52073 129.65046
#> 78  103.42399 153.76677  72.10521 198.26892 167.84197 129.90179 122.34627
#> 79  169.05507 113.86305 158.98713 175.00616 205.86899  98.47398 190.24798
#> 80  130.12909 115.17899 125.03549 145.98751 169.38779 115.91470 125.37639
#> 81  160.09968 140.38253 149.46535 144.47170 185.45198 130.33729 161.08159
#> 82  138.57305 147.67656 100.61658 155.03339 179.03691 143.39381 213.11141
#> 83  191.56487 122.17536 133.01751 148.41751  84.11546 155.21220 172.61179
#> 84  171.39548 101.27194 152.82623 176.11022 141.89129 203.60517 116.00959
#> 85  158.45921 155.44049 155.99290 171.05713 182.62740 162.45252 120.20361
#> 86  121.55086 115.75882 162.33701 164.82374 162.02329 131.05803  96.08685
#> 87  167.06460 121.24016 145.50728 105.61641 198.34489 167.22498 122.63019
#> 88  137.59552 176.47984 136.34770 117.47321 129.31384 183.25417 146.68057
#> 89  127.21281 183.69521 147.56196 134.99073  98.52528 105.36027 151.19688
#> 90  165.85761 111.77011 152.53691 173.56145 155.76363 141.43017 103.72481
#> 91  119.41362  81.09047 128.28683 138.21117 106.82121  74.42306 162.72111
#> 92  132.03926  80.98886 126.50029 165.98969 137.36236 160.27492 144.13917
#> 93  121.32439 172.51400 155.52224 101.13448 225.48562 125.02370 116.44662
#> 94  147.27644 182.18071 128.39998 139.47997 158.29079 109.02563 115.77105
#> 95  155.64551 142.03857 169.34722 209.40176 218.66160 121.34873 167.48992
#> 96  124.11284 122.79766 152.50299 171.25302 118.96653 121.28842 143.11761
#> 97  144.06974 141.59501  97.70867 153.91647 120.46227 141.03634 128.72753
#> 98  164.26411  87.50790 177.38358 135.99995 121.40694 109.07647 136.33938
#> 99  153.24450 137.28058  83.81626 124.40907 152.30375 165.69718 140.81768
#> 100 139.89846 151.85631 130.10357 134.91099 119.35641 136.97283 131.14614
#>         B.4_1
#> 1   132.39173
#> 2   129.51588
#> 3   142.84036
#> 4   163.08249
#> 5   141.05726
#> 6   168.92848
#> 7   130.11868
#> 8   176.24591
#> 9   202.17020
#> 10  103.45608
#> 11  148.18308
#> 12  173.97610
#> 13  112.18012
#> 14  118.20225
#> 15  120.87899
#> 16  158.26484
#> 17  136.10800
#> 18  123.14027
#> 19  139.53864
#> 20  157.81970
#> 21  162.47434
#> 22  159.15085
#> 23  197.15953
#> 24   94.85225
#> 25  155.56589
#> 26  171.84747
#> 27  184.94640
#> 28  113.24324
#> 29  135.78521
#> 30  151.75938
#> 31  153.22100
#> 32  220.60906
#> 33  163.13813
#> 34  134.60048
#> 35  129.69487
#> 36  113.11808
#> 37  168.24041
#> 38  176.99027
#> 39  141.16839
#> 40  180.91893
#> 41  134.49146
#> 42  148.78064
#> 43  167.01927
#> 44  182.05710
#> 45  198.15768
#> 46  199.02056
#> 47  158.40933
#> 48  116.36716
#> 49  166.53732
#> 50  142.60889
#> 51  170.43087
#> 52  144.95860
#> 53  181.08145
#> 54  183.59349
#> 55  172.43498
#> 56  168.08193
#> 57  160.02636
#> 58  211.33474
#> 59  161.54836
#> 60  250.95720
#> 61  127.51326
#> 62  114.41173
#> 63  151.29548
#> 64  159.82675
#> 65  182.07838
#> 66  146.67823
#> 67  141.47566
#> 68  139.35288
#> 69  173.12125
#> 70  141.94970
#> 71  170.41626
#> 72  192.92574
#> 73  147.67777
#> 74  121.40713
#> 75  144.21409
#> 76  138.83370
#> 77  128.54708
#> 78  206.92827
#> 79  192.68905
#> 80  155.60363
#> 81  135.54214
#> 82  178.54324
#> 83  105.17098
#> 84  140.04877
#> 85  165.23861
#> 86  114.03127
#> 87  149.77164
#> 88  130.83738
#> 89  178.22827
#> 90  101.51779
#> 91  140.04936
#> 92  170.87391
#> 93  128.25883
#> 94  183.36308
#> 95  198.81484
#> 96  154.08126
#> 97  153.87181
#> 98  124.01341
#> 99  136.89705
#> 100 148.50741