Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10,10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(subset = rep(c("A","B"),50),
                   level = rep(1:4, each = 25))
rownames(meta) <- 1:100
scRNA <- Seurat::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA, by = c("subset","level"),
               min.cells = 10, max.cells = 20)
#>         A.1_1     B.1_1     A.2_1     B.2_1     A.3_1     B.3_1     A.4_1
#> 1   150.65435 166.28800 175.40383 139.38398 157.84066 130.87018 162.66136
#> 2   156.49051  92.28544 124.51867 164.19610 161.01439 154.05611 164.12993
#> 3   127.37391 129.86723 128.22895 129.69522 144.96697 136.34747 135.38209
#> 4   171.44335 134.43697  90.80749 148.87860 173.04789 133.94490 156.69009
#> 5   170.08813 179.17704 122.08962 129.21266 152.14289 124.15994 134.43778
#> 6   159.70877 159.28336 155.37013 133.18655 183.85275 138.15073 190.06095
#> 7   133.86652 120.47885  96.41241 111.10319 146.77452 133.44013 133.31112
#> 8   105.67566 164.86872 151.87285 130.88987 128.68989 159.65469  79.71113
#> 9   200.16302 128.72693 101.55246 135.45677 164.90270 113.52367  78.06155
#> 10  127.66408 166.44525 117.59235 170.29006 179.93051 119.43283 116.37071
#> 11  148.15773 153.91165 212.86954 105.21621 141.88645 156.48119 124.25996
#> 12  199.14260 106.31396 134.97504 180.69808 165.46386 158.01653 133.06497
#> 13  168.12455 103.92717 160.68103 193.28369 137.77849 127.46729 132.58755
#> 14  196.75740 127.73511 120.76392 137.35868 158.38078 137.25375 129.12498
#> 15  127.19467 132.89053 157.65481 112.34728 195.09877 154.78154 179.50372
#> 16  196.95133 122.66037 106.23308 204.59262 111.62298 121.36429 116.68986
#> 17  118.47411 127.07801 133.60449 160.11244 175.28976 158.90961 196.57309
#> 18  132.99288 131.29634 122.25766 150.68182 149.05120 158.86629 163.30958
#> 19  128.20823 122.16587 190.73060 114.74187 242.05483 124.94305  96.34172
#> 20  139.89154 190.69674 125.48145 146.36820 154.28044 194.13522 134.51270
#> 21  230.31300 136.78012 219.10770 136.50511 196.95835 133.13813 145.89944
#> 22  171.76579 190.74812 128.92053 178.77298 214.07183 120.35135 143.57230
#> 23  146.79624 139.61726 155.64435 195.90173 122.65939 174.64330 131.60341
#> 24  152.98244 121.42364 159.08109 177.18920 129.46449  94.31645 180.67340
#> 25  122.26273 166.25484  95.67073 145.07813 130.97534 125.95174 179.29935
#> 26  228.88190 136.73964 147.59466 106.98343 166.75747 147.67774 157.71708
#> 27  165.62447 129.48918  79.40396 144.13608 166.41526 135.15911 144.67497
#> 28  164.91612 179.75986 142.13936 199.94811 184.88184 162.52478 137.18889
#> 29  123.08936 186.46073 168.32827 139.01339 170.08878 140.89055  85.62016
#> 30  189.43973 156.52818 156.19400 142.27019 102.65088  88.70415 110.81428
#> 31  121.67958 133.45802 140.49882 132.69120 119.75943 141.70095 149.91442
#> 32  147.71045 180.24011  76.23310 179.94532 198.46359 152.42480  97.37732
#> 33  141.77341 181.57390 166.87322 126.41126 150.17822 132.85692 128.17958
#> 34  153.02971 103.44111 123.12290 150.35804 119.78902 194.49082 157.38538
#> 35  129.73793 110.23338 122.56822 127.95396 226.99751 139.77648 182.68750
#> 36  155.92792 114.66564 146.55061 167.48080 110.99240 115.94055 147.92123
#> 37   69.82898 139.20040 131.90080  99.71076 155.93959 156.64052 166.60755
#> 38  124.33129 204.02241 120.84340 157.19149 148.74019 115.21869 151.43346
#> 39   97.29488  89.71840 151.94731 150.25909 113.01949 177.78133 114.52894
#> 40  116.00462  95.06669 199.41440 117.26706  98.95954 133.05542 186.46274
#> 41  146.57251 123.66187 158.08785 213.00966 101.51117 135.64376 173.64426
#> 42  107.32087 144.88999 205.48182 167.37423 164.03664 141.12671 143.24153
#> 43  134.55659  97.47482 116.96760 125.06075 163.60308 156.49810 102.98453
#> 44  133.23937 136.30846 125.17182 121.47877 194.60446 129.55574 140.63846
#> 45  154.36877 166.27363 136.32559 147.44594 122.68871 120.12857 139.32209
#> 46  170.10137 192.56317 146.62473 141.37474 168.56364  91.39542 150.29785
#> 47  172.64825 189.01144 168.93974 199.09434 159.99107 139.66452 153.77458
#> 48  113.97915 165.14946 130.00010 157.96755 103.60243 125.75206 141.34616
#> 49  128.25496 141.55902 151.99174 184.56987 115.64442 174.93501 146.74365
#> 50  150.86413 156.27136 130.95299 102.08239 184.83024 158.63216 113.12072
#> 51  137.25130 122.44601 127.89438  90.59865 146.27177 169.71077 178.32049
#> 52  125.02596 119.49649  77.91839 108.29336 171.33753 137.37902 200.87200
#> 53  125.72121 134.61084 167.47181 123.95932 169.29120 110.42087 103.81877
#> 54  155.27526 166.11386  93.09675 131.67577 194.64055 128.99133 161.17839
#> 55  163.03707 157.43711 165.72420 156.47824 118.97889 131.58911 125.49454
#> 56  120.64997 160.76951 151.37619 170.12723 129.42320 102.77110 125.75521
#> 57  196.15802 139.51668 179.78611 153.48956 156.88743 134.87700 159.63132
#> 58  139.62106 134.02901  98.22800 166.40246 135.58909  95.36358 174.78406
#> 59  139.05870 197.68776 176.44195 148.77023 161.75781 199.88382 166.65491
#> 60  159.18041 129.66596 127.13275 137.09031 141.94366 134.56120 154.27650
#> 61  149.42145 138.72912 144.73112 145.29955 119.06316 123.26298 116.59254
#> 62  147.89952 152.44584 134.56475 169.69380 175.49761 133.61044 145.74002
#> 63  145.81057 113.66236 166.16297 160.16407 142.26038 193.05266 174.82205
#> 64  194.09673 142.09676 108.73127 137.24174 201.23269 119.52657 162.18479
#> 65  146.96253 115.66127 129.01058 122.96601 122.29772 181.52141 145.09780
#> 66  106.80875 102.88479 125.46561 215.85413 173.50483 173.36670 172.54435
#> 67  176.25972 114.57836 117.86035 145.62571 184.75071 172.47929 129.02134
#> 68   95.35864 145.25579 174.88562 127.42427 159.31657  82.63501 159.59271
#> 69  163.39609 111.36662 127.87477 126.52186 163.18129 120.89959 195.92533
#> 70  158.56889 153.43907 143.06457 121.99698 164.59872 116.67292  84.28874
#> 71  125.92896 178.81286 145.82169 187.47985 159.02709 128.80187 136.61070
#> 72  188.44864 165.29176 166.95098 146.56544 203.88426  84.82477 135.19710
#> 73  146.60717 156.47674 113.26886 140.76784 131.33095 170.74737 184.17816
#> 74  152.21370 171.85870 163.49525 170.57249 159.07712 133.69050  84.11038
#> 75  134.83738 194.86433 112.23581 143.92449 125.98213 122.67119 163.04700
#> 76  121.96474 171.38000 121.41126 105.65984 132.88624 226.30416 184.63989
#> 77  136.89895 173.78556 139.47732 193.99797 185.82898 128.76247 140.71976
#> 78   93.47871 145.52120  97.94387 138.31782 170.56029 140.34807 163.54240
#> 79  174.34820 111.75928 110.37122 169.20292 155.93227 185.05019 148.33690
#> 80  102.14720 151.39972 205.64357 169.96207 174.37295 122.61981 182.16841
#> 81  115.38004 166.28232 163.25955 147.95946 144.27431 136.40819 153.05842
#> 82  165.30461 216.42436 135.65272 200.62198 165.98774 158.09795 206.16923
#> 83  199.20231 176.36876 115.22050 153.94678 145.10623 162.54824 161.87526
#> 84  165.24374 133.55754 135.99458 111.63452 190.19646 121.34380 139.67368
#> 85  182.18527 156.37942 121.50715 167.30850 159.28339 163.79439 173.62905
#> 86  152.31663 140.48794 133.37980 138.22400 180.39618 114.97749 146.29289
#> 87  171.60247  97.22235 166.82341 138.27291 133.91014 173.75570 110.86584
#> 88  166.89263 170.39539  93.99478 106.08100 191.59542  99.23617 130.81665
#> 89  171.00018 107.20915 123.63756 116.47396 134.27033 146.80576 108.09778
#> 90  214.78743 123.05015 171.89064 118.81707 221.36147 136.40000 156.84560
#> 91  119.91144 158.85518 169.38068 125.70771 137.64827 182.62591 138.18200
#> 92  192.10371 111.55901 161.84459 167.15020 217.72729 140.55695  99.73211
#> 93  189.40962 170.28167 211.30765 148.80778 170.11382 138.34896 171.88085
#> 94  144.17712 174.31360 123.04458 162.27652 167.44025 116.17342 113.62872
#> 95  152.12626 157.80645 113.33308 146.72835 152.95903 190.40572 135.52739
#> 96  145.09801 134.67786 123.63955 168.18928 150.57644 155.50278 175.61746
#> 97  128.27114 149.05347 117.05510 162.12915 114.58151 131.35634 146.04192
#> 98  128.15509 170.52604  93.27740 116.59315 146.97916 133.48098 143.17960
#> 99  137.53748 134.46225  89.09821 195.84152 113.11871 110.07553 135.19122
#> 100 185.50062 157.30340 165.99833 155.57918 162.96645 131.62703 125.53608
#>         B.4_1
#> 1    82.02803
#> 2   100.86074
#> 3   163.22654
#> 4   152.42774
#> 5    95.29958
#> 6   182.86349
#> 7   110.21470
#> 8   156.16176
#> 9   194.22628
#> 10  164.38893
#> 11  110.13647
#> 12  131.69224
#> 13  121.23994
#> 14  152.74289
#> 15   89.89478
#> 16  213.52115
#> 17  117.13257
#> 18  169.17080
#> 19  213.31271
#> 20  178.63674
#> 21  146.35016
#> 22  166.59296
#> 23  170.09962
#> 24  148.23449
#> 25  130.36882
#> 26  166.51989
#> 27  111.22595
#> 28  171.24170
#> 29  166.53909
#> 30  178.98684
#> 31  148.40486
#> 32  142.09101
#> 33  110.80553
#> 34  210.72875
#> 35  185.78697
#> 36  135.99848
#> 37  177.34041
#> 38  147.58488
#> 39   98.78717
#> 40  143.74302
#> 41  175.55166
#> 42  106.90276
#> 43  140.63393
#> 44   93.65525
#> 45  148.86199
#> 46  158.74153
#> 47  172.44838
#> 48  141.24352
#> 49  149.01197
#> 50  160.56164
#> 51  184.05250
#> 52  151.61424
#> 53  150.98038
#> 54  145.45008
#> 55  152.67313
#> 56  144.91461
#> 57  217.85414
#> 58  157.50709
#> 59  150.04463
#> 60  154.67360
#> 61  147.77191
#> 62  120.54910
#> 63  132.12695
#> 64  111.57867
#> 65  111.49164
#> 66  119.08982
#> 67  139.34232
#> 68  152.75680
#> 69  167.54504
#> 70  128.44980
#> 71  107.73317
#> 72  131.32207
#> 73  154.19817
#> 74  163.29178
#> 75  147.15402
#> 76  161.77020
#> 77  149.70317
#> 78  138.52920
#> 79  139.92459
#> 80  216.71584
#> 81  124.14036
#> 82  199.92183
#> 83  139.92405
#> 84  143.58784
#> 85  131.74989
#> 86  138.22846
#> 87  160.73724
#> 88  184.01592
#> 89  204.07545
#> 90  168.26975
#> 91  144.56097
#> 92  146.30320
#> 93  155.55755
#> 94  128.83112
#> 95  174.79799
#> 96  165.34696
#> 97  117.02097
#> 98  196.33157
#> 99  199.67468
#> 100 144.65725