Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10, 10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(
  subset = rep(c("A", "B"), 50),
  level = rep(1:4, each = 25)
)
rownames(meta) <- 1:100
scRNA <- Seurat::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA,
  by = c("subset", "level"),
  min.cells = 10, max.cells = 20
)
#>         A.1_1     B.1_1     A.2_1     B.2_1     A.3_1     B.3_1     A.4_1
#> 1   163.28044 148.26661 145.39275 161.76876 141.74428 130.81461 176.64002
#> 2   168.04990 121.31824 170.43297 151.71819 150.02693 186.10326 118.98979
#> 3    99.47382 113.06760 102.85578 129.97155 169.73044 127.48904 181.45496
#> 4   130.20402 107.05637 143.60396 154.74711 147.54773 124.53552 149.24985
#> 5   181.10656 148.11091 180.26043 127.90487 150.76624 171.45658 167.77355
#> 6   166.41501 146.28471 150.74958 143.66826 176.88857 170.29251 171.51624
#> 7   203.18346 193.60807 128.37040 169.11853 131.07287 160.10654 111.65879
#> 8   144.83338 150.95415 139.84546 114.14798 134.62612 177.08136 167.20400
#> 9    86.83969 173.99698 155.45849 141.02519 127.83444 149.75804 158.49664
#> 10  160.11115 190.10880 188.50947 118.14105 144.18557 140.54975  99.11295
#> 11  131.38979 203.80981 100.87105 205.02347 164.16516 113.16763 143.31896
#> 12  112.81476 159.90976 141.57236 161.77707 120.75210 128.11283 108.72822
#> 13  138.70063 118.91807 169.91516 106.64278 146.63935 151.18423 149.92397
#> 14  178.03568 146.55872 170.93328 157.44370 164.18925 147.02448 130.10535
#> 15  148.52309  89.69500  95.53071 101.11876 156.31879 138.27692 148.46175
#> 16  161.88074 163.14912 112.92716  93.96086 152.56034 124.23903 118.93283
#> 17  136.17053 205.01794 104.36917 152.31247 169.77762 172.76967 110.83875
#> 18  163.87557 146.59013 104.91014 126.70952 179.65010 142.96499 132.02558
#> 19  111.40763 156.24055 161.88093 154.79150 120.04855 122.76280  97.54192
#> 20  140.01148 129.83109  91.80539 167.06186 109.47535 136.89221 167.04991
#> 21  144.01155 113.93090 144.08944 140.49828 160.02884 126.92345 185.88745
#> 22  148.83421 121.46099 185.81300 153.38622 158.78803  93.11309 148.97350
#> 23  136.67266 126.59776 169.05501 145.91647 145.62367 140.66945 131.28377
#> 24  188.23646 120.01707 116.70768 157.35177 178.90508 246.05209  93.45485
#> 25  157.31008  81.82398 117.65061 142.09968 198.66108 171.38147 135.77426
#> 26  106.60398 199.13607 111.88000 126.92668 122.78468 119.37262 123.03446
#> 27  179.19165 134.06903 102.12796 120.88806 196.94132 105.05474 122.76767
#> 28  162.12232 130.79995 196.50500 157.59090 169.39905 151.84889 107.56005
#> 29  177.10230 141.23885 164.79116 146.05612 141.33583 148.84267 139.99622
#> 30  167.81589 153.13634 201.56684 128.98821 180.55908 120.91646 110.76798
#> 31  152.74542 164.62593 130.35449 134.48650 120.63983 201.82323 143.51081
#> 32  162.98360 119.27165 162.18368 180.42174 175.56987 128.31838 169.44051
#> 33  120.82253 117.97136 149.64059 145.53647 109.66631 133.35507 143.37597
#> 34  123.76574 121.51708 162.14640 177.23125 126.19483 114.68347 130.10294
#> 35  171.87254 102.34468 152.34785 202.00487 182.72415 106.91955 139.49289
#> 36  160.64352 174.33164  65.06077 126.79784 169.60599 146.24397 154.85899
#> 37  123.89105 150.25963  96.10456 151.61933 151.40958 146.74111 150.00951
#> 38  120.39651 154.99777 157.43096 154.18082 149.55222 113.01516 140.01314
#> 39  159.02191 165.42014  98.13988 161.52482  94.50201 136.48737 121.45447
#> 40   83.99908 119.97098 120.99381 112.28097 176.13674 175.65529 101.60284
#> 41   84.83913 176.50941 157.72926 198.31046 147.31472 124.56478 158.88655
#> 42  117.12265 135.36049  90.82714 129.42586 135.84722 107.95227 136.58079
#> 43  206.25461 118.79683 163.26560 179.72418 141.74081  95.34619 139.81539
#> 44  169.36769 119.78726 139.86729 122.61679 126.04850 124.77352 148.46813
#> 45  119.18911 160.36648 132.49620 135.65775 146.48023 133.29721 120.74740
#> 46  137.38680 153.33869 174.25117  97.77153  79.74394 120.55162 171.09462
#> 47  167.54128 133.49943 141.74278 111.02255 106.92015 152.55081 129.21705
#> 48  166.53465 150.92371 126.20546 211.25284 116.38271 168.46750 148.78238
#> 49  151.72370 123.69824 107.33111 111.96018 154.05373 118.79442  97.05652
#> 50  150.43581 131.60182 157.73592 160.24207 134.98157 176.50842  84.66455
#> 51  204.21190 161.23304  98.83461 147.02338 185.09441 162.65383 186.74513
#> 52  166.15162 132.47249  74.45111 108.30145 167.90793 146.21672  90.72542
#> 53  194.19133 130.94846 150.34693 155.26284 111.21455  84.77936  82.27740
#> 54  174.81979 136.04811 146.72972 159.88183 202.69405  91.35691 199.78479
#> 55  161.79829 119.05855 115.64215 122.02369 104.05070 133.95803 115.93860
#> 56  130.13904 141.96893 184.28889 156.17077 154.26204 145.91723 119.63114
#> 57  168.75703 111.52743 123.78975 166.30560 161.40009 135.67514 136.49502
#> 58  124.04680 164.30024 138.63283 199.63013 168.56780 139.05835 142.73149
#> 59  168.17845 181.92396 104.52133 200.72725 143.47579 171.13130 137.33615
#> 60  112.19955 144.31741 112.68262 149.25738 149.60408 135.86453 100.70936
#> 61  168.47913 118.94109 134.50189 156.79712 263.44556 158.73998 142.07089
#> 62  170.26623 131.41616 107.18571 150.20490 186.24692 138.98940 160.79112
#> 63  159.62861 119.50649 141.80649 144.88585 131.48009 109.22409 197.10886
#> 64  110.77832 141.58790 139.52578 139.04575 177.96529 134.87169 164.80439
#> 65  127.54518 143.16249 167.58526 124.71594 156.39567 130.18922 105.34310
#> 66  146.57026 165.30592 182.80168 188.83563 182.30597 124.96697 168.50911
#> 67  150.25567 118.56274 182.26868 127.40314 187.48834 181.68490  92.58982
#> 68  170.27957 135.78249 115.83006 123.72359 126.32318  99.91204 160.69907
#> 69  160.02000 136.24129 154.75030 103.82367 188.78994 149.32786 121.06322
#> 70  169.90111 156.12757 132.94708 119.08451 133.69333 149.87391 137.55532
#> 71   71.90267 125.95664  78.99920 164.92819 198.03905 123.55457 110.38161
#> 72  131.63034 164.19470 204.57395 148.41443 121.28563 156.68057 133.16865
#> 73  159.97012 122.70064  97.66219 149.20525 185.56853 134.74366 187.80237
#> 74  147.41943 109.59882 121.07548 185.30401 154.32820 170.88531 114.23912
#> 75  166.24620 177.87583 163.27380 153.74768 168.04656 155.39267 103.44594
#> 76  120.26841 127.81112 196.15097 146.50624 144.34213 160.82980 136.22952
#> 77  122.04661 167.50460 102.35364 109.51315 151.34785 179.02864 109.02813
#> 78  155.12332 125.85778 143.08898 133.62947 147.90367 138.41240  91.35923
#> 79  158.93687 148.59238 166.21929 130.20330 114.76340 161.53616 138.83077
#> 80  122.47275 164.72629 109.48001 103.05149 201.53689 152.53595  92.49957
#> 81  167.12532 148.22790 150.62656 157.63158 166.76692 167.28736 167.26638
#> 82  174.99110 165.92381 136.19966 163.88747 156.48551 144.55532 136.48045
#> 83  166.92208  79.10431 154.36187 143.85416 146.09918 130.77597 101.91934
#> 84  154.88325 117.15089 101.27241 154.43657 154.40880  96.06415  88.58357
#> 85  233.48913 131.10468 150.27377 164.43301 177.09524  98.38535 171.83562
#> 86  140.52466 155.04927 137.25228 129.77134 175.79185 168.36412 137.15613
#> 87  119.80679 125.86276 116.28665 149.35854 172.23655 171.96022 152.14058
#> 88  148.14160 183.53128 154.90893 136.93823 115.35483 168.30204 173.51578
#> 89  122.68896 173.61810 111.72413 117.14050 189.66253 202.54477 132.96044
#> 90  175.85061 204.70548 111.14368 148.07394 145.11448 150.12245 119.64400
#> 91  165.99409 150.45382 187.31107 104.74338 151.67762 118.62211 117.63021
#> 92   93.64557 123.90522 127.12267 157.63554 136.34023 140.50017 194.04191
#> 93  179.48516 132.32278 154.80456 160.28490  94.28081 220.24458 117.69550
#> 94  103.97010  93.47482 140.06092 160.15747 179.06899 129.48858 142.10241
#> 95  258.69346 110.67036 167.39575 159.37284 177.95368 110.14426 149.26810
#> 96  120.96327 128.94796 166.75725 173.35725 168.71951 116.56159 128.90450
#> 97  100.13878 106.48656 147.66865 144.41789 181.72768 127.93323 148.20490
#> 98  112.54851 136.69401 143.74948 137.90326 180.65791 156.41617 112.44581
#> 99  177.54447 105.24111 122.65564 170.04222 133.83152 165.24767 159.96176
#> 100 162.19060 122.00274 176.98060 129.79987 185.36808 145.83470 143.54769
#>         B.4_1
#> 1   148.55397
#> 2   115.70500
#> 3   119.72562
#> 4   186.24388
#> 5   165.57230
#> 6   165.74920
#> 7   195.68355
#> 8   148.80416
#> 9   122.05014
#> 10  130.21009
#> 11  163.13860
#> 12  217.42761
#> 13  168.93108
#> 14  167.06703
#> 15  135.15126
#> 16  180.57242
#> 17  159.07020
#> 18  177.91181
#> 19  172.11782
#> 20  150.93983
#> 21  170.20733
#> 22  138.85897
#> 23  142.88742
#> 24  193.79143
#> 25  136.04875
#> 26  166.00452
#> 27  170.63656
#> 28  101.15974
#> 29  195.48129
#> 30  162.29877
#> 31   84.00992
#> 32  123.66864
#> 33  155.43845
#> 34  130.46213
#> 35  169.52329
#> 36  176.65061
#> 37  139.73175
#> 38  157.39737
#> 39  138.78853
#> 40  161.64819
#> 41  111.93281
#> 42  148.43368
#> 43  190.17036
#> 44  204.00877
#> 45  126.42594
#> 46  199.23393
#> 47  134.64392
#> 48  117.37814
#> 49  183.34360
#> 50  108.77860
#> 51  135.58123
#> 52  149.27140
#> 53  201.77697
#> 54  174.80830
#> 55  171.43753
#> 56  145.41997
#> 57  193.38726
#> 58  185.69704
#> 59  106.74360
#> 60  133.25803
#> 61  149.02580
#> 62  152.40376
#> 63  151.54808
#> 64  141.22010
#> 65  192.74455
#> 66  113.70946
#> 67  162.69925
#> 68  150.45577
#> 69  184.26615
#> 70  169.29807
#> 71  167.43505
#> 72  178.44573
#> 73  165.73882
#> 74  144.52960
#> 75  126.32449
#> 76  168.11049
#> 77  187.20978
#> 78  101.15670
#> 79  171.95552
#> 80  161.05670
#> 81  191.49930
#> 82  176.87453
#> 83  121.52570
#> 84  113.24951
#> 85  176.21196
#> 86  170.75515
#> 87  160.25365
#> 88  177.24214
#> 89  178.57990
#> 90  177.36278
#> 91  139.78098
#> 92  197.99198
#> 93  187.68382
#> 94  204.71892
#> 95  145.08528
#> 96  146.63541
#> 97  118.59706
#> 98  157.75240
#> 99  105.55742
#> 100 164.25954