Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10,10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(subset = rep(c("A","B"),50),
                   level = rep(1:4, each = 25))
rownames(meta) <- 1:100
scRNA <- Seurat::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA, by = c("subset","level"),
               min.cells = 10, max.cells = 20)
#>         A.1_1     B.1_1     A.2_1     B.2_1     A.3_1     B.3_1     A.4_1
#> 1   146.36663 194.01972 129.70486 145.36936 133.28465 169.34965 127.20323
#> 2   128.34164 144.73275 148.86545 149.32307 186.44447 144.18099 175.00444
#> 3   157.92349 116.14754 137.44480 123.35123 179.65267 114.04389 186.33118
#> 4   169.59414 117.52555 205.40754 175.23249 178.22130 182.15777 134.45175
#> 5    85.51575 165.72316  91.84038 141.13470 156.14727 207.06479 159.76174
#> 6   133.36940 133.22477 186.72809 141.67152 127.88427 125.93484 108.95810
#> 7   125.62776 146.07932  92.87755 226.37352 147.11146 199.64085 119.23880
#> 8   201.78356 175.68863 183.04188 131.17158 148.89012 135.77505 142.76565
#> 9   164.97177 123.20460 119.40726 162.03444 169.06272 178.63780 112.27350
#> 10  151.09023 103.77745 136.29441 122.63776 119.72020 103.11560 164.65302
#> 11  148.96044 172.98075 108.61333 154.67722 169.41037  92.58993 148.91573
#> 12  160.34304 169.84629 149.63736 132.26770 194.44349 133.33686 143.66151
#> 13  134.65901 116.09071 157.02731 212.04132 145.76190 117.42483 134.11668
#> 14  158.12610 137.05796 130.40639 136.16627 165.72498 162.26586  86.32822
#> 15  119.46817 113.86908 118.27144 159.87971 134.92299 139.13606 126.71563
#> 16  151.42047 173.15869  84.78085 162.59817 155.15085 190.06267 111.58734
#> 17  118.39961 138.47113 119.14509 182.46002 169.44815 168.45311  68.23861
#> 18  230.59863 143.06800  98.98001 131.99029 160.16251 182.76638 121.06721
#> 19  148.30032 146.63359 147.48823 126.08951 151.85062 136.97606 223.87712
#> 20  173.06326 108.36202  97.98850 201.07231 126.30410 169.82798  71.30670
#> 21  126.32969 161.94552 128.73035 150.42425 156.24223 111.04715 149.75094
#> 22  126.07235 148.88672 137.82532 162.88751 158.32037 141.63860 131.32892
#> 23  123.11999 124.81392 166.63244 156.10903 173.92820 140.87431 157.74321
#> 24  166.83924 107.89434 147.25024 168.52565 165.53402 163.51930 119.29081
#> 25  157.24903 142.29856  98.02477 135.35810 119.41698 187.20243 150.02375
#> 26  212.82215 140.24243 154.12540 197.59927 219.60948 113.10078 121.62454
#> 27  113.68089 171.95400 138.45871 136.62625 124.98276 134.95149 125.37623
#> 28  128.45064 111.86154 147.56009  92.23792 164.71153 146.43466 150.08143
#> 29  137.68625 120.06654 134.83461 136.83830 173.47621 112.31633 155.73383
#> 30  125.34003 175.18659 127.89954 184.30116 142.52152 126.38006 126.22853
#> 31  176.17685 143.74178 183.52079 149.34157 149.55947 169.02000 164.40665
#> 32  179.50908  71.32062 108.54784 163.20425 153.65202 142.01900 135.20685
#> 33  148.10091 109.72106 136.41656 140.11330 189.65859 117.69830 152.41379
#> 34  152.54102 122.75527 136.79208 158.69761 171.40501 186.09156 128.42124
#> 35  111.28157  93.26072  68.76890 149.03426 183.27194 140.18806 122.65865
#> 36  166.00068 134.59175 126.65130 166.45240 185.01061 110.87598 154.59905
#> 37  147.81791 149.31507 127.71199 117.97699 150.14705 118.57635 190.27306
#> 38  162.89908  96.45494 173.33698 179.06902 124.26169 126.99110 144.96824
#> 39  142.28773 164.34567 111.58954 179.53760 168.81844 148.29331 125.13322
#> 40  150.44635 179.91735 114.14055 132.85385 112.01255  96.42028 110.06839
#> 41  164.60010  96.47697 123.88512  92.53755 144.45441 171.58968 171.80533
#> 42  197.26227 116.21987  94.74309 162.93425 206.98987 150.07428 140.90994
#> 43  104.86694 135.75002 192.92180 224.08326 129.13735 173.84387 165.55299
#> 44  146.31555 150.30100 187.77524 137.05815 198.93045 112.91475 157.58008
#> 45  154.43991 173.93627 158.58784 114.98334 200.89805 135.77032 161.80297
#> 46  201.49596 157.18930 138.65022 153.30727 174.01697  88.35161 150.23797
#> 47  174.78919 148.10733 149.10661 216.66399 163.11454 112.55125 137.37539
#> 48  169.78642 123.43665 116.69227 107.88814 203.35899 155.61793 144.96505
#> 49  158.73965  76.05778  98.40093 166.54373 130.98635 146.24697 115.08882
#> 50  110.85979 172.19124 162.45854 110.81541 176.39795 144.84449  80.80040
#> 51  140.61046 129.07101 145.72198 117.26834 156.73636 187.10140 142.00943
#> 52  151.68642 145.40155 186.27011 101.86562 137.87535 107.10837 181.51471
#> 53  129.17608 154.29148 143.91890 139.50903 104.77888 138.52596 147.37449
#> 54  130.04758 139.46735 127.30171 126.75433 168.27453 201.68594 136.61414
#> 55  179.87590 116.46640 129.94296 138.45409 118.79445  93.49753 109.59805
#> 56  151.66462 143.98432  84.64853 155.07025 177.27968 145.40815 105.71181
#> 57  125.70950 186.59345 173.72358 124.58502 166.78454 129.80287 139.44224
#> 58  157.90474 137.99630 134.79404 159.91817 130.26908 155.55733 159.10004
#> 59  191.24331 116.80298 147.66768 148.71557 130.81582 126.36316 109.58274
#> 60  161.27076 143.54692 186.26716 148.96995 126.17933 118.69059 100.76019
#> 61  138.45717 144.03548 152.55823 133.26884 177.48463 111.54252 141.53624
#> 62  153.78550 109.38395 139.75794 121.84560 121.79053  98.44290 142.21144
#> 63  211.82713 132.39090 162.78995 171.73285 141.55677 138.87119 131.60366
#> 64  137.34423 157.74713 177.73883 143.36250 152.27666 131.73155 123.81423
#> 65  164.55417 134.41283 115.99593 118.73711  88.25746 153.57218 202.00078
#> 66  158.97578 108.21110 150.09810 166.13827 133.67313 125.05807 153.92209
#> 67  195.97955 125.92682 113.62475 113.13296  95.05980 147.99639 159.64107
#> 68  112.69513 165.48681 167.11933 149.59380 117.33312 115.63547 158.54059
#> 69  136.62271 148.96923 125.14985 117.12959 135.11452 216.87428 100.23007
#> 70  149.73310 130.17645 133.86664 145.38505 137.81129  96.70922 146.34248
#> 71  104.07181 138.25990 114.26756 155.29487 139.02297 146.04515 152.50664
#> 72  143.20909 160.65240  99.53933 154.59467 130.63368 114.77062 128.05970
#> 73  172.09883 119.56778 177.66526 108.24506 156.15255 120.22563 131.65979
#> 74  169.65278  97.51202 128.73209 170.45009 164.88250 147.61835 168.87975
#> 75   91.67051 140.32786 150.94161 105.31249 209.03879 145.66284 119.24094
#> 76  135.08405 177.61485 142.47327 133.54657 176.31744 147.93370 158.05196
#> 77  185.78174  97.85302 144.57148 165.85079 142.21800 124.95528 156.43836
#> 78  172.25131 190.12752 109.36667 145.77425 163.37036 170.44929 129.45810
#> 79  174.82175 122.86166 180.69763 149.26816 136.77297 145.68431 120.15423
#> 80  157.41770 172.76108  64.52757 104.58303 185.66187 174.80350 118.77277
#> 81  153.63724 143.42679 154.11923 138.86152 157.54826 145.24556 148.61268
#> 82  124.70342  79.69033 141.71991 176.50702 145.64003 186.67266  86.92865
#> 83  167.57360 208.25019  90.75077 126.41389 147.81380 159.58857 124.15959
#> 84  154.72965 121.37695 174.21091 164.03221 151.23403 151.58471 168.30515
#> 85  140.16359  88.12459 164.93374 126.94189 210.76900 117.24147 156.32898
#> 86  142.32996 133.87211 112.17827 166.86371 193.76217 147.67622  91.38312
#> 87  198.50067 122.45338 167.06680 138.04801 160.23215 154.55371 129.01609
#> 88  195.33242 101.96828 163.84741 115.59706 185.90335 120.34808 153.90429
#> 89  146.48465 137.90577 115.00072 181.54765 124.01880 108.63810 141.83456
#> 90  143.26825 177.52675 157.92311 145.71135 175.37010 137.44912 147.72097
#> 91  130.91699 142.81407 170.88550 173.55119 182.68348 131.82093 125.66355
#> 92  150.58170 116.60909 162.52561 160.72783 138.03814 183.43539 148.87197
#> 93  106.40969 142.88655 114.78940 207.29184 162.85381 138.41253 191.51086
#> 94  146.18450 140.64200 189.11927 194.71177 199.71230 139.22619 105.55788
#> 95  183.38185 112.77972 167.71791 192.66933 178.64923 108.49669 101.96910
#> 96  137.90630  74.94273 123.90149 124.85356 160.93247 156.96866 135.21873
#> 97  182.18838 185.06143 135.80320  99.48523 119.87955 104.73081 148.43818
#> 98  200.61276 115.18024 192.99487 208.80968 169.85315 157.63509 141.53268
#> 99  170.36793 130.14001 147.52055 174.42194 127.72103 150.28984 162.51787
#> 100 178.34854 148.31974 150.30825 147.01527 129.59013  98.04514  99.74170
#>         B.4_1
#> 1   132.14629
#> 2   182.17113
#> 3   203.68474
#> 4   166.84791
#> 5   135.62819
#> 6   155.19034
#> 7   154.99086
#> 8   183.12560
#> 9   163.97844
#> 10  184.39537
#> 11  142.15620
#> 12  173.35511
#> 13  140.57028
#> 14  159.95304
#> 15  168.64842
#> 16  128.60991
#> 17  167.35239
#> 18  195.42666
#> 19  102.15260
#> 20   89.40082
#> 21  164.51615
#> 22  148.76715
#> 23  145.01425
#> 24  140.76918
#> 25  128.52803
#> 26  105.84920
#> 27  140.53211
#> 28  144.23552
#> 29  107.74954
#> 30  121.47299
#> 31  184.67081
#> 32  121.45133
#> 33  131.65962
#> 34  144.24578
#> 35  108.60307
#> 36  195.30377
#> 37  184.09647
#> 38  164.53931
#> 39  160.23036
#> 40  129.89191
#> 41  123.78287
#> 42  185.04405
#> 43  151.96042
#> 44  220.83880
#> 45  121.51686
#> 46   80.70282
#> 47  163.28776
#> 48  124.82783
#> 49  151.63073
#> 50  163.50763
#> 51  167.21404
#> 52  173.64921
#> 53  156.83773
#> 54  145.04555
#> 55  130.41469
#> 56  118.07696
#> 57  166.70903
#> 58  131.96838
#> 59  167.91212
#> 60  170.53714
#> 61   82.54426
#> 62  198.90515
#> 63  124.81138
#> 64  181.47570
#> 65  154.72319
#> 66  165.93918
#> 67  121.44823
#> 68  132.13449
#> 69  144.36948
#> 70  125.26873
#> 71  134.66166
#> 72  150.74890
#> 73   97.26699
#> 74  137.25371
#> 75  135.50468
#> 76   97.21435
#> 77  135.38114
#> 78  160.74575
#> 79  144.95724
#> 80  105.10572
#> 81  163.21897
#> 82  160.93166
#> 83  152.74302
#> 84  194.81929
#> 85  132.37638
#> 86  182.32136
#> 87   93.03202
#> 88  145.44099
#> 89  122.95905
#> 90  176.89152
#> 91  186.49923
#> 92  181.45053
#> 93  148.32393
#> 94  145.17492
#> 95  124.46793
#> 96  135.19017
#> 97   97.67595
#> 98  174.06898
#> 99  127.37999
#> 100 149.47695