Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10, 10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(
  subset = rep(c("A", "B"), 50),
  level = rep(1:4, each = 25)
)
rownames(meta) <- 1:100
scRNA <- SeuratObject::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA,
  by = c("subset", "level"),
  min.cells = 10, max.cells = 20
)
#>         A.1_1     B.1_1     A.2_1     B.2_1     A.3_1     B.3_1     A.4_1
#> 1   146.93285 154.60465 168.55117 244.21399 221.25672 137.91624 116.75219
#> 2   146.87906 185.41188 122.82453 151.35832 175.83305 139.45936 123.89991
#> 3   134.55294 100.08562 147.46039 173.33216 144.03688 137.87872 155.40592
#> 4   146.72652  94.89513  95.51482 118.93691 128.68513 129.99809 178.59713
#> 5   122.55481 149.70122 134.14199 127.40607 168.30035  79.84315  92.44476
#> 6   167.52414 175.65576 133.86676 140.90515 149.57993 141.53129 119.58886
#> 7   121.02500 132.92145 119.53397  97.48146 126.06499 157.48862 135.89723
#> 8   126.78109  94.53157 159.44365 133.50071 106.74134 161.53782 108.18903
#> 9   126.27461 136.91464 172.97434 154.90294 157.01944 147.35801 125.73619
#> 10  122.78243 141.20244 140.97430 103.63420 203.52936 132.42969 135.60307
#> 11  148.78735 174.24992 129.57532 160.69006 132.86463 146.15611 115.27205
#> 12  152.54142 103.58203 155.96935 121.85645  93.61049  95.91495  91.73615
#> 13  153.63673 147.44929 137.29524 140.07749 159.92648 136.78051 128.15041
#> 14  152.30254 126.95051 165.94220 120.37510 175.96773 111.43448 144.74258
#> 15  151.69134 140.11127 155.46188 152.85494  73.86970 120.78691 144.88651
#> 16  171.26171 126.08032 150.41920  96.49276 145.44960 130.42753 161.71795
#> 17  197.92614 162.05071 136.57832 169.78949 118.14392 144.12445 113.94488
#> 18  134.02376 154.08814 149.89251 186.72054 177.74116  90.50486 156.95086
#> 19  128.15338 160.58280 106.67214 157.56893 164.47990 122.51605 171.99186
#> 20  116.99553  96.65319 150.62242 177.95635 194.14909 128.33937 199.05990
#> 21  168.07246  89.24514 141.84432 157.54843 170.55185 130.59508 127.08383
#> 22  125.97903 172.63608 129.46058 199.06621 186.55121 134.16066 179.78465
#> 23  187.39245 111.99741 125.84094 159.58379 108.77100 115.53983 167.32149
#> 24  188.52351 127.85453 160.83722 123.55150 117.43151 113.28537 124.16515
#> 25  144.39624 114.80324  82.24043 143.79109 156.50599 155.03046 178.31266
#> 26  126.98691 140.67242 129.14108 201.23830 170.20251 143.89055 130.96265
#> 27  122.57843 170.76262 207.56069 148.30657 175.30562 147.44396 170.11277
#> 28  137.39853 125.29012 146.88959 154.99980 127.67821  82.29659  99.43564
#> 29  148.06788 128.48461 108.17794 140.01522 151.72420 117.47680  99.62497
#> 30  136.15680 148.77378 119.43709 198.11262 155.72550 117.40410 164.09326
#> 31  159.01726 193.97941 136.89610 131.07700 101.37469 149.84759 163.21191
#> 32  164.68478 112.85230 146.56826 159.06918 181.92492 128.23439 101.52667
#> 33  184.41072 229.53498 126.32999 149.06327  65.41217 115.55792 126.90936
#> 34  120.96614 141.42790 160.98602 144.91856 173.36315 104.76914 119.39907
#> 35  191.35833 118.74304 143.25338 164.61719 140.94719 170.26426 142.09448
#> 36  157.46600 133.15315 133.87237 156.67658 142.92696 122.69033 108.28397
#> 37  152.34212 137.86857 158.03756 132.75742 162.35282 144.59670 177.25247
#> 38   85.28608 149.80510 123.96618 115.09131 156.78073 107.41873 171.47593
#> 39  130.51874 101.90498 145.25942 179.67172 165.06707  97.94215 121.00656
#> 40  172.12375 177.03889 103.79574 148.18734 165.48420 133.97170 136.62462
#> 41  101.16711 127.09272  99.08667 115.41249 191.13484 117.27791 131.28662
#> 42  131.58661 111.92707 132.55043 164.54207 156.01283 118.02854 169.06299
#> 43  164.77004 133.06693 151.34766 143.27075 126.21185 143.09266 124.49681
#> 44  158.21564 162.25240 142.44753 114.97586 115.10872 114.96424 101.59322
#> 45  194.87439 138.10676 148.66003 125.64066 143.43086 127.37967 129.90479
#> 46  123.11026 112.59098 145.47331 146.81074 171.19623 147.02739 149.28035
#> 47  141.44307 146.98639 173.08698  97.17020 126.16119 153.26031 107.99449
#> 48  150.50705 155.77724 126.33424 109.47865 155.39946 110.16126 137.66898
#> 49  168.98360 127.42310  97.24143 163.17588  93.70820 107.98941 144.52968
#> 50  178.06258 105.69470  96.13428 143.40598 182.45660 181.52444 112.77394
#> 51  164.82782 111.33388 193.10623 179.95425 168.83251 135.66185 123.60797
#> 52  149.40923 182.19766 154.64659 118.30929 156.13310 202.39289  94.85071
#> 53  153.21068 116.96628 124.82130 128.95315 174.08856 200.14041 126.82724
#> 54  145.21504 160.54802 102.51157 156.01494 173.00791 124.15712 115.74320
#> 55  251.37165 173.98322 113.91156 139.73094 141.86350 118.76869 134.58989
#> 56  153.36425 146.65013 151.54993 101.91089 132.50909 122.30552 196.12996
#> 57  132.79659 104.06829 103.22768 158.05338 114.57992 176.35980 120.78218
#> 58  165.63609 187.57718 168.84161 129.02906 132.91998 129.93723 128.22880
#> 59  165.87959 138.39151 134.05030 139.25625 112.14854  81.99187 157.18536
#> 60  161.11790 162.82688 141.79561 173.86229 111.18033 100.28985  99.49695
#> 61  180.59281 109.01641 170.54809 102.46521 156.75523 159.12452 115.00174
#> 62  158.56045 153.82611 187.01284 187.37963 162.41346 211.25390  86.88469
#> 63  148.23078 113.94161 127.62503 225.14543 177.66736 146.13782 181.70363
#> 64  196.23105 189.14868 167.87831 174.01395 135.30615 132.53734 149.87900
#> 65  165.92513 171.09436 180.29930 152.66960 112.52962 124.27869 112.91628
#> 66  207.20226  75.11867 156.74376 157.14010 198.19658 135.44782 160.69923
#> 67  195.78673 161.46807 129.26126 178.17846 183.03116 129.24186 135.23135
#> 68  141.00266 163.27838 125.68160 133.87648 169.26237 121.38017 113.10322
#> 69   92.83314 182.48408 103.78589 115.96825 186.99975 152.38187 226.90623
#> 70  180.96896 169.01968  97.58897 145.19876 112.72505 132.41340 168.33154
#> 71  162.15280 142.05191  99.17192 159.61410 142.80379 147.51314 147.88722
#> 72  132.30190 232.20722 164.83591 118.45478 112.57819 139.26704 136.86136
#> 73  113.82494 108.94864 133.73396 179.46498 178.91975 142.99763 139.46941
#> 74  183.84176 160.74044  93.85228  93.64186 141.18790 113.75546 143.26991
#> 75  169.12575 126.74344 179.84228 170.16250 124.25705 146.96573 141.24728
#> 76  154.18884 175.27311 163.81716 151.81947 126.91628 146.66057 162.07953
#> 77  197.15715 179.53611 153.80201 155.11795 168.39352 127.98675  91.41026
#> 78   90.43390 147.00129 118.16720 126.39659 138.17538 129.42681 155.56810
#> 79  187.43859 156.35042 169.94670 122.29851 190.84894 168.84841 125.21624
#> 80  180.88085 107.54492 101.11451  86.16612 141.84954 144.79730 124.19122
#> 81  192.95249 154.37450 148.82571 148.92503 174.01302 115.93888 135.23487
#> 82  143.18229 171.92308 112.69448 152.21245 173.44371 176.03978  93.71097
#> 83  181.33269 150.53226 150.03932 202.85109  95.41795 151.55828 114.90596
#> 84  214.17547 130.85795 164.02455 148.44949 194.77773 174.72926  79.43364
#> 85  116.69518 136.30341 139.97620 161.20977 198.92676 111.73518 175.54673
#> 86  134.75145 141.54878 163.94254 143.84687 160.23828 164.83514 133.33321
#> 87  137.26971 119.34659  93.18764 156.50091 167.99826 122.95991 167.92490
#> 88  148.77561 144.01478 144.42815 128.67026 137.22591 138.71780 141.08476
#> 89  139.67882 130.01246 124.50280 160.38197 149.48433 107.56182  95.08691
#> 90  124.66320 150.81547 108.06002 166.79966 168.23227 122.79490 131.81784
#> 91  130.11881 120.58528  79.28420 216.22509 143.74472  78.50983 102.00527
#> 92  142.73744 159.97066 128.72559 164.29437 162.25086 158.74241 147.14597
#> 93  162.74634 174.70557 129.96769 125.34435 145.16586 192.41807 132.63386
#> 94  172.35017 138.83773 142.69417 164.88120 178.53125 129.16993 169.33814
#> 95  143.40147 130.28916 164.01694 145.69107 163.39741 156.79808 145.20577
#> 96  178.36440 121.58151 123.70877 176.37551 149.31031 163.93946 139.09801
#> 97  128.29477 118.46183 153.97203 207.97285 135.16779 128.00609 131.60656
#> 98  165.14579 108.28371 132.28971 162.98043 187.89448 137.39312 119.17418
#> 99  154.22958 112.46335 159.52042 118.10009 170.91631  99.83068 162.55349
#> 100 146.05621 120.81752 121.99262 104.11523 152.91125 125.71684 126.92158
#>         B.4_1
#> 1   107.87551
#> 2   136.30801
#> 3   175.27940
#> 4   146.25845
#> 5   145.12487
#> 6   137.14392
#> 7   141.09707
#> 8   153.03300
#> 9   133.71760
#> 10  133.69634
#> 11  121.30719
#> 12  154.63821
#> 13  179.38875
#> 14  123.33212
#> 15  180.08795
#> 16  142.72491
#> 17  122.42701
#> 18  113.38621
#> 19  131.31206
#> 20  169.10821
#> 21  181.66550
#> 22  165.40525
#> 23  141.80441
#> 24  129.59866
#> 25  142.73009
#> 26  184.21917
#> 27  146.93829
#> 28  126.87494
#> 29  143.87184
#> 30  193.72919
#> 31  103.71134
#> 32  139.88446
#> 33  156.81591
#> 34  136.91513
#> 35  164.67441
#> 36  151.02681
#> 37  168.75117
#> 38  191.27414
#> 39  157.76807
#> 40  114.75199
#> 41  159.87851
#> 42  161.09456
#> 43  128.41957
#> 44  132.40164
#> 45  114.14299
#> 46  144.30165
#> 47  148.14417
#> 48  131.85089
#> 49  186.55030
#> 50  168.73387
#> 51  172.32490
#> 52  139.90202
#> 53  159.16688
#> 54  138.62638
#> 55  167.91371
#> 56   91.64114
#> 57   98.32015
#> 58   94.52240
#> 59  180.89119
#> 60  207.29650
#> 61  168.55481
#> 62  158.27718
#> 63  131.36124
#> 64  206.60401
#> 65  129.88513
#> 66  166.66263
#> 67  217.28293
#> 68  158.04913
#> 69  142.23195
#> 70  168.53830
#> 71  126.67619
#> 72  102.53588
#> 73  163.60683
#> 74  192.00132
#> 75  128.69020
#> 76  152.80969
#> 77  162.81700
#> 78  160.88386
#> 79  156.30663
#> 80  159.78325
#> 81  156.03817
#> 82  189.19469
#> 83  201.13609
#> 84  130.01972
#> 85  153.46738
#> 86  144.76423
#> 87  150.20659
#> 88  147.38282
#> 89  170.81416
#> 90  191.10365
#> 91  142.54153
#> 92  150.52379
#> 93  142.67619
#> 94  156.91326
#> 95  148.28713
#> 96  178.63818
#> 97  191.28104
#> 98  177.51878
#> 99  160.65505
#> 100 127.68190