Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10,10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(subset = rep(c("A","B"),50),
                   level = rep(1:4, each = 25))
rownames(meta) <- 1:100
scRNA <- Seurat::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA, by = c("subset","level"),
               min.cells = 10, max.cells = 20)
#>         A.1_1     B.1_1     A.2_1     B.2_1     A.3_1     B.3_1     A.4_1
#> 1   152.32004 119.01246 172.35548 173.22738 155.89377 116.61709 101.74277
#> 2   183.39867 133.36247 139.10557 175.37316 150.61840 136.65935 167.30894
#> 3   153.69570 131.73956 113.40739 131.69153  85.87608 137.26186 117.45706
#> 4   187.14480 128.71036 233.28157 156.71931 137.35332 145.44047 154.78853
#> 5   158.51939 143.64105 175.12132 178.31642 110.66865 181.14907 157.25614
#> 6   190.54368  97.69476  89.14125 179.08744 135.17251 163.00395 161.74444
#> 7   134.55646 138.50164  97.64704  99.88182 107.60993 160.97786 160.10467
#> 8    93.15794 123.63782 102.43515 117.93448 144.80170 184.78769 117.11155
#> 9   180.25022 134.10493 183.32321 105.27810 198.52007 138.19279 107.48581
#> 10  122.44358 152.50356 132.64034 145.58028 113.22436 172.29337 152.59471
#> 11  137.82543  95.65319 165.11024 160.98684 111.06402 136.31900 156.83227
#> 12  128.62180 160.03776 169.72428 166.54902 178.22979 152.83512 139.10347
#> 13  127.81292 101.92521 115.27243 103.45900 162.00560 153.94219 157.85687
#> 14  160.50590  97.87918 178.50634 126.12968 157.41516 113.33069 152.94494
#> 15  143.47247 155.25004  99.73502 142.84469  97.84811  99.38904 114.27900
#> 16  159.14765 133.64444 162.14312 154.77560 200.55836 131.28444 149.96684
#> 17  159.81783 173.41350 199.60773 144.72730 148.61826 166.07559 137.73055
#> 18  170.78959 110.71618 120.42260 146.21831 179.06964 192.41820 143.05988
#> 19  115.94444 135.02053 151.27930 144.87997 118.91936 128.34873 148.43103
#> 20  122.42051 175.78671 163.83843 141.74241 150.62440 133.47074  83.57025
#> 21  159.85358 123.47601 119.95136 194.36034  71.85273 144.19390 137.36275
#> 22  126.27229 160.58717 166.64731 144.76734 137.35480 146.09402 124.69302
#> 23  219.64203 161.01449 111.42396 144.68830 201.52661 189.84734 116.31785
#> 24  109.93098 116.75119 142.85519 182.75922 138.95877 126.95399 114.84332
#> 25  132.98523 149.01245 109.41269 193.38958 176.27752 136.98837 112.13652
#> 26  186.23240 153.70026 105.87292 141.70136 179.96374 164.88139 130.25599
#> 27  181.85347 118.25410 156.25202 148.12000 145.58775 141.24234 128.43892
#> 28  133.36060  79.81372 124.62405 164.08647 166.23751 164.30007 150.19396
#> 29  181.05185 225.52332 139.32818 113.83440 150.93475 166.91783 124.90222
#> 30  208.02122 169.20604 136.56734 135.58623 136.15538 194.05598 139.60382
#> 31  116.94032 123.77794 123.30400 141.49036 176.61079 124.78760 113.26937
#> 32  108.43778 145.12088 116.89177 156.18814 115.43230 185.73075 153.29762
#> 33  143.41852 112.35426 138.16563 179.63317 184.31667  98.68856 111.26284
#> 34  166.10042 116.16533 153.79522 180.64794 135.89622 179.86879 126.86043
#> 35  174.84774 117.60299 160.49185 206.30061 169.46720 126.65418 144.95942
#> 36  160.09844 111.59274 167.66783 146.83989 147.50980 166.15110 157.70988
#> 37  155.54077 134.61333 149.40325 158.42998 148.65170 125.35653 127.72871
#> 38  153.08863 144.62777 135.63480  86.65776 119.83145 151.26519 138.66258
#> 39  161.05017  80.06030 107.17677 160.70249 127.99083 111.18021 182.30804
#> 40  117.58705 127.57514 182.95034 193.64318 155.87084 163.77492 137.18077
#> 41  164.69189 125.39033 143.07476 150.57985 164.24232 142.70767  86.12176
#> 42  115.06883 156.74794 157.33237 139.12817 145.52256 174.07422 144.57892
#> 43  155.43868 147.49915 119.59425 210.76810 118.69466 182.38773 144.39411
#> 44  163.03187 142.82314 152.00668 196.02754 168.51257 110.65967 149.49571
#> 45  163.21101 148.17089 104.17005  97.49647 109.00986 117.44625 100.43833
#> 46  198.47680 105.98654 188.68579 157.35737 118.11708 146.66464 127.99563
#> 47  198.11781 193.30539 138.57450 168.96385 117.52134 164.62636 147.72148
#> 48  113.94033 129.50632  93.26769 159.93239 117.22373 133.40754 116.36788
#> 49  157.04277 152.24968 105.79983 152.29692 169.82447 108.93619 143.85625
#> 50  191.79516 142.70423 144.13589 198.05574 172.18973 107.92141 113.11507
#> 51  133.46124 128.72827 125.14906 190.81062 133.58301 159.49508 155.61111
#> 52  121.75749 134.67344 193.57236  97.89718 114.91749 141.05970 170.75996
#> 53  161.82439 103.65110 133.98617 153.56870 196.98818 124.95306 124.55206
#> 54  109.49937 134.70430 173.11843 186.10300 136.81529 141.91383 122.83861
#> 55  119.06358 190.06499 128.70686 164.54905 150.39090 118.51022 124.65499
#> 56  179.76299 134.68932 140.05662 189.97522 125.08899 144.35532 157.24696
#> 57  154.44180 126.60159  98.59223 189.84285 119.71473 131.47182 159.33016
#> 58  162.20355 113.79362 190.24915 224.36694 175.14870 108.91281 119.87288
#> 59  145.17350 191.83349 130.27979 170.82722 170.85743 129.49093 138.12819
#> 60  187.24277 206.50253 155.16377  97.31079 109.71967 148.87841 144.77280
#> 61  190.26125  91.43283 129.51458 147.02746 113.17372 143.50794 150.67379
#> 62  200.51416 151.78742 198.77300 163.96306 205.06504 122.01562 144.40542
#> 63  193.76063 160.36851 117.70334 149.26440  99.32265 133.57383 162.74407
#> 64  148.55564 116.61888 126.35405 139.74930 173.29463 143.04906 138.74593
#> 65  122.95917 137.52030 131.55905 126.03211 178.09147 128.84041 181.10462
#> 66  190.96221 125.36093 117.64793 166.02735 114.15533 117.07921 118.94398
#> 67   92.46752 143.69618 155.13289 136.59383 152.63535 127.39081 153.48759
#> 68   96.92071 140.63322 155.25163 123.95775 135.65976 156.19077 121.90447
#> 69  186.01733 157.58209 151.36676 158.27230  93.24033 156.26911 173.40262
#> 70  166.16576 171.54201 112.70766 176.52907 177.43313 117.04315 161.22476
#> 71  147.20580 101.28031 138.67368 104.44665 122.70502 137.63031 107.03068
#> 72  123.26078 148.23009 162.29245 115.16650 122.94165 134.17699 153.90927
#> 73  180.50206 179.46629 154.81515 182.15090 183.02629 162.65859 173.33812
#> 74  174.57260 167.72123 177.62904 115.72076 132.02013 149.76437 102.58100
#> 75  162.38501 132.55153 150.78371 153.84630 111.30488 144.86924  94.18542
#> 76  138.60904 162.99253 149.13515 132.25740 153.70636 186.33080 143.64282
#> 77  169.68780 159.13544 103.48286 148.87154 223.45476 125.61840 168.88700
#> 78  132.61187 135.41056 172.95759 180.99125 120.68649 101.74223  88.22948
#> 79  133.80930 115.90958 123.31271 129.56907 118.31176 135.01218 150.44267
#> 80  146.16810 190.31554  85.22294 148.47171 161.32741  79.00626 149.53965
#> 81  165.20835 122.15040 158.99168 148.89046 136.46215 104.18169 124.53582
#> 82  149.29300 121.53243 191.97090 148.36634 161.72685 143.94514 133.82503
#> 83  210.12583 112.35829 165.30438 187.76718 172.19353 144.36580 134.42151
#> 84  139.43628  79.04211 118.08099 133.38302 131.89917 103.55032 157.37597
#> 85  162.11448 177.28008 113.14603 177.35312 197.45718 136.06844 161.08257
#> 86  155.63844 141.85131 120.96728 132.61204 131.52431 166.97691 117.14354
#> 87  174.81870 140.01346 138.89140 121.28459 198.77043 129.95161 141.29823
#> 88  184.41265 187.30607 147.17044 151.60062  82.97619 210.65165 117.51331
#> 89  155.26635 162.85068 142.67279 178.64247 145.82745 160.34052 142.45951
#> 90   90.63912 143.00674 165.49605 155.63863 131.01444 155.07752 193.39431
#> 91  154.67046 156.04436 102.78204 182.20174 167.63791 183.59517 131.99729
#> 92  178.72853 161.61878 141.42647 162.78155 171.59624 117.06279 132.00465
#> 93  164.29642 105.24717 173.77891 223.47651 165.37777 117.30630 147.33785
#> 94  171.43478 131.19653  98.02423 155.33362 157.48797 128.75675 145.81477
#> 95  150.87199 117.78200 174.45844  98.77106 137.16169 151.76912 135.26631
#> 96  173.64455 101.11796 147.58055 137.03790 138.64708 146.27712 129.57411
#> 97  218.75171 143.05834 110.42359 143.47405 121.92202 155.61154 148.25970
#> 98  115.38046 136.54112 158.23328 165.60350 160.50098 117.08735 105.55012
#> 99  176.31116 166.74698 145.90543 150.22345 169.15186 143.65937 117.26239
#> 100 157.37856 121.67292 136.95559 170.50619 151.03071 183.73020 132.06572
#>         B.4_1
#> 1   139.96643
#> 2   157.03800
#> 3   137.79826
#> 4   178.41737
#> 5   151.37529
#> 6   149.49071
#> 7   211.42273
#> 8   140.38266
#> 9   103.64908
#> 10  170.32374
#> 11  158.78616
#> 12  159.87041
#> 13   78.07695
#> 14  141.62820
#> 15  154.23703
#> 16  165.76058
#> 17  166.34940
#> 18  159.31722
#> 19  117.51670
#> 20  103.51849
#> 21  157.12147
#> 22  151.74521
#> 23  137.19584
#> 24  107.33794
#> 25  161.10677
#> 26  163.32319
#> 27  140.34745
#> 28  141.93211
#> 29  169.71307
#> 30  126.20103
#> 31  169.69665
#> 32  167.34287
#> 33  165.99155
#> 34  165.63709
#> 35  157.71636
#> 36  138.98418
#> 37  168.01277
#> 38  178.32292
#> 39  139.46382
#> 40  141.88861
#> 41  153.29657
#> 42  115.77000
#> 43  168.46099
#> 44  179.95874
#> 45  140.31578
#> 46  151.77584
#> 47  128.20410
#> 48  148.28607
#> 49  116.27569
#> 50  164.78435
#> 51  172.43316
#> 52  148.25114
#> 53  173.41409
#> 54  102.34910
#> 55  121.64487
#> 56  196.48111
#> 57  147.18486
#> 58  142.55995
#> 59  112.27664
#> 60  150.54563
#> 61  149.79738
#> 62  155.13264
#> 63  130.23799
#> 64  145.55605
#> 65  136.02867
#> 66  140.55273
#> 67  127.91893
#> 68  174.78271
#> 69   83.67958
#> 70  154.48359
#> 71  182.05982
#> 72  145.08287
#> 73  120.19183
#> 74  153.62620
#> 75  155.38778
#> 76  177.91768
#> 77  205.75762
#> 78  106.62732
#> 79  138.20606
#> 80  136.44803
#> 81  126.06339
#> 82  162.28265
#> 83  138.40367
#> 84  150.39014
#> 85  142.60121
#> 86  201.11997
#> 87  193.71683
#> 88  122.69316
#> 89   89.96337
#> 90  168.21161
#> 91   97.11396
#> 92  168.93242
#> 93  157.39154
#> 94  153.17649
#> 95  148.15666
#> 96  105.25000
#> 97  168.22057
#> 98  120.27573
#> 99  137.31168
#> 100 154.56444