Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10,10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(subset = rep(c("A","B"),50),
                   level = rep(1:4, each = 25))
rownames(meta) <- 1:100
scRNA <- Seurat::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA, by = c("subset","level"),
               min.cells = 10, max.cells = 20)
#>         A.1_1     B.1_1     A.2_1     B.2_1    A.3_1     B.3_1     A.4_1
#> 1   109.62152 132.11708 145.08997 153.31864 124.0469 170.83204 140.17399
#> 2   192.95477  75.73589 128.23376 130.25363 165.1747 122.34716 113.88151
#> 3   136.16416 117.74278 163.21200 153.99149 110.6917 144.92106 157.52080
#> 4   147.70123 157.87307 108.97127 170.79876 138.6679 189.51002 116.85973
#> 5   140.24600 114.75017 145.44767 222.48436 170.3545 139.01034 163.72863
#> 6   127.20716 150.77342 112.15544  99.00244 166.6883 143.30890 167.93445
#> 7   166.80976  73.75998 109.45610 179.24879 184.7384 116.20768 124.84253
#> 8   168.14727 188.78105 130.94634 146.01486 195.4161 128.85356 131.14027
#> 9   173.58867 129.70803 136.56352 143.53064 141.7933 122.59949 143.76429
#> 10  154.08393 136.90330 144.24178 140.43664 156.7028 123.06985 121.10139
#> 11  168.23534 148.71796 146.03269 189.48416 190.5427 172.56328 139.75601
#> 12  123.02964 150.84269 195.35143 173.31690 147.4376 158.32119 144.19133
#> 13  162.35518 101.03440 111.71882 101.75663 117.5120 128.87821  93.71428
#> 14  126.95580 142.36013 135.78577 129.49106 147.7082 157.96343 116.22957
#> 15  152.08503 133.81895 187.95642 176.73589 121.7504 113.08063 139.95719
#> 16  123.47349 158.03268 167.67608 153.96395 159.6055 141.44104 149.11153
#> 17  190.96930 144.72536  98.87917 149.71567 177.0407 119.34494 166.62811
#> 18  216.14952 155.24592 137.95049 139.69940 139.0931 155.14324 134.86022
#> 19  183.99801 152.49155 150.80320 216.53585 124.3502 104.82087 167.54829
#> 20  132.19551 169.17709 124.48641 155.01770 166.1449 143.50373 105.66743
#> 21  113.94085 128.92860 138.66566 141.14977 139.5571 126.82001 109.99080
#> 22  110.89558 122.27553 137.10988 167.97226 155.6000 100.38065 105.83071
#> 23  175.64909 125.22958 127.16973 163.20699 139.7795 150.95192 131.14273
#> 24  168.24869 163.61922 108.77469 134.51197 130.1312 115.87442 143.08292
#> 25  144.88900  86.18027 100.01207 175.74877 135.0963 165.63135 141.28764
#> 26  143.96108 107.04058 118.14454 135.06928 176.3464 140.04217 129.19935
#> 27  114.86224 122.76122 139.90540 155.91790 162.4000 144.35248 147.41549
#> 28  108.88271 217.30525 186.23952 238.67618 143.5449  85.57989 177.42147
#> 29  185.28074 119.50095 164.07456 128.02039 181.5240 132.93041 163.68143
#> 30  164.88614 183.95583 137.21675 148.39315 170.3822 181.96202 106.17715
#> 31  182.79495 150.25465 117.26486 167.99990 193.2450  81.09627 122.92468
#> 32  124.01371 151.11456 131.78844 171.76681 148.4974 135.48781 134.02248
#> 33  146.15788 175.72506 120.13747 129.75536 158.8618 101.67427  88.21986
#> 34  168.47713  96.73815 151.29485 139.91099  96.3418 178.12471  96.82946
#> 35   91.95477 184.48983 163.90630 160.21301 211.1629  94.78639 123.02880
#> 36  178.11960 102.82775 118.72288 142.92159 169.0216 130.41649  95.14115
#> 37  152.79536 140.36891 112.54451 172.52362 188.2773 152.88591 102.76340
#> 38  149.80447  89.60235 136.38226  98.44480 196.6141 154.64179 141.30192
#> 39  186.66430 192.85462 182.41513 201.93135 170.3610 182.78543 113.38647
#> 40  161.93125 108.40107 152.85266 120.10711 120.3512 127.48504 108.57426
#> 41  149.56288 228.27258 125.79738 145.11253 176.3892 103.13644 155.73653
#> 42  136.41055 131.04607 138.48402 141.00747 157.8612 166.07117 160.88291
#> 43  147.24410 144.38223 128.08377 142.39154 161.8512 150.57244  99.24259
#> 44  141.22635 187.18172 196.07449 138.57682 173.6316 139.79513 136.55276
#> 45  161.72663 159.78200 180.03519 133.56012 149.1883 134.85197 149.11578
#> 46  119.95172 172.52911 156.11733 171.69652 157.9999 191.59765 138.92447
#> 47  206.03936 209.11463  77.44704 178.98502 170.6543 158.22377 140.62310
#> 48  103.84632 129.62184 186.06627 122.41170 149.9578  90.37868 122.62994
#> 49  153.66905 118.18939 111.01838 120.03287 116.6970 124.06074 169.87780
#> 50  135.67509 162.69561 117.57289 178.36519 150.7809 128.01000 153.44813
#> 51  152.94941 120.71808 142.25575 170.26852 172.1451 153.45472 142.94247
#> 52  117.19992 123.19183 178.49029 114.88945 216.3765 156.24279 137.18760
#> 53  121.72493 151.91367 166.27750 166.45938 183.8419 144.19586 164.96692
#> 54   99.43193 206.59609 189.71284 117.55560 184.9136 164.82889 135.89759
#> 55  163.60857 152.86861 161.66727 212.22270 139.6001 178.44775 150.13593
#> 56  195.53408 134.81792  95.11667 129.40828 135.8924 135.48566 104.03867
#> 57  112.61818 128.83396 127.95155 177.87569 102.6440 156.68737 138.25945
#> 58  145.05269 132.53723 188.59452 156.75597 164.4811 146.98331 169.70498
#> 59  134.35192 196.03253 146.18543 121.40531 161.4771 107.05037 154.81046
#> 60  134.46744 131.39071 172.91066  93.93669 159.4564 168.57844 219.50369
#> 61  181.89168  60.92194 179.50567 144.48097 136.9840 120.36265 163.42673
#> 62  158.06688 143.74062 156.70477 183.28583 133.9008 110.41257 141.81159
#> 63  154.33060 130.15510 181.76814 151.37274 145.2846 115.77297  96.88060
#> 64   99.69998 165.22132 135.52474 192.35259 145.2125 141.92457 105.81876
#> 65  168.59660 113.55667 135.76551  77.60276 130.1587 100.94208 129.04832
#> 66  202.39567 165.29518 113.35310 132.54694 101.2173 137.87747 113.66153
#> 67  109.69123 125.67925 154.87750 126.61960 183.3492 186.11778 143.94283
#> 68  134.89682 166.70353  99.08407 135.52139 178.1478 144.46853 135.17410
#> 69  158.30356 144.27657 171.79700  81.14691 131.7767 122.89383 111.49280
#> 70  177.57511 170.95529 137.15002 140.44403 146.6010 153.44643 138.49501
#> 71  176.67845 174.27282 108.90445 111.44705 155.7793 166.02804 160.32013
#> 72   95.21926 142.29342 143.88743 176.30237 158.1199 149.57152 122.86469
#> 73  126.51812 107.01656 196.80383 149.55986 169.5107 124.07772  95.05622
#> 74  160.72274 133.88803 126.85435 125.06869 225.1409 151.71260  79.41066
#> 75  109.78465 168.90981 126.18520 148.86708 141.6426 153.90179 193.00348
#> 76  154.60806 177.10322 133.82130 143.00503 162.8347 124.75540 109.01870
#> 77  140.77412 146.07827  88.03715 138.71941 182.6791 109.60660 210.43491
#> 78  201.54315 156.22788 119.02812 121.66460 158.2901  90.09748 115.64371
#> 79  154.85174 109.50770 113.21614 198.81649 178.0514 158.33439 171.45648
#> 80  181.21974 175.36545 117.37546 129.93252 146.8963 103.84737 132.25010
#> 81  123.16696 133.39194 120.14336 186.39761 166.8072  86.98980 175.72774
#> 82  146.19180 124.13007 150.96314 116.38432 139.1390  95.77491 129.76404
#> 83  130.02246 159.44829  74.35881 157.74585 149.3640 143.42517 152.06060
#> 84  148.30792 125.49929 165.96934 142.48212 166.7693 136.92080 168.10361
#> 85  131.30364 159.80058 125.89748 146.48354 145.7004 113.98531  94.74233
#> 86  165.92335 132.15260 117.16039 207.07773 181.5541 154.34238 136.81243
#> 87  179.40764 131.52841 143.93225 132.18692 133.1140 130.67047 134.89616
#> 88  168.02531 121.25311 159.79362 168.37579 132.7624 155.54368 159.96415
#> 89  194.05027 155.24647 167.79276 168.49085 154.9435 105.19740 102.75757
#> 90  192.76740 104.16502 138.24660 177.46017 115.6428 130.12067 113.04700
#> 91   95.41337 119.91693 114.42772 115.47666 140.4876 136.28996 121.10425
#> 92  147.60505 127.01848 190.55446 113.61870 130.0046 148.40825 118.10390
#> 93  116.43220 130.38910 144.28521 162.19780 151.1245 136.77247 145.50600
#> 94  182.24762 158.54946  84.47339 242.17803 140.0418 133.70212 186.77173
#> 95  165.33721  93.09269 199.67507 179.49393 129.7400 112.01530 141.60087
#> 96  164.32394 142.52138 145.86239 153.04269 161.9371 128.09630 184.29983
#> 97   81.13129 114.98325 149.72212 118.92964 154.7849 145.24698 115.85320
#> 98  188.70304 130.77824 112.04649 113.44237 156.2800 181.88651 153.71141
#> 99  152.02104 136.84406 168.96961 122.11398 155.6578 126.02003 139.78749
#> 100 176.98091 176.20018 141.21644 101.89703 165.5541 155.13786 128.18476
#>         B.4_1
#> 1   149.21961
#> 2   131.23157
#> 3   135.01214
#> 4   146.53523
#> 5   144.81227
#> 6   112.20424
#> 7   180.68401
#> 8   169.60414
#> 9   154.55946
#> 10  185.02338
#> 11  141.69374
#> 12  145.80250
#> 13  136.92476
#> 14  157.18477
#> 15  156.96774
#> 16  131.16721
#> 17  158.81586
#> 18  139.98937
#> 19  135.68971
#> 20  146.64971
#> 21  160.89346
#> 22  154.70484
#> 23  161.54018
#> 24  100.22254
#> 25  145.60190
#> 26  159.02218
#> 27  173.52830
#> 28  175.43628
#> 29  120.77479
#> 30  141.84379
#> 31  159.10919
#> 32  204.58417
#> 33  131.78001
#> 34  177.26731
#> 35  129.64482
#> 36  147.36839
#> 37  107.18081
#> 38  127.31727
#> 39  141.67435
#> 40  140.10171
#> 41  207.22204
#> 42  162.57528
#> 43  193.83668
#> 44  163.29421
#> 45  126.47169
#> 46  154.36067
#> 47  194.20143
#> 48  162.13170
#> 49  158.09133
#> 50  188.26432
#> 51  126.65024
#> 52  149.05409
#> 53  133.13410
#> 54  130.88130
#> 55  131.56293
#> 56  126.17890
#> 57  133.01322
#> 58  156.42602
#> 59  124.65561
#> 60  149.45807
#> 61  147.98647
#> 62  141.69249
#> 63  147.86444
#> 64  139.33254
#> 65  153.05140
#> 66  149.93845
#> 67  150.15432
#> 68  191.81808
#> 69  160.33870
#> 70  155.64264
#> 71  183.90794
#> 72  156.32818
#> 73  118.82408
#> 74  157.52443
#> 75  136.36630
#> 76  121.05018
#> 77  120.59397
#> 78  164.85742
#> 79  164.34463
#> 80  109.42007
#> 81  145.23009
#> 82  164.42521
#> 83  174.46018
#> 84  108.37569
#> 85  149.95875
#> 86  204.36938
#> 87  175.76917
#> 88  133.59557
#> 89  104.69773
#> 90   90.35018
#> 91  151.87303
#> 92  167.65003
#> 93  147.67820
#> 94  161.60353
#> 95  140.80687
#> 96  212.91213
#> 97  147.27538
#> 98   96.49131
#> 99  190.21547
#> 100 140.68905