Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10, 10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(
  subset = rep(c("A", "B"), 50),
  level = rep(1:4, each = 25)
)
rownames(meta) <- 1:100
scRNA <- Seurat::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA,
  by = c("subset", "level"),
  min.cells = 10, max.cells = 20
)
#>         A.1_1     B.1_1     A.2_1     B.2_1    A.3_1     B.3_1     A.4_1
#> 1   153.73331 114.00921 120.78608 126.55318 178.1544  98.52342 127.71706
#> 2   156.90680 181.87851 120.06258 174.96344 175.9204 109.63053 147.96253
#> 3   121.09773 149.74982 172.37824 115.65250 157.8901 159.59164 201.14639
#> 4   176.93826 172.67623 116.76214 152.20281 192.7953 153.47473 168.37625
#> 5   185.73186 148.21155 121.81119 143.21233 220.4436  96.05664 153.25245
#> 6    96.90603 137.29603  99.88383 166.87028 132.5158  97.70589 132.32187
#> 7   172.02964 177.44236 137.16619 152.44289 122.0314 128.72535 125.95450
#> 8   154.79023 122.47039 119.67285 138.79863 175.2919 151.43463 154.65641
#> 9   136.46060 190.16000 123.86228 142.04218 125.0533 119.13871 103.79324
#> 10  160.69258 184.86636 120.34698 157.81163 151.8444 162.94369 105.75308
#> 11  147.80642 103.77622 121.36451 159.18420 128.4227  98.92021 161.58000
#> 12  166.06018 119.24458 160.42113 182.77291 203.1449 165.26514 108.15088
#> 13  139.25101 136.31967 157.59330 144.86642 122.7483 124.92646 164.23851
#> 14  173.08148 176.92800 100.52600 152.23849 103.2080 157.79519 166.89950
#> 15  118.41219 137.32579 110.76539 145.42433 165.6557  93.45433 173.42350
#> 16  172.99469 168.89023 115.26469 160.03123 172.5601 138.26681 196.95334
#> 17  132.24691 147.86354 108.94344 201.32766 155.0750 129.42614 159.48411
#> 18  179.87032 139.51826 166.59570 148.07481 198.1767 132.62478 105.84117
#> 19  157.68949 162.32430 167.45651 122.87925 129.1490  89.81559 114.48128
#> 20  133.04410 118.15473 106.28202 135.08734 162.9377 129.13366 111.22511
#> 21  129.96250 182.43004 120.28815 147.53235 173.7006 133.79822 122.66949
#> 22  173.02826 138.25312 126.19269 196.04310 168.1828 120.59514 170.63119
#> 23  176.95159 184.85531 164.53891 120.91056 163.7838 128.84207 133.45656
#> 24  151.25532 139.73824 137.77925 164.93036 159.9324 207.87777 101.47251
#> 25  148.53719 143.53122 100.53956 157.76498 149.0441 115.82647 138.61495
#> 26  120.15743 131.67984 125.76459 169.88548 156.1793 184.27115 111.42275
#> 27  156.28195 165.93779 192.67204 129.42043 129.9590 111.09503 165.13273
#> 28  155.11913 139.88356 131.37700 152.77326 138.5442 141.22680 122.57105
#> 29  133.44383 112.77890 165.16248 115.43500 154.9779 157.29683 194.99441
#> 30  121.60099 143.71123  95.68437 113.17626 164.5415 135.85227 199.91013
#> 31  105.31200 119.02370 107.76112 159.35951 145.8349 114.19126 155.84639
#> 32  124.40600 131.94334 197.94893 194.35705 118.3172 141.50630 126.99254
#> 33  135.12436 156.10823 141.32637 137.24302 122.5957 157.02913 144.40216
#> 34  201.00936 116.95393 117.16719 118.96522 157.5005 142.60519 126.15594
#> 35  136.35712 159.62281 122.86380 172.53354 152.4838 160.21299 149.50908
#> 36  170.41934 114.19845 108.72727 162.52982 145.7990 174.19550 161.93754
#> 37  165.32435 183.49649 124.73354 183.57378 162.4950 148.94004 148.31555
#> 38  135.12365 155.27476 177.25251 196.00342 177.1225 167.18686 182.08022
#> 39  114.62874 115.12560  90.90185 122.17760 149.0447  98.87335 122.17691
#> 40  120.62680 133.99580 150.32344 158.11213 164.6373 126.72658 138.53351
#> 41  189.37989 123.31755 167.18370 148.10960 203.5019 160.54768 160.42232
#> 42  125.25582 169.29967 107.50874 160.31335 123.1212  98.12106 170.66254
#> 43  127.39117 159.86040 149.12039 140.67671 151.8271 114.90161  96.20247
#> 44  204.30262 172.81222 138.52755 152.42498 190.5786 159.25185 170.79296
#> 45  195.01447 144.45472 145.42332 202.08505 179.1626 117.53121 135.45839
#> 46  180.74814  97.97893 125.21994 121.61435 156.2546 127.68736 132.41855
#> 47  155.06376 102.89141 160.45022 116.77214 171.0652 153.75774 137.16922
#> 48  192.56301 128.22990 110.47853 138.21052 125.4176 144.73542 149.90479
#> 49  159.89514 142.67765 176.29898 130.66072 129.1305 110.97963 142.03546
#> 50  138.53070  84.73881 149.92556 145.82422 112.1373 158.04689 160.27404
#> 51  220.87557 157.58067 148.18793 155.08341 163.6884 109.89986 128.05239
#> 52  179.55370 120.79128 116.95146 167.08935 158.6842 124.77469 149.00481
#> 53  146.05990 131.79453 149.31536 191.49500 159.4259 152.73250 186.88782
#> 54  108.83950 118.38505 140.84297 159.66509 195.9300 134.36992  85.92978
#> 55  111.41780 130.15272 150.44466 156.01298 170.4472 127.65654 125.12458
#> 56  119.04153 156.32970  98.29594 185.85320 162.6404 160.79287 153.62242
#> 57  164.11664 114.34209 186.43239 103.28945 128.1559 112.02288 154.76297
#> 58  168.93723 115.59094 124.33944 153.00842 129.4499 152.08018 150.22838
#> 59  134.61395 112.24083 140.16037 152.96652 166.5678 136.33942 152.13518
#> 60  115.08778 156.02487  94.61835 137.03137 147.7862 104.96944 120.63725
#> 61  135.45037 126.66060 135.59922 172.40679 114.0623 131.89610 150.02300
#> 62  229.87726 183.58993 126.60260 186.02109 141.8721 165.83608 165.54108
#> 63  117.21144 153.33598 173.22917 130.61109 143.0231 141.94416 144.47595
#> 64  130.82901 174.64762 180.97851 154.40983 150.1595 109.19070 133.44748
#> 65  117.35840 111.12331 167.11130 154.25503 187.7188 161.39969 157.09013
#> 66  118.52538 126.36777 141.61089 110.45776 204.5751 122.54812 109.97967
#> 67  144.06640 132.05975 150.16539 124.98998 129.9201 214.63613  78.94559
#> 68  188.18104 186.11935 128.23861 172.10199 157.4396 165.59322 115.55906
#> 69  139.64518 127.15961 164.21194 131.48842 184.5456 126.78131 148.21426
#> 70  133.99653 144.79975 176.97981 166.11520 122.8513 143.69634 162.39101
#> 71  140.22019 128.44364 171.20064 106.93884 148.1979 115.15498 140.02753
#> 72  116.52324  92.06125 199.89339 201.39232 165.2456 125.71258 119.75298
#> 73  167.18588 169.81703 138.21157 109.71302 126.0189 127.88507 147.15640
#> 74  175.00246  97.27904 106.15505 121.62592 171.6359 137.13585 165.25923
#> 75  189.58749 162.61425 137.82764 136.92954 207.5360 147.20230 150.33349
#> 76  145.50583 153.87378 146.90481 179.59931 206.1473 158.50409 120.61384
#> 77  103.16249 137.70258 169.41633 142.11808 169.3411 124.35510 137.52759
#> 78  134.87004 145.78941 140.65912 136.46896 159.2327 105.36854  85.08547
#> 79  136.96062 112.41294 150.82791 160.89806 154.6226 169.38209 159.42151
#> 80  155.47788 148.02598 124.09440 184.61016 144.2798 118.19452 147.39576
#> 81  166.87251 183.81954 198.75723 223.84247 152.3605 112.25367 159.80025
#> 82  125.08451 122.36774  96.35014 187.88478 172.4331 166.46852 117.06224
#> 83  137.73153 140.29857 122.47447 169.73771 119.9030 136.03142 115.58444
#> 84  120.40961 140.04387  84.28579 113.71318 192.2462 180.32424 136.23609
#> 85   93.28172 116.99132 168.70785  80.32839 164.3442 142.69647 116.84616
#> 86  109.98244 110.30337 126.57580 142.02316 128.1856 188.99395 129.73879
#> 87  189.05733 110.86302 150.96266 152.09820 122.6085 144.11110  91.28170
#> 88  139.74190 146.28198 105.20367 161.77403 141.6665 131.43786 120.61307
#> 89  130.87125 128.15283 154.19865 182.24182 130.9118 125.70055 149.26490
#> 90  161.61931 153.38341 124.33261 123.84477 121.1143 162.03145 139.26367
#> 91  170.24201 143.15236 232.96443 177.62522 191.1099 158.70335 101.10565
#> 92  105.38963 158.10919 148.67634 119.66216 101.5308 131.87823 152.89938
#> 93  133.52618 150.25760 130.14740  85.97431 122.0998 138.53151 123.38716
#> 94  183.83404  90.14124 187.20100 178.79885 196.1217 135.40495 121.19841
#> 95  140.30818 110.03608 101.80405 139.83112 159.1098 110.06853 150.20892
#> 96  135.19726 130.26185 161.58756 165.28168 156.4455 200.69918 141.82100
#> 97  156.45511 171.32810 159.90593 151.37971 174.4210 139.08636 121.99284
#> 98  104.13033 193.69878 169.51248  95.06391 192.9678 148.73303 116.92685
#> 99  142.81713 219.07505 119.21750 105.69398 134.7031 151.13838 138.53224
#> 100 131.28420 164.45579 136.43058 181.26790 158.1526 104.89582 205.49733
#>         B.4_1
#> 1   122.27209
#> 2   142.82073
#> 3   172.02917
#> 4   181.33505
#> 5   183.80917
#> 6   119.94457
#> 7   150.72709
#> 8   115.82242
#> 9   125.92095
#> 10  174.50812
#> 11  104.85591
#> 12  148.55571
#> 13  169.72508
#> 14  204.27370
#> 15  153.01823
#> 16  177.24331
#> 17  138.31029
#> 18  102.48770
#> 19  124.65518
#> 20  175.74914
#> 21  136.85781
#> 22  173.94255
#> 23  165.31497
#> 24  151.87464
#> 25  115.53023
#> 26  122.40260
#> 27  153.50120
#> 28  126.85060
#> 29  164.38368
#> 30  118.86733
#> 31  145.01071
#> 32  184.46373
#> 33  117.00903
#> 34  164.74708
#> 35  142.64918
#> 36  121.24702
#> 37  163.36412
#> 38  162.43540
#> 39  166.40149
#> 40  123.71742
#> 41  189.68532
#> 42  187.71313
#> 43  128.95223
#> 44  108.27765
#> 45  119.01583
#> 46  176.09359
#> 47  173.86097
#> 48  203.75046
#> 49   99.06153
#> 50  169.69471
#> 51  162.12518
#> 52  123.79921
#> 53  187.36550
#> 54  112.05454
#> 55  192.56240
#> 56  158.08630
#> 57  185.84865
#> 58  147.88397
#> 59  200.80702
#> 60  142.45177
#> 61  103.10021
#> 62  149.02967
#> 63  159.36301
#> 64  128.81947
#> 65  171.55486
#> 66  113.45989
#> 67   91.96884
#> 68  188.52646
#> 69  129.13285
#> 70   98.09190
#> 71  174.57135
#> 72  177.85144
#> 73  134.58333
#> 74  112.76902
#> 75  116.20095
#> 76  190.97173
#> 77  146.30151
#> 78  157.68263
#> 79  181.45327
#> 80  153.56553
#> 81  144.93823
#> 82  114.04276
#> 83  164.83757
#> 84  162.52413
#> 85  191.00362
#> 86  115.57106
#> 87  151.20440
#> 88  109.49721
#> 89  184.12186
#> 90  153.37990
#> 91  128.88837
#> 92  102.69117
#> 93  180.14713
#> 94  130.28346
#> 95  157.31401
#> 96  167.42933
#> 97  130.32666
#> 98  187.92957
#> 99  160.98425
#> 100 191.92046