Aggregate single cells to pseudo-samples according to specific factors

Gather cells for each group according to specified factors, then randomly assign and aggregate cells to each pseudo-samples with randomized cell size. (min.cells <= size <= max.cells)

pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,data.frame
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for matrix,vector
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for Seurat,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

# S4 method for SummarizedExperiment,character
pseudo_samples(
  data,
  by,
  fun = c("sum", "mean"),
  scale = NULL,
  min.cells = 0,
  max.cells = Inf,
  slot = "counts"
)

Arguments

data

a matrix or Seurat/SCE object containing expression and metadata

by

a vector of group names or dataframe for aggregation

fun

chr, methods used to aggregate cells, could be 'sum' or 'mean', default 'sum'

scale

a num or NULL, if to multiply a scale to the average expression

min.cells

num, default 300, the minimum size of cells aggregating to each pseudo-sample

max.cells

num, default 600, the maximum size of cells aggregating to each pseudo-sample

slot

chr, specify which slot of seurat object to aggregate, can be 'counts', 'data', 'scale.data'..., default is 'counts'

Value

An expression matrix after aggregating cells on specified factors

Examples

counts <- matrix(abs(rnorm(10000, 10, 10)), 100)
rownames(counts) <- 1:100
colnames(counts) <- 1:100
meta <- data.frame(
  subset = rep(c("A", "B"), 50),
  level = rep(1:4, each = 25)
)
rownames(meta) <- 1:100
scRNA <- Seurat::CreateSeuratObject(counts = counts, meta.data = meta)
pseudo_samples(scRNA,
  by = c("subset", "level"),
  min.cells = 10, max.cells = 20
)
#>         A.1_1     B.1_1     A.2_1     B.2_1     A.3_1     B.3_1     A.4_1
#> 1   127.63464 133.12791 168.09808 164.80270 175.59699 159.56114 107.56415
#> 2   160.31729 158.13191 118.23953 100.76523 191.49055  98.12376 163.46176
#> 3   124.17414 105.27126 139.18895 139.69177 135.93606  89.58795 124.59407
#> 4   142.17682 142.24411 103.29485 134.04524  89.50063 120.18201  97.70978
#> 5   145.62831 134.78986 145.74675 125.25986  91.48373 182.58325 124.91129
#> 6   155.88083 112.91491 157.67688 129.51333 200.87215 165.21446 113.41992
#> 7   136.63977 109.88800 193.04550 146.22493 109.09792 143.39153 132.00423
#> 8   130.62534 161.85819 166.56321 182.07048 142.80425 181.33967 191.48056
#> 9   174.92532 122.76083 114.88443 126.90526 129.93363 186.32343 116.43910
#> 10  123.03125 129.18327 149.58293 123.91151 198.77401 188.60653 152.63023
#> 11  170.28152 114.95110 130.20487 149.35497 207.62639 181.49804 146.38161
#> 12  178.21681 166.91507 132.90124 161.90238 139.51875 148.30349 122.20413
#> 13  164.34216 209.03030 173.90862 123.68842 153.58336 125.97681 141.98088
#> 14  149.37500 193.39078 159.80332 124.43785 162.48538 149.05572 172.96732
#> 15  116.92648 148.62996 169.91176 119.85915 135.54789 123.35178 139.00898
#> 16   63.92927 135.62543 130.08022 108.84595 186.23475 141.28944 189.63129
#> 17  173.24524 126.85445 134.09781  97.13819 121.97949 116.14471 128.82325
#> 18  150.47589 120.14615 175.87421 114.96262 201.03569 163.54412 122.98235
#> 19  157.82815 123.15939 126.88918 140.37023 175.03077 161.70722 184.24605
#> 20  116.19504 126.68276 131.60529 143.54513 180.69080 181.67417 152.71475
#> 21  167.63288 121.45484 166.58890 144.54336 141.15952 147.82007 188.81756
#> 22  148.71170 147.98447 134.38055 162.69042 109.98706  82.03675 108.06864
#> 23  141.73049 139.60690 142.30795 137.14911 128.14107 189.83120 176.50193
#> 24  165.15822 131.22416 100.77799 116.04122 138.61965  95.09913 172.66568
#> 25  135.15714 102.96337 155.48002 216.40751 181.04892 151.33081 134.65659
#> 26  167.80280 144.89726 112.43000 131.06688 147.29421 108.43443 115.35773
#> 27  134.57499 124.51878 157.95710 137.86287 116.00195 148.75663 136.89532
#> 28  185.96057 128.08536 157.65139 151.61766 194.16082 126.99467 120.70776
#> 29  131.35287 170.00583 132.80152 119.16847  92.85613 118.05685 130.59827
#> 30  146.19371 114.18008 177.98636 154.81645 116.38171 147.24145 119.54113
#> 31  110.13185 160.42236 144.27192 204.34551 171.77650 159.72192 107.21596
#> 32  192.26578 117.19496 129.21466 160.34440 186.63654 154.50965 135.09555
#> 33  164.08983 145.89349 181.18892 154.51989 170.17693 173.04028 118.74214
#> 34  102.68477 121.41246 142.43899 171.21849 135.95357 100.56150 123.89661
#> 35  130.81352 158.44093 132.41144 211.84127 106.64214 119.43847 171.82837
#> 36  164.14753 104.31482 156.85941 150.74123 137.46208 150.15363 174.95177
#> 37  141.74566 128.50885 121.25076 155.46150 142.40454 132.60305 137.21640
#> 38  128.38037 110.52979 160.15449 166.32500 176.22053 153.47124 107.25929
#> 39  183.47765 140.06801 148.25230 140.93100 134.07993 193.11501 127.06035
#> 40  190.88958 126.31265 129.29537 179.23361 191.04961 128.73276  76.17813
#> 41  158.06761 134.00800 126.42814 118.88880 177.72500 132.78364 153.27790
#> 42  144.79309  83.78664 134.64264 164.50944 143.26242 126.35667 136.42038
#> 43   81.61022 142.15623  89.62123 138.13470 188.24523 110.96976 128.19401
#> 44  198.65287 219.43516 202.01133 157.39132 148.34141 102.41151 167.07785
#> 45  131.45142 130.67461  93.90339 125.62197 111.17054 181.94368  85.46782
#> 46  147.76906 115.88304 178.64455 118.37639 165.86191  91.21976 129.59274
#> 47  117.44412  97.94134 156.92197 140.67216 145.62752 133.24706 134.70265
#> 48  206.43786 126.73686 158.13917 141.47774 169.67410 195.82307 161.66676
#> 49  166.19962 134.35871 102.50496 112.40727 195.90532 172.06466 140.92731
#> 50  113.01910 164.69697 136.91213 142.14386 138.26135 128.17192 211.32127
#> 51  156.15561 124.65721 148.22833 154.98356 178.89334  88.15046 155.39567
#> 52  125.15520 192.05733 196.19511 151.73681 169.21630 174.40856 103.58816
#> 53  113.15076 149.35535 145.18261 123.52402 138.15210 171.19772 141.80172
#> 54  197.26675 144.54416 143.42025 163.38874 178.32712 130.79440 173.85269
#> 55  180.93016 156.15093 152.32984 202.71946 142.05109 171.79037 125.62796
#> 56  168.13429 189.80517 165.10587 138.87009 134.04436 141.34054 125.42936
#> 57  161.82303 146.02608 100.33960  93.03852 189.94481  87.36575 104.27793
#> 58  146.25577 200.42330 115.23005 124.37217 150.60290 155.52065 133.83973
#> 59  133.12538 122.25582 154.26248 147.78254 180.10879 147.46002 138.69859
#> 60  146.44139 127.45055  90.80409 179.91969 166.77173 112.60119 167.08567
#> 61  162.39656 171.08734 133.10609 141.66252 138.97249 153.62703 127.88698
#> 62  138.92453 116.15336 158.05014 159.07603 195.33266 131.49692 156.24156
#> 63  159.17272 117.35703 116.53195 174.44907 155.56558 111.20477 128.85132
#> 64   86.66414 136.01406 149.02030 125.49590 140.57602 188.87250 138.60883
#> 65  150.74731 128.95932 171.35987 104.72170 198.82273 155.12201 120.12194
#> 66  195.30422 185.78977  93.38144 182.00828 128.37860 180.06133 174.99170
#> 67  153.09964 159.72550  79.95922 126.10199 142.70133 175.33947 105.91618
#> 68  128.34832 168.79569  97.38702 168.39454 139.45660 101.09864 116.74669
#> 69  150.54293 183.52369 122.32578 162.41852 148.54865 107.14914 123.29340
#> 70  166.15326 148.32657 161.78505 210.42261 147.68975  97.48099 136.68830
#> 71  113.54889 139.69559 115.08513 178.29847 205.98440 192.80378 159.92050
#> 72   88.93156  93.99246 130.91459 155.27607 131.45922 136.09660 130.89181
#> 73  180.37850 139.90115 140.23152 158.16477 184.49359 201.07809 148.20724
#> 74  184.38939 126.40032 177.50045 117.28651 140.89184 149.82243 147.70833
#> 75  139.30467 145.79145 150.80847 104.05163 124.93448  93.07885 111.68744
#> 76  163.70003 185.50552 120.63488 156.19037 164.41685 147.68897 136.14908
#> 77  147.46942 120.53383  88.90365 176.63982 152.10167 188.65420 169.95657
#> 78  151.87090 117.45186 136.43935 142.04961 145.46660 111.98160 131.82104
#> 79  119.32840 106.18878 146.07645 161.12468 117.38950 164.39332 146.36086
#> 80  143.32603 145.81174 141.04441 155.34858 176.92801  89.02799 192.32121
#> 81  226.48263 157.70441 143.60964 162.05470 194.32670 140.22424 127.79169
#> 82  185.78382 154.30506 137.37662 238.76430 173.30453 147.85859 169.31848
#> 83  124.71450 110.28118 173.32582 198.86099 102.28439 177.40914 145.23321
#> 84  173.58825 205.80281 135.01002 158.77940 132.75224 159.72773 143.49375
#> 85  134.31548 124.27244 167.54248 164.94629 132.45146 189.95890 136.46124
#> 86   90.97685 110.05999 169.43135  93.06967 148.85483 137.91430 138.78685
#> 87  155.60446 121.50543 162.30942 119.78243 122.43365 138.88098 165.71276
#> 88  110.45981 129.78421 101.50348 191.88306 139.24687 122.74576 137.03718
#> 89  191.33348 154.53368 122.83557 164.03328 191.45136 175.63084 137.26156
#> 90  134.59680 235.56808 156.24368 186.43589 148.11845  99.17680  87.54055
#> 91  120.89244 207.94674 148.47900 114.78251 146.87110 120.91127 144.37007
#> 92  146.62938 226.41996 141.92472 137.24013 152.12782 166.77753 130.56853
#> 93  156.24705 137.14283 153.90094 121.20784 171.18954 107.39734  91.45600
#> 94  100.83938 155.02081 162.00393  95.54236 221.67799  98.34368 144.62121
#> 95  162.57695 130.82439 162.07730 205.86988 171.31950 146.26316 157.14138
#> 96  110.81149 179.48960 132.67128 173.08751 153.48845 196.10652 118.36417
#> 97  159.71490 192.04059 108.63383 189.76275 159.00988 131.95416 126.84847
#> 98  116.81870 131.33224 121.90928 142.29457 169.50432 117.92838 123.68870
#> 99  129.30636  98.37527 143.24417 132.72494 206.27609 174.73136 133.84058
#> 100 139.51204 123.00965 153.21391 170.31311 132.51629 144.66730 153.00239
#>         B.4_1
#> 1   144.65219
#> 2   149.50632
#> 3    95.69831
#> 4   122.70457
#> 5   120.38257
#> 6   189.10771
#> 7   133.74847
#> 8   171.68540
#> 9   166.83937
#> 10  171.98333
#> 11  116.38273
#> 12  146.09997
#> 13  137.37380
#> 14  119.53464
#> 15   90.23908
#> 16  137.49140
#> 17  156.99798
#> 18  170.18564
#> 19  130.17900
#> 20  136.73929
#> 21  111.69403
#> 22  125.45135
#> 23  166.58798
#> 24  113.64545
#> 25  191.09572
#> 26  173.14867
#> 27  148.70209
#> 28  188.10551
#> 29  136.25835
#> 30  134.82477
#> 31  228.71516
#> 32  172.32160
#> 33  119.52807
#> 34  125.35199
#> 35  172.52441
#> 36  196.05345
#> 37  129.51648
#> 38  153.79668
#> 39  122.90461
#> 40  147.71971
#> 41  113.90735
#> 42  116.42151
#> 43  109.90983
#> 44  122.48110
#> 45  139.02194
#> 46  145.24588
#> 47  159.42747
#> 48  174.04684
#> 49  133.66546
#> 50  140.84688
#> 51  152.19647
#> 52  135.80134
#> 53  142.28022
#> 54  129.26961
#> 55  148.52932
#> 56  163.22557
#> 57  208.72824
#> 58  125.99255
#> 59  144.82749
#> 60  121.45702
#> 61  240.82283
#> 62  110.78177
#> 63  145.39458
#> 64  131.82409
#> 65  169.32920
#> 66  122.57208
#> 67  124.44553
#> 68   98.10299
#> 69  142.39540
#> 70   84.25420
#> 71  105.47009
#> 72  114.74464
#> 73  195.78917
#> 74  153.33830
#> 75  161.18491
#> 76  182.66164
#> 77  167.81536
#> 78  143.25156
#> 79  114.52665
#> 80  147.76058
#> 81  134.30827
#> 82   92.10894
#> 83  132.63791
#> 84  121.49251
#> 85  104.88094
#> 86  195.98901
#> 87  172.60821
#> 88  124.07968
#> 89  139.84375
#> 90  131.65443
#> 91  144.05377
#> 92  150.87675
#> 93  155.00242
#> 94  153.15135
#> 95  102.42265
#> 96  169.61012
#> 97  177.56128
#> 98  195.91009
#> 99  198.87236
#> 100 136.38806